Manage multiple CDISCDataConnector, CDISCDatasetConnector and CDISCDataset objects.

Class manages CDISCDataConnector, CDISCDatasetConnector and CDISCDataset objects and aggregate them in one collection. Class also decides whether to launch app before initialize teal application.

Super classes

teal::DataAbstract -> teal::RelationalData -> CDISCData

Methods

Public methods

CDISCData$new()

CDISCData$get_parent()

CDISCData$check_metadata()

CDISCData$clone()

Inherited methods

teal::DataAbstract$check()

teal::DataAbstract$check_reproducibility()

teal::DataAbstract$execute_mutate()

teal::DataAbstract$get_check_result()

teal::DataAbstract$get_code()

teal::DataAbstract$get_code_class()

teal::DataAbstract$get_dataset()

teal::DataAbstract$get_datasets()

teal::DataAbstract$is_pulled()

teal::DataAbstract$mutate()

teal::DataAbstract$mutate_dataset()

teal::DataAbstract$set_check()

teal::DataAbstract$set_pull_code()

teal::RelationalData$get_connectors()

teal::RelationalData$get_datanames()

teal::RelationalData$get_items()

teal::RelationalData$get_join_keys()

teal::RelationalData$get_server()

teal::RelationalData$get_ui()

teal::RelationalData$launch()

teal::RelationalData$mutate_join_keys()

teal::RelationalData$print()

Method new()

Create a new object of CDISCData class

Usage

CDISCData$new(..., check = FALSE, join_keys)

Arguments

...

(RelationalDataConnector, Dataset or DatasetConnector) elements to include where ADSL data is mandatory.

check

(logical) reproducibility check - whether evaluated preprocessing code gives the same objects as provided in arguments. Check is run only if flag is true and preprocessing code is not empty.

join_keys

(JoinKeys) or a single (JoinKeySet)
(optional) object with datasets column names used for joining. If empty then it would be automatically derived basing on intersection of datasets primary keys

Method get_parent()

Get all datasets parent names

Usage

CDISCData$get_parent()

Returns

(named list) with dataset name and its corresponding parent dataset name

Method check_metadata()

Check correctness of stored joining keys and presence of keys to parent

Usage

CDISCData$check_metadata()

Returns

raise and error or invisible TRUE

Method clone()

The objects of this class are cloneable with this method.

Usage

CDISCData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.