Manage multiple and DatasetConnector of the same type.

Class manages DatasetConnector to specify additional dynamic arguments and to open/close connection.

Super classes

teal::DataAbstract -> teal::RelationalDataConnector -> CDISCDataConnector

Methods

Public methods

CDISCDataConnector$new()

CDISCDataConnector$get_parent()

CDISCDataConnector$clone()

Inherited methods

teal::DataAbstract$check()

teal::DataAbstract$check_reproducibility()

teal::DataAbstract$execute_mutate()

teal::DataAbstract$get_check_result()

teal::DataAbstract$get_code()

teal::DataAbstract$get_datanames()

teal::DataAbstract$get_dataset()

teal::DataAbstract$get_datasets()

teal::DataAbstract$get_items()

teal::DataAbstract$is_pulled()

teal::DataAbstract$mutate_dataset()

teal::DataAbstract$reassign_datasets_vars()

teal::DataAbstract$set_check()

teal::DataAbstract$set_pull_code()

teal::RelationalDataConnector$get_code_class()

teal::RelationalDataConnector$get_connection()

teal::RelationalDataConnector$get_preopen_server()

teal::RelationalDataConnector$get_server()

teal::RelationalDataConnector$get_ui()

teal::RelationalDataConnector$is_failed()

teal::RelationalDataConnector$launch()

teal::RelationalDataConnector$mutate()

teal::RelationalDataConnector$print()

teal::RelationalDataConnector$pull()

teal::RelationalDataConnector$set_preopen_server()

teal::RelationalDataConnector$set_pull_args()

teal::RelationalDataConnector$set_server()

teal::RelationalDataConnector$set_ui()

Method new()

Create a new CDISCDataConnector object

Usage

CDISCDataConnector$new(connection, connectors)

Arguments

connection

(DataConnection)
connection to data source

connectors

(list of DatasetConnector elements)
list with dataset connectors

Method get_parent()

Get all datasets parent names

Usage

CDISCDataConnector$get_parent()

Returns

(named list) with dataset name and its corresponding parent dataset name

Method clone()

The objects of this class are cloneable with this method.

Usage

CDISCDataConnector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.