R6 Class representing a dataset with parent attribute

Any data.frame object can be stored inside this object.

The difference compared to the inherited class is a parent field that indicates name of the parent dataset. Note that the parent field might be empty (i.e. character(0)).

Super class

teal::Dataset -> CDISCDataset

Methods

Public methods

CDISCDataset$new()

CDISCDataset$recreate()

CDISCDataset$get_attrs()

CDISCDataset$get_parent()

CDISCDataset$set_parent()

CDISCDataset$clone()

Inherited methods

teal::Dataset$check()

teal::Dataset$check_keys()

teal::Dataset$get_character_colnames()

teal::Dataset$get_code()

teal::Dataset$get_code_class()

teal::Dataset$get_colnames()

teal::Dataset$get_column_labels()

teal::Dataset$get_dataname()

teal::Dataset$get_datanames()

teal::Dataset$get_dataset_label()

teal::Dataset$get_factor_colnames()

teal::Dataset$get_keys()

teal::Dataset$get_numeric_colnames()

teal::Dataset$get_raw_data()

teal::Dataset$get_row_labels()

teal::Dataset$get_rownames()

teal::Dataset$is_pulled()

teal::Dataset$mutate()

teal::Dataset$set_code()

teal::Dataset$set_dataset_label()

teal::Dataset$set_keys()

teal::Dataset$set_vars()

Method new()

Create a new object of CDISCDataset class

Usage

CDISCDataset$new(
  dataname,
  x,
  keys,
  parent,
  code = character(0),
  label = character(0),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(data.frame)

keys

(character)
vector with primary keys

parent

optional, (character)
parent dataset name

code

(character)
A character string defining the code needed to produce the data set in x

label

(character)
Label to describe the dataset

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Method recreate()

Recreate a dataset with its current attributes This is useful way to have access to class initialize method basing on class object

Usage

CDISCDataset$recreate(
  dataname = self$get_dataname(),
  x = self$get_raw_data(),
  keys = self$get_keys(),
  parent = self$get_parent(),
  code = self$get_code_class(),
  label = self$get_dataset_label(),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(data.frame)

keys

(character)
vector with primary keys

parent

optional, (character)
parent dataset name

code

(character)
A character string defining the code needed to produce the data set in x

label

(character)
Label to describe the dataset

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Returns

a new object of CDISCDataset class

Method get_attrs()

Get all dataset attributes

Usage

CDISCDataset$get_attrs()

Returns

(named list) with dataset attributes

Method get_parent()

Get parent dataset name

Usage

CDISCDataset$get_parent()

Returns

(character) indicating parent dataname

Method set_parent()

Set parent dataset name

Usage

CDISCDataset$set_parent(parent)

Arguments

parent

(character) indicating parent dataname

Returns

(self) invisibly for chaining

Method clone()

The objects of this class are cloneable with this method.

Usage

CDISCDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- cdisc_dataset(
  dataname = "XYZ",
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = "y",
  parent = "ABC",
  code = "XYZ <- data.frame(x = c(1, 2), y = c('aa', 'bb'),
                            stringsAsFactors = FALSE)"
)

x$ncol

#> [1] 2

x$get_code()

#> [1] "XYZ <- data.frame(x = c(1, 2), y = c(\"aa\", \"bb\"), stringsAsFactors = FALSE)"

x$get_dataname()

#> [1] "XYZ"

x$get_keys()

#> [1] "y"

x$get_parent()

#> [1] "ABC"