Manage multiple CDISCTealDataConnector, CDISCTealDatasetConnector and CDISCTealDataset objects.

Class manages CDISCTealDataConnector, CDISCTealDatasetConnector and CDISCTealDataset objects and aggregate them in one collection. Class also decides whether to launch app before initialize teal application.

Super classes

teal::TealDataAbstract -> teal::TealData -> CDISCTealData

Methods

Public methods

CDISCTealData$new()

CDISCTealData$get_parent()

CDISCTealData$check_metadata()

CDISCTealData$clone()

Inherited methods

teal::TealDataAbstract$check()

teal::TealDataAbstract$check_reproducibility()

teal::TealDataAbstract$execute_mutate()

teal::TealDataAbstract$get_check_result()

teal::TealDataAbstract$get_code()

teal::TealDataAbstract$get_code_class()

teal::TealDataAbstract$get_dataset()

teal::TealDataAbstract$get_datasets()

teal::TealDataAbstract$is_pulled()

teal::TealDataAbstract$mutate()

teal::TealDataAbstract$mutate_dataset()

teal::TealDataAbstract$reassign_datasets_vars()

teal::TealDataAbstract$set_check()

teal::TealDataAbstract$set_pull_code()

teal::TealData$copy()

teal::TealData$get_connectors()

teal::TealData$get_datanames()

teal::TealData$get_items()

teal::TealData$get_join_keys()

teal::TealData$get_server()

teal::TealData$get_ui()

teal::TealData$launch()

teal::TealData$mutate_join_keys()

teal::TealData$print()

Method new()

Create a new object of CDISCTealData class

Usage

CDISCTealData$new(..., check = FALSE, join_keys)

Arguments

...

(TealDataConnector, TealDataset or TealDatasetConnector) elements to include where ADSL data is mandatory.

check

(logical) reproducibility check - whether evaluated preprocessing code gives the same objects as provided in arguments. Check is run only if flag is true and preprocessing code is not empty.

join_keys

(JoinKeys) or a single (JoinKeySet)
(optional) object with datasets column names used for joining. If empty then it would be automatically derived basing on intersection of datasets primary keys

Method get_parent()

Get all datasets parent names

Usage

CDISCTealData$get_parent()

Returns

(named list) with dataset name and its corresponding parent dataset name

Method check_metadata()

Check correctness of stored joining keys and presence of keys to parent

Usage

CDISCTealData$check_metadata()

Returns

raise and error or invisible TRUE

Method clone()

The objects of this class are cloneable with this method.

Usage

CDISCTealData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.