R6 Class representing a dataset with parent attribute

Any data.frame object can be stored inside this object.

The difference compared to the inherited class is a parent field that indicates name of the parent dataset. Note that the parent field might be empty (i.e. character(0)).

Super class

teal::TealDataset -> CDISCTealDataset

Methods

Public methods

CDISCTealDataset$new()

CDISCTealDataset$recreate()

CDISCTealDataset$get_attrs()

CDISCTealDataset$get_parent()

CDISCTealDataset$set_parent()

CDISCTealDataset$clone()

Inherited methods

teal::TealDataset$check()

teal::TealDataset$check_keys()

teal::TealDataset$get_character_colnames()

teal::TealDataset$get_code()

teal::TealDataset$get_code_class()

teal::TealDataset$get_colnames()

teal::TealDataset$get_column_labels()

teal::TealDataset$get_dataname()

teal::TealDataset$get_datanames()

teal::TealDataset$get_dataset()

teal::TealDataset$get_dataset_label()

teal::TealDataset$get_factor_colnames()

teal::TealDataset$get_hash()

teal::TealDataset$get_join_keys()

teal::TealDataset$get_keys()

teal::TealDataset$get_mutate_code_class()

teal::TealDataset$get_mutate_vars()

teal::TealDataset$get_numeric_colnames()

teal::TealDataset$get_raw_data()

teal::TealDataset$get_row_labels()

teal::TealDataset$get_rownames()

teal::TealDataset$get_var_r6()

teal::TealDataset$get_vars()

teal::TealDataset$is_mutate_delayed()

teal::TealDataset$is_pulled()

teal::TealDataset$merge_join_keys()

teal::TealDataset$mutate()

teal::TealDataset$mutate_join_keys()

teal::TealDataset$print()

teal::TealDataset$reassign_datasets_vars()

teal::TealDataset$set_code()

teal::TealDataset$set_dataset_label()

teal::TealDataset$set_join_keys()

teal::TealDataset$set_keys()

teal::TealDataset$set_vars()

Method new()

Create a new object of CDISCTealDataset class

Usage

CDISCTealDataset$new(
  dataname,
  x,
  keys,
  parent,
  code = character(0),
  label = character(0),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(data.frame)

keys

(character)
vector with primary keys

parent

optional, (character)
parent dataset name

code

(character)
A character string defining the code needed to produce the data set in x

label

(character)
Label to describe the dataset

vars

(named list))
In case when this object code depends on other TealDataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include TealDataset or TealDatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Method recreate()

Recreate a dataset with its current attributes This is useful way to have access to class initialize method basing on class object

Usage

CDISCTealDataset$recreate(
  dataname = self$get_dataname(),
  x = self$get_raw_data(),
  keys = self$get_keys(),
  parent = self$get_parent(),
  code = private$code,
  label = self$get_dataset_label(),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(data.frame)

keys

(character)
vector with primary keys

parent

optional, (character)
parent dataset name

code

(character)
A character string defining the code needed to produce the data set in x

label

(character)
Label to describe the dataset

vars

(named list))
In case when this object code depends on other TealDataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include TealDataset or TealDatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Returns

a new object of CDISCTealDataset class

Method get_attrs()

Get all dataset attributes

Usage

CDISCTealDataset$get_attrs()

Returns

(named list) with dataset attributes

Method get_parent()

Get parent dataset name

Usage

CDISCTealDataset$get_parent()

Returns

(character) indicating parent dataname

Method set_parent()

Set parent dataset name

Usage

CDISCTealDataset$set_parent(parent)

Arguments

parent

(character) indicating parent dataname

Returns

(self) invisibly for chaining

Method clone()

The objects of this class are cloneable with this method.

Usage

CDISCTealDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- cdisc_dataset(
  dataname = "XYZ",
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = "y",
  parent = "ABC",
  code = "XYZ <- data.frame(x = c(1, 2), y = c('aa', 'bb'),
                            stringsAsFactors = FALSE)"
)

x$ncol
#> [1] 2
x$get_code()
#> [1] "XYZ <- data.frame(x = c(1, 2), y = c(\"aa\", \"bb\"), stringsAsFactors = FALSE)"
x$get_dataname()
#> [1] "XYZ"
x$get_keys()
#> [1] "y"
x$get_parent()
#> [1] "ABC"