A CallableFunction class of objects

A CallableFunction class of objects

A CallableFunction class of objects

Details

Object that stores function name with it's arguments. Methods to get call and run it.

Methods

Public methods

CallableFunction$new()

CallableFunction$run()

CallableFunction$get_call()

CallableFunction$refresh()

CallableFunction$set_args()

CallableFunction$set_arg_value()

CallableFunction$clone()

Method new()

Create a new CallableFunction object

Usage

CallableFunction$new(fun)

Arguments

fun

(function) function to set up

Returns

new CallableFunction object

Method run()

Run function

Usage

CallableFunction$run(return = TRUE, args = NULL, try = FALSE)

Arguments

return

(logical) whether to return an object

args

(NULL or named list) dynamic arguments to function

try

(logical) whether perform function evaluation inside try clause

Returns

nothing or output from function depending on return argument

Method get_call()

Get function call with substituted arguments in args. These arguments will not be stored in the object.

Usage

CallableFunction$get_call(deparse = TRUE, args = NULL)

Arguments

deparse

(logical) whether to return a deparsed version of call

args

(NULL or named list) dynamic arguments to function

Returns

call or character depending on deparse argument

Method refresh()

Refresh call with function name and saved arguments

Usage

CallableFunction$refresh()

Returns

nothing

Method set_args()

Set up function arguments

Usage

CallableFunction$set_args(args)

Arguments

args

(NULL or named list) dynamic arguments to function

Returns

nothing

Method set_arg_value()

Set up single function argument with value

Usage

CallableFunction$set_arg_value(name, value)

Arguments

name

(character) argument name

value

argument value

Returns

nothing

Method clone()

The objects of this class are cloneable with this method.

Usage

CallableFunction$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.