FilterState object for factor or character variable

Note

Casts the passed object to character before validating the input making it possible to pass any object coercible to character to this method.

Super class

teal::FilterState -> ChoicesFilterState

Methods

Public methods

ChoicesFilterState$new()

ChoicesFilterState$get_call()

ChoicesFilterState$ui()

ChoicesFilterState$server()

ChoicesFilterState$set_state()

ChoicesFilterState$set_selected()

ChoicesFilterState$clone()

Inherited methods

teal::FilterState$destroy_observers()

teal::FilterState$get_dataname()

teal::FilterState$get_keep_na()

teal::FilterState$get_selected()

teal::FilterState$get_varlabel()

teal::FilterState$get_varname()

teal::FilterState$set_keep_na()

Method new()

Initialize a FilterState object

Usage

ChoicesFilterState$new(
  x,
  varname,
  varlabel = character(0),
  input_dataname = NULL,
  extract_type = character(0)
)

Arguments

x

(character or factor)
values of the variable used in filter

varname

(character, name)
name of the variable

varlabel

(character(1))
label of the variable (optional).

input_dataname

(name or call)
name of dataset where x is taken from

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <input_dataname>$<varname>

"matrix" varname in the condition call will be returned as <input_dataname>[, <varname>]

Method get_call()

Returns reproducible condition call for current selection. For this class returned call looks like <varname> %in% c(<values selected>) with optional is.na(<varname>).

Usage

ChoicesFilterState$get_call()

Returns

(call)

Method ui()

UI Module for ChoicesFilterState. This UI element contains available choices selection and checkbox whether to keep or not keep the NA values.

Usage

ChoicesFilterState$ui(id)

Arguments

id

(character(1))
id of shiny element

Method server()

Server module

Usage

ChoicesFilterState$server(id)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

Returns

moduleServer function which returns NULL

Method set_state()

Set state

Usage

ChoicesFilterState$set_state(state)

Arguments

state

(list)
contains fields relevant for a specific class

selected defines initial selection

keep_na (logical) defines whether to keep or remove NA values

Method set_selected()

Sets the selected values of this ChoicesFilterState.

Usage

ChoicesFilterState$set_selected(value)

Arguments

value

(character) the array of the selected choices. Must not contain NA values.

Returns

invisibly NULL

Examples

filter <- teal:::ChoicesFilterState$new(c("a", "b", "c"), varname = "name")
filter$set_selected(c("c", "a"))

Method clone()

The objects of this class are cloneable with this method.

Usage

ChoicesFilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

filter_state <- teal:::init_filter_state(
  c(LETTERS, NA),
  varname = "x",
  input_dataname = as.name("data"),
  extract_type = character(0)
)
isolate(filter_state$get_call())

#> x %in% c("A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", 
#>     "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", 
#>     "X", "Y", "Z")

isolate(filter_state$set_selected("B"))
isolate(filter_state$set_keep_na(TRUE))
isolate(filter_state$get_call())

#> is.na(x) | x == "B"

## ------------------------------------------------
## Method `ChoicesFilterState$set_selected`
## ------------------------------------------------

filter <- teal:::ChoicesFilterState$new(c("a", "b", "c"), varname = "name")
filter$set_selected(c("c", "a"))