A DatasetConnector class of objects Objects of this class store the connection function to fetch a single dataset. Note that for some specific connection types (e.g. RICE or SAICE), an object of class DataConnection must be provided. Data can be pulled via the pull method and accessed directly through the dataset active binding. Pulled data inherits from the class Dataset

Methods

Public methods

DatasetConnector$new()

DatasetConnector$get_dataname()

DatasetConnector$get_datanames()

DatasetConnector$get_dataset_label()

DatasetConnector$get_keys()

DatasetConnector$get_code()

DatasetConnector$get_code_class()

DatasetConnector$get_pull_args()

DatasetConnector$get_dataset()

DatasetConnector$get_error_message()

DatasetConnector$get_pull_callable()

DatasetConnector$get_raw_data()

DatasetConnector$set_dataset_label()

DatasetConnector$set_keys()

DatasetConnector$pull()

DatasetConnector$set_args()

DatasetConnector$mutate()

DatasetConnector$is_failed()

DatasetConnector$is_pulled()

DatasetConnector$check()

DatasetConnector$set_ui_input()

DatasetConnector$get_ui()

DatasetConnector$get_server()

DatasetConnector$launch()

DatasetConnector$clone()

Method new()

Create a new DatasetConnector object. Set the pulling function CallableFunction which returns a data.frame, e.g. by reading from a function or creating it on the fly.

Usage

DatasetConnector$new(
  dataname,
  pull_callable,
  keys = character(0),
  label = character(0),
  code = character(0),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

pull_callable

(CallableFunction)
function with necessary arguments set to fetch data from connection.

keys

optional, (character)
vector of dataset primary keys column names

label

(character)
Label to describe the dataset.

code

(character)
A character string defining code to modify raw_data from this dataset. To modify current dataset code should contain at least one assignment to object defined in dataname argument. For example if dataname = ADSL example code should contain ADSL <- <some R code>. Can't be used simultaneously with script

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Method get_dataname()

Get dataname of dataset

Usage

DatasetConnector$get_dataname()

Returns

dataname of the dataset

Method get_datanames()

Get dataname of dataset

Usage

DatasetConnector$get_datanames()

Returns

character dataname of the dataset

Method get_dataset_label()

Get label of dataset

Usage

DatasetConnector$get_dataset_label()

Returns

character dataset label

Method get_keys()

Get primary keys of dataset

Usage

DatasetConnector$get_keys()

Returns

character vector with dataset primary keys

Method get_code()

Get code to get data

Usage

DatasetConnector$get_code(deparse = TRUE)

Arguments

deparse

(logical)
whether return deparsed form of a call

Returns

optionally deparsed call object

Method get_code_class()

Get internal CodeClass object

Usage

DatasetConnector$get_code_class()

Returns

\code{CodeClass}

Method get_pull_args()

Derive the arguments this connector will pull with

Usage

DatasetConnector$get_pull_args()

Returns

list of pull function fixed arguments

Method get_dataset()

Get dataset

Usage

DatasetConnector$get_dataset()

Returns

dataset (Dataset)

Method get_error_message()

Get error message from last pull

Usage

DatasetConnector$get_error_message()

Returns

character object with error message or character(0) if last pull was successful.

Method get_pull_callable()

Get pull function

Usage

DatasetConnector$get_pull_callable()

Returns

CallableFunction

Method get_raw_data()

Get raw data from dataset

Usage

DatasetConnector$get_raw_data()

Returns

data.frame data

Method set_dataset_label()

Set label of the dataset object

Usage

DatasetConnector$set_dataset_label(label)

Arguments

label

(character)
Label to describe the dataset.

Returns

(self) invisibly for chaining

Method set_keys()

Set new keys

Usage

DatasetConnector$set_keys(keys)

Arguments

keys

optional, (character)
vector of dataset primary keys column names

Returns

(self) invisibly for chaining.

Method pull()

Pull the data

Read or create data using pull_callable specified in the constructor.

Usage

DatasetConnector$pull(args = NULL, try = FALSE)

Arguments

args

(NULL or named list)
additional dynamic arguments for pull function. args can be omitted if pull_callable from constructor already contains all necessary arguments to pull data. One can try to execute pull_callable directly by x$pull_callable$run() or to get code using x$pull_callable$get_code(). args specified in pull are used temporary to get data but not saved in code.

try

(logical value)
whether perform function evaluation inside try clause

Returns

(self) if successful.

Method set_args()

Set arguments to the pulling function

Usage

DatasetConnector$set_args(args)

Arguments

args

(NULL or named list) dynamic arguments to function

Returns

(self) invisibly for chaining

Method mutate()

Mutate dataset by code

Either code or script must be provided, but not both.

Usage

DatasetConnector$mutate(code, vars = list())

Arguments

code

(character)
A character string defining code to modify raw_data from this dataset. To modify current dataset code should contain at least one assignment to object defined in dataname argument. For example if dataname = ADSL example code should contain ADSL <- <some R code>. Can't be used simultaneously with script

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Returns

(self) invisibly for chaining.

Method is_failed()

Check if pull has not failed.

Usage

DatasetConnector$is_failed()

Returns

TRUE if pull failed, else FALSE

Method is_pulled()

Check if dataset has already been pulled.

Usage

DatasetConnector$is_pulled()

Returns

TRUE if connector has been already pulled, else FALSE

Method check()

Check to determine if the raw data is reproducible from the get_code() code.

Usage

DatasetConnector$check()

Returns

TRUE always for all connectors to avoid evaluating the same code multiple times.

Method set_ui_input()

Sets the shiny UI according to the given inputs. Inputs must provide only scalar (length of 1) variables.

Usage

DatasetConnector$set_ui_input(inputs = NULL)

Arguments

inputs

(function) A shiny module UI function with single argument ns. This function needs to return a list of shiny inputs with their inputId wrapped in function ns. The inputId must match exactly the argument name to be set. See example. Nested lists are not allowed.

Returns

(self) invisibly for chaining.

Examples

ds &lt;- dataset_connector("xyz", pull_callable = callable_function(data.frame))
ds$set_ui_input(
  function(ns) {
    list(sliderInput(ns("colA"), "Select value for colA", min = 0, max = 10, value = 3),
         sliderInput(ns("colB"), "Select value for colB", min = 0, max = 10, value = 7))
  }
)
\dontrun{
ds$launch()
}

Method get_ui()

Get shiny ui function

Usage

DatasetConnector$get_ui(id)

Arguments

id

(character) namespace id

Returns

shiny UI in given namespace id

Method get_server()

Get shiny server function

Usage

DatasetConnector$get_server()

Returns

shiny server function

Method launch()

Launches a shiny app.

Usage

DatasetConnector$launch()

Returns

Shiny app

Examples

ds &lt;- dataset_connector("xyz", pull_callable = callable_function(data.frame))
ds$set_ui_input(
  function(ns) {
    list(sliderInput(ns("colA"), "Select value for colA", min = 0, max = 10, value = 3),
         sliderInput(ns("colB"), "Select value for colB", min = 0, max = 10, value = 7))
  }
)
\dontrun{
ds$launch()
}

Method clone()

The objects of this class are cloneable with this method.

Usage

DatasetConnector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `DatasetConnector$set_ui_input`
## ------------------------------------------------

ds <- dataset_connector("xyz", pull_callable = callable_function(data.frame))
ds$set_ui_input(
  function(ns) {
    list(sliderInput(ns("colA"), "Select value for colA", min = 0, max = 10, value = 3),
         sliderInput(ns("colB"), "Select value for colB", min = 0, max = 10, value = 7))
  }
)
if (FALSE) {
ds$launch()
}

## ------------------------------------------------
## Method `DatasetConnector$launch`
## ------------------------------------------------

ds <- dataset_connector("xyz", pull_callable = callable_function(data.frame))
ds$set_ui_input(
  function(ns) {
    list(sliderInput(ns("colA"), "Select value for colA", min = 0, max = 10, value = 3),
         sliderInput(ns("colB"), "Select value for colB", min = 0, max = 10, value = 7))
  }
)
if (FALSE) {
ds$launch()
}