FilterState object for POSIXct variable

Note

Casts the passed object to POSIXct before validating the input making it possible to pass any object coercible to POSIXct to this method.

Super class

teal::FilterState -> DatetimeFilterState

Methods

Public methods

DatetimeFilterState$new()

DatetimeFilterState$get_call()

DatetimeFilterState$ui()

DatetimeFilterState$server()

DatetimeFilterState$set_selected()

DatetimeFilterState$clone()

Inherited methods

teal::FilterState$destroy_observers()

teal::FilterState$get_dataname()

teal::FilterState$get_keep_na()

teal::FilterState$get_selected()

teal::FilterState$get_varlabel()

teal::FilterState$get_varname()

teal::FilterState$set_keep_na()

teal::FilterState$set_state()

Method new()

Initialize a FilterState object. This class has an extra field, private$timezone, which is set to Sys.timezone() by default. However, in case when using this module in teal app, one needs timezone of the app user. App user timezone is taken from session$userData$timezone and is set only if object is initialized in shiny.

Usage

DatetimeFilterState$new(
  x,
  varname,
  varlabel = character(0),
  input_dataname = NULL,
  extract_type = character(0)
)

Arguments

x

(POSIXct or POSIXlt)
values of the variable used in filter

varname

(character, name)
name of the variable

varlabel

(character(1))
label of the variable (optional).

input_dataname

(name or call)
name of dataset where x is taken from

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <input_dataname>$<varname>

"matrix" varname in the condition call will be returned as <input_dataname>[, <varname>]

Method get_call()

Returns reproducible condition call for current selection. For this class returned call looks like <varname> >= as.POSIXct(<min>, tz = <timezone>) & <varname> <= <max>, tz = <timezone>) with optional is.na(<varname>).

Usage

DatetimeFilterState$get_call()

Method ui()

UI Module for DatetimeFilterState. This UI element contains two date-time selections for min and max of the range and a checkbox whether to keep the NA values.

Usage

DatetimeFilterState$ui(id)

Arguments

id

(character(1))
id of shiny element

Method server()

Server module

Usage

DatetimeFilterState$server(id)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

Returns

moduleServer function which returns NULL

Method set_selected()

Sets the selected time frame of this DatetimeFilterState.

Usage

DatetimeFilterState$set_selected(value)

Arguments

value

(POSIX(2)) the lower and the upper bound of the selected time frame. Must not contain NA values.

Returns

invisibly NULL.

Examples

date <- as.POSIXct(1, origin = "01/01/1970")
filter <- teal:::DatetimeFilterState$new(
  c(date, date + 1, date + 2, date + 3),
  varname = "name"
)
filter$set_selected(c(date + 1, date + 2))

Method clone()

The objects of this class are cloneable with this method.

Usage

DatetimeFilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

filter_state <- teal:::init_filter_state(
  c(Sys.time() + seq(0, by = 3600, length.out = 10), NA),
  varname = "x",
  input_dataname = as.name("data"),
  extract_type = character(0)
)

isolate(filter_state$get_call())

#> x >= as.POSIXct("2021-10-28 09:23:34", tz = "Etc/UTC") & x < 
#>     as.POSIXct("2021-10-28 18:23:35", tz = "Etc/UTC")

isolate(filter_state$set_selected(c(Sys.time() + 3L, Sys.time() + 8L)))
isolate(filter_state$set_keep_na(TRUE))
isolate(filter_state$get_call())

#> is.na(x) | x >= as.POSIXct("2021-10-28 09:23:37", tz = "Etc/UTC") & 
#>     x < as.POSIXct("2021-10-28 09:23:43", tz = "Etc/UTC")

## ------------------------------------------------
## Method `DatetimeFilterState$set_selected`
## ------------------------------------------------

date <- as.POSIXct(1, origin = "01/01/1970")
filter <- teal:::DatetimeFilterState$new(
  c(date, date + 1, date + 2, date + 3),
  varname = "name"
)
filter$set_selected(c(date + 1, date + 2))