FilterState object for empty variable

Super class

teal::FilterState -> EmptyFilterState

Methods

Public methods

EmptyFilterState$new()

EmptyFilterState$get_call()

EmptyFilterState$ui()

EmptyFilterState$server()

EmptyFilterState$set_state()

EmptyFilterState$clone()

Inherited methods

teal::FilterState$destroy_observers()

teal::FilterState$get_dataname()

teal::FilterState$get_keep_na()

teal::FilterState$get_selected()

teal::FilterState$get_varlabel()

teal::FilterState$get_varname()

teal::FilterState$set_keep_na()

teal::FilterState$set_na_rm()

teal::FilterState$set_selected()

Method new()

Initialize EmptyFilterState object

Usage

EmptyFilterState$new(
  x,
  varname,
  varlabel = character(0),
  input_dataname = NULL,
  extract_type = character(0)
)

Arguments

x

(vector)
values of the variable used in filter

varname

(character, name)
name of the variable

varlabel

(character(1))
label of the variable (optional).

input_dataname

(name or call)
name of dataset where x is taken from

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <input_dataname>$<varname>

"matrix" varname in the condition call will be returned as <input_dataname>[, <varname>]

Method get_call()

Returns reproducible condition call for current selection relevant for selected variable type. Method is using internal reactive values which makes it reactive and must be executed in reactive or isolated context.

Usage

EmptyFilterState$get_call()

Method ui()

UI Module for EmptyFilterState. This UI element contains checkbox input to filter or keep missing values.

Usage

EmptyFilterState$ui(id)

Arguments

id

(character(1))
id of shiny element

Method server()

Controls selection of keep_na checkbox input

Usage

EmptyFilterState$server(id)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

Returns

moduleServer function which returns NULL

Method set_state()

Set state

Usage

EmptyFilterState$set_state(state)

Arguments

state

(list)
contains fields relevant for a specific class

keep_na (logical) defines whether to keep or remove NA values

Method clone()

The objects of this class are cloneable with this method.

Usage

EmptyFilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

filter_state <- teal:::EmptyFilterState$new(
  NA,
  varname = "x",
  input_dataname = as.name("data"),
  extract_type = character(0)
)
isolate(filter_state$get_call())
isolate(filter_state$set_selected(TRUE))
isolate(filter_state$set_keep_na(TRUE))
isolate(filter_state$get_call())