Abstract class to encapsulate filter states

Details

This class is responsible for managing single filter item within FilteredData class. Filter states depend on the variable type: (logical, integer, numeric, factor, character, Date, POSIXct, POSIXlt) and returns FilterState object with class corresponding to input variable. Class controls single filter entry in module_single_filter_item and returns code relevant to selected values.

factor, character: class = ChoicesFilterState

numeric: class = RangeFilterState

logical: class = LogicalFilterState

Date: class = DateFilterState

POSIXct, POSIXlt: class = DatetimeFilterState

all NA entries: class: FilterState, cannot be filtered

default: FilterState, cannot be filtered Each variable's filter state is a R6 object which contains choices, selected, varname, dataname, labels, na_count, keep_na and other variable type specific fields (keep_inf, inf_count, timezone). Object contains also shiny module (ui and server) which manages state of the filter through reactive values selected, keep_na, keep_inf which trigger get_call() and every R function call up in reactive chain.

Methods

Public methods

FilterState$new()

FilterState$destroy_observers()

FilterState$get_call()

FilterState$get_dataname()

FilterState$get_keep_na()

FilterState$get_varlabel()

FilterState$get_varname()

FilterState$get_selected()

FilterState$set_keep_na()

FilterState$set_na_rm()

FilterState$set_selected()

FilterState$set_state()

FilterState$server()

FilterState$ui()

FilterState$clone()

Method new()

Initialize a FilterState object

Usage

FilterState$new(
  x,
  varname,
  varlabel = character(0),
  input_dataname = NULL,
  extract_type = character(0)
)

Arguments

x

(vector)
values of the variable used in filter

varname

(character, name)
name of the variable

varlabel

(character(1))
label of the variable (optional).

input_dataname

(name or call)
name of dataset where x is taken from. Must be specified if extract_type argument is not empty.

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <input_dataname>$<varname>

"matrix" varname in the condition call will be returned as <input_dataname>[, <varname>]

Method destroy_observers()

Destroy observers stored in private$observers.

Usage

FilterState$destroy_observers()

Method get_call()

Returns reproducible condition call for current selection relevant for selected variable type. Method is using internal reactive values which makes it reactive and must be executed in reactive or isolated context.

Usage

FilterState$get_call()

Method get_dataname()

Returns dataname

Usage

FilterState$get_dataname(deparse = TRUE)

Arguments

deparse

(logical(1))
whether dataname should be deparsed. TRUE by default

Returns

(name or character(1))

Method get_keep_na()

Returns current keep_na selection

Usage

FilterState$get_keep_na()

Returns

(logical(1))

Method get_varlabel()

Returns variable label

Usage

FilterState$get_varlabel()

Returns

(character(1))

Method get_varname()

Get variable name

Usage

FilterState$get_varname(deparse = FALSE)

Arguments

deparse

(logical(1))
whether variable name should be deparsed. FALSE by default

Returns

(name or character(1))

Method get_selected()

Get selected values from FilterState

Usage

FilterState$get_selected()

Returns

class of the returned object depends of class of the FilterState

Method set_keep_na()

Set if NA should be kept

Usage

FilterState$set_keep_na(value)

Arguments

value

(logical(1))
value(s) which come from the filter selection. Value is set in server modules after selecting check-box-input in the shiny interface. Values are set to private$keep_na which is reactive.

Method set_na_rm()

Some methods needs additional !is.na(varame) condition to not include missing values. When private$na_rm = TRUE is set, self$get_call returns condition extended by !is.na condition.

Usage

FilterState$set_na_rm(value)

Arguments

value

(logical(1)) when TRUE, FilterState$get_call appends an expression removing NA values to the filter expression returned by get_call

Method set_selected()

Set selection

Usage

FilterState$set_selected(value)

Arguments

value

(vector)
value(s) which come from the filter selection. Values are set in server modules after choosing value in app interface. Values are set to private$selected which is reactive. Values type have to be the same as private$choices.

Method set_state()

Set state

Usage

FilterState$set_state(state)

Arguments

state

(list)
contains fields relevant for a specific class

selected defines initial selection

keep_na (logical) defines whether to keep or remove NA values

Method server()

Server module

Usage

FilterState$server(id)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

Returns

moduleServer function which returns NULL

Method ui()

UI Module

Usage

FilterState$ui(id)

Arguments

id

(character(1))
id of shiny element. UI for this class contains simple message informing that it's not supported description Adds is.na(varname) before existing condition calls if keep_na is selected. Otherwise, if missings are found in the variable !is.na will be added only if private$na_rm = TRUE return (call) description Prefixed (or not) variableReturn variable name needed to condition call. If isTRUE(private$use_dataset) variable is prefixed by dataname to be evaluated as extracted object, for example data$var return (name or call) Sets keep_na field according to observed input$keep_na If keep_na = TRUE is.na(varname) is added to the returned call. Otherwise returned call excludes NA when executed. Set choicesSet choices is supposed to be executed once in the constructor to define set/range which selection is made from. parameter choices (vector)
class of the vector depends on the FilterState class. return a NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

FilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.