FilterStates R6 class

FilteredDataset is a class which renders/controls FilterStates(s) for Dataset. Each FilteredDataset contains filter_states field - a list which contains one (data.frame) or multiple (MultiAssayExperiment) FilterStates objects. Each FilterStates is responsible for one filter/subset expression applied for specific components of the Dataset.

Methods

Public methods

FilteredDataset$new()

FilteredDataset$add_to_eval_env()

FilteredDataset$queues_empty()

FilteredDataset$get_call()

FilteredDataset$get_data()

FilteredDataset$get_data_reactive()

FilteredDataset$get_filter_states()

FilteredDataset$get_dataname()

FilteredDataset$get_dataset()

FilteredDataset$get_filter_overview_info()

FilteredDataset$get_hash()

FilteredDataset$get_keys()

FilteredDataset$get_join_keys()

FilteredDataset$get_varlabels()

FilteredDataset$get_varnames()

FilteredDataset$get_filtered_dataname()

FilteredDataset$get_filterable_varnames()

FilteredDataset$set_bookmark_state()

FilteredDataset$ui()

FilteredDataset$server()

FilteredDataset$ui_add_filter_state()

FilteredDataset$srv_add_filter_state()

FilteredDataset$clone()

Method new()

Initializes this FilteredDataset object

Usage

FilteredDataset$new(dataset)

Arguments

dataset

(Dataset)
single dataset for which filters are rendered

Method add_to_eval_env()

Adds objects to the filter call evaluation environment

Usage

FilteredDataset$add_to_eval_env(name, value)

Arguments

name

(character) object name

value

object value

Returns

invisibly this FilteredDataset

Method queues_empty()

Removes all active filter items applied to this dataset

Usage

FilteredDataset$queues_empty()

Returns

NULL

Method get_call()

Gets a filter expression

This functions returns filter calls equivalent to selected items within each of filter_states. Configuration of the calls is constant and depends on filter_states type and order which are set during initialization.

Usage

FilteredDataset$get_call()

Returns

filter call or list of filter calls

Method get_data()

Gets raw data of this dataset

Usage

FilteredDataset$get_data(filtered)

Arguments

filtered

(logical(1))
whether returned data should be filtered or not

Returns

type of returned object depending on a data stored in Dataset. Currently data.frame or MultiAssayExperiment

Method get_data_reactive()

Gets the reactive object which returns filtered data

Usage

FilteredDataset$get_data_reactive()

Returns

(reactive)

Method get_filter_states()

Gets the filter states

Usage

FilteredDataset$get_filter_states(id = character(0))

Arguments

id

(character(1), character(0))
the id of the private$filter_states list element where FilterStates is kept.

Returns

FilterStates or list of FilterStates objects.

Method get_dataname()

Get name of the dataset

Get name of the dataset

Usage

FilteredDataset$get_dataname()

Returns

character(1) as a name of this dataset

Method get_dataset()

Gets the dataset in this FilteredDataset

Usage

FilteredDataset$get_dataset()

Returns

Dataset

Method get_filter_overview_info()

Get filter overview rows of a dataset

Usage

FilteredDataset$get_filter_overview_info()

Returns

(matrix) matrix of observations and subjects

Method get_hash()

Returns the hash of the unfiltered dataset

Usage

FilteredDataset$get_hash()

Returns

(character(1)) the hash

Method get_keys()

Gets the keys for the dataset of this FilteredDataset

Usage

FilteredDataset$get_keys()

Returns

(character) the keys of dataset Gets join keys to join the dataset of this FilteredDataset with other Dataset objects.

Method get_join_keys()

Usage

FilteredDataset$get_join_keys()

Returns

list of keys

Method get_varlabels()

Gets labels of variables in the data

Variables are the column names of the data. Either, all labels must have been provided for all variables in set_data or NULL.

Usage

FilteredDataset$get_varlabels(variables = NULL)

Arguments

variables

(character vector) variables to get labels for; if NULL, for all variables in data

Returns

(character or NULL) variable labels, NULL if column_labels attribute does not exist for the data

Method get_varnames()

Gets variable names from dataset

Usage

FilteredDataset$get_varnames()

Returns

character the variable names

Method get_filtered_dataname()

Gets the suffixed dataname Used when filtering the data to get <dataname>_FILTERED, <dataname>_FILTERED_ALONE or any other name.

Usage

FilteredDataset$get_filtered_dataname(
  dataname = self$get_dataname(),
  suffix = "_FILTERED"
)

Arguments

dataname

(character(1)) dataname

suffix

(character(1)) string to be putted after dataname

Returns

character(1)

Method get_filterable_varnames()

Gets variable names for the filtering.

Usage

FilteredDataset$get_filterable_varnames()

Returns

(character vector) of variable names

Method set_bookmark_state()

Sets the bookmark state

Usage

FilteredDataset$set_bookmark_state(id, state)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

state

(named list)
containing values of the initial filter. Values should be relevant to the referred column.

Returns

moduleServer function.

Method ui()

UI module for dataset active filters

UI module containing dataset active filters along with title and remove button.

Usage

FilteredDataset$ui(id)

Arguments

id

(character(1))
identifier of the element - preferably containing dataset name

Returns

function - shiny UI module

Method server()

Server module for a dataset active filters

Server module managing a active filters.

Usage

FilteredDataset$server(id)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

Returns

moduleServer function which returns NULL

Method ui_add_filter_state()

UI module to add filter variable for this dataset

UI module to add filter variable for this dataset

Usage

FilteredDataset$ui_add_filter_state(id)

Arguments

id

(character(1))
identifier of the element - preferably containing dataset name

Returns

function - shiny UI module

Method srv_add_filter_state()

Server module to add filter variable for this dataset

Server module to add filter variable for this dataset

Usage

FilteredDataset$srv_add_filter_state(id, ...)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

...

ignored

Returns

moduleServer function.

Method clone()

The objects of this class are cloneable with this method.

Usage

FilteredDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.