MAEFilteredDataset R6 class

MAEFilteredDataset R6 class

MAEFilteredDataset R6 class

Super class

teal::FilteredDataset -> MAEFilteredDataset

Methods

Public methods

MAEFilteredDataset$new()

MAEFilteredDataset$get_call()

MAEFilteredDataset$get_data()

MAEFilteredDataset$get_filter_overview_info()

MAEFilteredDataset$get_filterable_varnames()

MAEFilteredDataset$set_bookmark_state()

MAEFilteredDataset$ui_add_filter_state()

MAEFilteredDataset$srv_add_filter_state()

MAEFilteredDataset$clone()

Inherited methods

teal::FilteredDataset$add_to_eval_env()

teal::FilteredDataset$get_data_reactive()

teal::FilteredDataset$get_dataname()

teal::FilteredDataset$get_dataset()

teal::FilteredDataset$get_filter_states()

teal::FilteredDataset$get_filtered_dataname()

teal::FilteredDataset$get_hash()

teal::FilteredDataset$get_join_keys()

teal::FilteredDataset$get_keys()

teal::FilteredDataset$get_varlabels()

teal::FilteredDataset$get_varnames()

teal::FilteredDataset$queues_empty()

teal::FilteredDataset$server()

teal::FilteredDataset$ui()

Method new()

Initialize MAEFilteredDataset object

Usage

MAEFilteredDataset$new(dataset)

Arguments

dataset

(MAEDataset)
single dataset for which filters are rendered

Method get_call()

Get filter expression

This functions returns filter calls equivalent to selected items within each of filter_states. Configuration of the calls is constant and depends on filter_states type and order which are set during initialization. This class contains multiple FilterStates:

colData(dataset)for this object single MAEFilterStates which returns subsetByColData call

experimentsfor each experiment single SEFilterStates and FilterStates_matrix, both returns subset call

Usage

MAEFilteredDataset$get_call()

Returns

filter call or list of filter calls

Method get_data()

Gets raw data of this dataset

Usage

MAEFilteredDataset$get_data(filtered)

Arguments

filtered

(logical(1))
whether returned data should be filtered or not

Returns

MultiAssayExperiment

Method get_filter_overview_info()

Get filter overview rows of a dataset

Usage

MAEFilteredDataset$get_filter_overview_info()

Returns

(matrix) matrix of observations and subjects

Method get_filterable_varnames()

Gets variable names for the filtering.

Usage

MAEFilteredDataset$get_filterable_varnames()

Returns

(character(0))

Method set_bookmark_state()

Set bookmark state

Usage

MAEFilteredDataset$set_bookmark_state(id, state)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

state

(named list)
names of the list should correspond to the names of the initialized FilterStates kept in private$filter_states. For this object they are "subjects" and names of the experiments. Values of initial state should be relevant to the referred column.

Returns

moduleServer function which returns NULL

Method ui_add_filter_state()

UI module to add filter variable for this dataset

UI module to add filter variable for this dataset

Usage

MAEFilteredDataset$ui_add_filter_state(id)

Arguments

id

(character(1))
identifier of the element - preferably containing dataset name

Returns

function - shiny UI module

Method srv_add_filter_state()

Server module to add filter variable for this dataset

Server module to add filter variable for this dataset. For this class srv_add_filter_state calls multiple modules of the same name from FilterStates as MAEFilteredDataset contains one FilterStates object for colData and one for each experiment.

Usage

MAEFilteredDataset$srv_add_filter_state(id, ...)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

...

ignored.

Returns

moduleServer function which returns NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

MAEFilteredDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.