R6 Class representing a dataset including code

Any data.frame or rtable object can be stored inside this object. Additionally there needs to be the code that was used to generate this object. additionally it is possible to label the data set by handing over a label argument to the new method

Super class

teal::RawDataset -> NamedDataset

Methods

Public methods

NamedDataset$new()

NamedDataset$get_dataname()

NamedDataset$get_datanames()

NamedDataset$get_dataset_label()

NamedDataset$set_dataset_label()

NamedDataset$get_code()

NamedDataset$get_code_class()

NamedDataset$mutate()

NamedDataset$set_vars()

NamedDataset$set_code()

NamedDataset$check()

NamedDataset$clone()

Inherited methods

teal::RawDataset$get_character_colnames()

teal::RawDataset$get_colnames()

teal::RawDataset$get_column_labels()

teal::RawDataset$get_factor_colnames()

teal::RawDataset$get_numeric_colnames()

teal::RawDataset$get_raw_data()

teal::RawDataset$get_row_labels()

teal::RawDataset$get_rownames()

teal::RawDataset$is_pulled()

Method new()

initialize a NamedDataset class object

Usage

NamedDataset$new(
  dataname,
  x,
  code = character(0),
  label = character(0),
  vars = list()
)

Arguments

dataname

(character) A given name for the dataset it may not contain spaces

x

(data.frame)

code

(character) A character string defining the code needed to produce the data set in x

label

(character) Label to describe the dataset

vars

(list)
In case when this object code depends on the raw_data from the other NamedDataset, NamedDatasetConnector object(s) or other constant value, this/these object(s) should be included

Returns

new NamedDataset object

Method get_dataname()

Derive the name which was formerly called dataname

Usage

NamedDataset$get_dataname()

Returns

character name of the dataset

Method get_datanames()

Derive the dataname

Usage

NamedDataset$get_datanames()

Returns

character name of the dataset

Method get_dataset_label()

Derive the label which was former called datalabel

Usage

NamedDataset$get_dataset_label()

Returns

character label of the dataset

Method set_dataset_label()

Set the label for the dataset

Usage

NamedDataset$set_dataset_label(label)

Arguments

label

(character) the new label

Returns

self invisibly for chaining

Method get_code()

Get code to get data

Usage

NamedDataset$get_code(deparse = TRUE)

Arguments

deparse

(logical) whether return deparsed form of a call

Returns

optionally deparsed call object

Method get_code_class()

Get internal CodeClass object

Usage

NamedDataset$get_code_class()

Returns

CodeClass

Method mutate()

Mutate dataset by code

Either code or script must be provided, but not both.

Usage

NamedDataset$mutate(code, vars = list())

Arguments

code

(character) Code to mutate the dataset. Must contain the dataname

vars

(list)
In case when this object code depends on the raw_data from the other RelationalDataset, RelationalDatasetConnector object(s) or other constant value, this/these object(s) should be included

Returns

self invisibly for chaining

Method set_vars()

Adds variables which code depends on

Usage

NamedDataset$set_vars(vars)

Arguments

vars

(list)
In case when this object code depends on the raw_data from the other RelationalDataset, RelationalDatasetConnector object(s) or other constant value, this/these object(s) should be included Sets reproducible code

Method set_code()

Usage

NamedDataset$set_code(code)

Arguments

code

(character) Code to mutate the dataset. Must contain the dataname

Method check()

Check to determine if the raw data is reproducible from the get_code() code.

Usage

NamedDataset$check()

Returns

TRUE if the dataset generated from evaluating the get_code() code is identical to the raw data, else FALSE.

Method clone()

The objects of this class are cloneable with this method.

Usage

NamedDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- named_dataset(
  dataname = "XY",
  x = data.frame(x = c(2, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  code = "XY <- data.frame(x = c(2, 2),
    y = c('a', 'b'), stringsAsFactors = FALSE)"
)
x$ncol

#> [1] 2

x$get_code()

#> [1] "XY <- data.frame(x = c(2, 2), y = c(\"a\", \"b\"), stringsAsFactors = FALSE)"

x$get_dataname()

#> [1] "XY"