A NamedDatasetConnector class of objects

A NamedDatasetConnector class of objects

A NamedDatasetConnector class of objects

Details

Objects of this class store the connection function to fetch a single dataset. Note that for some specific connection types (e.g. RICE or SAICE), an object of class DataConnection must be provided. Data can be pulled via the pull method and accessed directly through the dataset active binding. Pulled data inherits from the class NamedDataset.

Super class

teal::RawDatasetConnector -> NamedDatasetConnector

Methods

Public methods

NamedDatasetConnector$new()

NamedDatasetConnector$get_dataname()

NamedDatasetConnector$get_dataset()

NamedDatasetConnector$get_datanames()

NamedDatasetConnector$get_dataset_label()

NamedDatasetConnector$set_dataset_label()

NamedDatasetConnector$get_code_class()

NamedDatasetConnector$mutate()

NamedDatasetConnector$pull()

NamedDatasetConnector$get_pull_args()

NamedDatasetConnector$check()

NamedDatasetConnector$clone()

Inherited methods

teal::RawDatasetConnector$get_code()

teal::RawDatasetConnector$get_error_message()

teal::RawDatasetConnector$get_pull_fun()

teal::RawDatasetConnector$get_raw_data()

teal::RawDatasetConnector$get_server()

teal::RawDatasetConnector$get_ui()

teal::RawDatasetConnector$is_failed()

teal::RawDatasetConnector$is_pulled()

teal::RawDatasetConnector$launch()

teal::RawDatasetConnector$set_args()

teal::RawDatasetConnector$set_ui_input()

Method new()

Create a new NamedDatasetConnector object. Set the pulling function load the data. dataname will be used as name of object to be assigned.

Usage

NamedDatasetConnector$new(
  dataname,
  pull_fun,
  code = character(0),
  label = character(0),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset, it may not contain spaces.

pull_fun

(CallableFunction)
function to load the data, must return a data.frame.

code

(character)
A character string defining code to modify raw_data from this dataset. To modify current dataset code should contain at least one assignment to object defined in dataname argument. For example if dataname = ADSL example code should contain ADSL <- <some modification code>.

label

(character)
Label to describe the dataset.

vars

(list)
In case when this object code depends on the raw_data from the other NamedDataset, NamedDatasetConnector object(s) or other constant value, this/these object(s) should be included.

Returns

new NamedDatasetConnector object

Method get_dataname()

Get dataname of dataset

Usage

NamedDatasetConnector$get_dataname()

Returns

dataname of the dataset

Method get_dataset()

Get dataset

Usage

NamedDatasetConnector$get_dataset()

Returns

dataset (<...>Dataset)

Method get_datanames()

Get dataname of dataset

Usage

NamedDatasetConnector$get_datanames()

Returns

character dataname of the dataset

Method get_dataset_label()

Get label of dataset

Usage

NamedDatasetConnector$get_dataset_label()

Returns

character dataset label

Method set_dataset_label()

Set label of the dataset object

Usage

NamedDatasetConnector$set_dataset_label(label)

Arguments

label

(character)
Label to describe the dataset

Returns

self invisibly for chaining

Method get_code_class()

Get internal CodeClass object

Usage

NamedDatasetConnector$get_code_class()

Returns

CodeClass

Method mutate()

Mutate dataset by code

Either code or script must be provided, but not both.

Usage

NamedDatasetConnector$mutate(code, vars = list())

Arguments

code

(character) Code to mutate the dataset. Must contain the dataset$dataname

vars

(list)
In case when this object code depends on the raw_data from the other RelationalDataset, RelationalDatasetConnector object(s) or other constant value, this/these object(s) should be included

Returns

self invisibly for chaining.

Method pull()

Pull the data

Read or create the data using pull_fun specified in the constructor.

Usage

NamedDatasetConnector$pull(args = NULL, try = FALSE)

Arguments

args

(NULL or named list)
additional dynamic arguments for pull function. args can be omitted if pull_fun from constructor already contains all necessary arguments to pull data. One can try to execute pull_fun directly by x$pull_fun$run() or to get code using x$pull_fun$get_code(). args specified in pull are used temporary to get data but not saved in code.

try

(logical value)
whether perform function evaluation inside try clause

Returns

self invisibly for chaining.

Method get_pull_args()

Derive the arguments this connector will pull with

Usage

NamedDatasetConnector$get_pull_args()

Returns

list of pull function fixed arguments

Method check()

Check to determine if the raw data is reproducible from the get_code() code.

Usage

NamedDatasetConnector$check()

Returns

TRUE always for all connectors to avoid evaluating the same code multiple times.

Method clone()

The objects of this class are cloneable with this method.

Usage

NamedDatasetConnector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.