FilterState object for numeric variable

FilterState object for numeric variable

FilterState object for numeric variable

Value

invisibly NULL

Note

Casts the passed object to numeric before validating the input making it possible to pass any object coercible to numeric to this method.

Super class

teal::FilterState -> RangeFilterState

Methods

Public methods

RangeFilterState$new()

RangeFilterState$get_call()

RangeFilterState$get_keep_inf()

RangeFilterState$get_state()

RangeFilterState$ui()

RangeFilterState$server()

RangeFilterState$set_keep_inf()

RangeFilterState$set_keep_inf_reactive()

RangeFilterState$set_state()

RangeFilterState$set_state_reactive()

RangeFilterState$set_selected()

RangeFilterState$clone()

Inherited methods

teal::FilterState$destroy_observers()

teal::FilterState$get_dataname()

teal::FilterState$get_keep_na()

teal::FilterState$get_selected()

teal::FilterState$get_varlabel()

teal::FilterState$get_varname()

teal::FilterState$set_keep_na()

teal::FilterState$set_keep_na_reactive()

teal::FilterState$set_na_rm()

teal::FilterState$set_selected_reactive()

Method new()

Initialize a FilterState object

Usage

RangeFilterState$new(
  x,
  varname,
  varlabel = character(0),
  input_dataname = NULL,
  extract_type = character(0)
)

Arguments

x

(numeric)
values of the variable used in filter

varname

(character, name)
name of the variable

varlabel

(character(1))
label of the variable (optional).

input_dataname

(name or call)
name of dataset where x is taken from

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <input_dataname>$<varname>

"matrix" varname in the condition call will be returned as <input_dataname>[, <varname>]

Method get_call()

Returns reproducible condition call for current selection. For this class returned call looks like <varname> >= <min value> & <varname> <= <max value> with optional is.na(<varname>) and is.finite(<varname>).

Usage

RangeFilterState$get_call()

Returns

(call)

Method get_keep_inf()

Returns current keep_inf selection

Usage

RangeFilterState$get_keep_inf()

Returns

(logical(1))

Method get_state()

Returns the filtering state.

Usage

RangeFilterState$get_state()

Returns

list containing values taken from the reactive fields:

selected (numeric(2)) range of the filter.

keep_na (logical(1)) whether NA should be kept.

keep_inf (logical(1)) whether Inf should be kept. UI Module for RangeFilterState. This UI element contains two values for min and max of the range and two checkboxes whether to keep the NA or Inf values.

Method ui()

Usage

RangeFilterState$ui(id)

Arguments

id

(character(1))
id of shiny element

Method server()

Server module

Usage

RangeFilterState$server(id)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

Returns

moduleServer function which returns NULL

Method set_keep_inf()

Set if Inf should be kept

Usage

RangeFilterState$set_keep_inf(value)

Arguments

value

(logical(1))
Value(s) which come from the filter selection. Value is set in server modules after selecting check-box-input in the shiny interface. Values are set to private$keep_inf which is reactive.

Method set_keep_inf_reactive()

Set if Inf should be kept when passing filters using set_filter_state

Usage

RangeFilterState$set_keep_inf_reactive(value)

Arguments

value

(logical(1))
Value(s) which come from the filter set by the user. Value is set in server modules after setting the filters using set_filter_state. Values are set to private$keep_inf_reactive which is reactive.

Method set_state()

Set state

Usage

RangeFilterState$set_state(state)

Arguments

state

(list)
contains fields relevant for a specific class

selected defines initial selection

keep_na (logical) defines whether to keep or remove NA values

keep_inf (logical) defines whether to keep or remove Inf values

Method set_state_reactive()

Set state when using set_filter_state

Usage

RangeFilterState$set_state_reactive(state)

Arguments

state

(list)
contains fields relevant for a specific class

selected defines initial selection

keep_na (logical) defines whether to keep or remove NA values

keep_inf (logical) defines whether to keep or remove Inf values

Method set_selected()

Sets the selected values of this RangeFilterState.

Usage

RangeFilterState$set_selected(value)

Arguments

value

(numeric(2)) the two-elements array of the lower and upper bound of the selected range. Must not contain NA values.

Examples

filter <- teal:::RangeFilterState$new(c(1, 2, 3, 4), varname = "name")
filter$set_selected(c(2, 3))

Method clone()

The objects of this class are cloneable with this method.

Usage

RangeFilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

filter_state <- teal:::RangeFilterState$new(
  c(NA, Inf, seq(1:10)),
  varname = "x",
  input_dataname = as.name("data"),
  extract_type = character(0)
)
isolate(filter_state$get_call())
#> x >= 1 & x <= 10
isolate(filter_state$set_selected(c(3L, 8L)))
isolate(filter_state$set_keep_na(TRUE))
isolate(filter_state$set_keep_inf(TRUE))
isolate(filter_state$get_call())
#> is.na(x) | (is.infinite(x) | x >= 3 & x <= 8)

## ------------------------------------------------
## Method `RangeFilterState$set_selected`
## ------------------------------------------------

filter <- teal:::RangeFilterState$new(c(1, 2, 3, 4), varname = "name")
filter$set_selected(c(2, 3))