A RawDatasetConnector class of objects

Objects of this class store the connection function to fetch a single dataset. Note that for some specific connection types (e.g. RICE or SAICE), an object of class DataConnection must be provided. Data can be pulled via the pull method and accessed directly through the dataset active binding. Pulled data inherits from the class RelationalDataset.

Methods

Public methods

RawDatasetConnector$new()

RawDatasetConnector$set_args()

RawDatasetConnector$get_code()

RawDatasetConnector$get_code_class()

RawDatasetConnector$get_dataset()

RawDatasetConnector$get_error_message()

RawDatasetConnector$get_pull_callable()

RawDatasetConnector$get_raw_data()

RawDatasetConnector$pull()

RawDatasetConnector$is_failed()

RawDatasetConnector$is_pulled()

RawDatasetConnector$set_ui_input()

RawDatasetConnector$get_ui()

RawDatasetConnector$get_server()

RawDatasetConnector$launch()

RawDatasetConnector$clone()

Method new()

Create a new RawDatasetConnector object. Set the pulling function CallableFunction which returns a data.frame, e.g. by reading from a function or creating it on the fly.

Usage

RawDatasetConnector$new(pull_callable, vars = list())

Arguments

pull_callable

(CallableFunction)
function to pull the data.

vars

(list)
In case when this object code depends on the raw_data from the other RelationalDataset, RelationalDatasetConnector object(s) or other constant value, this/these object(s) should be included

Returns

new RawDatasetConnector object

Method set_args()

Set arguments to the pulling function

Usage

RawDatasetConnector$set_args(args)

Arguments

args

(NULL or named list) dynamic arguments to function

Returns

self invisibly for chaining

Method get_code()

Get code to get data

Usage

RawDatasetConnector$get_code(deparse = TRUE)

Arguments

deparse

(logical)
whether return deparsed form of a call

Returns

optionally deparsed call object

Method get_code_class()

Get internal CodeClass object

Usage

RawDatasetConnector$get_code_class()

Returns

CodeClass

Method get_dataset()

Get dataset

Usage

RawDatasetConnector$get_dataset()

Returns

dataset (RawDataset)

Method get_error_message()

Get error message from last pull

Usage

RawDatasetConnector$get_error_message()

Returns

character object with error message or character(0) if last pull was successful.

Method get_pull_callable()

Get pull function

Usage

RawDatasetConnector$get_pull_callable()

Returns

CallableFunction

Method get_raw_data()

Get raw data from dataset

Usage

RawDatasetConnector$get_raw_data()

Returns

data.frame data

Method pull()

Pull the data

Read or create data using pull_callable specified in the constructor.

Usage

RawDatasetConnector$pull(args = NULL, try = FALSE)

Arguments

args

(NULL or named list)
additional dynamic arguments for pull function. args can be omitted if pull_callable from constructor already contains all necessary arguments to pull data. One can try to execute pull_callable directly by x$pull_callable$run() or to get code using x$pull_callable$get_code(). args specified in pull are used temporary to get data but not saved in code.

try

(logical value)
whether perform function evaluation inside try clause

Returns

self if successful.

Method is_failed()

Check if pull has not failed.

Usage

RawDatasetConnector$is_failed()

Returns

TRUE if pull failed, else FALSE

Method is_pulled()

Check if dataset has already been pulled.

Usage

RawDatasetConnector$is_pulled()

Returns

TRUE if connector has been already pulled, else FALSE

Method set_ui_input()

Sets the shiny UI according to the given inputs. Inputs must provide only scalar (length of 1) variables.

Usage

RawDatasetConnector$set_ui_input(inputs = NULL)

Arguments

inputs

(function) A shiny module UI function with single argument ns. This function needs to return a list of shiny inputs with their inputId wrapped in function ns. The inputId must match exactly the argument name to be set. See example. Nested lists are not allowed.

Returns

self invisibly for chaining.

Examples

ds &lt;- raw_dataset_connector(pull_callable = callable_function(data.frame))
ds$set_ui_input(
  function(ns) {
    list(sliderInput(ns("colA"), "Select value for colA", min = 0, max = 10, value = 3),
         sliderInput(ns("colB"), "Select value for colB", min = 0, max = 10, value = 7))
  }
)
\dontrun{
ds$launch()
}

Method get_ui()

Get shiny ui function

Usage

RawDatasetConnector$get_ui(id)

Arguments

id

(character) namespace id

Returns

shiny UI in given namespace id

Method get_server()

Get shiny server function

Usage

RawDatasetConnector$get_server()

Returns

shiny server function

Method launch()

Launches a shiny app.

Usage

RawDatasetConnector$launch()

Returns

Shiny app

Examples

ds &lt;- raw_dataset_connector(pull_callable = callable_function(data.frame))
ds$set_ui_input(
  function(ns) {
    list(sliderInput(ns("colA"), "Select value for colA", min = 0, max = 10, value = 3),
         sliderInput(ns("colB"), "Select value for colB", min = 0, max = 10, value = 7))
  }
)
\dontrun{
ds$launch()
}

Method clone()

The objects of this class are cloneable with this method.

Usage

RawDatasetConnector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `RawDatasetConnector$set_ui_input`
## ------------------------------------------------

ds <- raw_dataset_connector(pull_callable = callable_function(data.frame))
ds$set_ui_input(
  function(ns) {
    list(sliderInput(ns("colA"), "Select value for colA", min = 0, max = 10, value = 3),
         sliderInput(ns("colB"), "Select value for colB", min = 0, max = 10, value = 7))
  }
)
if (FALSE) {
ds$launch()
}

## ------------------------------------------------
## Method `RawDatasetConnector$launch`
## ------------------------------------------------

ds <- raw_dataset_connector(pull_callable = callable_function(data.frame))
ds$set_ui_input(
  function(ns) {
    list(sliderInput(ns("colA"), "Select value for colA", min = 0, max = 10, value = 3),
         sliderInput(ns("colB"), "Select value for colB", min = 0, max = 10, value = 7))
  }
)
if (FALSE) {
ds$launch()
}