Reactive extension of Queue class

Implements the queue abstract data type. The last element added to this queue is the last one to be returned from it.

Super class

teal::Queue -> ReactiveQueue

Methods

Public methods

ReactiveQueue$new()

ReactiveQueue$get()

ReactiveQueue$clone()

Inherited methods

teal::Queue$empty()

teal::Queue$pop()

teal::Queue$print()

teal::Queue$push()

teal::Queue$remove()

teal::Queue$size()

Method new()

Initializes ReactiveQueue by setting empty reactiveVal

Usage

ReactiveQueue$new()

Method get()

Get queue

Usage

ReactiveQueue$get(reversed = FALSE)

Arguments

reversed

(logical(1))
whether order of elements in the queue should be reversed. FALSE by default

Returns

values stored in the queue

Method clone()

The objects of this class are cloneable with this method.

Usage

ReactiveQueue$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.