Manage multiple RelationalDataConnector, DatasetConnector and Dataset objects.

Class manages RelationalDataConnector, DatasetConnector and Dataset objects and aggregate them in one collection. Class also decides whether to launch app before initialize teal application.

Super class

teal::DataAbstract -> RelationalData

Methods

Public methods

RelationalData$new()

RelationalData$get_datanames()

RelationalData$get_join_keys()

RelationalData$get_connectors()

RelationalData$get_items()

RelationalData$get_ui()

RelationalData$get_server()

RelationalData$launch()

RelationalData$mutate_join_keys()

RelationalData$check_metadata()

RelationalData$clone()

Inherited methods

teal::DataAbstract$check()

teal::DataAbstract$check_reproducibility()

teal::DataAbstract$execute_mutate()

teal::DataAbstract$get_check_result()

teal::DataAbstract$get_code()

teal::DataAbstract$get_code_class()

teal::DataAbstract$get_dataset()

teal::DataAbstract$get_datasets()

teal::DataAbstract$is_pulled()

teal::DataAbstract$mutate()

teal::DataAbstract$mutate_dataset()

teal::DataAbstract$set_check()

teal::DataAbstract$set_pull_code()

Method new()

Create a new object of RelationalData class

Usage

RelationalData$new(..., check = FALSE, join_keys)

Arguments

...

(RelationalDataConnector, Dataset, DatasetConnector)
objects

check

(logical) reproducibility check - whether evaluated preprocessing code gives the same objects as provided in arguments. Check is run only if flag is true and preprocessing code is not empty.

join_keys

(JoinKeys) or a single (JoinKeySet)
(optional) object with dataset column relationships used for joining. If empty then no joins between pairs of objects

Method get_datanames()

Derive the names of all datasets

Usage

RelationalData$get_datanames()

Returns

(character vector) with names

Method get_join_keys()

Get JoinKeys object with keys used for joining.

Usage

RelationalData$get_join_keys()

Returns

(JoinKeys) Get data connectors.

Method get_connectors()

Usage

RelationalData$get_connectors()

Returns

(list) with all DatasetConnector or RelationalDataConnector objects.

Method get_items()

Get all datasets and all dataset connectors

name of dataset connector to be returned. If NULL, all connectors are returned.

Usage

RelationalData$get_items(dataname = NULL)

Arguments

dataname

(character value)

Returns

list with all datasets and all connectors

Method get_ui()

Get a shiny-module UI to render the necessary app to derive RelationalDataConnector object's data

Usage

RelationalData$get_ui(id)

Arguments

id

(character) item ID for the shiny module

Returns

the shiny ui function

Method get_server()

Get a shiny-module server to render the necessary app to derive RelationalDataConnector object's data

Usage

RelationalData$get_server()

Returns

shiny server module.

Method launch()

Launch an app that allows to run the user-interfaces of all DataConnector and DatasetConnector modules

This piece is mainly used for debugging.

Usage

RelationalData$launch()

Method mutate_join_keys()

Change join_keys for a given pair of dataset names

Usage

RelationalData$mutate_join_keys(dataset_1, dataset_2, val)

Arguments

dataset_1, dataset_2

(character) datasets for which join_keys are to be returned

val

(named character) column names used to join

Returns

(self) invisibly for chaining

Method check_metadata()

Check there is consistency between the datasets and join_keys

Usage

RelationalData$check_metadata()

Returns

raise and error or invisible TRUE

Method clone()

The objects of this class are cloneable with this method.

Usage

RelationalData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(scda)
adsl_cf <- callable_function(function() synthetic_cdisc_data("latest")$adsl)
adlb_cf <- callable_function(function() synthetic_cdisc_data("latest")$adlb)
adrs_cf <- callable_function(function() synthetic_cdisc_data("latest")$adrs)
adtte_cf <- callable_function(function() synthetic_cdisc_data("latest")$adtte)
x1 <- cdisc_dataset_connector("ADSL", adsl_cf, keys = get_cdisc_keys("ADSL"))
x2 <- cdisc_dataset_connector("ADRS", adrs_cf, keys = get_cdisc_keys("ADRS"))
x3 <- cdisc_dataset(
  dataname = "ADAE",
  x = synthetic_cdisc_data("latest")$adae,
  code = "library(scda)\nADAE <- synthetic_cdisc_data(\"latest\")$adae"
)
x4 <- cdisc_dataset_connector("ADTTE", adtte_cf, keys = get_cdisc_keys("ADTTE"))
tc <- teal:::RelationalData$new(x1, x2, x3, x4)
tc$get_datanames()

#> [1] "ADSL"  "ADRS"  "ADAE"  "ADTTE"

if (FALSE) {
tc$get_datasets()
tc$get_dataset("ADAE")
}
tc$get_items()

#> $ADSL
#> <CDISCDatasetConnector>
#>   Inherits from: <DatasetConnector>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset: function () 
#>     get_dataset_label: function () 
#>     get_error_message: function () 
#>     get_join_keys: function () 
#>     get_keys: function () 
#>     get_parent: function () 
#>     get_pull_args: function () 
#>     get_pull_callable: function () 
#>     get_raw_data: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     get_var_r6: function () 
#>     initialize: function (dataname, pull_callable, keys, parent, code = character(0), 
#>     is_failed: function () 
#>     is_mutate_delayed: function () 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     mutate_join_keys: function (dataset, val) 
#>     pull: function (args = NULL, try = FALSE) 
#>     set_args: function (args) 
#>     set_dataset_label: function (label) 
#>     set_join_keys: function (x) 
#>     set_keys: function (keys) 
#>     set_ui_input: function (inputs = NULL) 
#>   Private:
#>     dataname: ADSL
#>     dataset: NULL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     get_pull_code_class: function (args = NULL) 
#>     is_pulled_flag: FALSE
#>     join_keys: JoinKeys, R6
#>     keys: STUDYID USUBJID
#>     parent: 
#>     pull_callable: CallableFunction, Callable, R6
#>     pull_internal: function (args = NULL, try = FALSE) 
#>     pull_vars: list
#>     server: function (input, output, session, data_args = NULL) 
#>     set_dataname: function (dataname) 
#>     set_failure: function (res) 
#>     set_parent: function (parent) 
#>     set_pull_callable: function (pull_callable) 
#>     set_pull_vars: function (pull_vars) 
#>     set_ui: function (ui_args = NULL) 
#>     set_var_r6: function (vars) 
#>     ui: function (id) 
#>     ui_input: NULL
#>     var_r6: list
#> 
#> $ADRS
#> <CDISCDatasetConnector>
#>   Inherits from: <DatasetConnector>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset: function () 
#>     get_dataset_label: function () 
#>     get_error_message: function () 
#>     get_join_keys: function () 
#>     get_keys: function () 
#>     get_parent: function () 
#>     get_pull_args: function () 
#>     get_pull_callable: function () 
#>     get_raw_data: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     get_var_r6: function () 
#>     initialize: function (dataname, pull_callable, keys, parent, code = character(0), 
#>     is_failed: function () 
#>     is_mutate_delayed: function () 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     mutate_join_keys: function (dataset, val) 
#>     pull: function (args = NULL, try = FALSE) 
#>     set_args: function (args) 
#>     set_dataset_label: function (label) 
#>     set_join_keys: function (x) 
#>     set_keys: function (keys) 
#>     set_ui_input: function (inputs = NULL) 
#>   Private:
#>     dataname: ADRS
#>     dataset: NULL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     get_pull_code_class: function (args = NULL) 
#>     is_pulled_flag: FALSE
#>     join_keys: JoinKeys, R6
#>     keys: STUDYID USUBJID PARAMCD AVISIT
#>     parent: ADSL
#>     pull_callable: CallableFunction, Callable, R6
#>     pull_internal: function (args = NULL, try = FALSE) 
#>     pull_vars: list
#>     server: function (input, output, session, data_args = NULL) 
#>     set_dataname: function (dataname) 
#>     set_failure: function (res) 
#>     set_parent: function (parent) 
#>     set_pull_callable: function (pull_callable) 
#>     set_pull_vars: function (pull_vars) 
#>     set_ui: function (ui_args = NULL) 
#>     set_var_r6: function (vars) 
#>     ui: function (id) 
#>     ui_input: NULL
#>     var_r6: list
#> 
#> $ADAE
#> A Dataset object containing the following data.frame (1934 rows and 66 columns):
#>   STUDYID              USUBJID SUBJID SITEID AGE SEX  RACE COUNTRY DTHFL
#> 1 AB12345 AB12345-BRA-1-id-134 id-134  BRA-1  47   M WHITE     BRA     N
#> 2 AB12345 AB12345-BRA-1-id-134 id-134  BRA-1  47   M WHITE     BRA     N
#> 3 AB12345 AB12345-BRA-1-id-134 id-134  BRA-1  47   M WHITE     BRA     N
#> 4 AB12345 AB12345-BRA-1-id-134 id-134  BRA-1  47   M WHITE     BRA     N
#> 5 AB12345 AB12345-BRA-1-id-141 id-141  BRA-1  35   F WHITE     BRA     N
#> 6 AB12345 AB12345-BRA-1-id-141 id-141  BRA-1  35   F WHITE     BRA     N
#>          INVID        INVNAM            ARM ARMCD         ACTARM ACTARMCD
#> 1 INV ID BRA-1 Dr. BRA-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 2 INV ID BRA-1 Dr. BRA-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 3 INV ID BRA-1 Dr. BRA-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 4 INV ID BRA-1 Dr. BRA-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 5 INV ID BRA-1 Dr. BRA-1 Doe C: Combination ARM C C: Combination    ARM C
#> 6 INV ID BRA-1 Dr. BRA-1 Doe C: Combination ARM C C: Combination    ARM C
#>         REGION1 STRATA1 STRATA2   BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL
#> 1 South America       B      S2 6.462991    LOW     Y     Y        Y       N
#> 2 South America       B      S2 6.462991    LOW     Y     Y        Y       N
#> 3 South America       B      S2 6.462991    LOW     Y     Y        Y       N
#> 4 South America       B      S2 6.462991    LOW     Y     Y        Y       N
#> 5 South America       B      S1 7.516076   HIGH     Y     Y        Y       Y
#> 6 South America       B      S1 7.516076   HIGH     Y     Y        Y       Y
#>       RANDDT             TRTSDTM             TRTEDTM    EOSSTT      EOSDT EOSDY
#> 1 2020-11-03 2020-11-04 03:50:33 2022-11-04 15:28:57 COMPLETED 2022-11-04   731
#> 2 2020-11-03 2020-11-04 03:50:33 2022-11-04 15:28:57 COMPLETED 2022-11-04   731
#> 3 2020-11-03 2020-11-04 03:50:33 2022-11-04 15:28:57 COMPLETED 2022-11-04   731
#> 4 2020-11-03 2020-11-04 03:50:33 2022-11-04 15:28:57 COMPLETED 2022-11-04   731
#> 5 2020-07-22 2020-07-25 14:10:56 2022-07-26 01:49:20 COMPLETED 2022-07-26   731
#> 6 2020-07-22 2020-07-25 14:10:56 2022-07-26 01:49:20 COMPLETED 2022-07-26   731
#>   DCSREAS DTHDT DTHCAUS DTHCAT LDDTHELD LDDTHGR1   LSTALVDT study_duration_secs
#> 1    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2022-11-15            63113904
#> 2    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2022-11-15            63113904
#> 3    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2022-11-15            63113904
#> 4    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2022-11-15            63113904
#> 5    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2022-08-10            63113904
#> 6    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2022-08-10            63113904
#>   ASEQ AESEQ        AETERM         AELLT       AEDECOD       AEHLT     AEHLGT
#> 1    1     1 trm B.2.1.2.1 llt B.2.1.2.1 dcd B.2.1.2.1 hlt B.2.1.2 hlgt B.2.1
#> 2    2     2 trm D.1.1.4.2 llt D.1.1.4.2 dcd D.1.1.4.2 hlt D.1.1.4 hlgt D.1.1
#> 3    3     3 trm A.1.1.1.2 llt A.1.1.1.2 dcd A.1.1.1.2 hlt A.1.1.1 hlgt A.1.1
#> 4    4     4 trm A.1.1.1.2 llt A.1.1.1.2 dcd A.1.1.1.2 hlt A.1.1.1 hlgt A.1.1
#> 5    1     1 trm B.2.1.2.1 llt B.2.1.2.1 dcd B.2.1.2.1 hlt B.2.1.2 hlgt B.2.1
#> 6    2     2 trm D.2.1.5.3 llt D.2.1.5.3 dcd D.2.1.5.3 hlt D.2.1.5 hlgt D.2.1
#>   AEBODSYS AESOC    AESEV AESER            AEACN AEREL
#> 1   cl B.2  cl B MODERATE     N DOSE NOT CHANGED     N
#> 2   cl D.1  cl D MODERATE     N DOSE NOT CHANGED     N
#> 3   cl A.1  cl A MODERATE     Y DOSE NOT CHANGED     N
#> 4   cl A.1  cl A MODERATE     Y DOSE NOT CHANGED     N
#> 5   cl B.2  cl B MODERATE     N   DOSE INCREASED     N
#> 6   cl D.2  cl D     MILD     N          UNKNOWN     Y
#>                              AEOUT AESDTH TRTEMFL AECONTRT     ASTDTM
#> 1               RECOVERED/RESOLVED      N       Y        N 2021-07-13
#> 2             RECOVERING/RESOLVING      N       Y        N 2021-09-04
#> 3 RECOVERED/RESOLVED WITH SEQUELAE      N       Y        Y 2022-03-15
#> 4             RECOVERING/RESOLVING      N       Y        N 2022-07-05
#> 5               RECOVERED/RESOLVED      N       Y        N 2021-04-10
#> 6       NOT RECOVERED/NOT RESOLVED      N       Y        Y 2021-05-24
#>       AENDTM ASTDY AENDY AETOXGR SMQ01NAM SMQ02NAM SMQ01SC SMQ02SC
#> 1 2022-04-05   251   517       3     <NA>     <NA>    <NA>    <NA>
#> 2 2022-05-16   304   558       3     <NA>     <NA>    <NA>    <NA>
#> 3 2022-10-29   496   724       2     <NA>     <NA>    <NA>    <NA>
#> 4 2022-09-08   608   673       2     <NA>     <NA>    <NA>    <NA>
#> 5 2021-04-17   259   266       3     <NA>     <NA>    <NA>    <NA>
#> 6 2022-07-18   303   723       1     <NA>     <NA>    <NA>    <NA>
#>                    CQ01NAM ANL01FL
#> 1                     <NA>       Y
#> 2                     <NA>       Y
#> 3                     <NA>       Y
#> 4                     <NA>       Y
#> 5                     <NA>       Y
#> 6 D.2.1.5.3/A.1.1.1.1 AESI       Y
#> 
#> ...
#>      STUDYID              USUBJID SUBJID SITEID AGE SEX  RACE COUNTRY DTHFL
#> 1929 AB12345 AB12345-USA-8-id-206 id-206  USA-8  34   F ASIAN     USA     N
#> 1930 AB12345 AB12345-USA-8-id-206 id-206  USA-8  34   F ASIAN     USA     N
#> 1931 AB12345 AB12345-USA-8-id-206 id-206  USA-8  34   F ASIAN     USA     N
#> 1932 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#> 1933 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#> 1934 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#>             INVID        INVNAM            ARM ARMCD         ACTARM ACTARMCD
#> 1929 INV ID USA-8 Dr. USA-8 Doe C: Combination ARM C C: Combination    ARM C
#> 1930 INV ID USA-8 Dr. USA-8 Doe C: Combination ARM C C: Combination    ARM C
#> 1931 INV ID USA-8 Dr. USA-8 Doe C: Combination ARM C C: Combination    ARM C
#> 1932 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 1933 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 1934 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#>            REGION1 STRATA1 STRATA2    BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL
#> 1929 North America       B      S1 14.478057   HIGH     Y     Y        Y
#> 1930 North America       B      S1 14.478057   HIGH     Y     Y        Y
#> 1931 North America       B      S1 14.478057   HIGH     Y     Y        Y
#> 1932 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#> 1933 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#> 1934 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#>      BEP01FL     RANDDT             TRTSDTM             TRTEDTM    EOSSTT
#> 1929       Y 2019-07-19 2019-07-22 10:19:01 2021-07-21 21:57:25 COMPLETED
#> 1930       Y 2019-07-19 2019-07-22 10:19:01 2021-07-21 21:57:25 COMPLETED
#> 1931       Y 2019-07-19 2019-07-22 10:19:01 2021-07-21 21:57:25 COMPLETED
#> 1932       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#> 1933       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#> 1934       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#>           EOSDT EOSDY DCSREAS DTHDT DTHCAUS DTHCAT LDDTHELD LDDTHGR1   LSTALVDT
#> 1929 2021-07-21   731    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2021-08-01
#> 1930 2021-07-21   731    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2021-08-01
#> 1931 2021-07-21   731    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2021-08-01
#> 1932 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2022-08-17
#> 1933 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2022-08-17
#> 1934 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA> 2022-08-17
#>      study_duration_secs ASEQ AESEQ        AETERM         AELLT       AEDECOD
#> 1929            63113904    6     6 trm B.2.2.3.1 llt B.2.2.3.1 dcd B.2.2.3.1
#> 1930            63113904    7     7 trm B.2.2.3.1 llt B.2.2.3.1 dcd B.2.2.3.1
#> 1931            63113904    8     8 trm C.2.1.2.1 llt C.2.1.2.1 dcd C.2.1.2.1
#> 1932            63113904    1     1 trm D.1.1.1.1 llt D.1.1.1.1 dcd D.1.1.1.1
#> 1933            63113904    2     2 trm D.1.1.4.2 llt D.1.1.4.2 dcd D.1.1.4.2
#> 1934            63113904    3     3 trm D.1.1.4.2 llt D.1.1.4.2 dcd D.1.1.4.2
#>            AEHLT     AEHLGT AEBODSYS AESOC    AESEV AESER            AEACN
#> 1929 hlt B.2.2.3 hlgt B.2.2   cl B.2  cl B     MILD     Y DOSE NOT CHANGED
#> 1930 hlt B.2.2.3 hlgt B.2.2   cl B.2  cl B     MILD     Y DOSE NOT CHANGED
#> 1931 hlt C.2.1.2 hlgt C.2.1   cl C.2  cl C MODERATE     N DOSE NOT CHANGED
#> 1932 hlt D.1.1.1 hlgt D.1.1   cl D.1  cl D   SEVERE     Y    NOT EVALUABLE
#> 1933 hlt D.1.1.4 hlgt D.1.1   cl D.1  cl D MODERATE     N DOSE NOT CHANGED
#> 1934 hlt D.1.1.4 hlgt D.1.1   cl D.1  cl D MODERATE     N DOSE NOT CHANGED
#>      AEREL                            AEOUT AESDTH TRTEMFL AECONTRT     ASTDTM
#> 1929     N       NOT RECOVERED/NOT RESOLVED      N       Y        Y 2020-07-20
#> 1930     N             RECOVERING/RESOLVING      N       Y        N 2020-09-21
#> 1931     Y RECOVERED/RESOLVED WITH SEQUELAE      N       Y        N 2020-10-08
#> 1932     Y                            FATAL      Y       Y        N 2020-08-15
#> 1933     N RECOVERED/RESOLVED WITH SEQUELAE      N       Y        N 2022-01-06
#> 1934     N             RECOVERING/RESOLVING      N       Y        Y 2022-05-23
#>          AENDTM ASTDY AENDY AETOXGR                 SMQ01NAM SMQ02NAM SMQ01SC
#> 1929 2021-06-22   364   701       1 C.1.1.1.3/B.2.2.3.1 AESI     <NA>   BROAD
#> 1930 2021-04-29   427   647       1 C.1.1.1.3/B.2.2.3.1 AESI     <NA>   BROAD
#> 1931 2021-03-19   444   606       2                     <NA>     <NA>    <NA>
#> 1932 2021-05-17    28   303       5                     <NA>     <NA>    <NA>
#> 1933 2022-02-13   537   575       3                     <NA>     <NA>    <NA>
#> 1934 2022-05-24   674   675       3                     <NA>     <NA>    <NA>
#>      SMQ02SC CQ01NAM ANL01FL
#> 1929    <NA>    <NA>       Y
#> 1930    <NA>    <NA>       Y
#> 1931    <NA>    <NA>       Y
#> 1932    <NA>    <NA>       Y
#> 1933    <NA>    <NA>       Y
#> 1934    <NA>    <NA>       Y
#> 
#> $ADTTE
#> <CDISCDatasetConnector>
#>   Inherits from: <DatasetConnector>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset: function () 
#>     get_dataset_label: function () 
#>     get_error_message: function () 
#>     get_join_keys: function () 
#>     get_keys: function () 
#>     get_parent: function () 
#>     get_pull_args: function () 
#>     get_pull_callable: function () 
#>     get_raw_data: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     get_var_r6: function () 
#>     initialize: function (dataname, pull_callable, keys, parent, code = character(0), 
#>     is_failed: function () 
#>     is_mutate_delayed: function () 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     mutate_join_keys: function (dataset, val) 
#>     pull: function (args = NULL, try = FALSE) 
#>     set_args: function (args) 
#>     set_dataset_label: function (label) 
#>     set_join_keys: function (x) 
#>     set_keys: function (keys) 
#>     set_ui_input: function (inputs = NULL) 
#>   Private:
#>     dataname: ADTTE
#>     dataset: NULL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     get_pull_code_class: function (args = NULL) 
#>     is_pulled_flag: FALSE
#>     join_keys: JoinKeys, R6
#>     keys: STUDYID USUBJID PARAMCD
#>     parent: ADSL
#>     pull_callable: CallableFunction, Callable, R6
#>     pull_internal: function (args = NULL, try = FALSE) 
#>     pull_vars: list
#>     server: function (input, output, session, data_args = NULL) 
#>     set_dataname: function (dataname) 
#>     set_failure: function (res) 
#>     set_parent: function (parent) 
#>     set_pull_callable: function (pull_callable) 
#>     set_pull_vars: function (pull_vars) 
#>     set_ui: function (ui_args = NULL) 
#>     set_var_r6: function (vars) 
#>     ui: function (id) 
#>     ui_input: NULL
#>     var_r6: list
#> 

tc$get_code()

#> [1] "ADSL <- (function() synthetic_cdisc_data(\"latest\")$adsl)()\nADRS <- (function() synthetic_cdisc_data(\"latest\")$adrs)()\nlibrary(scda)\nADAE <- synthetic_cdisc_data(\"latest\")$adae\nADTTE <- (function() synthetic_cdisc_data(\"latest\")$adtte)()"

tc$get_code("ADAE")

#> [1] "library(scda)\nADAE <- synthetic_cdisc_data(\"latest\")$adae"

if (FALSE) {
tc$launch()
tc$get_datasets()
tc$get_dataset("ADAE")
tc$check()
}

library(scda)
x <- cdisc_dataset(
  dataname = "ADSL",
  x = synthetic_cdisc_data("latest")$adsl,
  code = "library(scda)\nADSL <- synthetic_cdisc_data(\"latest\")$adsl"
)

x2 <- cdisc_dataset_connector("ADTTE", adtte_cf, keys = get_cdisc_keys("ADTTE"))
tc <- teal:::RelationalData$new(x, x2)
if (FALSE) {
tc$get_datasets()
get_raw_data(tc)
tc$launch()
get_raw_data(tc)
}