RelationalDataCollection class

Abstract class containing code for handling set of datasets. This cannot be created directly. Use either RelationalDataConnector (for a set of datasets to be pulled from an external source) or RelationalData (for a set of datasets to be passed into a teal application)

Methods

Public methods

RelationalDataCollection$new()

RelationalDataCollection$check()

RelationalDataCollection$get_check_result()

RelationalDataCollection$get_code()

RelationalDataCollection$get_code_class()

RelationalDataCollection$get_datanames()

RelationalDataCollection$get_dataset()

RelationalDataCollection$get_datasets()

RelationalDataCollection$get_items()

RelationalDataCollection$is_pulled()

RelationalDataCollection$mutate()

RelationalDataCollection$mutate_dataset()

RelationalDataCollection$set_check()

RelationalDataCollection$set_pull_code()

RelationalDataCollection$clone()

Method new()

Cannot create a RelationalDataCollection object Create a RelationalData or
RelationalDataConnector object instead

Usage

RelationalDataCollection$new()

Returns

throws error

Method check()

Check if the object raw data is reproducible from the get_code() code.

Usage

RelationalDataCollection$check()

Returns

NULL if check step has been disabled TRUE if all the datasets generated from evaluating the get_code() code are identical to the raw data, else FALSE.

Method get_check_result()

Get result of reproducibility check

Usage

RelationalDataCollection$get_check_result()

Returns

NULL if check has not been called yet, TRUE / FALSE otherwise

Method get_code()

Get code for all datasets.

Usage

RelationalDataCollection$get_code(dataname = NULL, deparse = TRUE)

Arguments

dataname

(character) dataname or NULL for all datasets

deparse

(logical) whether to return the deparsed form of a call

Returns

(character) vector of code to generate datasets.

Method get_code_class()

Get internal CodeClass object

Usage

RelationalDataCollection$get_code_class(only_pull = FALSE)

Arguments

only_pull

(logical value)
if TRUE only code to pull datasets will be returned without mutate code.

Returns

CodeClass

Method get_datanames()

Get names of the datasets.

Usage

RelationalDataCollection$get_datanames()

Returns

character vector with names of all datasets.

Method get_dataset()

Get RelationalDataset object.

Usage

RelationalDataCollection$get_dataset(dataname = NULL)

Arguments

dataname

(character value)
name of dataset to be returned. If NULL, all datasets are returned.

Returns

RelationalDataset.

Method get_datasets()

Get list of RelationalDataset objects.

Usage

RelationalDataCollection$get_datasets()

Returns

list of RelationalDataset.

Method get_items()

Get all datasets and all dataset connectors

Usage

RelationalDataCollection$get_items(dataname = NULL)

Arguments

dataname

(character value)
name of dataset connector to be returned. If NULL, all connectors are returned.

Returns

list with all datasets and all connectors

Method is_pulled()

Check if dataset has already been pulled.

Usage

RelationalDataCollection$is_pulled()

Returns

TRUE if dataset has been already pulled, else FALSE

Method mutate()

Mutate data by code

Usage

RelationalDataCollection$mutate(code, vars = list())

Arguments

code

(character) Code to mutate the dataset. Must contain the dataset$dataname

vars

(list)
In case when this object code depends on the raw_data from the other RelationalDataset, RelationalDatasetConnector object(s) or other constant value, this/these object(s) should be included

Returns

self invisibly for chaining

Method mutate_dataset()

Mutate dataset by code

Usage

RelationalDataCollection$mutate_dataset(dataname, code, vars = list())

Arguments

dataname

(character) Dataname to be mutated

code

(character) Code to mutate the dataset. Must contain the dataset$dataname

vars

(list)
In case when this object code depends on the raw_data from the other RelationalDataset, RelationalDatasetConnector object(s) or other constant value, this/these object(s) should be included

Returns

self invisibly for chaining

Method set_check()

Set reproducibility check

Usage

RelationalDataCollection$set_check(check = FALSE)

Arguments

check

(logical) whether to perform reproducibility check.

Returns

self invisibly for chaining.

Method set_pull_code()

Set pull code

Usage

RelationalDataCollection$set_pull_code(code)

Arguments

code

(character value)
code to reproduce data in RawDataset objects. Can't be set if any dataset has code set already.

Returns

self invisibly for chaining.

Method clone()

The objects of this class are cloneable with this method.

Usage

RelationalDataCollection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.