Manage multiple and DatasetConnector of the same type.

Class manages DatasetConnector to specify additional dynamic arguments and to open/close connection.

Super class

teal::DataAbstract -> RelationalDataConnector

Methods

Public methods

RelationalDataConnector$new()

RelationalDataConnector$print()

RelationalDataConnector$get_connection()

RelationalDataConnector$get_code_class()

RelationalDataConnector$get_server()

RelationalDataConnector$get_preopen_server()

RelationalDataConnector$get_ui()

RelationalDataConnector$set_pull_args()

RelationalDataConnector$set_ui()

RelationalDataConnector$set_server()

RelationalDataConnector$set_preopen_server()

RelationalDataConnector$pull()

RelationalDataConnector$launch()

RelationalDataConnector$mutate()

RelationalDataConnector$is_failed()

RelationalDataConnector$clone()

Inherited methods

teal::DataAbstract$check()

teal::DataAbstract$check_reproducibility()

teal::DataAbstract$execute_mutate()

teal::DataAbstract$get_check_result()

teal::DataAbstract$get_code()

teal::DataAbstract$get_datanames()

teal::DataAbstract$get_dataset()

teal::DataAbstract$get_datasets()

teal::DataAbstract$get_items()

teal::DataAbstract$is_pulled()

teal::DataAbstract$mutate_dataset()

teal::DataAbstract$set_check()

teal::DataAbstract$set_pull_code()

Method new()

Create a new RelationalDataConnector object Prints this RelationalDataConnector.

Usage

RelationalDataConnector$new(connection, connectors)

Arguments

connection

(DataConnection)
connection to data source

connectors

(list of DatasetConnector elements)
list with dataset connectors

Method print()

Usage

RelationalDataConnector$print(...)

Arguments

...

additional arguments to the printing method

Returns

invisibly self

Method get_connection()

Get connection to data source

Usage

RelationalDataConnector$get_connection()

Returns

connector's connection

Method get_code_class()

Get internal CodeClass object

Usage

RelationalDataConnector$get_code_class()

Returns

\code{CodeClass}

Method get_server()

Usage

RelationalDataConnector$get_server()

Returns

the server function

Method get_preopen_server()

Usage

RelationalDataConnector$get_preopen_server()

Returns

the server function

Method get_ui()

Get Shiny module with inputs for all DatasetConnector objects

Usage

RelationalDataConnector$get_ui(id)

Arguments

id

character shiny element id

Returns

the ui function

Method set_pull_args()

Set argument to the pull_fun

Usage

RelationalDataConnector$set_pull_args(args)

Arguments

args

(named list)
arguments values as separate list elements named by argument name. These arguments are passed to each dataset.

Returns

nothing

Method set_ui()

Set connector UI function

Usage

RelationalDataConnector$set_ui(f)

Arguments

f

(function)
shiny module as function. Inputs specified in this ui are passed to server module defined by set_server method.

Returns

nothing

Method set_server()

Set connector server function

This function will be called after submit button will be hit. There is no possibility to specify some dynamic ui as server function is executed after hitting submit button.

Usage

RelationalDataConnector$set_server(f)

Arguments

f

(function)
A shiny module server function that should load data from all connectors

Returns

nothing

Method set_preopen_server()

Set connector pre-open server function

This function will be called before submit button will be hit.

Usage

RelationalDataConnector$set_preopen_server(f)

Arguments

f

(function)
A shiny module server function

Returns

nothing

Method pull()

Load data from each DatasetConnector

Usage

RelationalDataConnector$pull(con_args = NULL, args = NULL, try = TRUE)

Arguments

con_args

(NULL or named list)
additional dynamic arguments for connection function. args will be passed to each DatasetConnector object to evaluate CallableFunction assigned to this dataset. If args is null than default set of arguments will be used, otherwise call will be executed on these arguments only (arguments set before will be ignored). pull function doesn't update reproducible call, it's just evaluate function.

args

(NULL or named list)
additional dynamic arguments to pull dataset. args will be passed to each DatasetConnector object to evaluate CallableFunction assigned to this dataset. If args is null than default set of arguments will be used, otherwise call will be executed on these arguments only (arguments set before will be ignored). pull function doesn't update reproducible call, it's just evaluate function.

try

(logical value)
whether perform function evaluation inside try clause

Returns

(self) invisibly for chaining. In order to get the data please use get_datasets method.

Method launch()

Run simple application that uses its ui and server fields to pull data from connection.

Useful for debugging

Usage

RelationalDataConnector$launch()

Returns

An object that represents the app

Method mutate()

Mutate data by code.

Usage

RelationalDataConnector$mutate(...)

Arguments

...

parameters inherited from DataAbstract.

Returns

Informational message to not use mutate_data() with RelationalDataConnectors.

Method is_failed()

Check if pull or connection has not failed.

Usage

RelationalDataConnector$is_failed()

Returns

TRUE if pull or connection failed, else FALSE

Method clone()

The objects of this class are cloneable with this method.

Usage

RelationalDataConnector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(scda)
adsl <- scda_cdisc_dataset_connector(dataname = "ADSL", "adsl")
adlb <- scda_cdisc_dataset_connector(dataname = "ADLB", "adlb")

open_fun <- callable_function(library)
open_fun$set_args(list(package = "scda"))

con <- data_connection(open_fun = open_fun)
con$set_open_server(
  function(input, output, session, connection) {
    connection$open(try = TRUE)
    return(invisible(connection))
  }
)

x <- teal:::RelationalDataConnector$new(connection = con, connectors = list(adsl, adlb))

x$set_ui(
  function(id, connection, connectors) {
    ns <- NS(id)
    tagList(
      connection$get_open_ui(ns("open_connection")),
      textInput(ns("name"), p("Choose", code("scda data version")), value = "latest"),
      do.call(
        what = "tagList",
        args = lapply(
          connectors,
          function(connector) {
            div(
              connector$get_ui(
                id = ns(connector$get_dataname())
              ),
              br()
            )
          }
        )
      )
    )
  }
)

x$set_server(
  function(input, output, session, connection, connectors) {
    # opens connection
    callModule(connection$get_open_server(),
               id = "open_connection",
               connection = connection
    )
    if (connection$is_opened()) {
      for (connector in connectors) {
        set_args(connector, args = list(name = input$name))
        # pull each dataset
        callModule(connector$get_server(), id = connector$get_dataname())
        if (connector$is_failed()) {
          break
        }
      }
    }
  }
)
if (FALSE) {
x$launch()
x$get_datasets()
}