Initialize SEFilterStates object

Initialize SEFilterStates object

Initialize SEFilterStates object

Details

Initialize SEFilterStates object

Super class

teal::FilterStates -> SEFilterStates

Methods

Public methods

SEFilterStates$new()

SEFilterStates$server()

SEFilterStates$get_filter_state()

SEFilterStates$set_filter_state()

SEFilterStates$remove_filter_state()

SEFilterStates$ui_add_filter_state()

SEFilterStates$srv_add_filter_state()

SEFilterStates$clone()

Inherited methods

teal::FilterStates$get_call()

teal::FilterStates$get_fun()

teal::FilterStates$queue_empty()

teal::FilterStates$queue_get()

teal::FilterStates$queue_initialize()

teal::FilterStates$queue_push()

teal::FilterStates$queue_remove()

teal::FilterStates$ui()

Method new()

Usage

SEFilterStates$new(input_dataname, output_dataname, datalabel)

Arguments

input_dataname

(character(1) or name or call)
name of the data used on lhs of the expression specified to the function argument attached to this FilterStates.

output_dataname

(character(1) or name or call)
name of the output data on the lhs of the assignment expression.

datalabel

(character(0) or character(1))
text label value.

Method server()

Server module

Usage

SEFilterStates$server(id)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

Returns

moduleServer function which returns NULL

Method get_filter_state()

Gets the reactive values from the active FilterState objects.

Gets all active filters from this dataset in form of the nested list. The output list is a compatible input to self$set_filter_state.

Usage

SEFilterStates$get_filter_state()

Returns

list containing one or two lists depending on the number of ReactiveQueue object (I.e. if rowData and colData exist). Each list contains elements number equal to number of active filter variables.

Method set_filter_state()

Set filter state

Usage

SEFilterStates$set_filter_state(data, state, ...)

Arguments

data

(SummarizedExperiment)
data which are supposed to be filtered.

state

(named list)
this list should contain subset and select element where each should be a named list containing values as a selection in the FilterState. Names of each the list element in subset and select should correspond to the name of the column in rowData(data) and colData(data).

...

ignored.

Returns

NULL

Method remove_filter_state()

Remove a variable from the ReactiveQueue and its corresponding UI element.

Usage

SEFilterStates$remove_filter_state(element_id)

Arguments

element_id

(character(1))
name of ReactiveQueue element.

Returns

NULL

Method ui_add_filter_state()

Shiny UI module to add filter variable

Usage

SEFilterStates$ui_add_filter_state(id, data)

Arguments

id

(character(1))
id of shiny module

data

(SummarizedExperiment)
object containing colData and rowData which columns are used to choose filter variables. Column selection from colData and rowData are separate shiny entities.

Returns

shiny.tag

Method srv_add_filter_state()

Shiny server module to add filter variable

Module controls available choices to select as a filter variable. Selected filter variable is being removed from available choices. Removed filter variable gets back to available choices. This module unlike other FilterStates classes manages two sets of filter variables - one for colData and another for rowData.

Usage

SEFilterStates$srv_add_filter_state(id, data, ...)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

data

(SummarizedExperiment)
object containing colData and rowData which columns are used to choose filter variables. Column selection from colData and rowData are separate shiny entities.

...

ignored

Returns

moduleServer function which returns NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

SEFilterStates$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.