RcodeBlock

RcodeBlock

RcodeBlock

Super classes

teal.reporter::ContentBlock -> teal.reporter::TextBlock -> TealSlicesBlock

Methods

Public methods

TealSlicesBlock$new()

TealSlicesBlock$set_content()

TealSlicesBlock$from_list()

TealSlicesBlock$to_list()

TealSlicesBlock$clone()

Inherited methods

teal.reporter::ContentBlock$get_content()

teal.reporter::TextBlock$get_available_styles()

teal.reporter::TextBlock$get_style()

teal.reporter::TextBlock$set_style()

Method new()

Returns a TealSlicesBlock object.

Usage

TealSlicesBlock$new(content = teal_slices(), style = "verbatim")

Arguments

content

(teal_slices) object returned from teal_slices() function.

style

(character(1)) string specifying style to apply.

Details

Returns a TealSlicesBlock object with no content and no parameters.

Returns

Object of class TealSlicesBlock, invisibly.

Method set_content()

Sets content of this TealSlicesBlock. Sets content as YAML text which represents a list generated from teal_slices. The list displays limited number of fields from teal_slice objects, but this list is sufficient to conclude which filters were applied. When selected field in teal_slice object is a range, then it is displayed as a "min"

Usage

TealSlicesBlock$set_content(content)

Arguments

content

(teal_slices) object returned from teal_slices() function.

Returns

self, invisibly.

Method from_list()

Create the RcodeBlock from a list.

Usage

TealSlicesBlock$from_list(x)

Arguments

x

(named list) with two fields c("text", "params"). Use the get_available_params method to get all possible parameters.

Returns

self, invisibly.

Method to_list()

Convert the RcodeBlock to a list.

Usage

TealSlicesBlock$to_list()

Returns

named list with a text and params.

Method clone()

The objects of this class are cloneable with this method.

Usage

TealSlicesBlock$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.