Convert a Raw<...> object to a Relational<...> object

Convert a Raw<...> to a Relational<...>

as_relational(x, dataname, keys, label)

# S3 method for RawDataset
as_relational(x, dataname, keys, label = character(0))

# S3 method for NamedDataset
as_relational(x, dataname, keys, label = character(0))

# S3 method for RawDatasetConnector
as_relational(x, dataname, keys, label = character(0))

as_cdisc_relational(x, dataname, label = character(0))

Arguments

x
(RawDataset, NamedDataset, RawDatasetConnector)

dataname
(character)
A given name for the dataset it may not contain spaces

keys
(keys)
object of S3 class keys containing foreign, primary keys and parent information

label
(character)
Label to describe the dataset.

Value

(RelationalDataset or RelationalDatasetConnector) object

Details

as_cdisc_relational will derive the keys by the dataname and therefore does not need the keys argument to be specified

Examples

library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_raw <- raw_dataset(x = ADSL)

ADSL_relational <- as_relational(
  ADSL_raw,
  dataname = "ADSL",
  keys = keys(primary = c("USUBJID", "STUDYID"), foreign = NULL, parent = NULL)
)
get_raw_data(ADSL_relational)

#> # A tibble: 400 x 33
#>    STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY INVID INVNAM ARM  
#>    <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <chr> <chr>  <fct>
#>  1 AB12345 AB1234… id-128 CHN-3     32 M     ASIAN CHN     INV … Dr. C… A: D…
#>  2 AB12345 AB1234… id-262 CHN-15    35 M     BLAC… CHN     INV … Dr. C… C: C…
#>  3 AB12345 AB1234… id-378 RUS-3     30 F     ASIAN RUS     INV … Dr. R… C: C…
#>  4 AB12345 AB1234… id-220 CHN-11    26 F     ASIAN CHN     INV … Dr. C… B: P…
#>  5 AB12345 AB1234… id-267 CHN-7     40 M     ASIAN CHN     INV … Dr. C… B: P…
#>  6 AB12345 AB1234… id-201 CHN-15    49 M     ASIAN CHN     INV … Dr. C… C: C…
#>  7 AB12345 AB1234… id-45  USA-1     34 F     ASIAN USA     INV … Dr. U… A: D…
#>  8 AB12345 AB1234… id-261 USA-1     32 F     ASIAN USA     INV … Dr. U… B: P…
#>  9 AB12345 AB1234… id-173 NGA-11    24 F     BLAC… NGA     INV … Dr. N… C: C…
#> 10 AB12345 AB1234… id-307 CHN-1     24 M     ASIAN CHN     INV … Dr. C… B: P…
#> # … with 390 more rows, and 22 more variables: ARMCD <fct>, ACTARM <fct>,
#> #   ACTARMCD <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>, EOSSTT <fct>, EOSDT <date>,
#> #   EOSDY <int>, DCSREAS <fct>, DTHDT <date>, LSTALVDT <date>,
#> #   study_duration_secs <dbl>

ADSL_relational$get_keys()

#> $primary
#> [1] "USUBJID" "STUDYID"
#> 
#> $foreign
#> NULL
#> 
#> $parent
#> NULL
#> 
#> attr(,"class")
#> [1] "keys"

library(random.cdisc.data)
fun <- callable_function(radsl)
fun$set_args(list(N = 5, seed = 1, cached = TRUE))
x <- raw_dataset_connector(fun)

x2 <- as_relational(x, dataname = "ADSL", keys = get_cdisc_keys("ADSL"))