Dataflow

{r dataflow_mermaid1, echo=FALSE}
DiagrammeR::mermaid("
%%| fig-width: 7.5
graph
    autonumber
    participant data
    participant filters
    participant filtered data
    participant teal module
    data->filters: gets
    filters->>filtered data: becomes
    filtered data->>teal module: sent to
")

The sequence diagram above illustrates the different stages that data goes through within the teal framework, supported by the teal.slice R package:

Users (or app developers) select which filters to apply;

See the Filter panel vignette for details;

The R code is automatically generated from the selected filters;

Creates a qenv R object that contains the code that will be applied to the data (and used for reproducibility);

Apply the filters to the data (creating a filtered data instance);

A module can have a custom set of filters that is only applied to it;

The filtered data is sent to the teal modules as the main data source for analysis.

Whenever filters are added or removed, these stages are restarted, providing the teal module with new filtered data to conduct the required analysis.