Compare bookmarks.

Test if two bookmarks store identical state.

Usage

bookmarks_identical(book1, book2)

Arguments

book1, book2

bookmark directories stored in shiny_bookmarks/; default to the two most recently modified directories

Value

Invisible NULL if bookmarks are identical or if there are no bookmarks to test. FALSE if inconsistencies are detected.

Details

input environments are compared one variable at a time and if not identical, values in both bookmarks are reported. States of datatables are stripped of the time element before comparing because the time stamp is always different. The contents themselves are not printed as they are large and the contents are not informative. Elements present in one bookmark and absent in the other are also reported. Differences are printed as messages.

values environments are compared with all.equal.

How to use

Open an application, change relevant inputs (typically, all of them), and create a bookmark. Then open that bookmark and immediately create a bookmark of that. If restoring bookmarks occurred properly, the two bookmarks should store the same state.