Create a call using a function in a given namespace

The arguments in ... need to be quoted because they will be evaluated otherwise

call_with_colon(name, ..., unlist_args = list())

Arguments

name

character function name, possibly using namespace colon ::, also works with ::: (sometimes needed, but strongly discouraged)

...

arguments to pass to function with name name

unlist_args

list extra arguments passed in a single list, avoids the use of do.call with this function

Examples

print_call_and_eval <- function(x) {
  eval(print(x))
}

print_call_and_eval(
  call_with_colon("glue::glue", "x = {x}", x = 10)
)
#> glue::glue("x = {x}", x = 10)
#> x = 10
if (FALSE) {
# mtcars$cyl evaluated
print_call_and_eval(
  call_with_colon("dplyr::filter", as.name("mtcars"), mtcars$cyl == 6)
)

# mtcars$cyl argument not evaluated immediately (in call expression)
print_call_and_eval(
  call_with_colon("dplyr::filter", as.name("mtcars"), quote(cyl == 6))
)

# does not work because argument is evaluated and the
# non-dplyr filter does not look inside mtcars
# cannot eval becausee it does not pass checks because of non-standard evaluation
call("filter", as.name("mtcars"), quote(cyl == 6))
# works, but non-dplyr filter is taken
call("filter", as.name("mtcars"), mtcars$cyl == 6)

nb_args <- function(...) nargs()
print_call_and_eval(
  call_with_colon("nb_args", arg1 = 1, unlist_args = list(arg2 = 2, args3 = 3))
)
# duplicate arguments
print_call_and_eval(
  call_with_colon("nb_args", arg1 = 1, unlist_args = list(arg2 = 2, args2 = 2))
)
}