Data input for teal app

Function passes datasets to teal application with option to read preprocessing code and reproducibility checking.

cdisc_data(..., join_keys, code = "", check = FALSE)

Arguments

...
(RelationalDataConnector, Dataset or DatasetConnector) elements to include where ADSL data is mandatory.

join_keys
(JoinKeys) or a single (JoinKeySet)
(optional) object with datasets column names used for joining. If empty then it would be automatically derived basing on intersection of datasets primary keys

code
(character) code to reproduce the datasets.

check
(logical) reproducibility check - whether evaluated preprocessing code gives the same objects as provided in arguments. Check is run only if flag is true and preprocessing code is not empty.

Value

a CDISCData object

Details

This function checks if there were keys added to all data sets that shall be analyzed inside a teal app.

Note

This function does not automatically assign keys to Dataset and DatasetConnector objects passed to it. If the keys are needed they should be assigned before calling cdisc_data. See example:

Examples

library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl
ADTTE <- synthetic_cdisc_data("latest")$adtte

# basic example
cdisc_data(
  cdisc_dataset("ADSL", ADSL),
  cdisc_dataset("ADTTE", ADTTE),
  code = 'ADSL <- synthetic_cdisc_data("latest")$adsl
          ADTTE <- synthetic_cdisc_data("latest")$adtte',
  check = TRUE
)

#> <CDISCData>
#>   Inherits from: <RelationalData>
#>   Public:
#>     check: function () 
#>     check_metadata: function () 
#>     check_reproducibility: function () 
#>     clone: function (deep = FALSE) 
#>     execute_mutate: function () 
#>     get_check_result: function () 
#>     get_code: function (dataname = NULL, deparse = TRUE) 
#>     get_code_class: function (only_pull = FALSE) 
#>     get_connectors: function () 
#>     get_datanames: function () 
#>     get_dataset: function (dataname = NULL) 
#>     get_datasets: function () 
#>     get_items: function (dataname = NULL) 
#>     get_join_keys: function () 
#>     get_parent: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     id: 693632481
#>     initialize: function (..., check = FALSE, join_keys) 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     mutate_dataset: function (dataname, code, vars = list()) 
#>     mutate_join_keys: function (dataset_1, dataset_2, val) 
#>     set_check: function (check = FALSE) 
#>     set_pull_code: function (code) 
#>   Private:
#>     .check: TRUE
#>     check_combined_code: function () 
#>     check_names: function (x) 
#>     check_result: TRUE
#>     datasets: list
#>     deep_clone: function (name, value) 
#>     get_datasets_code_class: function () 
#>     get_mutate_code_class: function () 
#>     get_pull_code_class: function () 
#>     join_keys: NULL
#>     mutate_code: CodeClass, R6
#>     mutate_vars: list
#>     parent: list
#>     pull_code: CodeClass, R6
#>     server: function (input, output, session) 
#>     set_mutate_code: function (code, dataname = self$get_datanames(), deps = names(private_mutate_vars)) 
#>     set_mutate_vars: function (vars) 
#>     ui: function (id) 

# Example with keys
cdisc_data(
  cdisc_dataset("ADSL", ADSL,
    keys = c("STUDYID", "USUBJID")
  ),
  cdisc_dataset("ADTTE", ADTTE,
    keys = c("STUDYID", "USUBJID", "PARAMCD"),
    parent = "ADSL"
  ),
  join_keys = join_keys(
    join_key(
      "ADSL",
      "ADTTE",
      c("STUDYID" = "STUDYID", "USUBJID" = "USUBJID")
    )
  ),
  code = 'ADSL <- synthetic_cdisc_data("latest")$adsl
          ADTTE <- synthetic_cdisc_data("latest")$adtte',
  check = TRUE
)

#> <CDISCData>
#>   Inherits from: <RelationalData>
#>   Public:
#>     check: function () 
#>     check_metadata: function () 
#>     check_reproducibility: function () 
#>     clone: function (deep = FALSE) 
#>     execute_mutate: function () 
#>     get_check_result: function () 
#>     get_code: function (dataname = NULL, deparse = TRUE) 
#>     get_code_class: function (only_pull = FALSE) 
#>     get_connectors: function () 
#>     get_datanames: function () 
#>     get_dataset: function (dataname = NULL) 
#>     get_datasets: function () 
#>     get_items: function (dataname = NULL) 
#>     get_join_keys: function () 
#>     get_parent: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     id: 138623016
#>     initialize: function (..., check = FALSE, join_keys) 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     mutate_dataset: function (dataname, code, vars = list()) 
#>     mutate_join_keys: function (dataset_1, dataset_2, val) 
#>     set_check: function (check = FALSE) 
#>     set_pull_code: function (code) 
#>   Private:
#>     .check: TRUE
#>     check_combined_code: function () 
#>     check_names: function (x) 
#>     check_result: TRUE
#>     datasets: list
#>     deep_clone: function (name, value) 
#>     get_datasets_code_class: function () 
#>     get_mutate_code_class: function () 
#>     get_pull_code_class: function () 
#>     join_keys: NULL
#>     mutate_code: CodeClass, R6
#>     mutate_vars: list
#>     parent: list
#>     pull_code: CodeClass, R6
#>     server: function (input, output, session) 
#>     set_mutate_code: function (code, dataname = self$get_datanames(), deps = names(private_mutate_vars)) 
#>     set_mutate_vars: function (vars) 
#>     ui: function (id)