Create a new object of CDISCTealDataset class

Function that creates CDISCTealDataset object

cdisc_dataset(
  dataname,
  x,
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0) else "ADSL",
  label = data_label(x),
  code = character(0),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(data.frame or MultiAssayExperiment)

keys

optional, (character)
vector with primary keys

parent

optional, (character)
parent dataset name

label

(character)
Label to describe the dataset

code

(character)
A character string defining the code needed to produce the data set in x

vars

(named list))
In case when this object code depends on other TealDataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include TealDataset or TealDatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Value

(CDISCTealDataset) a dataset with connected metadata

Examples

library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl

cdisc_dataset("ADSL", ADSL)
#> A CDISCTealDataset object containing the following data.frame (400 rows and 39 columns):
#>   STUDYID               USUBJID SUBJID SITEID AGE SEX                      RACE
#> 1 AB12345  AB12345-CHN-3-id-128 id-128  CHN-3  32   M                     ASIAN
#> 2 AB12345 AB12345-CHN-15-id-262 id-262 CHN-15  35   M BLACK OR AFRICAN AMERICAN
#> 3 AB12345  AB12345-RUS-3-id-378 id-378  RUS-3  30   F                     ASIAN
#> 4 AB12345 AB12345-CHN-11-id-220 id-220 CHN-11  26   F                     ASIAN
#> 5 AB12345  AB12345-CHN-7-id-267 id-267  CHN-7  40   M                     ASIAN
#> 6 AB12345 AB12345-CHN-15-id-201 id-201 CHN-15  49   M                     ASIAN
#>   COUNTRY DTHFL         INVID         INVNAM            ARM ARMCD
#> 1     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A
#> 2     CHN     N INV ID CHN-15 Dr. CHN-15 Doe C: Combination ARM C
#> 3     RUS     N  INV ID RUS-3  Dr. RUS-3 Doe C: Combination ARM C
#> 4     CHN     N INV ID CHN-11 Dr. CHN-11 Doe     B: Placebo ARM B
#> 5     CHN     N  INV ID CHN-7  Dr. CHN-7 Doe     B: Placebo ARM B
#> 6     CHN     N INV ID CHN-15 Dr. CHN-15 Doe C: Combination ARM C
#>           ACTARM ACTARMCD REGION1 STRATA1 STRATA2    BMRKR1 BMRKR2 ITTFL SAFFL
#> 1      A: Drug X    ARM A    Asia       C      S2 14.424934 MEDIUM     Y     Y
#> 2 C: Combination    ARM C    Asia       C      S1  4.055463    LOW     Y     Y
#> 3 C: Combination    ARM C Eurasia       A      S1  2.803240   HIGH     Y     Y
#> 4     B: Placebo    ARM B    Asia       B      S2 10.262734 MEDIUM     Y     Y
#> 5     B: Placebo    ARM B    Asia       C      S1  6.206763    LOW     Y     Y
#> 6 C: Combination    ARM C    Asia       C      S2  6.906799 MEDIUM     Y     Y
#>   BMEASIFL BEP01FL     RANDDT             TRTSDTM             TRTEDTM
#> 1        Y       Y 2019-02-22 2019-02-24 11:09:18 2021-02-23 22:47:42
#> 2        N       N 2019-02-26 2019-02-26 09:05:00 2021-02-25 20:43:24
#> 3        Y       N 2019-02-24 2019-02-28 03:19:08 2021-02-27 14:57:32
#> 4        Y       Y 2019-02-27 2019-03-01 13:33:03 2021-03-01 01:11:27
#> 5        N       N 2019-03-01 2019-03-02 00:09:16 2021-03-01 11:47:40
#> 6        Y       N 2019-03-05 2019-03-05 15:23:44 2021-02-17 20:23:53
#>         EOSSTT       EOTSTT      EOSDT EOSDY          DCSREAS DTHDT DTHCAUS
#> 1    COMPLETED    COMPLETED 2021-02-23   731             <NA>  <NA>    <NA>
#> 2    COMPLETED    COMPLETED 2021-02-25   731             <NA>  <NA>    <NA>
#> 3    COMPLETED    COMPLETED 2021-02-27   731             <NA>  <NA>    <NA>
#> 4    COMPLETED    COMPLETED 2021-03-01   731             <NA>  <NA>    <NA>
#> 5    COMPLETED    COMPLETED 2021-03-01   731             <NA>  <NA>    <NA>
#> 6 DISCONTINUED DISCONTINUED 2021-02-17   716 LACK OF EFFICACY  <NA>    <NA>
#>   DTHCAT LDDTHELD LDDTHGR1   LSTALVDT study_duration_secs
#> 1   <NA>       NA     <NA> 2021-03-05            63113904
#> 2   <NA>       NA     <NA> 2021-03-15            63113904
#> 3   <NA>       NA     <NA> 2021-03-15            63113904
#> 4   <NA>       NA     <NA> 2021-03-17            63113904
#> 5   <NA>       NA     <NA> 2021-03-25            63113904
#> 6   <NA>       NA     <NA> 2021-03-01            63113904
#> 
#> ...
#>     STUDYID               USUBJID SUBJID SITEID AGE SEX
#> 395 AB12345   AB12345-CHN-9-id-11  id-11  CHN-9  28   F
#> 396 AB12345 AB12345-CHN-16-id-352 id-352 CHN-16  28   M
#> 397 AB12345  AB12345-CHN-1-id-186 id-186  CHN-1  27   M
#> 398 AB12345  AB12345-CHN-1-id-371 id-371  CHN-1  28   F
#> 399 AB12345  AB12345-CHN-1-id-233 id-233  CHN-1  36   F
#> 400 AB12345 AB12345-USA-12-id-131 id-131 USA-12  44   F
#>                                          RACE COUNTRY DTHFL         INVID
#> 395 NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER     CHN     N  INV ID CHN-9
#> 396                                     ASIAN     CHN     N INV ID CHN-16
#> 397                                     ASIAN     CHN     N  INV ID CHN-1
#> 398                                     ASIAN     CHN     Y  INV ID CHN-1
#> 399                 BLACK OR AFRICAN AMERICAN     CHN     N  INV ID CHN-1
#> 400          AMERICAN INDIAN OR ALASKA NATIVE     USA     N INV ID USA-12
#>             INVNAM            ARM ARMCD         ACTARM ACTARMCD       REGION1
#> 395  Dr. CHN-9 Doe     B: Placebo ARM B     B: Placebo    ARM B          Asia
#> 396 Dr. CHN-16 Doe C: Combination ARM C C: Combination    ARM C          Asia
#> 397  Dr. CHN-1 Doe     B: Placebo ARM B     B: Placebo    ARM B          Asia
#> 398  Dr. CHN-1 Doe      A: Drug X ARM A      A: Drug X    ARM A          Asia
#> 399  Dr. CHN-1 Doe C: Combination ARM C C: Combination    ARM C          Asia
#> 400 Dr. USA-12 Doe      A: Drug X ARM A      A: Drug X    ARM A North America
#>     STRATA1 STRATA2   BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL     RANDDT
#> 395       C      S1 4.267916 MEDIUM     Y     Y        Y       Y 2021-01-26
#> 396       A      S2 4.216812 MEDIUM     Y     Y        N       Y 2021-01-30
#> 397       A      S2 4.521179   HIGH     Y     Y        Y       N 2021-01-30
#> 398       A      S2 6.326863    LOW     Y     Y        N       Y 2021-02-07
#> 399       C      S1 4.076818    LOW     Y     Y        N       Y 2021-02-09
#> 400       B      S1 3.726935    LOW     Y     Y        Y       N 2021-02-14
#>                 TRTSDTM             TRTEDTM       EOSSTT       EOTSTT
#> 395 2021-01-27 01:57:07 2023-01-27 13:35:31    COMPLETED    COMPLETED
#> 396 2021-02-02 09:04:52 2023-02-02 20:43:16    COMPLETED    COMPLETED
#> 397 2021-02-02 19:14:46 2023-02-03 06:53:10    COMPLETED    COMPLETED
#> 398 2021-02-09 05:28:52 2021-02-19 19:06:01 DISCONTINUED DISCONTINUED
#> 399 2021-02-11 21:33:08 2021-02-17 09:12:06 DISCONTINUED DISCONTINUED
#> 400 2021-02-17 14:42:27 2023-02-18 02:20:51    COMPLETED    COMPLETED
#>          EOSDT EOSDY            DCSREAS      DTHDT DTHCAUS DTHCAT LDDTHELD
#> 395 2023-01-27   731               <NA>       <NA>    <NA>   <NA>       NA
#> 396 2023-02-02   731               <NA>       <NA>    <NA>   <NA>       NA
#> 397 2023-02-03   731               <NA>       <NA>    <NA>   <NA>       NA
#> 398 2021-02-19    11              DEATH 2021-03-29 MISSING  OTHER       38
#> 399 2021-02-17     6 PHYSICIAN DECISION       <NA>    <NA>   <NA>       NA
#> 400 2023-02-18   731               <NA>       <NA>    <NA>   <NA>       NA
#>     LDDTHGR1   LSTALVDT study_duration_secs
#> 395     <NA> 2023-02-22            63113904
#> 396     <NA> 2023-02-23            63113904
#> 397     <NA> 2023-03-03            63113904
#> 398      >30 2021-03-29            63113904
#> 399     <NA> 2021-03-17            63113904
#> 400     <NA> 2023-03-01            63113904