Ensure the ellipsis, ..., in method arguments are empty

Ellipsis, ..., are needed as part of method arguments to allow for its arguments to be different from its generic's arguments and for this to pass check(). Hence, ..., should always be empty. This function will check for this condition.

check_ellipsis(..., stop = FALSE, allowed_args = character(0))

Arguments

...

it should literally just be ...

stop

TRUE to raise an error; FALSE will output warning message

allowed_args

character vector naming arguments that are allowed in the .... to allow for unnamed arguments, let "" be one of the elements in this character vector.

Value

NULL if ... is empty

Examples

method.class <- function(a, b, c, ...) {
  check_ellipsis(...)
}
method.class <- function(a, b, c, ...) {
  check_ellipsis(..., allowed_args = c("y", "z"))
}