Check that one set is a subset of another

Raises an error message if not and says which elements are not in the allowed choices.

check_in_subset(subset, choices, pre_msg = "")

Arguments

subset
collection-like should be a subset of choices

choices
collection-like superset

pre_msg
character message to print before

Examples

check_in_subset <- teal:::check_in_subset
check_in_subset(c("a", "b"), c("a", "b", "c"))
if (FALSE) {
check_in_subset(c("a", "b"), c("b", "c"), pre_msg = "Error: ")
# truncated because too long
check_in_subset("a", LETTERS, pre_msg = "Error: ")
}