Extract raw values from choices

Extract raw values from choices

extract_raw_choices(choices, sep)

Arguments

choices
(choices_labeled or list or character) with choices

sep
(character) A separator string to split the choices or selected inputs into the values of the different columns

Value

choices simplified