Dummy module to show the filter calls generated by the right encoding panel

The Show R Code functionality is not available here because it is only defined in teal.devel.

filter_calls_module(label = "Filter Calls Module", active_datanames = "all")

Arguments

label

character label of module

active_datanames

character vector datanames shown in filter panel; can be "all" to mean all available datasets

Details

Please do not remove, this is useful for debugging teal without dependencies and simplifies \link[devtools]{load_all} which otherwise fails and avoids session restarts!

Examples

library(scda)
#> 
ADSL <- synthetic_cdisc_data("latest")$adsl
# by testing with NA values, we will see whether the filtering really works when
# we add and remove filters
ADSL$SEX[1:150] <- NA

app <- init(
  data = cdisc_data(
    teal.data::cdisc_dataset(
      dataname = "ADSL",
      x = ADSL
    ),
    code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"
  ),
  modules = modules(
    teal:::filter_calls_module()
  ),
  header = "Simple teal app"
)
if (FALSE) {
shinyApp(app$ui, app$server)
}