Function Dataset Connector

fun_dataset_connector(
  dataname,
  fun,
  fun_args = NULL,
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  func_name = substitute(fun),
  ...
)

fun_cdisc_dataset_connector(
  dataname,
  fun,
  fun_args = NULL,
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  label = character(0),
  code = character(0),
  script = character(0),
  func_name = substitute(fun),
  ...
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

fun

(function)
a custom function to obtain dataset.

fun_args

(list)
additional arguments for (func).

keys

optional, (character)
vector of dataset primary keys column names

label

(character)
Label to describe the dataset.

code

(character)
A character string defining code to modify raw_data from this dataset. To modify current dataset code should contain at least one assignment to object defined in dataname argument. For example if dataname = ADSL example code should contain ADSL <- <some R code>. Can't be used simultaneously with script

script

(character)
Alternatively to code - location of the file containing modification code. Can't be used simultaneously with script.

func_name

(name)
for internal purposes, please keep it default

...

Additional arguments applied to pull function. In case when this object code depends on the raw_data from the other TealDataset, TealDatasetConnector object(s) or other constant value, this/these object(s) should be included. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

parent

optional, (character)
parent dataset name

Details

Create a TealDatasetConnector from function and its arguments.

Create a CDISCTealDatasetConnector from function and its arguments with keys and parent name assigned automatically by dataname.

Examples

my_data <- function(...) {
  data.frame(
    ID = paste0("ABC_", seq_len(10)),
    var1 = rnorm(n = 10),
    var2 = rnorm(n = 10),
    var3 = rnorm(n = 10)
  )
}
y <- fun_dataset_connector(
  dataname = "XYZ",
  fun = my_data
)

y$get_code()
#> [1] "my_data <- function(...) {\n    data.frame(ID = paste0(\"ABC_\", seq_len(10)), var1 = rnorm(n = 10), var2 = rnorm(n = 10), var3 = rnorm(n = 10))\n}\nXYZ <- (function(...) {\n    data.frame(ID = paste0(\"ABC_\", seq_len(10)), var1 = rnorm(n = 10), var2 = rnorm(n = 10), var3 = rnorm(n = 10))\n})()"

y$pull()

get_raw_data(y)
#>        ID        var1       var2        var3
#> 1   ABC_1  1.24975052 -1.9535756  2.11073495
#> 2   ABC_2  0.25004102 -0.2224238 -1.73685556
#> 3   ABC_3 -0.28553523  0.7309128  0.76511834
#> 4   ABC_4 -0.11141620 -0.3464119 -1.28997875
#> 5   ABC_5 -1.26778732 -0.5272868  1.12745907
#> 6   ABC_6 -0.35278079  0.9019622  0.34002796
#> 7   ABC_7 -0.73574845  0.8280713  1.31621915
#> 8   ABC_8  0.62310309 -1.5066728  0.08218495
#> 9   ABC_9  1.59970393  0.3361594  0.81474384
#> 10 ABC_10 -0.03995177 -0.2534198  0.38820308