Creates a short information summary about the duplicate primary key values in a dataset

# S3 method for TealDataset
get_key_duplicates(dataset, keys = NULL)

Arguments

dataset
a TealDataset object, which will be used to detect duplicated primary keys

keys
character vector of variable names in dataset consisting the key or keys object, which does have a primary element with a vector of variable names in dataset consisting the key. Optional, default: NULL

Details

S3 method for get_key_duplicates. Uses the public API of TealDataset to read the primary key and the raw data.

If keys argument is provided, then checks against that, if it's NULL, then checks against the get_keys()$primary method of the dataset argument.

See also

Examples

library(scda)

adsl <- synthetic_cdisc_data("latest")$adsl
# Keys are automatically assigned, because the name
# is in the recognized ADAM nomenclature ("ADSL")
rel_adsl <- cdisc_dataset("ADSL", adsl)
get_key_duplicates(rel_adsl)
#> # A tibble: 0 × 4
#> # … with 4 variables: STUDYID <chr>, USUBJID <chr>, rows <chr>, n <int>