Creates a duplicate keys information summary.

Creates a duplicate keys information summary.

get_key_duplicates_util(dataframe, keys)

Arguments

dataframe

dataframe

keys

character list of variable names consisting the key to the data.frame

Value

tibble with a duplicate keys information summary

Details

Accepts a list of variable names - keys, which are treated as the key to the data.frame argument. An instance of duplicated key is defined as two rows, which have the same values in columns defined by keys. Per each key value with duplicates returns a row in a tibble. The return table has columns corresponding to the variable names passed in keys and two additional columns: row_no and duplicates_count, which provide information about row numbers of the original dataframe, which contain duplicated keys and total duplicates counts.

See also

Examples

df <- as.data.frame(
  list(a = c("a", "a", "b", "b", "c"), b = c(1, 2, 3, 3, 4), c = c(1, 2, 3, 4, 5))
)
res <- teal:::get_key_duplicates_util(df, keys = c("a", "b")) # duplicated keys are in rows 3 and 4
print(res) # outputs a tibble
#> # A tibble: 1 × 4
#>   a         b rows      n
#>   <chr> <dbl> <chr> <int>
#> 1 b         3 3,4       2