Generates header text for analysis items

get_rcode_header(title, description = NULL, libraries = NULL,
  git_repo = NULL, data = NULL)

Arguments

title
A character title of the module

description
A character description of the module with additional information not reflected in the title

libraries
A character vector of the names of libraries required for the module.

git_repo
A character vector of the link to the GitHub repository containing the R-code for the teal module. Optional argument

data
optionally either an FilteredData object, or a named list of datasets. The data sets need the import attribute and optionally the md5sum attribute.

Value

A character string for the header text

Examples

cat(get_rcode_header(
 title = "Cross-Table",
 description = "A cross-table counts ..."
))

#> # Cross-Table
#> # 
#> # A cross-table counts ...
#> # 
#> # Module Source: Not given
#> # Git-commit Shiny App: -
#> # Date: Fri Jan 19 15:04:21 2018
#> # 
#> # You can run this code interactively in http://r.roche.com
#> 

cat(get_rcode_header(
 title = "Cross-Table",
 description = "A cross-table counts ...",
 libraries = c('haven', 'ggplot2', 'dplyr')
))

#> # Cross-Table
#> # 
#> # A cross-table counts ...
#> # 
#> # Module Source: Not given
#> # Git-commit Shiny App: -
#> # Date: Fri Jan 19 15:04:21 2018
#> # 
#> # You can run this code interactively in http://r.roche.com
#> 
#> library(haven)
#> library(ggplot2)
#> library(dplyr)
#> 

cat(get_rcode_header(
 title = "Cross-Table",
 description = "A cross-table counts ...",
 libraries = c('haven', 'ggplot2', 'dplyr'),
 git_repo = "http://github.roche.com/Rpackages/teal/",
 data = list(
   ASL = structure(data.frame(a = 1), import = "haven::read_sas('/opt/BIOSTAT/asl.sas7bdat')"),
   ATE = structure(data.frame(a = 1),
                   import = "haven::read_sas('/opt/BIOSTAT/ate.sas7bdat')",
                   md5sum = "32sdf32fds324")
 )
))

#> # Cross-Table
#> # 
#> # A cross-table counts ...
#> # 
#> # Module Source: http://github.roche.com/Rpackages/teal/
#> # Git-commit Shiny App: -
#> # Date: Fri Jan 19 15:04:21 2018
#> # 
#> # You can run this code interactively in http://r.roche.com
#> 
#> library(haven)
#> library(ggplot2)
#> library(dplyr)
#> 
#> ASL <- haven::read_sas('/opt/BIOSTAT/asl.sas7bdat')  
#> ATE <- haven::read_sas('/opt/BIOSTAT/ate.sas7bdat')  # md5sum at time of analysis: 32sdf32fds324