Initializes FilterState

Initializes FilterState depending on a variable class.

init_filter_state(
  x,
  varname,
  varlabel = if_null(attr(x, "label"), character(0)),
  input_dataname = NULL,
  extract_type = character(0)
)

Arguments

x

(vector)
values of the variable used in filter

varname

(character(1), name)
name of the variable

varlabel

(character(1))
label of the variable (optional).

input_dataname

(name or call)
name of dataset where x is taken from. Must be specified if extract_type argument is not empty.

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <input_dataname>$<varname>

"matrix" varname in the condition call will be returned as <input_dataname>[, <varname>]

Value

FilterState object

Examples

filter_state <- teal:::RangeFilterState$new(
  c(1:10, NA, Inf),
  varname = "x",
  varlabel = "Pretty name",
  input_dataname = as.name("dataname"),
  extract_type = "matrix"
)

filter_state$get_varname()
#> x
filter_state$get_varlabel()
#> [1] "Pretty name"
isolate(filter_state$get_call())
#> dataname[, "x"] >= 1 & dataname[, "x"] <= 10
if (FALSE) {
shinyApp(
  ui = fluidPage(
    isolate(filter_state$ui(id = "app")),
    verbatimTextOutput("call")
  ),
  server = function(input, output, session) {
    filter_state$server("app")

    output$call <- renderText(
      deparse1(filter_state$get_call(), collapse = "\n")
    )
  }
)
}