Initialize FilterStates object

Initialize FilterStates object

init_filter_states(
  data,
  input_dataname,
  output_dataname = input_dataname,
  datalabel = character(0),
  ...
)

Arguments

data

(data.frame, MultiAssayExperiment, SummarizedExperiment, matrix)
the R object which subset function is applied on.

input_dataname

(character(1) or name or call)
name of the data used on lhs of the expression specified to the function argument attached to this FilterStates.

output_dataname

(character(1) or name or call)
name of the output data on the lhs of the assignment expression.

datalabel

(character(0) or character(1))
text label value.

...

(optional) additional arguments for specific classes: keys

Examples

df <- data.frame(
  character = letters,
  numeric = seq_along(letters),
  date = seq(Sys.Date(), length.out = length(letters), by = "1 day"),
  datetime = seq(Sys.time(), length.out = length(letters), by = "33.33 hours")
)
rf <- teal:::init_filter_states(
  data = df,
  input_dataname = "DF",
  output_dataname = "DF_FILTERED",
  varlabels = c(
    "character variable", "numeric variable", "date variable", "datetime variable"
  )
)
if (FALSE) {
shinyApp(
  ui = fluidPage(
    actionButton("clear", span(icon("times"), "Remove all filters")),
    rf$ui_add_filter_state(id = "add", data = df),
    rf$ui("states"),
    verbatimTextOutput("expr"),
  ),
  server = function(input, output, session) {
    rf$srv_add_filter_state(id = "add", data = df)
    rf$server(id = "states")
    output$expr <- renderText({
      deparse1(rf$get_call(), collapse = "\n")
    })
    observeEvent(input$clear, rf$queue_empty())
  }
)
}