Teal Application Usage Logging.

Analysis of Teal application utilization is a key component towards ongoing enhancements of the framework. The utilization logs provide source data for better understanding frequency and extent of use.

log_app_usage(
  ta,
  molecule,
  indication,
  anl_type,
  pkg_meta = c("Package", "Title", "Version", "RemoteRef")
)

Arguments

ta
therapeutic area name. required argument. e.g. "I2ON".

molecule
molecule name. required argument. e.g. "Lucentis".

indication
indication. required argument. e.g. "AMD".

anl_type
analysis type. required argument. e.g. "Interim Analysis".

pkg_meta
package metadata fields. default values: c("Package", "Title", "Version", "RemoteRef").

Details

Each Teal application can implement usage logging by calling this function in the application startup file(s) like app.R. The following data are captured by default: UNIX ID, System Date, Application Directory, Therapeutic Area, Indication, Package Name, Package Title, Package Version and Remotes Definition. Capture frequency is currently per user at session start only. Suggested values for ta: "Oncology" or "I2ON" Suggested values for anl_type: "Exploratory", "Interim Analysis", "CSR"

Examples

if (FALSE) {
log_app_usage(ta = "Oncology", molecule = "Tecentriq", ind = "NSCLC", anl_type = "Exploratory")
readLines(file.path("logs", "utilization.log"), warn = FALSE)
}