Creates a module with a new shiny page

This function embeds a shiny application inside a teal application.

module(label, server, ui, filters, server_args = NULL, ui_args = NULL)

# S3 method for teal_module
toString(x, indent = 0, ...)

# S3 method for teal_module
print(x, ...)

Arguments

label

(character) Label shown in the navigation item for the module.

server

(function) Shiny server module function (see callModule or moduleServer).

ui

(function) Shiny ui module function (see callModule or moduleServer) with additional teal-specific datasets argument.

filters

(character) A vector with datanames that are relevant for the item. The filter panel will automatically update the shown filters to include only filters in the listed datasets. NULL will hide the filter panel, and the keyword 'all' will show the filters of all datasets. The argument can be thought of as 'active_datanames' and may be renamed in future versions of teal.

server_args

(list) Named list with additional arguments passed on to the server function. Note that the FilteredDatasets object gets automatically passed to the server function as arguments datasets.

ui_args

(list) Named list with additional arguments passed on to the ui function.

x

teal_module

indent

(integer) indent level; each submodule is indented one level more

...

parameters passed to toString

Value

object of class teal_module.

Examples

library(shiny)

app <- init(
  data = teal_data(
    dataset("iris", iris)
  ),
  modules = list(
    module(
      label = "Module",
      server = function(id, datasets) {
        moduleServer(
          id,
          module = function(input, output, session) {
            output$data <- renderDataTable(datasets$get_data("iris"))
          }
        )
      },
      ui = function(id, datasets) {
        ns <- NS(id)
        tagList(dataTableOutput(ns("data")))
      },
      filters = NULL
    )
  )
)
if (FALSE) {
runApp(app)
}