Data summary

Module and its utils to display the number of rows and subjects in the filtered and unfiltered data.

Usage

ui_data_summary(id)

srv_data_summary(id, teal_data)

get_filter_overview(teal_data)

get_object_filter_overview(
  filtered_data,
  unfiltered_data,
  dataname,
  subject_keys
)

get_object_filter_overview_array(
  filtered_data,
  unfiltered_data,
  dataname,
  subject_keys
)

get_object_filter_overview_MultiAssayExperiment(
  filtered_data,
  unfiltered_data,
  dataname
)

Arguments

id

(character(1)) shiny module instance id.

teal_data

(reactive returning teal_data)

filtered_data

(list) of filtered objects

unfiltered_data

(list) of unfiltered objects

dataname

(character(1))

Value

NULL.

Details

Handling different data classes: get_object_filter_overview() is a pseudo S3 method which has variants for:

array (data.frame, DataFrame, array, Matrix and SummarizedExperiment): Method variant can be applied to any two-dimensional objects on which ncol() can be used.

MultiAssayExperiment: for which summary contains counts for colData and all experiments.