Data Module for teal

This module manages the data argument for srv_teal. The teal framework uses teal_data(), which can be provided in various ways:

Directly as a teal.data::teal_data() object. This will automatically convert it into a reactive teal_data.

As a reactive object that returns a teal.data::teal_data() object.

Usage

ui_init_data(id, data)

srv_init_data(id, data, modules, filter = teal_slices())

Arguments

id

(character) Optionally, a string specifying the shiny module id in cases it is used as a shiny module rather than a standalone shiny app. This is a legacy feature.

data

(teal_data, teal_data_module, or reactive returning teal_data) The ui component of this module does not require data if teal_data_module is not provided. The data argument in the ui is included solely for the $ui function of the teal_data_module. Otherwise, it can be disregarded, ensuring that ui_teal does not depend on the reactive data of the enclosing application.

modules

(list or teal_modules or teal_module) Nested list of teal_modules or teal_module objects or a single teal_modules or teal_module object. These are the specific output modules which will be displayed in the teal application. See modules() and module() for more details.

filter

(teal_slices) Optionally, specifies the initial filter using teal_slices().

Value

A reactive object that returns:

teal_data when the object is validated

shiny.silent.error when not validated.

Details

Reactive teal_data:

The data in the application can be reactively updated, prompting srv_teal() to rebuild the content accordingly. There are two methods for creating interactive teal_data:

Using a reactive object provided from outside the teal application. In this scenario, reactivity is controlled by an external module, and srv_teal responds to changes.

Using teal_data_module(), which is embedded within the teal application, allowing data to be resubmitted by the user as needed.

Since the server of teal_data_module() must return a reactive teal_data object, both methods (1 and 2) produce the same reactive behavior within a teal application. The distinction lies in data control: the first method involves external control, while the second method involves control from a custom module within the app.

For more details, see module_teal_data.