Calls all modules

On the UI side each teal_modules is translated to a tabsetPanel and each teal_module is a tabPanel. Both, UI and server are called recursively so that each tab is a separate module and reflect nested structure of modules argument.

Usage

ui_teal_module(id, modules, depth = 0L)

# Default S3 method
ui_teal_module(id, modules, depth = 0L)

# S3 method for class 'teal_modules'
ui_teal_module(id, modules, depth = 0L)

# S3 method for class 'shiny.tag'
ui_teal_module(id, modules, depth = 0L)

# S3 method for class 'teal_module'
ui_teal_module(id, modules, depth = 0L)

srv_teal_module(
  id,
  data_rv,
  modules,
  datasets = NULL,
  slices_global,
  reporter = teal.reporter::Reporter$new(),
  is_active = reactive(TRUE)
)

# Default S3 method
srv_teal_module(
  id,
  data_rv,
  modules,
  datasets = NULL,
  slices_global,
  reporter = teal.reporter::Reporter$new(),
  is_active = reactive(TRUE)
)

# S3 method for class 'teal_modules'
srv_teal_module(
  id,
  data_rv,
  modules,
  datasets = NULL,
  slices_global,
  reporter = teal.reporter::Reporter$new(),
  is_active = reactive(TRUE)
)

# S3 method for class 'teal_module'
srv_teal_module(
  id,
  data_rv,
  modules,
  datasets = NULL,
  slices_global,
  reporter = teal.reporter::Reporter$new(),
  is_active = reactive(TRUE)
)

Arguments

id

(character) Optionally, a string specifying the shiny module id in cases it is used as a shiny module rather than a standalone shiny app. This is a legacy feature.

modules

(list or teal_modules or teal_module) Nested list of teal_modules or teal_module objects or a single teal_modules or teal_module object. These are the specific output modules which will be displayed in the teal application. See modules() and module() for more details.

depth

(integer(1)) number which helps to determine depth of the modules nesting.

data_rv

(reactive returning teal_data)

datasets

(reactive returning FilteredData or NULL) When datasets is passed from the parent module (srv_teal) then dataset is a singleton which implies in filter-panel to be "global". When NULL then filter-panel is "module-specific".

slices_global

(reactiveVal returning modules_teal_slices) see module_filter_manager

Value

output of currently active module.

srv_teal_module.teal_module returns reactiveVal containing output of the called module.

srv_teal_module.teal_modules returns output of module selected by input$active_tab.