Calls all modules

Module calls modules with id according to their location in the tree. There are multiple modules managing this process:

ui/srv_teal_modules_nav container module where input$active_module_id exists.

ui_teal_modules_nav_dropdown creates drop-down menu to select active module.

ui/srv_teal_module creates tab containers for each module. Only one module is visible at the time as a result of selecting in ui_teal_modules_nav_dropdown.

Usage

srv_teal_modules_nav(
  id,
  data,
  modules,
  datasets = NULL,
  slices_global,
  reporter = teal.reporter::Reporter$new(),
  data_load_status = reactive("ok")
)

ui_teal_modules_nav_dropdown(id, modules, active_module_id)

# S3 method for class 'teal_modules'
ui_teal_modules_nav_dropdown(id, modules, active_module_id)

# S3 method for class 'teal_module'
ui_teal_modules_nav_dropdown(id, modules, active_module_id)

ui_teal_module(id, modules, active_module_id)

# Default S3 method
ui_teal_module(id, modules, active_module_id)

# S3 method for class 'teal_modules'
ui_teal_module(id, modules, active_module_id)

# S3 method for class 'teal_module'
ui_teal_module(id, modules, active_module_id)

srv_teal_module(
  id,
  data,
  modules,
  datasets = NULL,
  slices_global,
  reporter = teal.reporter::Reporter$new(),
  data_load_status = reactive("ok"),
  active_module_id = reactive(TRUE)
)

# Default S3 method
srv_teal_module(
  id,
  data,
  modules,
  datasets = NULL,
  slices_global,
  reporter = teal.reporter::Reporter$new(),
  data_load_status = reactive("ok"),
  active_module_id = reactive(TRUE)
)

# S3 method for class 'teal_modules'
srv_teal_module(
  id,
  data,
  modules,
  datasets = NULL,
  slices_global,
  reporter = teal.reporter::Reporter$new(),
  data_load_status = reactive("ok"),
  active_module_id = reactive(TRUE)
)

# S3 method for class 'teal_module'
srv_teal_module(
  id,
  data,
  modules,
  datasets = NULL,
  slices_global,
  reporter = teal.reporter::Reporter$new(),
  data_load_status = reactive("ok"),
  active_module_id = reactive(TRUE)
)

Arguments

id

(character(1)) shiny module instance id.

data

(reactive returning teal_data)

modules

(teal_modules) teal_modules object. These are the specific output modules which will be displayed in the teal application. See modules() and module() for more details.

datasets

(reactive returning FilteredData or NULL) When datasets is passed from the parent module (srv_teal) then dataset is a singleton which implies the filter-panel to be "global". When NULL then filter-panel is "module-specific".

slices_global

(reactiveVal returning modules_teal_slices) see module_filter_manager

data_load_status

(reactive returning character(1)) Determines action dependent on a data loading status:

"ok" when teal_data is returned from the data loading.

"teal_data_module failed" when teal_data_module() didn't return teal_data. Disables tab buttons.

"external failed" when a reactive passed to srv_teal(data) didn't return teal_data. Hides the whole tab panel.

active_module_id

(reactive returning character(1)) id of the currently active module. This helps to determine which module can listen to reactive events.

Value

Output of currently active module.

srv_teal_module.teal_module returns reactiveVal containing output of the called module.

srv_teal_module.teal_modules returns output of module selected by drop-down.

srv_teal_modules_nav bypasses output of srv_teal_module.teal_modules.

Details

UI

On the UI side, a drop-down is created that lists buttons labeled with the labels of all teal_modules. The buttons in the drop-down are grouped according to their respective groups. These UI components are created in a way to utilize navigation-bar functionality of shiny. To achieve this, the buttons need to be placed inside a <ul class = "nav shiny-tab-input">.

Each module button has following attributes:

href = "#<teal_module container id>" links button with the container of the specific teal_module

data-bs-toggle = "tab" tells Bootstrap to hide or show teal_module's content

data-value = "<teal_module$path>" This attribute is taken by shiny from all buttons of class "active" and used as input$active_module_id in the namespace of srv_teal_modules_nav().

Each module content is wrapped in a container with the class = "tab-pane" with a unique id. This tells Bootstrap library that clicking a button should toggle relevant "tab" with an id matching button's href attribute.

Server

On the server side, teal_module(s) are called with id respective to their label and group. input$active_module_id determines which module is currently active and only this module observes reactive changes.

Initialization and isolation of the teal_module(s)

Modules are initialized only when they are active. This speeds up app initialization and on startup only the first module is activated and its outputs are calculated. Only the active module is listening to reactive events. This way, modules are isolated and only one can run at any given time. This makes the app more efficient by reducing unnecessary computations on server side.