Create a collection of module and modules object

Modules collects a tree of module() and modules() objects. This is useful to define the navigation structure of a teal app.

modules(..., label = "root")

# S3 method for teal_modules
toString(x, indent = 0, ...)

# S3 method for teal_modules
print(x, ...)

Arguments

...

parameters passed to toString

label

(character(1))
label of modules collection (default "root"). If using the label argument then it must be explicitly named. For example modules("lab", ...) should be converted to modules(label = "lab", ...).

x

teal_modules

indent

(integer) indent level; each submodule is indented one level more

Value

object of class teal_modules. Object contains following fields

label: taken from label argument

children: list containing objects passed in .... List elements are named after their label attribute converted to valid shiny id.

(character)

Examples

library(shiny)

app <- init(
  data = teal_data(
    dataset("iris", iris)
  ),
  modules = modules(
    label = "Modules",
    modules(
      label = "Module",
      module(
        label = "Inner module",
        server = function(id, datasets) {
          moduleServer(
            id,
            module = function(input, output, session) {
              output$data <- renderDataTable(datasets$get_data("iris"))
            }
          )
        },
        ui = function(id, datasets) {
          ns <- NS(id)
          tagList(dataTableOutput(ns("data")))
        },
        filters = "all"
      )
    ),
    module(
      label = "Another module",
      server = function(id, datasets) {
        moduleServer(
          id,
          module = function(input, output, session) {
            output$text <- renderText("Another module")
          }
        )
      },
      ui = function(id, datasets) {
        ns <- NS(id)
        tagList(textOutput(ns("text")))
      },
      filters = NULL
    )
  )
)
if (FALSE) {
runApp(app)
}