Constructor for Dataset class

named_dataset(
  dataname,
  x,
  code = character(0),
  label = character(0),
  vars = list()
)

Arguments

dataname
(character)
A given name for the dataset it may not contain spaces

x
(data.frame or MultiAssayExperiment)

code
(character)
A character string defining the code needed to produce the data set in x

label
(character)
Label to describe the dataset

vars
(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Value

Dataset object

Examples

# Simple example
dataset("iris", iris)

#> A Dataset object containing the following data.frame (150 rows and 5 columns):
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa
#> 
#> ...
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
#> 145          6.7         3.3          5.7         2.5 virginica
#> 146          6.7         3.0          5.2         2.3 virginica
#> 147          6.3         2.5          5.0         1.9 virginica
#> 148          6.5         3.0          5.2         2.0 virginica
#> 149          6.2         3.4          5.4         2.3 virginica
#> 150          5.9         3.0          5.1         1.8 virginica

# Example with more arguments
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
ADSL_dataset <- dataset(dataname = "ADSL", x = ADSL)

ADSL_dataset$get_dataname()

#> [1] "ADSL"

ADSL_dataset <- dataset(
  dataname = "ADSL",
  x = ADSL,
  label = "AdAM subject-level dataset",
  code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"
)

ADSL_dataset$get_dataset_label()

#> [1] "AdAM subject-level dataset"

ADSL_dataset$get_code()

#> [1] "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"