Helper function to connect to Snowflake

This is used by snowflake_connection and does not need to be called directly

snowflake_connection_function(
  username = askpass::askpass("Please enter your username"),
  password = askpass::askpass("Please enter your password"),
  role,
  database,
  schema,
  warehouse,
  server,
  port = 443,
  driver = "SnowflakeDSIIDriver",
  token_provider
)

Arguments

username
the username used to collect the auth token to connect to snowflake.

password
the password used to collect the auth token to connect to snowflake

role
the user role used to connect to Snowflake.

database
the Snowflake database to connect to.

schema
the Snowflake schema to connect to.

warehouse
the Snowflake warehouse to connect to.

server
the Snowflakeserver to connect to.

port
the port to connect to the Snowflake instance.

driver
the driver to use to connect to the Snowflake instance.

token_provider
location of the auth token provider needed to access Snowflake.