Create a simple cross-table

Create a table with the table[base] function

tm_scatterplot(label, dataname, xvar, yvar, xvar_choices = xvar,
  yvar_choices = yvar, color_by = NULL, color_by_choices = color_by,
  plot_height = c(600, 200, 2000), alpha = c(1, 0, 1), size = c(4, 1, 12),
  pre_output = NULL, post_output = NULL)

Arguments

label
label shown in the navigation for the item

xvar
variable name of x varbiable

yvar
variable name of y variable

xvar_choices
vector with variable names of possible x variables. If missing or identincal to xvar then the plot will be fixed to the xvar.

yvar_choices
vector with variable names of possible y variables. If missing or identincal to xvar then the plot will be fixed to the yvar.

color_by
variable name of variable that defines the color encoding. If NULL then no color encoding option will be displayed. Note _none_ is a keyword and means that no color encoding should be used.

color_by_choices
vector with variable names that can be used for color encodings. If missing or identical to color_by then the color encoding of the scatterplot points will be fixed to the color_by variable.

plot_height
if scalar then the plot will have a fixed height. If a slider should be presented to adjust the plot height dynamically then it can be a vector of length three with vlaue, min and max.

alpha
if scalar then the plot points will have a fixed opacity. If a slider should be presented to adjust the plot point opacity dynamically then it can be a vector of length three with vlaue, min and max.

size
if scalar then the plot points sizes will have a fixed opacity. If a slider should be presented to adjust the plot point sizes dynamically then it can be a vector of length three with vlaue, min and max.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

Examples

# NOT RUN {
x <- teal::init(
  data = list(ASL = generate_sample_data('ASL'),
              AAE = generate_sample_data('AAE')),
  root_modules(
     tm_data_table(),
     tm_variable_browser(),
     tm_scatterplot("Scatterplot Choices",
                    dataname = 'AAE',
                    xvar = 'AESDY', yvar = 'AEEDY', xvar_choices =  c('AESDY',  'AEEDY'),
                    color_by = "_none_", color_by_choices = c("_none_", "AESTMF", "ANLFL")
     ),
     tm_scatterplot("Scatterplot No Color Choices",
                    dataname = 'ASL',
                    xvar = 'AGE', yvar = 'TRTDUR', size = 3, alpha = 1, plot_height=600
     )
  )
)

shinyApp(x$ui, x$server)

# }