Create a UI of nested tabs of teal_modules

Each teal_modules is translated to a tabsetPanel and each of its children is another tab-module called recursively. The UI of a teal_module is obtained by calling the ui function on it.

ui_nested_tabs(id, modules, datasets, depth = 0L)

# S3 method for default
ui_nested_tabs(id, modules, datasets, depth = 0L)

# S3 method for teal_modules
ui_nested_tabs(id, modules, datasets, depth = 0L)

# S3 method for teal_module
ui_nested_tabs(id, modules, datasets, depth = 0L)

Arguments

id

(character(1))
module id

modules

(list or teal_modules)
nested list of teal_modules or module objects. See modules() and module() for more details.

datasets

(FilteredData)
object to store filter state and filtered datasets, shared across modules. For more details see FilteredData

depth

(integer(1))
number which helps to determine depth of the modules nesting.

Value

depending on class of modules:

teal_module: instantiated UI of the module

teal_modules: tabsetPanel with each tab corresponding to recursively calling this function on it.

Details

The datasets argument is required to resolve the teal arguments in an isolated context (with respect to reactivity)

Examples

mods <- teal:::get_dummy_modules()
datasets <- teal:::get_dummy_datasets()
app <- shinyApp(
  ui = function() {
    tagList(
      teal:::include_teal_css_js(),
      textOutput("info"),
      fluidPage( # needed for nice tabs
        teal:::ui_nested_tabs("dummy", modules = mods, datasets = datasets)
      )
    )
  },
  server = function(input, output, session) {
    active_module <- teal:::srv_nested_tabs("dummy", datasets = datasets, modules = mods)
    output$info <- renderText({
      paste0("The currently active tab name is ", active_module()$label)
    })
  }
)
if (FALSE) {
runApp(app)
}