UI to filter a single filter variable

We pass in the initial state to avoid filtering twice that otherwise comes with the updateInput functions.

ui_single_filter_item(id, filter_info, filter_state, prelabel)

Arguments

id
module id

filter_info
filter_info returned by datasets class

filter_state
filter_state returned by datasets class

prelabel
character label to append before computed label of input

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
attr(ADSL, "keys") <- get_cdisc_keys("ADSL")

datasets <- teal:::FilteredData$new()
isolate({
  datasets$set_data("ADSL", ADSL)
  datasets$set_filter_state("ADSL", list(
    AGE = list(range = c(33, 44), keep_na = FALSE)
  ))
})

#> [1] TRUE

app <- shinyApp(ui = function() {
  tagList(
    teal:::include_teal_css_js(),
    isolate(teal:::ui_single_filter_item(
      "var_AGE",
      datasets$get_filter_info("ADSL", "AGE"),
      datasets$get_filter_state("ADSL", "AGE"),
      prelabel = "ADSL.AGE"
    ))
  )
}, server = function(input, output, session) {
  callModule(teal:::srv_single_filter_item, "var_AGE", datasets, "ADSL", "AGE")
})

if (FALSE) {
runApp(app)
}