UI to show a splash screen in the beginning, then delegate to srv_teal

The splash screen could be used to query for a password to fetch the data. init is a very thin wrapper around this module useful for end-users which assumes that it is a top-level module and cannot be embedded. This function instead adheres to the Shiny module conventions.

If data is obtained through delayed loading, its splash screen is used. Otherwise, a default splash screen is shown.

Please also refer to the doc of init().

ui_teal_with_splash(
  id,
  data,
  title,
  header = tags$p("Add Title Here"),
  footer = tags$p("Add Footer Here")
)

Arguments

id

(character(1))
module id

data

(TealData or TealDataset or TealDatasetConnector or list or data.frame)
R6 object as returned by cdisc_data(), teal_data(), cdisc_dataset(), dataset(), dataset_connector() or cdisc_dataset_connector() or a single data.frame or a list of the previous objects or function returning a named list. NOTE: teal does not guarantee reproducibility of the code when names of the list elements do not match the original object names. To ensure reproducibility please use teal_data() or cdisc_data() with check = TRUE enabled.

title

(NULL or character)
The browser window title (defaults to the host URL of the page).

header

(character or shiny.tag)
the header of the app. Note shiny code placed here (and in the footer argument) will be placed in the app's ui function so code which needs to be placed in the ui function (such as loading css via htmltools::htmlDependency()) should be included here.

footer

(character or shiny.tag)
the footer of the app