Class to encapsulate relational filtered datasets with its parents.

Details

The CDISCFilteredData class implements logic to filter a relational dataset by inheriting from FilteredData. A dataset can have up to one parent dataset. Rows are identified by the foreign key and only those rows that appear in the parent dataset are kept in the filtered dataset.

The relational model is attached to the data. When set_data is called with the data, its keys attribute has the following elements:

primary: columns based on which unique rows are counted

parent: parent dataset based on which the dataset is filtered (via inner_join)

foreign: keys to join on when filtering with parent dataset

The teal UI works with objects of class FilteredData which may mix CDISC and other data (e.g. iris).

It is not too difficult to make this class generalize to more than one parent dataset, provided each dataset then contains information about its parents together with the shared keys with each of them.

See also

Super class

teal::FilteredData -> CDISCFilteredData

Methods

Public methods

CDISCFilteredData$datanames()

CDISCFilteredData$get_call()

CDISCFilteredData$get_filterable_datanames()

CDISCFilteredData$get_filterable_varnames()

CDISCFilteredData$get_parentname()

CDISCFilteredData$set_dataset()

CDISCFilteredData$clone()

Inherited methods

teal::FilteredData$get_code()

teal::FilteredData$get_data()

teal::FilteredData$get_data_attr()

teal::FilteredData$get_datalabel()

teal::FilteredData$get_filter_overview()

teal::FilteredData$get_filter_state()

teal::FilteredData$get_filtered_dataset()

teal::FilteredData$get_join_keys()

teal::FilteredData$get_keys()

teal::FilteredData$get_varlabels()

teal::FilteredData$get_varnames()

teal::FilteredData$handle_active_datanames()

teal::FilteredData$initialize()

teal::FilteredData$remove_all_filter_states()

teal::FilteredData$remove_filter_state()

teal::FilteredData$restore_state_from_bookmark()

teal::FilteredData$set_code()

teal::FilteredData$set_filter_state()

teal::FilteredData$srv_filter_overview()

teal::FilteredData$srv_filter_panel()

teal::FilteredData$ui_filter_overview()

teal::FilteredData$ui_filter_panel()

Method datanames()

Get datanames

The datanames are returned in the order in which they must be evaluated (in case of dependencies).

Usage

CDISCFilteredData$datanames()

Returns

(character vector) of datanames

Method get_call()

Returns the filter call to filter a single dataset including the inner_join with its parent dataset. It assumes that the filtered datasets it depends on are available.

Usage

CDISCFilteredData$get_call(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(call or list of calls ) to filter dataset

Method get_filterable_datanames()

Get names of datasets available for filtering

Usage

CDISCFilteredData$get_filterable_datanames(dataname)

Arguments

dataname

(character vector) names of the dataset

Returns

(character vector) of dataset names

Method get_filterable_varnames()

Gets variable names of a given dataname for the filtering. This excludes parent dataset variable names.

Usage

CDISCFilteredData$get_filterable_varnames(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(character vector) of variable names

Method get_parentname()

Get parent dataset name

Usage

CDISCFilteredData$get_parentname(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(character) name of parent dataset

Method set_dataset()

Add dataset

Add dataset and preserve all attributes attached to this object. Technically set_dataset created FilteredDataset which keeps dataset for filtering purpose.

Usage

CDISCFilteredData$set_dataset(dataset)

Arguments

dataset

(TealDataset)
object containing data and attributes.

Returns

(self) object of this class

Method clone()

The objects of this class are cloneable with this method.

Usage

CDISCFilteredData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(scda)
#> 
library(shiny)

ADSL <- synthetic_cdisc_data("latest")$adsl
ADTTE <- synthetic_cdisc_data("latest")$adtte

adsl <- cdisc_dataset("ADSL", ADSL)
adtte <- cdisc_dataset("ADTTE", ADTTE)
data <- cdisc_data(adsl, adtte)
datasets <- teal:::CDISCFilteredData$new()

# to avoid using isolate(), you can provide a default isolate context by calling
# options(shiny.suppressMissingContextError = TRUE) #nolint
# don't forget to deactivate this option at the end
# options(shiny.suppressMissingContextError = FALSE) #nolint

# setting the data
isolate({
  datasets$set_dataset(adsl)
  datasets$set_dataset(adtte)
})

isolate({
  datasets$datanames()

  # number observations and subjects of filtered/non-filtered dataset
  datasets$get_filter_overview("ADSL")

  print(datasets$get_call("ADSL"))
  print(datasets$get_call("ADTTE"))

  df <- datasets$get_data("ADSL", filtered = FALSE)
  print(df)
})
#> $filter
#> ADSL_FILTERED <- ADSL
#> 
#> $filter
#> ADTTE_FILTERED_ALONE <- ADTTE
#> 
#> [[2]]
#> ADTTE_FILTERED <- dplyr::inner_join(x = ADTTE_FILTERED_ALONE, 
#>     y = ADSL_FILTERED[, c("STUDYID", "USUBJID"), drop = FALSE], 
#>     by = c("STUDYID", "USUBJID"))
#> 
#> # A tibble: 400 × 39
#>    STUDYID USUBJID    SUBJID SITEID   AGE SEX   RACE  COUNTRY DTHFL INVID INVNAM
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct>   <fct> <chr> <chr> 
#>  1 AB12345 AB12345-C… id-128 CHN-3     32 M     ASIAN CHN     N     INV … Dr. C…
#>  2 AB12345 AB12345-C… id-262 CHN-15    35 M     BLAC… CHN     N     INV … Dr. C…
#>  3 AB12345 AB12345-R… id-378 RUS-3     30 F     ASIAN RUS     N     INV … Dr. R…
#>  4 AB12345 AB12345-C… id-220 CHN-11    26 F     ASIAN CHN     N     INV … Dr. C…
#>  5 AB12345 AB12345-C… id-267 CHN-7     40 M     ASIAN CHN     N     INV … Dr. C…
#>  6 AB12345 AB12345-C… id-201 CHN-15    49 M     ASIAN CHN     N     INV … Dr. C…
#>  7 AB12345 AB12345-U… id-45  USA-1     34 F     ASIAN USA     N     INV … Dr. U…
#>  8 AB12345 AB12345-U… id-261 USA-1     32 F     ASIAN USA     N     INV … Dr. U…
#>  9 AB12345 AB12345-N… id-173 NGA-11    24 F     BLAC… NGA     N     INV … Dr. N…
#> 10 AB12345 AB12345-C… id-307 CHN-1     24 M     ASIAN CHN     N     INV … Dr. C…
#> # … with 390 more rows, and 28 more variables: ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>,
#> #   BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>,
#> #   BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>, EOSSTT <fct>,
#> #   EOTSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>, DTHDT <date>,
#> #   DTHCAUS <fct>, DTHCAT <fct>, LDDTHELD <int>, LDDTHGR1 <fct>,
#> #   LSTALVDT <date>, study_duration_secs <dbl>

filter_state_adtte <- teal:::init_filter_state(
  ADTTE$PARAMCD,
  varname = "PARAMCD",
  input_dataname = as.name("ADTTE"),
  extract_type = "list"
)
filter_state_adtte$set_selected("OS")

states <- datasets$get_filtered_dataset("ADTTE")$get_filter_states(1)
states$queue_push(filter_state_adtte, queue_index = 1L, element_id = "PARAMCD")

isolate(datasets$get_call("ADTTE"))
#> $filter
#> ADTTE_FILTERED_ALONE <- dplyr::filter(ADTTE, ADTTE$PARAMCD == 
#>     "OS")
#> 
#> [[2]]
#> ADTTE_FILTERED <- dplyr::inner_join(x = ADTTE_FILTERED_ALONE, 
#>     y = ADSL_FILTERED[, c("STUDYID", "USUBJID"), drop = FALSE], 
#>     by = c("STUDYID", "USUBJID"))
#> 

filter_state_adsl <- teal:::init_filter_state(
  ADSL$SEX,
  varname = "SEX",
  input_dataname = as.name("SEX"),
  extract_type = "list"
)
filter_state_adsl$set_selected("F")

states <- datasets$get_filtered_dataset("ADSL")$get_filter_states("filter")
states$queue_push(filter_state_adsl, queue_index = 1L, element_id = "SEX")

isolate(datasets$get_call("ADSL"))
#> $filter
#> ADSL_FILTERED <- dplyr::filter(ADSL, SEX$SEX == "F")
#> 
isolate(datasets$get_call("ADTTE"))
#> $filter
#> ADTTE_FILTERED_ALONE <- dplyr::filter(ADTTE, ADTTE$PARAMCD == 
#>     "OS")
#> 
#> [[2]]
#> ADTTE_FILTERED <- dplyr::inner_join(x = ADTTE_FILTERED_ALONE, 
#>     y = ADSL_FILTERED[, c("STUDYID", "USUBJID"), drop = FALSE], 
#>     by = c("STUDYID", "USUBJID"))
#>