Class to encapsulate relational filtered datasets with its parents.

Details

The CDISCFilteredData class implements logic to filter a relational dataset by inheriting from FilteredData. A dataset can have up to one parent dataset. Rows are identified by the foreign key and only those rows that appear in the parent dataset are kept in the filtered dataset.

The relational model is attached to the data. When set_data is called with the data, its keys attribute has the following elements:

primary: columns based on which unique rows are counted

parent: parent dataset based on which the dataset is filtered (via inner_join)

foreign: keys to join on when filtering with parent dataset

The teal UI works with objects of class FilteredData which may mix CDISC and other data (e.g. iris).

It is not too difficult to make this class generalize to more than one parent dataset, provided each dataset then contains information about its parents together with the shared keys with each of them.

See also

Super class

teal::FilteredData -> CDISCFilteredData

Methods

Public methods

CDISCFilteredData$new()

CDISCFilteredData$datanames()

CDISCFilteredData$get_filterable_datanames()

CDISCFilteredData$get_filterable_varnames()

CDISCFilteredData$get_filter_expr()

CDISCFilteredData$get_data_info()

CDISCFilteredData$get_parentname()

CDISCFilteredData$clone()

Inherited methods

teal::FilteredData$get_bookmark_state()

teal::FilteredData$get_code()

teal::FilteredData$get_data()

teal::FilteredData$get_data_attr()

teal::FilteredData$get_datalabel()

teal::FilteredData$get_default_filter_state()

teal::FilteredData$get_filter_info()

teal::FilteredData$get_filter_overview_tbl()

teal::FilteredData$get_filter_state()

teal::FilteredData$get_filter_type()

teal::FilteredData$get_join_keys()

teal::FilteredData$get_keys()

teal::FilteredData$get_primary_keys()

teal::FilteredData$get_variable_labels()

teal::FilteredData$get_varlabels()

teal::FilteredData$get_varnames()

teal::FilteredData$print()

teal::FilteredData$print_filter_info()

teal::FilteredData$restore_state_from_bookmark()

teal::FilteredData$set_code()

teal::FilteredData$set_data()

teal::FilteredData$set_data_attr()

teal::FilteredData$set_data_attrs()

teal::FilteredData$set_filter_state()

teal::FilteredData$set_join_keys()

Method new()

Initialize a CDISCFilteredData object

Usage

CDISCFilteredData$new()

Method datanames()

Get datanames

The datanames are returned in the order in which they must be evaluated (in case of dependencies).

Usage

CDISCFilteredData$datanames()

Returns

(character vector) of datanames

Method get_filterable_datanames()

Get dataset names of a given dataname for the filtering.

Usage

CDISCFilteredData$get_filterable_datanames(dataname)

Arguments

dataname

(character vector) names of the dataset

Returns

(character vector) of dataset names

Method get_filterable_varnames()

Get variable names of a given dataname for the filtering.

This method incorporates parent-child relations, i.e. parent variable names are removed from child variable names.

Usage

CDISCFilteredData$get_filterable_varnames(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(character vector) of variable names

Method get_filter_expr()

Returns the filter expression to filter a single dataset including the inner_join with its parent dataset. It assumes that the filtered datasets it depends on are available.

Usage

CDISCFilteredData$get_filter_expr(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(expression) to filter dataset

Method get_data_info()

Get info about dataname, i.e. number of rows, subjects.

Usage

CDISCFilteredData$get_data_info(dataname, filtered)

Arguments

dataname

(character) name of the dataset

filtered

(logical) whether to obtain this info for the filtered dataset

Returns

a named vector

Method get_parentname()

Get parent dataset name

Usage

CDISCFilteredData$get_parentname(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(character) name of parent dataset

Method clone()

The objects of this class are cloneable with this method.

Usage

CDISCFilteredData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
attr(ADSL, "keys") <- get_cdisc_keys("ADSL")
datasets <- teal:::CDISCFilteredData$new()

# to avoid using isolate(), you can provide a default isolate context by calling
# options(shiny.suppressMissingContextError = TRUE) #nolint
# don't forget to deactivate this option at the end
# options(shiny.suppressMissingContextError = FALSE) #nolint

isolate({
  datasets$datanames()
  datasets

  datasets$set_data("ADSL", ADSL)
  datasets$set_join_keys(join_keys())

  datasets$datanames()
  datasets

  datasets$datanames()
  datasets$get_data_info("ADSL", filtered = FALSE)
  # filters dataset to obtain information
  datasets$get_data_info("ADSL", filtered = TRUE)
  datasets$get_filter_info("ADSL")
  df <- datasets$get_data("ADSL", filtered = FALSE)
  # df

  datasets$get_filter_type("ADSL", "SEX")
  datasets$set_filter_state("ADSL", state = list(
    AGE = list(range = c(33, 44), keep_na = FALSE),
    SEX = list(choices = c("M", "F"), keep_na = FALSE)
  ))
  datasets
  datasets$get_filter_type("ADSL", "SEX")
  datasets$get_filter_info("ADSL")[["SEX"]]$type

  # will fail because of invalid range
  # datasets$set_filter_state("ADSL", list(
  #   AGE = list(range = c(3, 7), keep_na = FALSE)
  # ))
  datasets$set_filter_state("ADSL", list(
    AGE = list(range = c(33, 44), keep_na = FALSE)
  ))
  # wrapper functions
  teal:::set_single_filter_state(datasets, "ADSL", "AGE", c(35, 42))
  teal:::set_single_filter_state(
    datasets, "ADSL", "AGE",
    state = list(range = c(33, 44), keep_na = FALSE)
  )
  datasets$set_filter_state(
    "ADSL",
     state = list(
       SEX = list(choices = c("M", "F"), keep_na = FALSE)
    )
  )
  datasets$get_filter_type("ADSL", "SEX")

  datasets$get_filter_state("ADSL")

  datasets$print_filter_info("ADSL")

  # add ADAE to show dependency
  ADAE <- radae(cached = TRUE)
  ADAE_indep <- ADAE
  attr(ADAE, "keys") <- get_cdisc_keys("ADAE")
  attr(ADAE_indep, "keys") <- get_cdisc_keys("ADAE")
  attr(ADAE, "parent") <- "ADSL"

  datasets$set_data("ADAE", ADAE)
  datasets$set_data("ADAEIndep", ADAE_indep)

  datasets$set_join_keys(join_keys(join_key("ADSL", "ADAE", c("STUDYID", "USUBJID"))))

  datasets$get_data_info("ADSL", filtered = FALSE)
  datasets$get_data_info("ADSL", filtered = TRUE)
  datasets$get_data_info("ADAE", filtered = FALSE)
  datasets$get_data_info("ADAE", filtered = TRUE)
  datasets$get_data_info("ADAEIndep", filtered = FALSE)
  datasets$get_data_info("ADAEIndep", filtered = TRUE)

  # see generated code
  datasets$get_filter_expr("ADSL")
  datasets$get_filter_expr("ADAEIndep")
})

#> |Ordered datanames computed: 
#> |Ordered datanames computed: ADSL
#> |################################################# Refiltering dataset  ADSL
#> ==== ADSL =======================
#> STUDYID             has filter type choices   : AB12345
#>                                  \--> selected: 
#> USUBJID             has filter type choices   : AB12345-BRA-1-id-105, AB12345-BRA-1-id-134, AB12345-BRA-1-id-141, AB12345-BRA-1-id-236, AB12345-BRA-1-id-265, ...
#>                                  \--> selected: 
#> SUBJID              has filter type choices   : id-1, id-10, id-100, id-101, id-102, ...
#>                                  \--> selected: 
#> SITEID              has filter type choices   : BRA-1, BRA-11, BRA-12, BRA-13, BRA-14, ...
#>                                  \--> selected: 
#> AGE                 has filter type range     : 20 - 69
#>                                  \--> selected: 33 - 44
#> SEX                 has filter type choices   : F, M
#>                                  \--> selected: M, F
#> RACE                has filter type choices   : ASIAN, BLACK OR AFRICAN AMERICAN, WHITE, AMERICAN INDIAN OR ALASKA NATIVE, MULTIPLE, ...
#>                                  \--> selected: 
#> COUNTRY             has filter type choices   : CHN, USA, BRA, PAK, NGA, ...
#>                                  \--> selected: 
#> DTHFL               has filter type choices   : N, Y
#>                                  \--> selected: 
#> INVID               has filter type choices   : INV ID BRA-1, INV ID BRA-11, INV ID BRA-12, INV ID BRA-13, INV ID BRA-14, ...
#>                                  \--> selected: 
#> INVNAM              has filter type choices   : Dr. BRA-1 Doe, Dr. BRA-11 Doe, Dr. BRA-12 Doe, Dr. BRA-13 Doe, Dr. BRA-14 Doe, ...
#>                                  \--> selected: 
#> ARM                 has filter type choices   : A: Drug X, B: Placebo, C: Combination
#>                                  \--> selected: 
#> ARMCD               has filter type choices   : ARM A, ARM B, ARM C
#>                                  \--> selected: 
#> ACTARM              has filter type choices   : A: Drug X, B: Placebo, C: Combination
#>                                  \--> selected: 
#> ACTARMCD            has filter type choices   : ARM A, ARM B, ARM C
#>                                  \--> selected: 
#> REGION1             has filter type choices   : Africa, Asia, Eurasia, Europe, North America, ...
#>                                  \--> selected: 
#> STRATA1             has filter type choices   : A, B, C
#>                                  \--> selected: 
#> STRATA2             has filter type choices   : S1, S2
#>                                  \--> selected: 
#> BMRKR1              has filter type range     : 0.165447463531243 - 21.3933676083323
#>                                  \--> selected: 
#> BMRKR2              has filter type choices   : LOW, MEDIUM, HIGH
#>                                  \--> selected: 
#> ITTFL               has filter type choices   : Y
#>                                  \--> selected: 
#> SAFFL               has filter type choices   : Y
#>                                  \--> selected: 
#> BMEASIFL            has filter type choices   : Y, N
#>                                  \--> selected: 
#> BEP01FL             has filter type choices   : Y, N
#>                                  \--> selected: 
#> RANDDT              has filter type date      : 2019-02-22 - 2021-02-14
#>                                  \--> selected: 
#> TRTSDTM             has filter type date      : 2019-02-24 11:09:18 - 2021-02-17 14:42:27
#>                                  \--> selected: 
#> TRTEDTM             has filter type date      : 2021-02-11 23:04:23 - 2023-02-18 02:20:51
#>                                  \--> selected: 
#> EOSSTT              has filter type choices   : COMPLETED, DISCONTINUED, ONGOING
#>                                  \--> selected: 
#> EOSDT               has filter type date      : 2021-02-11 - 2023-02-18
#>                                  \--> selected: 
#> EOSDY               has filter type range     : 6 - 731
#>                                  \--> selected: 
#> DCSREAS             has filter type choices   : ADVERSE EVENT, DEATH, LACK OF EFFICACY, PHYSICIAN DECISION, PROTOCOL VIOLATION, ...
#>                                  \--> selected: 
#> DTHDT               has filter type date      : 2021-02-12 - 2021-04-05
#>                                  \--> selected: 
#> DTHCAUS             has filter type choices   : ADVERSE EVENT, DISEASE PROGRESSION, LOST TO FOLLOW UP, MISSING, SUICIDE, ...
#>                                  \--> selected: 
#> DTHCAT              has filter type choices   : ADVERSE EVENT, OTHER, PROGRESSIVE DISEASE
#>                                  \--> selected: 
#> LDDTHELD            has filter type range     : 0 - 49
#>                                  \--> selected: 
#> LDDTHGR1            has filter type choices   : <=30, >30
#>                                  \--> selected: 
#> LSTALVDT            has filter type date      : 2021-02-12 - 2023-03-03
#>                                  \--> selected: 
#> study_duration_secs has filter type choices   : 63113904
#>                                  \--> selected: 
#> ===========================
#> |Ordered datanames computed: ADSL, ADAE
#> |################################################# Refiltering dataset  ADSL
#> |################################################# Refiltering dataset  ADAE
#> |all elements in SMQ02NAM are NA
#> |all elements in SMQ02SC are NA
#> |################################################# Refiltering dataset  ADAEIndep
#> |all elements in SMQ02NAM are NA
#> |all elements in SMQ02SC are NA

#> {
#>     ADAEIndep_FILTERED <- ADAEIndep
#> }