A CallablePythonCode class of objects

Super classes

teal::Callable -> teal::CallableFunction -> CallablePythonCode

Methods

Public methods

CallablePythonCode$new()

CallablePythonCode$set_object()

CallablePythonCode$assign_to_env()

CallablePythonCode$run()

CallablePythonCode$clone()

Inherited methods

teal::Callable$get_error_message()

teal::Callable$is_failed()

teal::CallableFunction$get_args()

teal::CallableFunction$get_call()

teal::CallableFunction$set_arg_value()

teal::CallableFunction$set_args()

Method new()

Create a new CallablePythonCode object

Usage

CallablePythonCode$new(fun)

Arguments

fun

(function)
function to be evaluated in class. Function should be named

Returns

new CallablePythonCode object

Method set_object()

For scripts and code that contain multiple objects, save the name of the object that corresponds to the final dataset of interest. This is required for running python scripts with reticulate.

Usage

CallablePythonCode$set_object(x)

Arguments

x

(character) the name of the object produced by the code or script.

Returns

(self) invisibly for chaining.

Method assign_to_env()

Assigns x <- value object to .GlobalEnv.

Usage

CallablePythonCode$assign_to_env(x, value)

Arguments

x

(character value)
name of the variable in class environment

value

(data.frame)
object to be assigned to x

Details

Assignment is done to the global environment. Any variables that are overwritten are saved in private$duplicate_vars and restored after $run.

Returns

(self) invisibly for chaining.

Method run()

Execute Callable python code.

Usage

CallablePythonCode$run(args = NULL, try = FALSE)

Arguments

args

(NULL or named list)
supplied for callable functions only, these are dynamic arguments passed to reticulate::py_run_string or reticulate::py_run_file. Dynamic arguments are executed in this call and are not saved which means that self$get_call() won't include them later.

try

(logical value)
whether perform function evaluation inside try clause

Returns

nothing or output from function depending on return argument. If run fails it will return object of class simple-error error when try = TRUE or will stop if try = FALSE.

Method clone()

The objects of this class are cloneable with this method.

Usage

CallablePythonCode$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.