Code Class

Active bindings

code

(list) Derive the code of the dataset.

Methods

Public methods

CodeClass$new()

CodeClass$append()

CodeClass$set_code()

CodeClass$get_code()

CodeClass$eval()

CodeClass$clone()

Method new()

CodeClass constructor

Usage

CodeClass$new(
  code = character(0),
  dataname = character(0),
  deps = character(0)
)

Arguments

code

(character) vector of code text to be set

dataname

optional, (character) vector of datanames to assign code to. If empty then the code is considered to be "global"

deps

optional, (character) vector of datanames that given code depends on

Returns

object of class CodeClass

Method append()

Append CodeClass object to a given CodeClass object

Usage

CodeClass$append(x)

Arguments

x

(CodeClass) object to be appended

Returns

changed CodeClass object

Method set_code()

Set code in form of character

Usage

CodeClass$set_code(code, dataname = character(0), deps = character(0))

Arguments

code

(character) vector of code text to be set

dataname

optional, (character) vector of datanames to assign code to. If empty then the code is considered to be "global"

deps

optional, (character) vector of datanames that given code depends on

Returns

changed CodeClass object

Method get_code()

Get the code for a given data names

Usage

CodeClass$get_code(dataname = NULL, deparse = TRUE)

Arguments

dataname

optional, (character) vector of datanames for which the code is extracted. If NULL then get the code for all data names

deparse

optional, (logical) whether to return the deparsed form of a call

Returns

character or list of calls

Method eval()

Evaluates internal code within given environment

Usage

CodeClass$eval(envir = new.env(parent = parent.env(.GlobalEnv)))

Arguments

envir

(environment) environment in which code will be evaluated

Returns

invisibly NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

CodeClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

cc <- teal:::CodeClass$new()
cc$set_code(c("foo <- function() {1}", "foo2 <- function() {2}"))
cc$get_code()

#> [1] "foo <- function() {\n    1\n}\nfoo2 <- function() {\n    2\n}"

cc$get_code(deparse = FALSE)

#> [[1]]
#> foo <- function() {
#>     1
#> }
#> 
#> [[2]]
#> foo2 <- function() {
#>     2
#> }
#> 

cc$set_code(c("ADSL <- radsl()", "ADSL$var <- 1"), "ADSL")
cc$set_code("ADSL$a <- foo()", "ADSL")

# dependent dataset
cc$set_code("ADAE <- radae(ADSL = ADSL)", "ADAE", deps = "ADSL")

cc$set_code("baz <- function() {2}")
cc$set_code("ADAE$a <- baz()", "ADAE")

cc$get_code()

#> [1] "foo <- function() {\n    1\n}\nfoo2 <- function() {\n    2\n}\nADSL <- radsl()\nADSL$var <- 1\nADSL$a <- foo()\nADAE <- radae(ADSL = ADSL)\nbaz <- function() {\n    2\n}\nADAE$a <- baz()"

cc$get_code("ADSL")

#> [1] "foo <- function() {\n    1\n}\nfoo2 <- function() {\n    2\n}\nADSL <- radsl()\nADSL$var <- 1\nADSL$a <- foo()"

cc$get_code("ADAE")

#> [1] "foo <- function() {\n    1\n}\nfoo2 <- function() {\n    2\n}\nADSL <- radsl()\nADSL$var <- 1\nADSL$a <- foo()\nADAE <- radae(ADSL = ADSL)\nbaz <- function() {\n    2\n}\nADAE$a <- baz()"

#########################################
#########################################
#########################################

x1 <- teal:::CodeClass$new()
x1$set_code("ADSL <- radsl(cached = TRUE)", "ADSL")
x1$get_code()

#> [1] "ADSL <- radsl(cached = TRUE)"

x2 <- teal:::CodeClass$new()
x2$set_code("ADAE <- radae(ADSL = ADSL)", "ADAE", "ADSL")
x2$get_code()

#> [1] "ADAE <- radae(ADSL = ADSL)"

x <- teal:::CodeClass$new()
x$append(x1)
x$append(x2)

x$get_code()

#> [1] "ADSL <- radsl(cached = TRUE)\nADAE <- radae(ADSL = ADSL)"

x$get_code("ADSL")

#> [1] "ADSL <- radsl(cached = TRUE)"

x$get_code("ADAE")

#> [1] "ADSL <- radsl(cached = TRUE)\nADAE <- radae(ADSL = ADSL)"

x$get_code(c("ADSL", "ADAE"))

#> [1] "ADSL <- radsl(cached = TRUE)\nADAE <- radae(ADSL = ADSL)"

x3 <- teal:::CodeClass$new()
x3$set_code("ADRS <- radae(cached = TRUE)", "ADRS")
x3$get_code()

#> [1] "ADRS <- radae(cached = TRUE)"

x$append(x3)
x$get_code("ADRS")

#> [1] "ADRS <- radae(cached = TRUE)"

# mutation simulation
x$set_code("ADRS$x <- foo(ADSL$x)", "ADRS", deps = "ADSL")
x$get_code("ADRS")

#> [1] "ADSL <- radsl(cached = TRUE)\nADRS <- radae(cached = TRUE)\nADRS$x <- foo(ADSL$x)"