A DataConnection class of objects

A DataConnection class of objects

A DataConnection class of objects

Details

Objects of this class store the connection to a data source. It can be a database or server (RICE or SAICE) connection.

Methods

Public methods

DataConnection$new()

DataConnection$is_opened()

DataConnection$is_failed()

DataConnection$open()

DataConnection$get_conn()

DataConnection$get_open_call()

DataConnection$get_open_error_message()

DataConnection$get_open_server()

DataConnection$get_open_ui()

DataConnection$is_open_failed()

DataConnection$set_open_args()

DataConnection$set_open_server()

DataConnection$set_open_ui()

DataConnection$close()

DataConnection$get_close_call()

DataConnection$get_close_error_message()

DataConnection$get_close_server()

DataConnection$get_close_ui()

DataConnection$is_close_failed()

DataConnection$set_close_args()

DataConnection$set_close_ui()

DataConnection$set_close_server()

DataConnection$clone()

Method new()

Create a new DataConnection object

Usage

DataConnection$new(
  open_fun = NULL,
  close_fun = NULL,
  ping_fun = NULL,
  if_conn_obj = FALSE
)

Arguments

open_fun

(CallableFunction) function to open connection

close_fun

(CallableFunction) function to close connection

ping_fun

(CallableFunction) function to ping connection

if_conn_obj

optional, (logical) whether to store conn object returned from opening connection

Returns

new DataConnection object

Method is_opened()

If connection is opened

If open connection has been successfully evaluated

Usage

DataConnection$is_opened()

Returns

logical if connection is open

Method is_failed()

Check if connection has not failed.

Usage

DataConnection$is_failed()

Returns

TRUE if connection failed, else FALSE

Method open()

Open the connection.

Note that if the connection is already opened then it does nothing.

Usage

DataConnection$open(args = NULL, silent = FALSE, try = FALSE)

Arguments

args

(NULL or named list) additional arguments not set up previously

silent

(logical) whether convert all "missing function" errors to messages

try

(logical) whether perform function evaluation inside try clause

Returns

if try = TRUE then try-error on error, NULL otherwise

Method get_conn()

Get internal connection object

Usage

DataConnection$get_conn()

Returns

connection object

Method get_open_call()

Get executed open connection call

Usage

DataConnection$get_open_call(deparse = TRUE, args = NULL, silent = FALSE)

Arguments

deparse

(logical) whether return deparsed form of a call

args

(NULL or named list) additional arguments not set up previously

silent

(logical) whether convert all "missing function" errors to messages

Returns

optionally deparsed call object

Method get_open_error_message()

Get error message from last connection

Usage

DataConnection$get_open_error_message()

Returns

try-error object with error message or character(0) if last connection was successful.

Method get_open_server()

Get shiny server module to open connection.

Usage

DataConnection$get_open_server()

Returns

the server function to open connection.

Method get_open_ui()

Get Shiny module with inputs to open connection

Usage

DataConnection$get_open_ui(id)

Arguments

id

character shiny element id

Returns

the ui function to set arguments to open connection function.

Method is_open_failed()

Check if open connection has not failed.

Usage

DataConnection$is_open_failed()

Returns

TRUE if open connection failed, else FALSE

Method set_open_args()

Set open connection function argument

Usage

DataConnection$set_open_args(args, silent = FALSE)

Arguments

args

(NULL or named list) with values where list names are argument names

silent

(logical) whether convert all "missing function" errors to messages

Returns

self invisibly for chaining.

Method set_open_server()

Set open-connection server function

This function will be called after submit button will be hit. There is no possibility to specify some dynamic ui as server function is executed after hitting submit button.

Usage

DataConnection$set_open_server(open_module)

Arguments

open_module

(function)
A shiny module server function that should load data from all connectors

Returns

self invisibly for chaining.

Method set_open_ui()

Set open connection UI function

Usage

DataConnection$set_open_ui(open_module)

Arguments

open_module

(function)
shiny module as function. Inputs specified in this ui are passed to server module defined by set_open_server method.

Returns

self invisibly for chaining.

Method close()

Close the connection.

Usage

DataConnection$close(silent = FALSE, try = FALSE)

Arguments

silent

(logical) whether convert all "missing function" errors to messages

try

(logical) whether perform function evaluation inside try clause

Returns

if try = TRUE then try-error on error, NULL otherwise

Method get_close_call()

Get executed close connection call

Usage

DataConnection$get_close_call(deparse = TRUE, silent = FALSE)

Arguments

deparse

(logical) whether return deparsed form of a call

silent

(logical) whether convert all "missing function" errors to messages

Returns

optionally deparsed call object

Method get_close_error_message()

Get error message from last connection

Usage

DataConnection$get_close_error_message()

Returns

try-error object with error message or character(0) if last connection was successful.

Method get_close_server()

Get shiny server module to close connection.

Usage

DataConnection$get_close_server()

Returns

the server function to close connection.

Method get_close_ui()

Get Shiny module with inputs to close connection

Usage

DataConnection$get_close_ui(id)

Arguments

id

character shiny element id

Returns

the ui function to set arguments to close connection function.

Method is_close_failed()

Check if close connection has not failed.

Usage

DataConnection$is_close_failed()

Returns

TRUE if close connection failed, else FALSE

Method set_close_args()

Set close connection function argument

Usage

DataConnection$set_close_args(args, silent = FALSE)

Arguments

args

(named list) with values where list names are argument names

silent

(logical) whether convert all "missing function" errors to messages

Returns

self invisibly for chaining.

Method set_close_ui()

Set close connection UI function

Usage

DataConnection$set_close_ui(close_module)

Arguments

close_module

(function)
shiny module as function. Inputs specified in this ui are passed to server module defined by set_close_server method.

Returns

self invisibly for chaining.

Method set_close_server()

Set close-connection server function

This function will be called after submit button will be hit. There is no possibility to specify some dynamic ui as server function is executed after hitting submit button.

Usage

DataConnection$set_close_server(close_module)

Arguments

close_module

(function)
A shiny module server function that should load data from all connectors

Returns

self invisibly for chaining.

Method clone()

The objects of this class are cloneable with this method.

Usage

DataConnection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
open_fun <- callable_function(data.frame) # define opening function
open_fun$set_args(list(x = 1:5)) # define fixed arguments to opening function

close_fun <- callable_function(print) # define closing function
close_fun$set_args(list(x = "Hi there")) # define fixed arguments to closing function

x <- DataConnection$new() # define connection
x$set_ping_fun(ping_fun) # able to set an optional ping function - eg. rice::rice_session_active
x$set_open_fun(open_fun) # define opening function
x$set_close_fun(close_fun) # define closing function

x$set_open_args(args = list(y = letters[1:5])) # define additional arguments if necessary

x$open() # call opening function

x$open(args = list(x = 1:5, y = letters[1:5])) # able to call opening function with arguments

x$get_open_call() # check reproducible R code
# get data from connection via DataConnector$get_dataset()

x$close() # call closing function
}