A DataConnection class of objects

A DataConnection class of objects

A DataConnection class of objects

A DatasetConnector class of objects

A DatasetConnector class of objects

Details

Objects of this class store connection to data source. It can be DB or server (RICE or SAICE) connection.

Objects of this class store connection function to single dataset. Note that for some specific connection type (e.g. RAICE or SAICE), pre-requisite object of class DataConnection is required. Data can be pulled via pull method and returned via get_data method.

Methods

Public methods

DataConnection$init()

DataConnection$open()

DataConnection$get_open_call()

DataConnection$close()

DataConnection$get_close_call()

DataConnection$set_open_fun()

DataConnection$set_open_args()

DataConnection$set_close_fun()

DataConnection$set_close_args()

DataConnection$refresh_calls()

DataConnection$clone()

Method init()

Create a new DataConnection object

Usage

DataConnection$init()

Returns

new DataConnection object

Method open()

Open the connection.

Note that if the connection is already opened then it does nothing.

Usage

DataConnection$open(args = NULL, silent = FALSE, try = FALSE)

Arguments

args

(NULL or named list) additional arguments not set up previously

silent

(logical) whether convert all "missing function" errors to messages

try

(logical) whether perform function evaluation inside try clause

Returns

if try = TRUE then try-error on error, NULL otherwise

Method get_open_call()

Get executed open connection call

Usage

DataConnection$get_open_call(deparse = TRUE, args = NULL, silent = FALSE)

Arguments

deparse

(logical) whether return deparsed form of a call

args

(NULL or named list) additional arguments not set up previously

silent

(logical) whether convert all "missing function" errors to messages

Returns

optionally deparsed call object

Method close()

Close the connection.

Usage

DataConnection$close(silent = FALSE, try = FALSE)

Arguments

silent

(logical) whether convert all "missing function" errors to messages

try

(logical) whether perform function evaluation inside try clause

Returns

if try = TRUE then try-error on error, NULL otherwise

Method get_close_call()

Get executed close connection call

Usage

DataConnection$get_close_call(deparse = TRUE, silent = FALSE)

Arguments

deparse

(logical) whether return deparsed form of a call

silent

(logical) whether convert all "missing function" errors to messages

Returns

optionally deparsed call object

Method set_open_fun()

Set open connection function

Usage

DataConnection$set_open_fun(fun)

Arguments

fun

(CallableFunction) function to open connection

Returns

nothing

Method set_open_args()

Set open connection function argument

Usage

DataConnection$set_open_args(args, silent = FALSE)

Arguments

args

(NULL or named list) with values where list names are argument names

silent

(logical) whether convert all "missing function" errors to messages

Returns

nothing

Method set_close_fun()

Set close connection function

Usage

DataConnection$set_close_fun(fun)

Arguments

fun

(CallableFunction) function to close connection

Returns

nothing

Method set_close_args()

Set close connection function argument

Usage

DataConnection$set_close_args(args, silent = FALSE)

Arguments

args

(named list) with values where list names are argument names

silent

(logical) whether convert all "missing function" errors to messages

Returns

nothing

Method refresh_calls()

Refresh both open and close calls

Usage

DataConnection$refresh_calls()

Returns

nothing

Method clone()

The objects of this class are cloneable with this method.

Usage

DataConnection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Methods

Public methods

DatasetConnector$new()

DatasetConnector$get_call()

DatasetConnector$get_data()

DatasetConnector$get_dataname()

DatasetConnector$get_keys()

DatasetConnector$get_path()

DatasetConnector$pull()

DatasetConnector$set_dataname()

DatasetConnector$set_keys()

DatasetConnector$set_path()

DatasetConnector$set_pull_arg_value()

DatasetConnector$set_pull_args()

DatasetConnector$set_pull_fun()

DatasetConnector$clone()

Method new()

Create a new DatasetConnector object

Usage

DatasetConnector$new()

Returns

new DatasetConnector object

Method get_call()

Get executed call

Usage

DatasetConnector$get_call(deparse = TRUE, args = NULL, silent = FALSE)

Arguments

deparse

(logical) whether return deparsed form of a call

args

(NULL or named list) additional dynamic arguments for pull function

silent

(logical) whether convert all "missing function" errors to messages

Returns

optionally deparsed call object

Method get_data()

Get the data from connection

Pull (if it is not already pulled) and return data

Usage

DatasetConnector$get_data(args = NULL, silent = FALSE, try = FALSE)

Arguments

args

(NULL or named list) additional dynamic arguments for pull function

silent

(logical) whether convert all "missing function" errors to messages

try

(logical) whether perform function evaluation inside try clause

Returns

if try = TRUE then try-error on error, object returned from connection function

Method get_dataname()

Get connection dataname

Usage

DatasetConnector$get_dataname()

Returns

(character) connection data dataname

Method get_keys()

Get keys of the dataset

Usage

DatasetConnector$get_keys()

Returns

(list) of keys

Method get_path()

Get path of the dataset

Usage

DatasetConnector$get_path()

Returns

(character)

Method pull()

Pull the data

Force execution of connection function

Usage

DatasetConnector$pull(args = NULL, silent = FALSE, try = FALSE)

Arguments

args

(NULL or named list) additional dynamic arguments for pull function

silent

(logical) whether convert all "missing function" errors to messages

try

(logical) whether perform function evaluation inside try clause

Returns

nothing, in order to get the data please use get_data method

Method set_dataname()

Set dataname

Usage

DatasetConnector$set_dataname(dataname)

Arguments

dataname

(character) dataname

Returns

nothing

Method set_keys()

Set dataset keys

Usage

DatasetConnector$set_keys(keys)

Arguments

keys

(keys) object

Returns

nothing

Method set_path()

Set connection path

Usage

DatasetConnector$set_path(path)

Arguments

path

(character) path

Returns

nothing

Method set_pull_arg_value()

Set connection function single argument with value

Usage

DatasetConnector$set_pull_arg_value(name, value, silent = FALSE)

Arguments

name

(character) function argument name

value

value to pass as a name

silent

(logical) whether convert all "missing function" errors to messages

Returns

nothing

Method set_pull_args()

Set connection function arguments with values

Usage

DatasetConnector$set_pull_args(args, silent = FALSE)

Arguments

args

(NULL or named list) with values where list names are argument names

silent

(logical) whether convert all "missing function" errors to messages

Returns

nothing

Method set_pull_fun()

Set the pulling function, which itself defines the connection through which the data gets pulled.

Usage

DatasetConnector$set_pull_fun(fun)

Arguments

fun

(CallableFunction) function to pull the data @return nothing

Method clone()

The objects of this class are cloneable with this method.

Usage

DatasetConnector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
open_fun <- CallableFunction$new(data.frame)  # define opening function
open_fun$set_args(list(x = 1:5))   # define fixed arguments to opening function

close_fun <- CallableFunction$new(print) # define closing function
close_fun$set_args(list(x = "Hi there"))  # define fixed arguments to closing function

x <- DataConnection$new() # define connection
x$set_open_fun(open_fun) # define opening function
x$set_close_fun(close_fun) # define closing function

x$set_open_args(args = list(y = letters[1:5])) # define additional arguments if necessary

x$open() # call opening function

x$open(args = list(x = 1:5, y = letters[1:5])) # able to call opening function with arguments

x$get_open_call() # check reroducible R code
# get data from connection via DataConnector$get_data()

x$close() # call closing function
}