FilterStates R6 class

Implements ReactiveQueue to the teal filters. Class manages adding and removing FilterState to the reactive queue and returns reproducible R expression relevant to specific FilterStates subclass. Currently data.frame, MultiAssayExperiment, SummarizedExperiment and matrix are available.

Note

throws an exception if the length of x does not match the length of element_id

Methods

Public methods

FilterStates$new()

FilterStates$get_call()

FilterStates$get_fun()

FilterStates$queue_empty()

FilterStates$queue_get()

FilterStates$queue_initialize()

FilterStates$queue_push()

FilterStates$queue_remove()

FilterStates$ui()

FilterStates$get_filter_state()

FilterStates$set_filter_state()

FilterStates$remove_filter_state()

FilterStates$ui_add_filter_state()

FilterStates$srv_add_filter_state()

FilterStates$clone()

Method new()

Initializes this FilterStates object.

Usage

FilterStates$new(input_dataname, output_dataname, datalabel)

Arguments

input_dataname

(character(1) or name or call)
name of the data used on lhs of the expression specified to the function argument attached to this FilterStates.

output_dataname

(character(1) or name or call)
name of the output data on the lhs of the assignment expression.

datalabel

(character(0) or character(1))
text label value.

Method get_call()

Filter call

Makes a subset function call based on condition calls from FilterState objects selection. lhs of the call is private$output_dataname and in rhs self$get_fun() with private$input_dataname as argument and list of condition calls from FilterState. If input and output data-names are the same and no filters applied, method returns NULL to avoid no-op call such as x <- x.

Usage

FilterStates$get_call()

Returns

call or NULL

Method get_fun()

Gets the name of the function used to filter the data in this FilterStates.

Get function name used to create filter call. By default it's a "subset" but can be overridden by child class method.

Usage

FilterStates$get_fun()

Returns

character(1) the name of the function

Method queue_empty()

Remove all FilterState objects from all queues in this FilterStates.

Usage

FilterStates$queue_empty()

Returns

NULL

Method queue_get()

Returns a list of FilterState objects stored in this FilterStates.

Usage

FilterStates$queue_get(queue_index, element_id = character(0))

Arguments

queue_index

(character(1), integer(1))
index of the private$queue list where ReactiveQueue are kept.

element_id

(character(1))
name of ReactiveQueue element.

Returns

list of FilterState objects

Method queue_initialize()

Sets ReactiveQueue objects.

Usage

FilterStates$queue_initialize(x)

Arguments

x

(list of ReactiveQueue)
must be a list even if single ReactiveQueue is set.

Method queue_push()

Adds a new FilterState object to this FilterStates

Usage

FilterStates$queue_push(x, queue_index, element_id)

Arguments

x

(FilterState)
object to be added to the queue

queue_index

(character(1), integer(1))
index of the private$queue list where ReactiveQueue are kept.

element_id

(character(1))
name of the ReactiveQueue element.

Method queue_remove()

Removes a single filter state

Removes a single filter state with all shiny elements associated with this state. It removes:

particular FilterState from private$queue

UI card created for this filter

observers listening selection and remove button

Usage

FilterStates$queue_remove(queue_index, element_id)

Arguments

queue_index

(character(1), logical(1))
index of the private$queue list where ReactiveQueue are kept.

element_id

(character(1))
name of ReactiveQueue element.

Method ui()

Shiny UI module

Shiny UI element being a container for FilterState elements. Content of this container is created using renderUI in server module

Usage

FilterStates$ui(id)

Arguments

id

(character(1))
id of the shiny element

Returns

shiny.tag

Method get_filter_state()

Gets the reactive values from the active FilterState objects.

Get active filter state from the FilterState objects kept in ReactiveQueue(s). The output list is a compatible input to self$set_filter_state.

Usage

FilterStates$get_filter_state()

Returns

list containing list per each FilterState in the ReactiveQueue

Method set_filter_state()

Sets active FilterState objects.

Usage

FilterStates$set_filter_state(data, state)

Arguments

data

(data.frame)
data which are supposed to be filtered

state

(named list)
should contain values which are initial selection in the FilterState. Names of the list element should correspond to the name of the column in data.

Returns

function which throws an error

Method remove_filter_state()

Remove a single FilterState from the ReactiveQueue.

Usage

FilterStates$remove_filter_state(element_id)

Arguments

element_id

(character)
Name of variable to remove its FilterState.

Returns

NULL

Method ui_add_filter_state()

Shiny UI module to add filter variable.

Usage

FilterStates$ui_add_filter_state(id, data)

Arguments

id

(character(1))
id of shiny module

data

(data.frame, MultiAssayExperiment, SummarizedExperiment, matrix)
object containing columns to be used as filter variables.

Returns

shiny.tag

Method srv_add_filter_state()

Shiny server module to add filter variable

Usage

FilterStates$srv_add_filter_state(id, data, ...)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

data

(data.frame, MultiAssayExperiment, SummarizedExperiment, matrix)
object containing columns to be used as filter variables.

...

ignored

Returns

moduleServer function which returns NULL Module to insert/remove FilterState UI

This module adds the shiny UI of the newly added FilterState object to queue to the Active Filter Variables, calls FilterState modules and creates observer to remove state parameter filter_state (FilterState)

parameter queue_index (character(1), logical(1))
index of the private$queue list where ReactiveQueue are kept. parameter element_id (character(1))
name of ReactiveQueue element. return moduleServer function which returns NULL observeEvent for remove-filter-state is set and also from FilteredDataset level, where shiny-session-namespace is different. That is why it's important to remove shiny elements from anywhere. In add_filter_state session$ns(NULL) is equivalent to private$ns(queue_index). This means that

Method clone()

The objects of this class are cloneable with this method.

Usage

FilterStates$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(shiny)
filter_states <- teal:::DFFilterStates$new(
  input_dataname = "data",
  output_dataname = "data_filtered",
  varlabels = c(x = "x variable", SEX = "Sex"),
  datalabel = character(0),
  keys = character(0)
)
filter_state <- teal:::RangeFilterState$new(
  c(NA, Inf, seq(1:10)),
  varname = "x",
  varlabel = "x variable",
  input_dataname = as.name("data"),
  extract_type = "list"
)
isolate(filter_state$set_selected(c(3L, 8L)))

filter_states$queue_push(
  x = filter_state,
  queue_index = 1L,
  element_id = "x"
)
isolate(filter_states$get_call())
#> data_filtered <- dplyr::filter(data, data$x >= 3 & data$x <= 
#>     8)