Class to encapsulate filtered datasets

Details

The main purpose of this class is to provide a collection of reactive datasets, each dataset having a filter state that determines how it is filtered.

For each dataset, get_filter_expr returns the call to filter the dataset according to the filter state. The data itself can be obtained through get_data. Other classes take care of actually merging together all the datasets.

The datasets are filtered lazily, i.e. only when requested / needed in the Shiny app.

By the design of the class (and reactiveValues), any dataname set through set_data cannot be removed because other code may already depend on it. As a workaround, the underlying data can be set to NULL.

The class currently supports variables of the following types within datasets:

choices: variable of type factor, e.g. ADSL$COUNTRY, iris$Species zero or more options can be selected, when the variable is a factor

logical: variable of type logical, e.g. ADSL$TRT_FLAG exactly one option must be selected, TRUE or FALSE

ranges: variable of type numeric, e.g. ADSL$AGE, iris$Sepal.Length numerical range, a range within this range can be selected Other variables cannot be filtered for.

Common arguments are:

filtered: filtered dataset or not

dataname: one of the datasets

varname: one of the columns in a dataset

Methods

Public methods

FilteredData$new()

FilteredData$datanames()

FilteredData$get_varnames()

FilteredData$get_filterable_datanames()

FilteredData$get_filterable_varnames()

FilteredData$get_data()

FilteredData$set_data()

FilteredData$get_code()

FilteredData$set_code()

FilteredData$set_join_keys()

FilteredData$get_join_keys()

FilteredData$get_primary_keys()

FilteredData$get_data_attr()

FilteredData$set_data_attr()

FilteredData$set_data_attrs()

FilteredData$get_variable_labels()

FilteredData$get_varlabels()

FilteredData$get_keys()

FilteredData$get_datalabel()

FilteredData$get_filter_info()

FilteredData$get_filter_overview_tbl()

FilteredData$get_filter_type()

FilteredData$get_filter_state()

FilteredData$set_filter_state()

FilteredData$get_default_filter_state()

FilteredData$get_filter_expr()

FilteredData$print()

FilteredData$print_filter_info()

FilteredData$get_data_info()

FilteredData$get_bookmark_state()

FilteredData$restore_state_from_bookmark()

FilteredData$clone()

Method new()

Initialize a FilteredData object

Usage

FilteredData$new()

Method datanames()

Get datanames

The datanames are returned in the order in which they must be evaluated (in case of dependencies).

Usage

FilteredData$datanames()

Returns

(character vector) of datanames

Method get_varnames()

Get variable names

Usage

FilteredData$get_varnames(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(character vector) of variable names

Method get_filterable_datanames()

Get dataset names of a given dataname for the filtering.

Usage

FilteredData$get_filterable_datanames(dataname)

Arguments

dataname

(character vector) names of the dataset

Returns

(character vector) of dataset names

Method get_filterable_varnames()

Get variable names of a given dataname for the filtering.

Usage

FilteredData$get_filterable_varnames(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(character vector) of variable names

Method get_data()

Get filtered or unfiltered dataset

For filtered = FALSE, the original data set with set_data is returned including all attributes.

Usage

FilteredData$get_data(dataname, filtered = TRUE)

Arguments

dataname

(character) name of the dataset

filtered

(logical) whether to return filtered or unfiltered dataset

Method set_data()

Add data

Will also add the hash sum of the data.

Note: due to the nature of reactiveValues(), once a dataname is added, it cannot be removed anymore because some code may depend on it. You can try to achieve this by setting the data to NULL and the observers must then know that NULL means that the dataset was removed.

Any attributes attached to the data are kept in the unfiltered data.

Usage

FilteredData$set_data(dataname, data)

Arguments

dataname

(character) name of the dataset, without spaces

data

(data.frame) data that corresponds to dataname

Returns

(self) object of this class

Method get_code()

Get the R preprocessing code string that generates the unfiltered datasets

Usage

FilteredData$get_code(dataname = self$datanames())

Arguments

dataname

(character) name(s) of dataset(s)

Returns

(character) deparsed code

Method set_code()

Set the R preprocessing code for single dataset

Usage

FilteredData$set_code(code)

Arguments

code

CodeClass preprocessing code that can be parsed to generate the unfiltered datasets

Returns

(self)

Method set_join_keys()

Set join keys object

Usage

FilteredData$set_join_keys(x)

Arguments

x

(JoinKeys)

Returns

(self) invisibly for chaining

Method get_join_keys()

Get join keys between two datasets.

Usage

FilteredData$get_join_keys(dataset_1, dataset_2)

Arguments

dataset_1

(character) one dataset name

dataset_2

(character) other dataset name

Returns

(named character) vector with column names

Method get_primary_keys()

Get dataset primary keys

Usage

FilteredData$get_primary_keys(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(character) vector with column names

Method get_data_attr()

Get data attribute for the dataset

sets and gets the data attribute on unfiltered data as it is never modified as attributes

Usage

FilteredData$get_data_attr(dataname, attr)

Arguments

dataname

(character) name of the dataset

attr

(character) attribute to get from the data attributes of the dataset

Returns

value of attribute, may be NULL if it does not exist

Method set_data_attr()

Set data attribute for the dataset

Usage

FilteredData$set_data_attr(dataname, attr, value)

Arguments

dataname

(character) name of the dataset

attr

attribute to get from the data attributes of the dataset

value

value to set attribute to

Returns

(self) invisibly for chaining

Method set_data_attrs()

Set data attributes for the dataset

Usage

FilteredData$set_data_attrs(dataname, attrs)

Arguments

dataname

(character) name of the dataset

attrs

(named list) data attributes of the dataset

Returns

(self) invisibly for chaining

Method get_variable_labels()

Get labels of variables in the data

Usage

FilteredData$get_variable_labels(dataname, variables = NULL)

Arguments

dataname

(character) name of the dataset

variables

(character vector) variables to get labels for; if NULL, for all variables in data

Method get_varlabels()

Get labels of variables in the data

Variables are the column names of the data. Either, all labels must have been provided for all variables in set_data or NULL.

Usage

FilteredData$get_varlabels(dataname, variables = NULL)

Arguments

dataname

(character) name of the dataset

variables

(character vector) variables to get labels for; if NULL, for all variables in data

Returns

(character or NULL) variable labels, NULL if column_labels attribute does not exist for the data Get keys for the dataset

Method get_keys()

Usage

FilteredData$get_keys(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(character) keys of dataset Get data label for the dataset

Useful to display in Show R Code.

Method get_datalabel()

Usage

FilteredData$get_datalabel(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(character) keys of dataset

Method get_filter_info()

Return filter characteristics for a variable in a dataset e.g. range for a numeric variable, choices for a factor variable, logical for a logical variable.

This can be used to see filtering conditions of a variable or a vector of variables. This method looks into private$filter_infos and returns a list of multiple filter information in case varname is a list or a named list of a single filter information in case varname is a singular name. The include_unknown argument is considered if and only when the varname is NULL.

Usage

FilteredData$get_filter_info(dataname, varname = NULL, include_unknown = TRUE)

Arguments

dataname

(character) name of the dataset

varname

(character) vector of variable names to include in the result. Default: all variables in dataname

include_unknown

(logical) whether to include variables with unknown filter type. Is ignored when varname is not NULL. Default: TRUE

Returns

a (named list) with filter information of varname variables. If include_unknown is TRUE, then also return filter information of variables with filter information of type unknown.

Examples

library(random.cdisc.data)

ADSL &lt;- radsl(cached = TRUE)
attr(ADSL, "keys") &lt;- get_cdisc_keys("ADSL")
datasets &lt;- teal:::FilteredData$new()

# to avoid using isolate(), you can provide a default isolate context by calling
# options(shiny.suppressMissingContextError = TRUE) #nolint
# don't forget to deactivate this option at the end
# options(shiny.suppressMissingContextError = FALSE) #nolint

isolate({
  datasets

  datasets$set_data("ADSL", ADSL)
  datasets
  datasets$get_data_info("ADSL", filtered = FALSE)
  datasets$datanames()
  # filters dataset to obtain information
  datasets$get_data_info("ADSL", filtered = TRUE)
  datasets$get_filter_info("ADSL")
  df &lt;- datasets$get_data("ADSL", filtered = FALSE)
  # df

  datasets$get_filter_type("ADSL", "SEX")
  datasets$set_filter_state("ADSL", state = list(
    AGE = list(range = c(33, 44), keep_na = FALSE),
    SEX = list(choices = c("M", "F"), keep_na = FALSE)
  ))

  infos &lt;- datasets$get_filter_info("ADSL") # will return a named list of all active filters
  names(infos) # notice that names correspond to variable names

  single_info &lt;- datasets$get_filter_info("ADSL", "AGE") # will return a list of AGE filter infos
  names(single_info) # note that this contains filter informations, not variable names as names

  single_info2 &lt;- datasets$get_filter_info("ADSL", include_unknown = TRUE)
  single_info2 # will include active filters of type unknown

  single_info3 &lt;- datasets$get_filter_info("ADSL",
    varname = c("AGE", "SEX"),
    include_unknown = FALSE)
  single_info3 # make note that include_unknown will be ignored, when varname is not NULL
})

Method get_filter_overview_tbl()

Get filter overview table in form of X (filtered) / Y (non-filtered)

This is intended to be presented in the application. The content for each of the data names is defined in get_data_info method.

Usage

FilteredData$get_filter_overview_tbl(datanames)

Arguments

datanames

(character vector) names of the dataset

Returns

(data.frame)

Method get_filter_type()

See get_filter_info, only returns filter type (choices, range, logical).

Usage

FilteredData$get_filter_type(dataname, varname)

Arguments

dataname

(character) name of the dataset

varname

(character) column within the dataset, must be provided

Returns

(character) filter type

Method get_filter_state()

Returns active filter state for a dataset (or only a variable within it).

Usage

FilteredData$get_filter_state(dataname, varname = NULL, include_unknown = TRUE)

Arguments

dataname

(character) name of the dataset

varname

(character) column within the dataset; if NULL, return all variables in a list

include_unknown

(logical) whether to include the "unknown" type of variables (TRUE) or not to include them (FALSE). Default: TRUE

Returns

(character vector) filter state or list of these

Method set_filter_state()

Set the filter state for a dataname

The state is only updated and triggers reactive behavior when it actually changes.

To set a state for a single variable, you can write: set_filter_state(dataname, setNames(list(single_state), varname)) or use the wrapper function set_single_filter_state. To remove a filter for a variable, pass in NULL, e.g. state = list(SEX = NULL, ..). To remove all filters for a dataset, call set_filter_state(dataname, state = list(), remove_omitted = TRUE).

Usage

FilteredData$set_filter_state(dataname, state, remove_omitted = FALSE)

Arguments

dataname

(character) name of the dataset to set filter state for

state

(named list) of new state to set; for each varname, state[[varname]] is the new state to set for varname

remove_omitted

(logical) whether to remove the filters for the variables that are not present in the named list state

Returns

(logical) if the state for the dataname was changed

Method get_default_filter_state()

Get the default filter state (useful for initial UI state) This is different to NULL state which means no filter applied. Instead, this filter is a sensible default, e.g. filter out NAs. for numerics, full range; for factors, all choices etc.

Usage

FilteredData$get_default_filter_state(dataname, varname)

Arguments

dataname

(character) name of the dataset

varname

(character) column within the dataset; must be provided

Returns

(character vector) default filter state for this variable

Method get_filter_expr()

Get an expression to filter the dataset according to the filter state

Helper function for reactive_filtered_dataset. The latter actually makes sure the dependent datasets are available and then evaluates the dataset.

It returns an expression to filter the dataset only, assuming the other (filtered) datasets it depends on are available.

Together with self$datanames() which returns the datasets in the correct evaluation order, this generates the whole filter code, see the function FilteredData$get_filter_code.

For the return type, note that rlang::is_expression returns TRUE on the return type, both for base R expressions and calls (single expression, capturing a function call).

The filtered dataset has the name given by self$filtered_dataname(dataname)

This can be used for the Show R Code generation.

Usage

FilteredData$get_filter_expr(dataname)

Arguments

dataname

(character) name of the dataset

Returns

(expression) which returns the filtered dataset

Method print()

Print method

To obtain the default printing of R6 objects with method listing, use cat(format(x)).

Usage

FilteredData$print(...)

Arguments

...

unused arguments

Method print_filter_info()

Prints filter characteristics of each variable for a given dataset to the console.

Usage

FilteredData$print_filter_info(
  dataname,
  filtered_vars_only = FALSE,
  variables = NULL
)

Arguments

dataname

(character) name of the dataset

filtered_vars_only

(logical) whether to only consider filtered vars or unfiltered as well

variables

(character vector) variables to consider; if NULL, takes all

Returns

NULL, only prints

Method get_data_info()

Get info about dataname, i.e. number of rows

Usage

FilteredData$get_data_info(dataname, filtered)

Arguments

dataname

(character) name of the dataset

filtered

logical whether to obtain this info for the filtered dataset

Returns

a named vector

Method get_bookmark_state()

Returns the state to be bookmarked

hash sums of datasets, filter_states and preproc_code are bookmarked.

Usage

FilteredData$get_bookmark_state()

Returns

named list

Method restore_state_from_bookmark()

Set this object from a bookmarked state

Only sets the filter state, does not set the data and the preprocessing code. The data should already have been set. Also checks the preprocessing code is identical if provided in the state.

Since this function is used from the end-user part, its error messages are more verbose. We don't call the Shiny modals from here because this class may be used outside of a Shiny app.

Usage

FilteredData$restore_state_from_bookmark(state, check_data_hash = TRUE)

Arguments

state

(list) containing fields data_hash, filter_states and preproc_code.

check_data_hash

(logical) whether to check that md5sums agree for the data; may not make sense with randomly generated data per session

Method clone()

The objects of this class are cloneable with this method.

Usage

FilteredData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

datasets <- teal:::FilteredData$new()

# to avoid using isolate(), you can provide a default isolate context by calling
# options(shiny.suppressMissingContextError = TRUE) #nolint
# don't forget to deactivate this option at the end
# options(shiny.suppressMissingContextError = FALSE) #nolint
isolate({
  datasets
  # we don't set keys because datasets are independent
  datasets$set_data("iris", iris)
  datasets$set_data("mtcars", mtcars)

  datasets

  datasets$datanames()
  datasets$get_data_info("iris", filtered = FALSE)

  datasets$get_filter_type("iris", "Species")
  datasets$set_filter_state("iris", state = list(
    Sepal.Length = list(range = c(4.5, 4.9), keep_na = FALSE),
    Species = list(choices = c("setosa", "virginica"), keep_na = FALSE)
  ))
  datasets
  datasets$get_filter_type("iris", "Species")
  datasets$get_filter_info("iris")[["Species"]]

  datasets$get_data("iris")

  # will fail because of invalid range
  # datasets$set_filter_state("iris", list(
  #   Sepal.Length = list(range = c(3, 7), keep_na = FALSE)
  # ))

  # wrapper functions
  teal:::set_single_filter_state(datasets, "mtcars", "cyl", c(4, 8))
  teal:::set_single_filter_state(datasets, "mtcars", "qsec", c(14.5, 16))
  datasets$get_data("mtcars")

  datasets$get_filter_state("iris")
  datasets$get_filter_state("mtcars")

  datasets$get_default_filter_state("iris", "Sepal.Length")
  datasets$get_filter_expr("iris")

  datasets$print_filter_info("mtcars")
})

#> |################################################# Refiltering dataset  iris
#> |################################################# Refiltering dataset  mtcars
#> ==== mtcars =======================
#> mpg  has filter type range     : 10.4 - 33.9
#>                   \--> selected: 
#> cyl  has filter type choices   : 4, 6, 8
#>                   \--> selected: 4, 8
#> disp has filter type range     : 71.1 - 472
#>                   \--> selected: 
#> hp   has filter type range     : 52 - 335
#>                   \--> selected: 
#> drat has filter type range     : 2.76 - 4.93
#>                   \--> selected: 
#> wt   has filter type range     : 1.513 - 5.424
#>                   \--> selected: 
#> qsec has filter type range     : 14.5 - 22.9
#>                   \--> selected: 14.5 - 16
#> vs   has filter type choices   : 0, 1
#>                   \--> selected: 
#> am   has filter type choices   : 0, 1
#>                   \--> selected: 
#> gear has filter type choices   : 3, 4, 5
#>                   \--> selected: 
#> carb has filter type range     : 1 - 8
#>                   \--> selected: 
#> ===========================

## ------------------------------------------------
## Method `FilteredData$get_filter_info`
## ------------------------------------------------

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
attr(ADSL, "keys") <- get_cdisc_keys("ADSL")
datasets <- teal:::FilteredData$new()

# to avoid using isolate(), you can provide a default isolate context by calling
# options(shiny.suppressMissingContextError = TRUE) #nolint
# don't forget to deactivate this option at the end
# options(shiny.suppressMissingContextError = FALSE) #nolint

isolate({
  datasets

  datasets$set_data("ADSL", ADSL)
  datasets
  datasets$get_data_info("ADSL", filtered = FALSE)
  datasets$datanames()
  # filters dataset to obtain information
  datasets$get_data_info("ADSL", filtered = TRUE)
  datasets$get_filter_info("ADSL")
  df <- datasets$get_data("ADSL", filtered = FALSE)
  # df

  datasets$get_filter_type("ADSL", "SEX")
  datasets$set_filter_state("ADSL", state = list(
    AGE = list(range = c(33, 44), keep_na = FALSE),
    SEX = list(choices = c("M", "F"), keep_na = FALSE)
  ))

  infos <- datasets$get_filter_info("ADSL") # will return a named list of all active filters
  names(infos) # notice that names correspond to variable names

  single_info <- datasets$get_filter_info("ADSL", "AGE") # will return a list of AGE filter infos
  names(single_info) # note that this contains filter informations, not variable names as names

  single_info2 <- datasets$get_filter_info("ADSL", include_unknown = TRUE)
  single_info2 # will include active filters of type unknown

  single_info3 <- datasets$get_filter_info("ADSL",
    varname = c("AGE", "SEX"),
    include_unknown = FALSE)
  single_info3 # make note that include_unknown will be ignored, when varname is not NULL
})

#> |################################################# Refiltering dataset  ADSL

#> $AGE
#> $AGE$type
#> [1] "range"
#> 
#> $AGE$label
#> [1] "Age"
#> 
#> $AGE$range
#> [1] 20 69
#> 
#> $AGE$histogram_data
#>            x            y
#> 1   13.94261 9.117225e-06
#> 2   14.55993 2.321839e-05
#> 3   15.17726 5.439041e-05
#> 4   15.79458 1.174687e-04
#> 5   16.41190 2.346172e-04
#> 6   17.02922 4.351502e-04
#> 7   17.64654 7.536781e-04
#> 8   18.26386 1.227863e-03
#> 9   18.88118 1.899402e-03
#> 10  19.49850 2.815356e-03
#> 11  20.11582 4.035965e-03
#> 12  20.73314 5.635405e-03
#> 13  21.35046 7.690577e-03
#> 14  21.96779 1.025643e-02
#> 15  22.58511 1.333701e-02
#> 16  23.20243 1.686859e-02
#> 17  23.81975 2.072756e-02
#> 18  24.43707 2.476110e-02
#> 19  25.05439 2.882359e-02
#> 20  25.67171 3.279805e-02
#> 21  26.28903 3.659320e-02
#> 22  26.90635 4.013048e-02
#> 23  27.52367 4.333371e-02
#> 24  28.14100 4.613345e-02
#> 25  28.75832 4.847642e-02
#> 26  29.37564 5.032419e-02
#> 27  29.99296 5.164218e-02
#> 28  30.61028 5.239834e-02
#> 29  31.22760 5.258938e-02
#> 30  31.84492 5.228558e-02
#> 31  32.46224 5.165638e-02
#> 32  33.07956 5.094021e-02
#> 33  33.69688 5.035932e-02
#> 34  34.31420 5.001961e-02
#> 35  34.93153 4.986149e-02
#> 36  35.54885 4.969468e-02
#> 37  36.16617 4.929268e-02
#> 38  36.78349 4.849310e-02
#> 39  37.40081 4.724777e-02
#> 40  38.01813 4.560162e-02
#> 41  38.63545 4.362576e-02
#> 42  39.25277 4.135668e-02
#> 43  39.87009 3.878472e-02
#> 44  40.48741 3.589678e-02
#> 45  41.10473 3.273844e-02
#> 46  41.72206 2.945108e-02
#> 47  42.33938 2.624860e-02
#> 48  42.95670 2.335132e-02
#> 49  43.57402 2.091440e-02
#> 50  44.19134 1.898506e-02
#> 51  44.80866 1.750354e-02
#> 52  45.42598 1.633832e-02
#> 53  46.04330 1.533383e-02
#> 54  46.66062 1.435199e-02
#> 55  47.27794 1.329741e-02
#> 56  47.89527 1.212520e-02
#> 57  48.51259 1.083516e-02
#> 58  49.12991 9.460259e-03
#> 59  49.74723 8.053976e-03
#> 60  50.36455 6.680121e-03
#> 61  50.98187 5.404770e-03
#> 62  51.59919 4.286256e-03
#> 63  52.21651 3.364286e-03
#> 64  52.83383 2.653147e-03
#> 65  53.45115 2.143430e-03
#> 66  54.06847 1.810992e-03
#> 67  54.68580 1.626298e-03
#> 68  55.30312 1.557612e-03
#> 69  55.92044 1.568024e-03
#> 70  56.53776 1.612380e-03
#> 71  57.15508 1.642427e-03
#> 72  57.77240 1.620205e-03
#> 73  58.38972 1.531670e-03
#> 74  59.00704 1.391728e-03
#> 75  59.62436 1.235834e-03
#> 76  60.24168 1.102216e-03
#> 77  60.85900 1.014455e-03
#> 78  61.47633 9.731437e-04
#> 79  62.09365 9.592634e-04
#> 80  62.71097 9.455134e-04
#> 81  63.32829 9.090432e-04
#> 82  63.94561 8.403178e-04
#> 83  64.56293 7.459863e-04
#> 84  65.18025 6.446621e-04
#> 85  65.79757 5.589640e-04
#> 86  66.41489 5.060852e-04
#> 87  67.03221 4.905017e-04
#> 88  67.64954 5.019032e-04
#> 89  68.26686 5.194932e-04
#> 90  68.88418 5.208616e-04
#> 91  69.50150 4.912473e-04
#> 92  70.11882 4.287656e-04
#> 93  70.73614 3.434570e-04
#> 94  71.35346 2.515189e-04
#> 95  71.97078 1.681267e-04
#> 96  72.58810 1.024938e-04
#> 97  73.20542 5.695910e-05
#> 98  73.82274 2.884829e-05
#> 99  74.44007 1.331316e-05
#> 100 75.05739 5.597063e-06
#> 
#> $AGE$inf_count
#> [1] 0
#> 
#> $AGE$is_integer
#> [1] TRUE
#> 
#> $AGE$na_count
#> [1] 0
#> 
#> 
#> $SEX
#> $SEX$type
#> [1] "choices"
#> 
#> $SEX$label
#> [1] "Sex"
#> 
#> $SEX$choices
#> $SEX$choices$`F (231)`
#> [1] "F"
#> 
#> $SEX$choices$`M (169)`
#> [1] "M"
#> 
#> 
#> $SEX$histogram_data
#>   x   y
#> 1 F 231
#> 2 M 169
#> 
#> $SEX$na_count
#> [1] 0
#> 
#>