Class to encapsulate filtered datasets

Class to encapsulate filtered datasets

Class to encapsulate filtered datasets

Details

The main purpose of this class is to provide a collection of reactive datasets, each of which can be filtered through the right filter panel of teal apps. For each dataset, get_filter_expr returns the call to filter the dataset according to the UI user selection. Other classes then take care of actually merging together all the datasets.

This class is ADSL-centric in the sense that ADSL is required to apply the filters. Every dataset, in addition to its own filter, is filtered to make sure that it only contains keys present in ADSL (defaulting to (USUBJID, STUDYID)). Once the ADSL dataset is set, the filters are applied.

The datasets are filtered lazily, i.e. only when requested / needed in the Shiny app.

By the design of the class (and reactiveValues), any dataname set through set_data cannot be removed because other code may already depend on it. As a workaround, the underlying data can be set to NULL.

The class currently supports variables of the following types within datasets:

choices: variable of type factor, e.g. ADSL$COUNTRY zero or more options can be selected, when the variable is a factor

logical: variable of type logical, e.g. ADSL$TRT_FLAG exactly one option must be selected, TRUE or FALSE

ranges: variable of type numeric, e.g. ADSL$AGE numerical range, a range within this range can be selected Other variables cannot be filtered for.

Common arguments are:

filtered: filtered dataset or not

dataname: one of the datasets

varname: one of the columns in a dataset

Methods

Public methods

FilteredData$new()

FilteredData$datanames()

FilteredData$get_data()

FilteredData$set_data()

FilteredData$get_code()

FilteredData$set_code()

FilteredData$get_data_attr()

FilteredData$set_data_attr()

FilteredData$get_variable_labels()

FilteredData$get_keys()

FilteredData$get_datalabel()

FilteredData$get_filter_info()

FilteredData$get_filter_type()

FilteredData$get_filter_state()

FilteredData$set_filter_state()

FilteredData$get_default_filter_state()

FilteredData$is_filterable()

FilteredData$get_filter_expr()

FilteredData$print()

FilteredData$print_filter_info()

FilteredData$get_data_info()

FilteredData$get_bookmark_state()

FilteredData$restore_state_from_bookmark()

FilteredData$clone()

Method new()

Usage

FilteredData$new()

Details

Initialize a FilteredData object

Method datanames()

Usage

FilteredData$datanames()

Details

Get datanames, ADSL appears first

Returns

character vector of datanames

Method get_data()

Usage

FilteredData$get_data(dataname, filtered = TRUE)

Arguments

dataname

character name of the dataset

filtered

logical whether to return filtered or unfiltered dataset

Details

Get filtered or unfiltered dataset

For filtered = FALSE, the original data set with set_data is returned including all attributes.

Method set_data()

Usage

FilteredData$set_data(dataname, data)

Arguments

dataname

character name of the dataset, without spaces

data

data.frame data that corresponds to dataname

Details

Add data

Will also add the md5 sum of the data.

Note: due to the nature of reactiveValues(), once a dataname is added, it cannot be removed anymore because some code may depend on it. You can try to achieve this by setting the data to NULL and the observers must then know that NULL means that the dataset was removed.

Any attributes attached to the data are kept in the unfiltered data.

Returns

self object of this class

Method get_code()

Usage

FilteredData$get_code(dataname = self$datanames())

Arguments

dataname

character name(s) of dataset(s)

Details

Get the R preprocessing code string that generates the unfiltered datasets

Returns

character deparsed code

Method set_code()

Usage

FilteredData$set_code(code)

Arguments

code

CodeClass preprocessing code that can be parsed to generate the unfiltered datasets

Details

Set the R preprocessing code for single dataset

Returns

self

Method get_data_attr()

Usage

FilteredData$get_data_attr(dataname, attr)

Arguments

dataname

character name of the dataset

attr

character attribute to get from the data attributes of the dataset

Details

Get data attribute for the dataset

sets and gets the data attribute on unfiltered data as it is never modified as attributes

Returns

value of attribute, may be NULL if it does not exist

Method set_data_attr()

Usage

FilteredData$set_data_attr(dataname, attr, value)

Arguments

dataname

character name of the dataset

attr

attribute to get from the data attributes of the dataset

value

value to set attribute to

Details

Set data attribute for the dataset

Returns

self for chaining

Method get_variable_labels()

Usage

FilteredData$get_variable_labels(dataname, variables = NULL)

Arguments

dataname

character name of the dataset

variables

(character vector) variables to get labels for; if NULL, for all variables in data

Details

Get labels of variables in the data

Variables are the column names of the data. Either, all labels must have been provided for all variables in set_data or NULL.

Returns

character or NULL variable labels, NULL if column_labels attribute does not exist for the data Get keys for the dataset

Method get_keys()

Usage

FilteredData$get_keys(dataname)

Arguments

dataname

character name of the dataset

Returns

character keys of dataset Get data label for the dataset

Useful to display in Show R Code.

Method get_datalabel()

Usage

FilteredData$get_datalabel(dataname)

Arguments

dataname

character name of the dataset

Returns

character keys of dataset

Method get_filter_info()

Usage

FilteredData$get_filter_info(dataname, varname = NULL, all_vars = FALSE)

Arguments

dataname

character name of the dataset

varname

character column within the dataset, if NULL, return all (filtered) variables

all_vars

logical (only applies if varname is NULL) whether to include non-filtered variables; attention: only filterable variables are included

Details

Return filter characteristics for a variable in a dataset e.g. range for a numeric variable, choices for a factor variable, logical for a logical variable.

This can be used to see how the variable must be filtered.

Method get_filter_type()

Usage

FilteredData$get_filter_type(dataname, varname)

Arguments

dataname

character name of the dataset

varname

character column within the dataset, must be provided

Details

See get_filter_info, only returns filter type (choices, range, logical).

Returns

character filter type

Method get_filter_state()

Usage

FilteredData$get_filter_state(dataname, varname = NULL)

Arguments

dataname

character name of the dataset

varname

character column within the dataset; if NULL, return all variables in a list

Details

Filter state for a dataset (or only a variable within it).

Returns

character vector filter state or list of these

Method set_filter_state()

Usage

FilteredData$set_filter_state(dataname, state, remove_omitted = FALSE)

Arguments

dataname

character name of the dataset to set filter state for

state

named list of new state to set; for each varname, state[[varname]] is the new state to set for varname

remove_omitted

logical whether to remove the filters for the variables that are not present in the named list state

Details

Set the filter state for a dataname

The state is only updated and triggers reactive behavior when it actually changes.

To set a state for a single variable, you can write: set_filter_state(dataname, setNames(list(single_state), varname)) or use the wrapper function set_single_filter_state. To remove a filter for a variable, pass in NULL, e.g. state = list(SEX = NULL, ..). To remove all filters for a dataset, call set_filter_state(dataname, state = list(), remove_omitted = FALSE).

Returns

logical if the state for the dataname was changed

Method get_default_filter_state()

Usage

FilteredData$get_default_filter_state(dataname, varname)

Arguments

dataname

character name of the dataset

varname

character column within the dataset; must be provided

Details

Get the default filter state (useful for initial UI state) This is different to NULL state which means no filter applied. Instead, this filter is a sensible default, e.g. filter out NAs. for numerics, full range; for factors, all choices etc.

Returns

character vector default filter state for this variable

Method is_filterable()

Usage

FilteredData$is_filterable(dataname, varname)

Arguments

dataname

character name of the dataset

varname

character column within the dataset; must be provided

Details

Check whether the variable can be filtered, i.e. not of unknown type

Returns

whether the variable can be filtered (type not unknown)

Method get_filter_expr()

Usage

FilteredData$get_filter_expr(dataname)

Arguments

dataname

character name of the dataset

Details

Get the call to filter the dataset according to the filter state

This can be used for the Show R Code generation.

When the dataname is ADSL, it is simply the filtered ADSL. Otherwise, it is the filtered dataset that only contains subjects which are also in the filtered ADSL.

Note: The dplyr::inner_join() function in returned call will fail if the corresponding dataset is NULL.

For the return type, note that rlang::is_expression returns TRUE on the return type, both for base R expressions and calls (single expression, capturing a function call).

Returns

expression to filter dataset (taking out patients not in ADSL)

Method print()

Usage

FilteredData$print(...)

Arguments

...

unused arguments

Details

Print method

To obtain the default printing of R6 objects with method listing, use cat(format(x)).

Method print_filter_info()

Usage

FilteredData$print_filter_info(
  dataname,
  filtered_vars_only = FALSE,
  variables = NULL
)

Arguments

dataname

character name of the dataset

filtered_vars_only

logical whether to only consider filtered vars or unfiltered as well

variables

(character vector) variables to consider; if NULL, takes all

Details

Prints filter characteristics of each variable for a given dataset to the console.

Returns

NULL, only prints

Method get_data_info()

Usage

FilteredData$get_data_info(dataname, filtered)

Arguments

dataname

character name of the dataset

filtered

logical whether to obtain this info for the filtered dataset

Details

Get info about dataname, e.g. dimensions and number of patients (in long datasets different from ADSL, a patient can occur in zero or more rows)

Returns

a list with names dim, patients Returns the state to be bookmarked

md5 sums of datasets, filter_states and preproc_code are bookmarked.

Method get_bookmark_state()

Usage

FilteredData$get_bookmark_state()

Returns

named list Set this object from a bookmarked state

Only sets the filter state, does not set the data and the preprocessing code. Also checks the preprocessing code is identical if provided in the state.

Since this function is used from the end-user part, its error messages are more verbose. We don't call the Shiny modals from here because this class may be used outside of a Shiny app.

Method restore_state_from_bookmark()

Usage

FilteredData$restore_state_from_bookmark(state, check_data_md5sums = TRUE)

Arguments

state

list containing fields data_md5sums, filter_states and preproc_code.

check_data_md5sums

logical whether to check that md5sums agree for the data; may not make sense with randomly generated data per session

Method clone()

The objects of this class are cloneable with this method.

Usage

FilteredData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
attr(ADSL, "keys") <- get_cdisc_keys("ADSL")
datasets <- teal:::FilteredData$new()

# to avoid using isolate(), you can provide a default isolate context by calling
# options(shiny.suppressMissingContextError = TRUE) #nolint
# don't forget to deactivate this option at the end
# options(shiny.suppressMissingContextError = FALSE) #nolint

isolate({
  datasets

  datasets$set_data("ADSL", ADSL)
  datasets

  datasets$datanames()
  datasets$get_data_info("ADSL", filtered = FALSE)
  # filters dataset to obtain information
  datasets$get_data_info("ADSL", filtered = TRUE)
  datasets$get_filter_info("ADSL")
  df <- datasets$get_data("ADSL", filtered = FALSE)
  # df

  datasets$get_filter_type("ADSL", "SEX")
  datasets$set_filter_state("ADSL", state = list(
    AGE = list(range = c(33, 44), keep_na = FALSE),
    SEX = list(choices = c("M", "F"), keep_na = FALSE)
  ))
  datasets
  datasets$get_filter_type("ADSL", "SEX")
  datasets$get_filter_info("ADSL")[["SEX"]]$type

  # will fail because of invalid range
  # datasets$set_filter_state("ADSL", list(
  #   AGE = list(range = c(3, 7), keep_na = FALSE)
  # ))
  datasets$set_filter_state("ADSL", list(
    AGE = list(range = c(33, 44), keep_na = FALSE)
  ))
  # wrapper functions
  teal:::set_single_filter_state(datasets, "ADSL", "AGE", c(35, 42))
  teal:::set_single_filter_state(
    datasets, "ADSL", "AGE",
    state = list(range = c(33, 44), keep_na = FALSE)
  )
  datasets$set_filter_state(
    "ADSL",
    state = list(
      SEX = list(choices = c("M", "F"), keep_na = FALSE)
    )
  )
  datasets$get_filter_type("ADSL", "SEX")

  datasets$get_filter_state("ADSL")

  datasets$print_filter_info("ADSL")
})

#> |################################################# Refiltering dataset  ADSL
#> |variable 'RANDDT' is of class 'Date' which has currently no filter UI element
#> |variable 'TRTSDTM' is of class 'POSIXct, POSIXt' which has currently no filter UI element
#> |variable 'TRTEDTM' is of class 'POSIXct, POSIXt' which has currently no filter UI element
#> |variable 'EOSDT' is of class 'Date' which has currently no filter UI element
#> |variable 'DTHDT' is of class 'Date' which has currently no filter UI element
#> |variable 'LSTALVDT' is of class 'Date' which has currently no filter UI element
#> ==== ADSL =======================
#> STUDYID             has filter type choices   : AB12345
#>                                  \--> selected: 
#> USUBJID             has filter type choices   : AB12345-BRA-1-id-105, AB12345-BRA-1-id-134, AB12345-BRA-1-id-141, AB12345-BRA-1-id-236, AB12345-BRA-1-id-265, ...
#>                                  \--> selected: 
#> SUBJID              has filter type choices   : id-1, id-10, id-100, id-101, id-102, ...
#>                                  \--> selected: 
#> SITEID              has filter type choices   : BRA-1, BRA-11, BRA-12, BRA-13, BRA-14, ...
#>                                  \--> selected: 
#> AGE                 has filter type range     : 20 - 69
#>                                  \--> selected: 33 - 44
#> SEX                 has filter type choices   : F, M
#>                                  \--> selected: M, F
#> RACE                has filter type choices   : ASIAN, BLACK OR AFRICAN AMERICAN, WHITE, AMERICAN INDIAN OR ALASKA NATIVE, MULTIPLE, ...
#>                                  \--> selected: 
#> COUNTRY             has filter type choices   : CHN, USA, BRA, PAK, NGA, ...
#>                                  \--> selected: 
#> INVID               has filter type choices   : INV ID BRA-1, INV ID BRA-11, INV ID BRA-12, INV ID BRA-13, INV ID BRA-14, ...
#>                                  \--> selected: 
#> INVNAM              has filter type choices   : Dr. BRA-1 Doe, Dr. BRA-11 Doe, Dr. BRA-12 Doe, Dr. BRA-13 Doe, Dr. BRA-14 Doe, ...
#>                                  \--> selected: 
#> ARM                 has filter type choices   : A: Drug X, B: Placebo, C: Combination
#>                                  \--> selected: 
#> ARMCD               has filter type choices   : ARM A, ARM B, ARM C
#>                                  \--> selected: 
#> ACTARM              has filter type choices   : A: Drug X, B: Placebo, C: Combination
#>                                  \--> selected: 
#> ACTARMCD            has filter type choices   : ARM A, ARM B, ARM C
#>                                  \--> selected: 
#> REGION1             has filter type choices   : Africa, Asia, Eurasia, Europe, North America, ...
#>                                  \--> selected: 
#> STRATA1             has filter type choices   : A, B, C
#>                                  \--> selected: 
#> STRATA2             has filter type choices   : S1, S2
#>                                  \--> selected: 
#> BMRKR1              has filter type range     : 0.165447463531243 - 21.3933676083323
#>                                  \--> selected: 
#> BMRKR2              has filter type choices   : LOW, MEDIUM, HIGH
#>                                  \--> selected: 
#> ITTFL               has filter type choices   : Y
#>                                  \--> selected: 
#> SAFFL               has filter type choices   : Y
#>                                  \--> selected: 
#> BMEASIFL            has filter type choices   : Y, N
#>                                  \--> selected: 
#> BEP01FL             has filter type choices   : Y, N
#>                                  \--> selected: 
#> RANDDT              has filter type unknown   : 
#>                                  \--> selected: 
#> TRTSDTM             has filter type unknown   : 
#>                                  \--> selected: 
#> TRTEDTM             has filter type unknown   : 
#>                                  \--> selected: 
#> EOSSTT              has filter type choices   : COMPLETED, DISCONTINUED, ONGOING
#>                                  \--> selected: 
#> EOSDT               has filter type unknown   : 
#>                                  \--> selected: 
#> EOSDY               has filter type range     : 6 - 731
#>                                  \--> selected: 
#> DCSREAS             has filter type choices   : ADVERSE EVENT, LACK OF EFFICACY, PHYSICIAN DECISION, PROTOCOL VIOLATION, WITHDRAWAL BY PARENT/GUARDIAN, ...
#>                                  \--> selected: 
#> DTHDT               has filter type unknown   : 
#>                                  \--> selected: 
#> LSTALVDT            has filter type unknown   : 
#>                                  \--> selected: 
#> study_duration_secs has filter type choices   : 63113904
#>                                  \--> selected: 
#> ===========================