Initialize MAEFilterStates object

Initialize MAEFilterStates object

Initialize MAEFilterStates object

Details

Initialize MAEFilterStates object

Super class

teal::FilterStates -> MAEFilterStates

Methods

Public methods

MAEFilterStates$new()

MAEFilterStates$get_fun()

MAEFilterStates$set_bookmark_state()

MAEFilterStates$ui_add_filter_state()

MAEFilterStates$srv_add_filter_state()

MAEFilterStates$clone()

Inherited methods

teal::FilterStates$get_call()

teal::FilterStates$queue_empty()

teal::FilterStates$queue_get()

teal::FilterStates$queue_initialize()

teal::FilterStates$queue_push()

teal::FilterStates$queue_remove()

teal::FilterStates$ui()

Method new()

Usage

MAEFilterStates$new(
  input_dataname,
  output_dataname,
  datalabel,
  varlabels,
  keys
)

Arguments

input_dataname

(character(1) or name or call)
name of the data used on lhs of the expression specified to the function argument attached to this FilterStates.

output_dataname

(character(1) or name or call)
name of the output data on the lhs of the assignment expression.

datalabel

(character(0) or character(1))
text label value.

varlabels

(character)
labels of the variables used in this object

keys

(character)
key columns names Get function nameGet function name used to create filter call. For MAEFilterStates MultiAssayExperiment::subsetByColData is used.

Method get_fun()

Usage

MAEFilterStates$get_fun()

Returns

character(1)

Method set_bookmark_state()

Set bookmark state

Usage

MAEFilterStates$set_bookmark_state(id, data, state)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

data

(MultiAssayExperiment)
data which are supposed to be filtered

state

(named list)
should contain values which are initial selection in the FilterState. Names of the list element should correspond to the name of the column in colData(data)

Returns

moduleServer function which returns NULL

Method ui_add_filter_state()

Shiny UI module to add filter variable

Usage

MAEFilterStates$ui_add_filter_state(id, data)

Arguments

id

(character(1))
id of shiny module

data

(MultiAssayExperiment)
object containing colData which columns are used to be used to choose filter variables.

Returns

shiny.tag

Method srv_add_filter_state()

Shiny server module to add filter variable.

Module controls available choices to select as a filter variable. Selected filter variable is being removed from available choices. Removed filter variable gets back to available choices.

Usage

MAEFilterStates$srv_add_filter_state(id, data, ...)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

data

(MultiAssayExperiment)
object containing colData which columns are used to be used to choose filter variables. in optionalSelectInput

...

ignored

Returns

moduleServer function which returns NULL description Get label of specific variable. In case when variable label is missing name of the variable is returned. parameter variable (character(1))
name of the variable for which label should be returned return character(1)

Method clone()

The objects of this class are cloneable with this method.

Usage

MAEFilterStates$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.