R6 Class representing a raw data set

Any data.frame or rtable object can be stored inside this object. Some attributes like colnames, dimension or column names for a specific type will be automatically derived.

Active bindings

ncol

Number of columns

nrow

Number of rows

dim

Dimension c(x, y)

colnames

The column names of the data

rownames

The rownames of the data

raw_data

The data.frame behind this R6 class

data

The data.frame behind this R6 class

column_labels

for backwards compatibility will be deprecated in future

var_names

The column names of the data

row_labels

Row labels (can have spaces)

Methods

Public methods

RawDataset$new()

RawDataset$get_raw_data()

RawDataset$get_numeric_colnames()

RawDataset$get_character_colnames()

RawDataset$get_factor_colnames()

RawDataset$get_colnames()

RawDataset$get_column_labels()

RawDataset$get_rownames()

RawDataset$get_row_labels()

RawDataset$is_pulled()

RawDataset$clone()

Method new()

initialize a RawDataset class object

Usage

RawDataset$new(x)

Arguments

x

(data.frame) object

Returns

RawDataset object

Method get_raw_data()

Derive the raw data frame inside this object

Usage

RawDataset$get_raw_data()

Returns

data.frame or rtable

Method get_numeric_colnames()

Derive the names of all numeric columns

Usage

RawDataset$get_numeric_colnames(include_factors = FALSE)

Arguments

include_factors

logical Whether to include factor variables

Returns

character vector.

Method get_character_colnames()

Derive the names of all character columns

Usage

RawDataset$get_character_colnames()

Returns

character vector.

Method get_factor_colnames()

Derive the names of all factor columns

Usage

RawDataset$get_factor_colnames()

Returns

character vector.

Method get_colnames()

Derive the column names

Usage

RawDataset$get_colnames()

Returns

character vector.

Method get_column_labels()

Derive the column labels

Usage

RawDataset$get_column_labels()

Returns

character vector.

Method get_rownames()

Derive the row names

Usage

RawDataset$get_rownames()

Returns

character vector.

Method get_row_labels()

Derive the row labels

Usage

RawDataset$get_row_labels()

Returns

character vector.

Method is_pulled()

Check if dataset has already been pulled.

Usage

RawDataset$is_pulled()

Returns

TRUE if dataset has been already pulled, else FALSE

Method clone()

The objects of this class are cloneable with this method.

Usage

RawDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.