Manage multiple RelationalDataConnector, DatasetConnector and Dataset objects.

Class manages RelationalDataConnector, DatasetConnector and Dataset objects and aggregate them in one collection. Class also decides whether to launch app before initialize teal application.

Super class

teal::DataAbstract -> RelationalData

Methods

Public methods

RelationalData$new()

RelationalData$get_datanames()

RelationalData$get_join_keys()

RelationalData$get_connectors()

RelationalData$get_items()

RelationalData$get_ui()

RelationalData$get_server()

RelationalData$launch()

RelationalData$mutate_join_keys()

RelationalData$check_metadata()

RelationalData$clone()

Inherited methods

teal::DataAbstract$check()

teal::DataAbstract$check_reproducibility()

teal::DataAbstract$execute_mutate()

teal::DataAbstract$get_check_result()

teal::DataAbstract$get_code()

teal::DataAbstract$get_code_class()

teal::DataAbstract$get_dataset()

teal::DataAbstract$get_datasets()

teal::DataAbstract$is_pulled()

teal::DataAbstract$mutate()

teal::DataAbstract$mutate_dataset()

teal::DataAbstract$set_check()

teal::DataAbstract$set_pull_code()

Method new()

Create a new object of RelationalData class

Usage

RelationalData$new(..., check = FALSE, join_keys)

Arguments

...

(RelationalDataConnector, Dataset, DatasetConnector)
objects

check

(logical) reproducibility check - whether evaluated preprocessing code gives the same objects as provided in arguments. Check is run only if flag is true and preprocessing code is not empty.

join_keys

(JoinKeys) or a single (JoinKeySet)
(optional) object with dataset column relationships used for joining. If empty then no joins between pairs of objects

Method get_datanames()

Derive the names of all datasets

Usage

RelationalData$get_datanames()

Returns

(character vector) with names

Method get_join_keys()

Get JoinKeys object with keys used for joining.

Usage

RelationalData$get_join_keys()

Returns

(JoinKeys) Get data connectors.

Method get_connectors()

Usage

RelationalData$get_connectors()

Returns

(list) with all DatasetConnector or RelationalDataConnector objects.

Method get_items()

Get all datasets and all dataset connectors

name of dataset connector to be returned. If NULL, all connectors are returned.

Usage

RelationalData$get_items(dataname = NULL)

Arguments

dataname

(character value)

Returns

list with all datasets and all connectors

Method get_ui()

Get a shiny-module UI to render the necessary app to derive RelationalDataConnector object's data

Usage

RelationalData$get_ui(id)

Arguments

id

(character) item ID for the shiny module

Returns

the shiny ui function

Method get_server()

Get a shiny-module server to render the necessary app to derive RelationalDataConnector object's data

Usage

RelationalData$get_server()

Returns

shiny server module.

Method launch()

Launch an app that allows to run the user-interfaces of all DataConnector and DatasetConnector modules

This piece is mainly used for debugging.

Usage

RelationalData$launch()

Method mutate_join_keys()

Change join_keys for a given pair of dataset names

Usage

RelationalData$mutate_join_keys(dataset_1, dataset_2, val)

Arguments

dataset_1, dataset_2

(character) datasets for which join_keys are to be returned

val

(named character) column names used to join

Returns

(self) invisibly for chaining

Method check_metadata()

Check there is consistency between the datasets and join_keys

Usage

RelationalData$check_metadata()

Returns

raise and error or invisible TRUE

Method clone()

The objects of this class are cloneable with this method.

Usage

RelationalData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(random.cdisc.data)
x1 <- rcd_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADLB", radlb, cached = TRUE)
)
x2 <- rcd_data(
  rcd_cdisc_dataset_connector("ADRS", radrs, cached = TRUE)
)
x3 <- cdisc_dataset(
  dataname = "ADAE",
  x = radae(cached = TRUE),
  code = "library(random.cdisc.data)\nADAE <- radae(cached = TRUE)"
)

x4 <- rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
tc <- teal:::RelationalData$new(x1, x2, x3, x4)
tc$get_datanames()

#>    ADSL    ADLB    ADRS                 
#>  "ADSL"  "ADLB"  "ADRS"  "ADAE" "ADTTE" 

if (FALSE) {
tc$get_datasets()
tc$get_dataset("ADAE")
}
tc$get_items()

#> $ADSL
#> <CDISCDatasetConnector>
#>   Inherits from: <DatasetConnector>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset: function () 
#>     get_dataset_label: function () 
#>     get_error_message: function () 
#>     get_keys: function () 
#>     get_parent: function () 
#>     get_pull_args: function () 
#>     get_pull_callable: function () 
#>     get_raw_data: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     initialize: function (dataname, pull_callable, keys, parent, code = character(0), 
#>     is_failed: function () 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     pull: function (args = NULL, try = FALSE) 
#>     set_args: function (args) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_ui_input: function (inputs = NULL) 
#>   Private:
#>     dataname: ADSL
#>     dataset: NULL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     get_mutate_code_class: function () 
#>     get_pull_code_class: function (args = NULL) 
#>     keys: STUDYID USUBJID
#>     mutate_code: CodeClass, R6
#>     mutate_vars: list
#>     parent: 
#>     pull_callable: CallableFunction, Callable, R6
#>     pull_internal: function (args = NULL, try = FALSE) 
#>     pull_vars: list
#>     server: function (input, output, session, data_args = NULL) 
#>     set_dataname: function (dataname) 
#>     set_failure: function (res) 
#>     set_mutate_code: function (code) 
#>     set_mutate_vars: function (vars) 
#>     set_parent: function (parent) 
#>     set_pull_callable: function (pull_callable) 
#>     set_pull_vars: function (pull_vars) 
#>     set_ui: function (ui_args = NULL) 
#>     ui: function (id) 
#>     ui_input: NULL
#> 
#> $ADLB
#> <CDISCDatasetConnector>
#>   Inherits from: <DatasetConnector>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset: function () 
#>     get_dataset_label: function () 
#>     get_error_message: function () 
#>     get_keys: function () 
#>     get_parent: function () 
#>     get_pull_args: function () 
#>     get_pull_callable: function () 
#>     get_raw_data: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     initialize: function (dataname, pull_callable, keys, parent, code = character(0), 
#>     is_failed: function () 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     pull: function (args = NULL, try = FALSE) 
#>     set_args: function (args) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_ui_input: function (inputs = NULL) 
#>   Private:
#>     dataname: ADLB
#>     dataset: NULL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     get_mutate_code_class: function () 
#>     get_pull_code_class: function (args = NULL) 
#>     keys: STUDYID USUBJID PARAMCD AVISIT
#>     mutate_code: CodeClass, R6
#>     mutate_vars: list
#>     parent: ADSL
#>     pull_callable: CallableFunction, Callable, R6
#>     pull_internal: function (args = NULL, try = FALSE) 
#>     pull_vars: list
#>     server: function (input, output, session, data_args = NULL) 
#>     set_dataname: function (dataname) 
#>     set_failure: function (res) 
#>     set_mutate_code: function (code) 
#>     set_mutate_vars: function (vars) 
#>     set_parent: function (parent) 
#>     set_pull_callable: function (pull_callable) 
#>     set_pull_vars: function (pull_vars) 
#>     set_ui: function (ui_args = NULL) 
#>     ui: function (id) 
#>     ui_input: NULL
#> 
#> $ADRS
#> <CDISCDatasetConnector>
#>   Inherits from: <DatasetConnector>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset: function () 
#>     get_dataset_label: function () 
#>     get_error_message: function () 
#>     get_keys: function () 
#>     get_parent: function () 
#>     get_pull_args: function () 
#>     get_pull_callable: function () 
#>     get_raw_data: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     initialize: function (dataname, pull_callable, keys, parent, code = character(0), 
#>     is_failed: function () 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     pull: function (args = NULL, try = FALSE) 
#>     set_args: function (args) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_ui_input: function (inputs = NULL) 
#>   Private:
#>     dataname: ADRS
#>     dataset: NULL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     get_mutate_code_class: function () 
#>     get_pull_code_class: function (args = NULL) 
#>     keys: STUDYID USUBJID PARAMCD AVISIT
#>     mutate_code: CodeClass, R6
#>     mutate_vars: list
#>     parent: ADSL
#>     pull_callable: CallableFunction, Callable, R6
#>     pull_internal: function (args = NULL, try = FALSE) 
#>     pull_vars: list
#>     server: function (input, output, session, data_args = NULL) 
#>     set_dataname: function (dataname) 
#>     set_failure: function (res) 
#>     set_mutate_code: function (code) 
#>     set_mutate_vars: function (vars) 
#>     set_parent: function (parent) 
#>     set_pull_callable: function (pull_callable) 
#>     set_pull_vars: function (pull_vars) 
#>     set_ui: function (ui_args = NULL) 
#>     ui: function (id) 
#>     ui_input: NULL
#> 
#> $ADAE
#> <CDISCDataset>
#>   Inherits from: <Dataset>
#>   Public:
#>     check: function () 
#>     check_keys: function (keys = private$.keys) 
#>     clone: function (deep = FALSE) 
#>     colnames: active binding
#>     column_labels: active binding
#>     data: active binding
#>     dim: active binding
#>     get_attrs: function () 
#>     get_character_colnames: function () 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_colnames: function () 
#>     get_column_labels: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset_label: function () 
#>     get_factor_colnames: function () 
#>     get_keys: function () 
#>     get_numeric_colnames: function () 
#>     get_parent: function () 
#>     get_raw_data: function () 
#>     get_row_labels: function () 
#>     get_rownames: function () 
#>     initialize: function (dataname, x, keys, parent, code = character(0), label = character(0), 
#>     is_pulled: function () 
#>     mutate: function (code, vars = list()) 
#>     ncol: active binding
#>     nrow: active binding
#>     raw_data: active binding
#>     recreate: function (dataname = self$get_dataname(), x = self$get_raw_data(), 
#>     row_labels: active binding
#>     rownames: active binding
#>     set_code: function (code) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_parent: function (parent) 
#>     set_vars: function (vars) 
#>     var_names: active binding
#>   Private:
#>     .col_labels: Study Identifier Unique Subject Identifier Subject Ident ...
#>     .colnames: STUDYID USUBJID SUBJID SITEID AGE SEX RACE COUNTRY DTHFL ...
#>     .dim: 1934 65
#>     .keys: STUDYID USUBJID ASTDTM AETERM AESEQ
#>     .ncol: 65
#>     .nrow: 1934
#>     .raw_data: tbl_df, tbl, data.frame
#>     .row_labels: NULL
#>     .rownames: 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 ...
#>     code: CodeClass, R6
#>     dataname: ADAE
#>     dataset_label: Adverse Events Analysis Dataset
#>     deep_clone: function (name, value) 
#>     execute_code: function (code, vars = list()) 
#>     get_class_colnames: function (class_type = "character") 
#>     parent: ADSL
#>     set_dataname: function (dataname) 
#>     vars: list
#> 
#> $ADTTE
#> <CDISCDatasetConnector>
#>   Inherits from: <DatasetConnector>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset: function () 
#>     get_dataset_label: function () 
#>     get_error_message: function () 
#>     get_keys: function () 
#>     get_parent: function () 
#>     get_pull_args: function () 
#>     get_pull_callable: function () 
#>     get_raw_data: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     initialize: function (dataname, pull_callable, keys, parent, code = character(0), 
#>     is_failed: function () 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     pull: function (args = NULL, try = FALSE) 
#>     set_args: function (args) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_ui_input: function (inputs = NULL) 
#>   Private:
#>     dataname: ADTTE
#>     dataset: NULL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     get_mutate_code_class: function () 
#>     get_pull_code_class: function (args = NULL) 
#>     keys: STUDYID USUBJID PARAMCD
#>     mutate_code: CodeClass, R6
#>     mutate_vars: list
#>     parent: ADSL
#>     pull_callable: CallableFunction, Callable, R6
#>     pull_internal: function (args = NULL, try = FALSE) 
#>     pull_vars: list
#>     server: function (input, output, session, data_args = NULL) 
#>     set_dataname: function (dataname) 
#>     set_failure: function (res) 
#>     set_mutate_code: function (code) 
#>     set_mutate_vars: function (vars) 
#>     set_parent: function (parent) 
#>     set_pull_callable: function (pull_callable) 
#>     set_pull_vars: function (pull_vars) 
#>     set_ui: function (ui_args = NULL) 
#>     ui: function (id) 
#>     ui_input: NULL
#> 

tc$get_code()

#> [1] "library(package = \"random.cdisc.data\")\nADSL <- radsl(cached = TRUE)\nADLB <- radlb(cached = TRUE)\nADRS <- radrs(cached = TRUE)\nlibrary(random.cdisc.data)\nADAE <- radae(cached = TRUE)\nADTTE <- radtte(cached = TRUE)"

tc$get_code("ADAE")

#> [1] "library(package = \"random.cdisc.data\")\nlibrary(random.cdisc.data)\nADAE <- radae(cached = TRUE)"

if (FALSE) {
tc$launch()
tc$get_datasets()
tc$get_dataset("ADAE")
tc$check()
}

library(random.cdisc.data)
x <- cdisc_dataset(
  dataname = "ADSL",
  x = radsl(cached = TRUE),
  code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)"
)

x2 <- rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
tc <- teal:::RelationalData$new(x, x2)
if (FALSE) {
tc$get_datasets()
get_raw_data(tc)
tc$launch()
get_raw_data(tc)
}