RelationalData class

Class combines multiple RelationalDataset objects.

Methods

Public methods

RelationalData$new()

RelationalData$check()

RelationalData$get_check_result()

RelationalData$get_code()

RelationalData$get_code_class()

RelationalData$get_datanames()

RelationalData$get_dataset()

RelationalData$get_datasets()

RelationalData$get_items()

RelationalData$is_pulled()

RelationalData$mutate()

RelationalData$mutate_dataset()

RelationalData$set_check()

RelationalData$set_pull_code()

RelationalData$clone()

Method new()

Create a new RelationalData object from multiple RelationalDataset objects.

Usage

RelationalData$new(...)

Arguments

...

(RelationalDataset)
at least one object.

Returns

new RelationalData object

Method check()

Check if the object raw data is reproducible from the get_code() code.

Usage

RelationalData$check()

Returns

NULL if check step has been disabled TRUE if all the datasets generated from evaluating the get_code() code are identical to the raw data, else FALSE.

Method get_check_result()

Get result of reproducibility check

Usage

RelationalData$get_check_result()

Returns

NULL if check has not been called yet, TRUE / FALSE otherwise

Method get_code()

Get code for all datasets.

Usage

RelationalData$get_code(dataname = NULL, deparse = TRUE)

Arguments

dataname

(character) dataname or NULL for all datasets

deparse

(logical) whether to return the deparsed form of a call

Returns

(character) vector of code to generate datasets.

Method get_code_class()

Get internal CodeClass object

Usage

RelationalData$get_code_class(only_pull = FALSE)

Arguments

only_pull

(logical value)
if TRUE only code to pull datasets will be returned without mutate code.

Returns

CodeClass

Method get_datanames()

Get names of the datasets.

Usage

RelationalData$get_datanames()

Returns

character vector with names of all datasets.

Method get_dataset()

Get RelationalDataset object.

Usage

RelationalData$get_dataset(dataname = NULL)

Arguments

dataname

(character value)
name of dataset to be returned. If NULL, all datasets are returned.

Returns

RelationalDataset.

Method get_datasets()

Get list of RelationalDataset objects.

Usage

RelationalData$get_datasets()

Returns

list of RelationalDataset.

Method get_items()

Get all datasets and all dataset connectors

name of dataset connector to be returned. If NULL, all connectors are returned.

Usage

RelationalData$get_items(dataname = NULL)

Arguments

dataname

(character value)

Returns

list with all datasets and all connectors

Method is_pulled()

Check if dataset has already been pulled.

Usage

RelationalData$is_pulled()

Returns

TRUE if dataset has been already pulled, else FALSE

Method mutate()

Mutate data by code

Usage

RelationalData$mutate(code, vars = list())

Arguments

code

(character) Code to mutate the dataset. Must contain the dataset$dataname

vars

(list)
In case when this object code depends on the raw_data from the other RelationalDataset, RelationalDatasetConnector object(s) or other constant value, this/these object(s) should be included

Returns

self invisibly for chaining

Method mutate_dataset()

Mutate dataset by code

Usage

RelationalData$mutate_dataset(dataname, code, vars = list())

Arguments

dataname

(character) Dataname to be mutated

code

(character) Code to mutate the dataset. Must contain the dataset$dataname

vars

(list)
In case when this object code depends on the raw_data from the other RelationalDataset, RelationalDatasetConnector object(s) or other constant value, this/these object(s) should be included

Returns

self invisibly for chaining

Method set_check()

Set reproducibility check

Usage

RelationalData$set_check(check = FALSE)

Arguments

check

(logical) whether to perform reproducibility check.

Returns

self invisibly for chaining.

Method set_pull_code()

Set pull code

Usage

RelationalData$set_pull_code(code)

Arguments

code

(character value)
code to reproduce data in RawDataset objects. Can't be set if any dataset has code set already.

Returns

self invisibly for chaining.

Method clone()

The objects of this class are cloneable with this method.

Usage

RelationalData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- relational_dataset(
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = keys(primary = "y", foreign = NULL, parent = NULL),
  dataname = "XY",
  code = "XY <- data.frame(x = c(1, 2), y = c('a', 'b'),
                           stringsAsFactors = FALSE)",
  label = character(0)
)

x2 <- relational_dataset(
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = keys(primary = "y", foreign = NULL, parent = NULL),
  dataname = "XYZ",
  code = "XYZ <- data.frame(x = c(1, 2), y = c('a', 'b'),
                           stringsAsFactors = FALSE)",
  label = character(0)
)

rd <- teal:::RelationalData$new(x, x2)