Manage multiple and RelationalDatasetConnector of the same type.

Class manages RelationalDatasetConnector to specify additional dynamic arguments and to open/close connection.

Super class

teal::RelationalData -> RelationalDataConnector

Methods

Public methods

RelationalDataConnector$new()

RelationalDataConnector$get_code_class()

RelationalDataConnector$get_connection()

RelationalDataConnector$get_server()

RelationalDataConnector$get_ui()

RelationalDataConnector$is_failed()

RelationalDataConnector$launch()

RelationalDataConnector$pull()

RelationalDataConnector$set_ui()

RelationalDataConnector$set_server()

RelationalDataConnector$clone()

Inherited methods

teal::RelationalData$check()

teal::RelationalData$get_check_result()

teal::RelationalData$get_code()

teal::RelationalData$get_datanames()

teal::RelationalData$get_dataset()

teal::RelationalData$get_datasets()

teal::RelationalData$get_items()

teal::RelationalData$is_pulled()

teal::RelationalData$mutate()

teal::RelationalData$mutate_dataset()

teal::RelationalData$set_check()

teal::RelationalData$set_pull_code()

Method new()

Create a new RelationalDataConnector object

Usage

RelationalDataConnector$new(connection, connectors)

Arguments

connection

(DataConnection) connection to data source

connectors

(list of RelationalDatasetConnector elements) list with dataset connectors

Returns

new RelationalDataConnector object

Method get_code_class()

Get internal CodeClass object

Usage

RelationalDataConnector$get_code_class()

Returns

CodeClass

Method get_connection()

Get connection to data source

Usage

RelationalDataConnector$get_connection()

Returns

connector's connection

Method get_server()

Usage

RelationalDataConnector$get_server()

Returns

the server function of the RelationalDataConnector

Method get_ui()

Get Shiny module with inputs for all RelationalDatasetConnector objects

Usage

RelationalDataConnector$get_ui(id)

Arguments

id

character shiny element id

Returns

the ui function of the RelationalDataConnector

Method is_failed()

Check if pull or connection has not failed.

Usage

RelationalDataConnector$is_failed()

Returns

TRUE if pull or connection failed, else FALSE

Method launch()

Run simple application that uses its ui and server fields to pull data from connection.

Useful for debugging

Usage

RelationalDataConnector$launch()

Returns

An object that represents the app

Method pull()

Load data from each RelationalDatasetConnector

Usage

RelationalDataConnector$pull(con_args = NULL, args = NULL, try = TRUE)

Arguments

con_args

(NULL or named list)
additional dynamic arguments for connection function. args will be passed to each RelationalDatasetConnector object to evaluate CallableFunction assigned to this dataset. If args is null than default set of arguments will be used, otherwise call will be executed on these arguments only (arguments set before will be ignored). pull function doesn't update reproducible call, it's just evaluate function.

args

(NULL or named list)
additional dynamic arguments to pull dataset. args will be passed to each RelationalDatasetConnector object to evaluate CallableFunction assigned to this dataset. If args is null than default set of arguments will be used, otherwise call will be executed on these arguments only (arguments set before will be ignored). pull function doesn't update reproducible call, it's just evaluate function.

try

(logical value)
whether perform function evaluation inside try clause

Returns

self invisibly for chaining. In order to get the data please use get_datasets method.

Method set_ui()

Set connector UI function

Usage

RelationalDataConnector$set_ui(data_input)

Arguments

data_input

(function)
shiny module as function. Inputs specified in this ui are passed to server module defined by set_server method.

Returns

nothing

Method set_server()

Set connector server function

This function will be called after submit button will be hit. There is no possibility to specify some dynamic ui as server function is executed after hitting submit button.

Usage

RelationalDataConnector$set_server(data_module)

Arguments

data_module

(function)
A shiny module server function that should load data from all connectors

Returns

NULL or self if data is loaded.

Method clone()

The objects of this class are cloneable with this method.

Usage

RelationalDataConnector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector("ADSL", radsl)
adlb <- rcd_cdisc_dataset_connector("ADLB", radlb, ADSL = adsl)
con <- teal:::rcd_connection()

ui <- function(id) {
  ns <- NS(id)
  tagList(
    numericInput(ns("seed"), "Choose seed", min = 1, max = 1000, value = 1),
    sliderInput(ns("N"), "Choose number of observations", min = 1, max = 400, value = 10)
  )
}

x <- teal:::RelationalDataConnector$new(connection = con, connectors = list(adsl, adlb))

x$set_ui(ui)
x$set_server(function(input, output, session, connectors, connection) {
  lapply(connectors, function(connector) {
    if (get_dataname(connector) == "ADSL") {
      set_args(connector, args = list(seed = input$seed, N = input$N))
    } else {
      set_args(connector, args = list(seed = input$seed))
    }
    connector$pull(try = TRUE)
  })
})
if (FALSE) {
x$launch()
x$get_datasets()
}