TealReportCard

Child class of teal.reporter::ReportCard that is used for teal specific applications. In addition to the parent methods, it supports rendering teal specific elements such as the source code, the encodings panel content and the filter panel content as part of the meta data.

Super class

teal.reporter::ReportCard -> TealReportCard

Methods

Public methods

TealReportCard$append_src()

TealReportCard$append_fs()

TealReportCard$append_encodings()

TealReportCard$clone()

Inherited methods

teal.reporter::ReportCard$append_content()

teal.reporter::ReportCard$append_html()

teal.reporter::ReportCard$append_metadata()

teal.reporter::ReportCard$append_plot()

teal.reporter::ReportCard$append_rcode()

teal.reporter::ReportCard$append_table()

teal.reporter::ReportCard$append_text()

teal.reporter::ReportCard$from_list()

teal.reporter::ReportCard$get_content()

teal.reporter::ReportCard$get_metadata()

teal.reporter::ReportCard$get_name()

teal.reporter::ReportCard$initialize()

teal.reporter::ReportCard$reset()

teal.reporter::ReportCard$set_name()

teal.reporter::ReportCard$to_list()

Method append_src()

Appends the source code to the content meta data of this TealReportCard.

Usage

TealReportCard$append_src(src, ...)

Arguments

src

(character(1)) code as text.

...

any rmarkdown R chunk parameter and its value. But eval parameter is always set to FALSE.

Returns

Object of class TealReportCard, invisibly.

Examples

card <- TealReportCard$new()$append_src(
  "plot(iris)"
)
card$get_content()[[1]]$get_content()

Method append_fs()

Appends the filter state list to the content and metadata of this TealReportCard. If the filter state list has an attribute named formatted, it appends it to the card otherwise it uses the default yaml::as.yaml to format the list. If the filter state list is empty, nothing is appended to the content.

Usage

TealReportCard$append_fs(fs)

Arguments

fs

(teal_slices) object returned from teal_slices() function.

Returns

self, invisibly.

Method append_encodings()

Appends the encodings list to the content and metadata of this TealReportCard.

Usage

TealReportCard$append_encodings(encodings)

Arguments

encodings

(list) list of encodings selections of the teal app.

Returns

self, invisibly.

Examples

card <- TealReportCard$new()$append_encodings(list(variable1 = "X"))
card$get_content()[[1]]$get_content()

Method clone()

The objects of this class are cloneable with this method.

Usage

TealReportCard$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `TealReportCard$append_src`
## ------------------------------------------------

card <- TealReportCard$new()$append_src(
  "plot(iris)"
)
card$get_content()[[1]]$get_content()
#> [1] "plot(iris)"

## ------------------------------------------------
## Method `TealReportCard$append_encodings`
## ------------------------------------------------

card <- TealReportCard$new()$append_encodings(list(variable1 = "X"))
card$get_content()[[1]]$get_content()
#> [1] "Selected Options"