Data input for teal app

Function passes datasets to teal application with option to read preprocessing code and reproducibility checking.

cdisc_data(..., join_keys, code = "", check = FALSE)

Arguments

...

(TealDataConnector, TealDataset or TealDatasetConnector) elements to include where ADSL data is mandatory.

join_keys

(JoinKeys) or a single (JoinKeySet)
(optional) object with datasets column names used for joining. If empty then it would be automatically derived basing on intersection of datasets primary keys

code

(character) code to reproduce the datasets.

check

(logical) reproducibility check - whether evaluated preprocessing code gives the same objects as provided in arguments. Check is run only if flag is true and preprocessing code is not empty.

Value

a CDISCTealData object

Details

This function checks if there were keys added to all data sets that shall be analyzed inside a teal app.

Note

This function does not automatically assign keys to TealDataset and TealDatasetConnector objects passed to it. If the keys are needed they should be assigned before calling cdisc_data. See example:

Examples

library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl
ADTTE <- synthetic_cdisc_data("latest")$adtte

# basic example
cdisc_data(
  cdisc_dataset("ADSL", ADSL),
  cdisc_dataset("ADTTE", ADTTE),
  code = 'ADSL <- synthetic_cdisc_data("latest")$adsl
          ADTTE <- synthetic_cdisc_data("latest")$adtte',
  check = TRUE
)
#> A CDISCTealData object containing 2 TealDataset/TealDatasetConnector object(s) as element(s):
#> --> Element 1:
#> A TealDataset object containing the following data.frame (400 rows and 39 columns):
#>   STUDYID               USUBJID SUBJID SITEID AGE SEX                      RACE
#> 1 AB12345  AB12345-CHN-3-id-128 id-128  CHN-3  32   M                     ASIAN
#> 2 AB12345 AB12345-CHN-15-id-262 id-262 CHN-15  35   M BLACK OR AFRICAN AMERICAN
#> 3 AB12345  AB12345-RUS-3-id-378 id-378  RUS-3  30   F                     ASIAN
#> 4 AB12345 AB12345-CHN-11-id-220 id-220 CHN-11  26   F                     ASIAN
#> 5 AB12345  AB12345-CHN-7-id-267 id-267  CHN-7  40   M                     ASIAN
#> 6 AB12345 AB12345-CHN-15-id-201 id-201 CHN-15  49   M                     ASIAN
#>   COUNTRY DTHFL         INVID         INVNAM            ARM ARMCD
#> 1     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A
#> 2     CHN     N INV ID CHN-15 Dr. CHN-15 Doe C: Combination ARM C
#> 3     RUS     N  INV ID RUS-3  Dr. RUS-3 Doe C: Combination ARM C
#> 4     CHN     N INV ID CHN-11 Dr. CHN-11 Doe     B: Placebo ARM B
#> 5     CHN     N  INV ID CHN-7  Dr. CHN-7 Doe     B: Placebo ARM B
#> 6     CHN     N INV ID CHN-15 Dr. CHN-15 Doe C: Combination ARM C
#>           ACTARM ACTARMCD REGION1 STRATA1 STRATA2    BMRKR1 BMRKR2 ITTFL SAFFL
#> 1      A: Drug X    ARM A    Asia       C      S2 14.424934 MEDIUM     Y     Y
#> 2 C: Combination    ARM C    Asia       C      S1  4.055463    LOW     Y     Y
#> 3 C: Combination    ARM C Eurasia       A      S1  2.803240   HIGH     Y     Y
#> 4     B: Placebo    ARM B    Asia       B      S2 10.262734 MEDIUM     Y     Y
#> 5     B: Placebo    ARM B    Asia       C      S1  6.206763    LOW     Y     Y
#> 6 C: Combination    ARM C    Asia       C      S2  6.906799 MEDIUM     Y     Y
#>   BMEASIFL BEP01FL     RANDDT             TRTSDTM             TRTEDTM
#> 1        Y       Y 2019-02-22 2019-02-24 11:09:18 2021-02-23 22:47:42
#> 2        N       N 2019-02-26 2019-02-26 09:05:00 2021-02-25 20:43:24
#> 3        Y       N 2019-02-24 2019-02-28 03:19:08 2021-02-27 14:57:32
#> 4        Y       Y 2019-02-27 2019-03-01 13:33:03 2021-03-01 01:11:27
#> 5        N       N 2019-03-01 2019-03-02 00:09:16 2021-03-01 11:47:40
#> 6        Y       N 2019-03-05 2019-03-05 15:23:44 2021-02-17 20:23:53
#>         EOSSTT       EOTSTT      EOSDT EOSDY          DCSREAS DTHDT DTHCAUS
#> 1    COMPLETED    COMPLETED 2021-02-23   731             <NA>  <NA>    <NA>
#> 2    COMPLETED    COMPLETED 2021-02-25   731             <NA>  <NA>    <NA>
#> 3    COMPLETED    COMPLETED 2021-02-27   731             <NA>  <NA>    <NA>
#> 4    COMPLETED    COMPLETED 2021-03-01   731             <NA>  <NA>    <NA>
#> 5    COMPLETED    COMPLETED 2021-03-01   731             <NA>  <NA>    <NA>
#> 6 DISCONTINUED DISCONTINUED 2021-02-17   716 LACK OF EFFICACY  <NA>    <NA>
#>   DTHCAT LDDTHELD LDDTHGR1   LSTALVDT study_duration_secs
#> 1   <NA>       NA     <NA> 2021-03-05            63113904
#> 2   <NA>       NA     <NA> 2021-03-15            63113904
#> 3   <NA>       NA     <NA> 2021-03-15            63113904
#> 4   <NA>       NA     <NA> 2021-03-17            63113904
#> 5   <NA>       NA     <NA> 2021-03-25            63113904
#> 6   <NA>       NA     <NA> 2021-03-01            63113904
#> 
#> ...
#>     STUDYID               USUBJID SUBJID SITEID AGE SEX
#> 395 AB12345   AB12345-CHN-9-id-11  id-11  CHN-9  28   F
#> 396 AB12345 AB12345-CHN-16-id-352 id-352 CHN-16  28   M
#> 397 AB12345  AB12345-CHN-1-id-186 id-186  CHN-1  27   M
#> 398 AB12345  AB12345-CHN-1-id-371 id-371  CHN-1  28   F
#> 399 AB12345  AB12345-CHN-1-id-233 id-233  CHN-1  36   F
#> 400 AB12345 AB12345-USA-12-id-131 id-131 USA-12  44   F
#>                                          RACE COUNTRY DTHFL         INVID
#> 395 NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER     CHN     N  INV ID CHN-9
#> 396                                     ASIAN     CHN     N INV ID CHN-16
#> 397                                     ASIAN     CHN     N  INV ID CHN-1
#> 398                                     ASIAN     CHN     Y  INV ID CHN-1
#> 399                 BLACK OR AFRICAN AMERICAN     CHN     N  INV ID CHN-1
#> 400          AMERICAN INDIAN OR ALASKA NATIVE     USA     N INV ID USA-12
#>             INVNAM            ARM ARMCD         ACTARM ACTARMCD       REGION1
#> 395  Dr. CHN-9 Doe     B: Placebo ARM B     B: Placebo    ARM B          Asia
#> 396 Dr. CHN-16 Doe C: Combination ARM C C: Combination    ARM C          Asia
#> 397  Dr. CHN-1 Doe     B: Placebo ARM B     B: Placebo    ARM B          Asia
#> 398  Dr. CHN-1 Doe      A: Drug X ARM A      A: Drug X    ARM A          Asia
#> 399  Dr. CHN-1 Doe C: Combination ARM C C: Combination    ARM C          Asia
#> 400 Dr. USA-12 Doe      A: Drug X ARM A      A: Drug X    ARM A North America
#>     STRATA1 STRATA2   BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL     RANDDT
#> 395       C      S1 4.267916 MEDIUM     Y     Y        Y       Y 2021-01-26
#> 396       A      S2 4.216812 MEDIUM     Y     Y        N       Y 2021-01-30
#> 397       A      S2 4.521179   HIGH     Y     Y        Y       N 2021-01-30
#> 398       A      S2 6.326863    LOW     Y     Y        N       Y 2021-02-07
#> 399       C      S1 4.076818    LOW     Y     Y        N       Y 2021-02-09
#> 400       B      S1 3.726935    LOW     Y     Y        Y       N 2021-02-14
#>                 TRTSDTM             TRTEDTM       EOSSTT       EOTSTT
#> 395 2021-01-27 01:57:07 2023-01-27 13:35:31    COMPLETED    COMPLETED
#> 396 2021-02-02 09:04:52 2023-02-02 20:43:16    COMPLETED    COMPLETED
#> 397 2021-02-02 19:14:46 2023-02-03 06:53:10    COMPLETED    COMPLETED
#> 398 2021-02-09 05:28:52 2021-02-19 19:06:01 DISCONTINUED DISCONTINUED
#> 399 2021-02-11 21:33:08 2021-02-17 09:12:06 DISCONTINUED DISCONTINUED
#> 400 2021-02-17 14:42:27 2023-02-18 02:20:51    COMPLETED    COMPLETED
#>          EOSDT EOSDY            DCSREAS      DTHDT DTHCAUS DTHCAT LDDTHELD
#> 395 2023-01-27   731               <NA>       <NA>    <NA>   <NA>       NA
#> 396 2023-02-02   731               <NA>       <NA>    <NA>   <NA>       NA
#> 397 2023-02-03   731               <NA>       <NA>    <NA>   <NA>       NA
#> 398 2021-02-19    11              DEATH 2021-03-29 MISSING  OTHER       38
#> 399 2021-02-17     6 PHYSICIAN DECISION       <NA>    <NA>   <NA>       NA
#> 400 2023-02-18   731               <NA>       <NA>    <NA>   <NA>       NA
#>     LDDTHGR1   LSTALVDT study_duration_secs
#> 395     <NA> 2023-02-22            63113904
#> 396     <NA> 2023-02-23            63113904
#> 397     <NA> 2023-03-03            63113904
#> 398      >30 2021-03-29            63113904
#> 399     <NA> 2021-03-17            63113904
#> 400     <NA> 2023-03-01            63113904
#> --> Element 2:
#> A TealDataset object containing the following data.frame (1600 rows and 50 columns):
#>   STUDYID              USUBJID SUBJID SITEID AGE SEX                      RACE
#> 1 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38   M BLACK OR AFRICAN AMERICAN
#> 2 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38   M BLACK OR AFRICAN AMERICAN
#> 3 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38   M BLACK OR AFRICAN AMERICAN
#> 4 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38   M BLACK OR AFRICAN AMERICAN
#> 5 AB12345 AB12345-BRA-1-id-134 id-134  BRA-1  47   M                     WHITE
#> 6 AB12345 AB12345-BRA-1-id-134 id-134  BRA-1  47   M                     WHITE
#>   COUNTRY DTHFL        INVID        INVNAM       ARM ARMCD    ACTARM ACTARMCD
#> 1     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#> 2     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#> 3     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#> 4     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#> 5     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#> 6     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#>         REGION1 STRATA1 STRATA2   BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL
#> 1 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 2 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 3 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 4 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 5 South America       B      S2 6.462991    LOW     Y     Y        Y       N
#> 6 South America       B      S2 6.462991    LOW     Y     Y        Y       N
#>       RANDDT             TRTSDTM             TRTEDTM       EOSSTT       EOTSTT
#> 1 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:47 DISCONTINUED DISCONTINUED
#> 2 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:47 DISCONTINUED DISCONTINUED
#> 3 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:47 DISCONTINUED DISCONTINUED
#> 4 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:47 DISCONTINUED DISCONTINUED
#> 5 2020-11-03 2020-11-04 03:50:33 2022-11-04 15:28:57    COMPLETED    COMPLETED
#> 6 2020-11-03 2020-11-04 03:50:33 2022-11-04 15:28:57    COMPLETED    COMPLETED
#>        EOSDT EOSDY       DCSREAS DTHDT DTHCAUS DTHCAT LDDTHELD LDDTHGR1
#> 1 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 2 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 3 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 4 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 5 2022-11-04   731          <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 6 2022-11-04   731          <NA>  <NA>    <NA>   <NA>       NA     <NA>
#>     LSTALVDT study_duration_secs ASEQ TTESEQ                          PARAM
#> 1 2021-03-07            63113904    2      2 Duration of Confirmed Response
#> 2 2021-03-07            63113904    4      4            Event Free Survival
#> 3 2021-03-07            63113904    3      3               Overall Survival
#> 4 2021-03-07            63113904    1      1      Progression Free Survival
#> 5 2022-11-15            63113904    4      4 Duration of Confirmed Response
#> 6 2022-11-15            63113904    2      2            Event Free Survival
#>   PARAMCD      AVAL AVALU       ADTM ADY CNSR                    EVNTDESC
#> 1    CRSD 1273.8498  DAYS 2020-08-13 158    1 Last Date Known To Be Alive
#> 2     EFS  135.9787  DAYS 2021-02-16 345    0                       Death
#> 3      OS 1759.8678  DAYS 2020-11-05 242    0                       Death
#> 4     PFS  241.1912  DAYS 2020-07-23 137    0         Disease Progression
#> 5    CRSD  544.5728  DAYS 2021-10-25 355    0                       Death
#> 6     EFS  401.2725  DAYS 2021-05-05 182    1 Last Date Known To Be Alive
#>                        CNSDTDSC
#> 1    End of AE Reporting Period
#> 2                              
#> 3                              
#> 4                              
#> 5                              
#> 6 Completion or Discontinuation
#> 
#> ...
#>      STUDYID              USUBJID SUBJID SITEID AGE SEX  RACE COUNTRY DTHFL
#> 1595 AB12345 AB12345-USA-8-id-206 id-206  USA-8  34   F ASIAN     USA     N
#> 1596 AB12345 AB12345-USA-8-id-206 id-206  USA-8  34   F ASIAN     USA     N
#> 1597 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#> 1598 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#> 1599 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#> 1600 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#>             INVID        INVNAM            ARM ARMCD         ACTARM ACTARMCD
#> 1595 INV ID USA-8 Dr. USA-8 Doe C: Combination ARM C C: Combination    ARM C
#> 1596 INV ID USA-8 Dr. USA-8 Doe C: Combination ARM C C: Combination    ARM C
#> 1597 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 1598 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 1599 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 1600 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#>            REGION1 STRATA1 STRATA2    BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL
#> 1595 North America       B      S1 14.478057   HIGH     Y     Y        Y
#> 1596 North America       B      S1 14.478057   HIGH     Y     Y        Y
#> 1597 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#> 1598 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#> 1599 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#> 1600 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#>      BEP01FL     RANDDT             TRTSDTM             TRTEDTM    EOSSTT
#> 1595       Y 2019-07-19 2019-07-22 10:19:01 2021-07-21 21:57:25 COMPLETED
#> 1596       Y 2019-07-19 2019-07-22 10:19:01 2021-07-21 21:57:25 COMPLETED
#> 1597       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#> 1598       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#> 1599       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#> 1600       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#>         EOTSTT      EOSDT EOSDY DCSREAS DTHDT DTHCAUS DTHCAT LDDTHELD LDDTHGR1
#> 1595 COMPLETED 2021-07-21   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 1596 COMPLETED 2021-07-21   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 1597 COMPLETED 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 1598 COMPLETED 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 1599 COMPLETED 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 1600 COMPLETED 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#>        LSTALVDT study_duration_secs ASEQ TTESEQ                          PARAM
#> 1595 2021-08-01            63113904    1      1               Overall Survival
#> 1596 2021-08-01            63113904    2      2      Progression Free Survival
#> 1597 2022-08-17            63113904    4      4 Duration of Confirmed Response
#> 1598 2022-08-17            63113904    3      3            Event Free Survival
#> 1599 2022-08-17            63113904    1      1               Overall Survival
#> 1600 2022-08-17            63113904    2      2      Progression Free Survival
#>      PARAMCD      AVAL AVALU       ADTM ADY CNSR                    EVNTDESC
#> 1595      OS 1312.4084  DAYS 2019-11-09 110    0                       Death
#> 1596     PFS  429.6270  DAYS 2020-05-03 286    1 Last Date Known To Be Alive
#> 1597    CRSD  832.4600  DAYS 2021-01-23 189    0                       Death
#> 1598     EFS  471.2179  DAYS 2021-01-18 184    1 Last Date Known To Be Alive
#> 1599      OS 1475.7467  DAYS 2020-08-03  16    1       Last Tumor Assessment
#> 1600     PFS  147.4837  DAYS 2020-08-25  38    1       Last Tumor Assessment
#>                           CNSDTDSC
#> 1595                              
#> 1596              Clinical Cut Off
#> 1597                              
#> 1598    End of AE Reporting Period
#> 1599 Completion or Discontinuation
#> 1600 Completion or Discontinuation

# Example with keys
cdisc_data(
  cdisc_dataset("ADSL", ADSL,
    keys = c("STUDYID", "USUBJID")
  ),
  cdisc_dataset("ADTTE", ADTTE,
    keys = c("STUDYID", "USUBJID", "PARAMCD"),
    parent = "ADSL"
  ),
  join_keys = join_keys(
    join_key(
      "ADSL",
      "ADTTE",
      c("STUDYID" = "STUDYID", "USUBJID" = "USUBJID")
    )
  ),
  code = 'ADSL <- synthetic_cdisc_data("latest")$adsl
          ADTTE <- synthetic_cdisc_data("latest")$adtte',
  check = TRUE
)
#> A CDISCTealData object containing 2 TealDataset/TealDatasetConnector object(s) as element(s):
#> --> Element 1:
#> A TealDataset object containing the following data.frame (400 rows and 39 columns):
#>   STUDYID               USUBJID SUBJID SITEID AGE SEX                      RACE
#> 1 AB12345  AB12345-CHN-3-id-128 id-128  CHN-3  32   M                     ASIAN
#> 2 AB12345 AB12345-CHN-15-id-262 id-262 CHN-15  35   M BLACK OR AFRICAN AMERICAN
#> 3 AB12345  AB12345-RUS-3-id-378 id-378  RUS-3  30   F                     ASIAN
#> 4 AB12345 AB12345-CHN-11-id-220 id-220 CHN-11  26   F                     ASIAN
#> 5 AB12345  AB12345-CHN-7-id-267 id-267  CHN-7  40   M                     ASIAN
#> 6 AB12345 AB12345-CHN-15-id-201 id-201 CHN-15  49   M                     ASIAN
#>   COUNTRY DTHFL         INVID         INVNAM            ARM ARMCD
#> 1     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A
#> 2     CHN     N INV ID CHN-15 Dr. CHN-15 Doe C: Combination ARM C
#> 3     RUS     N  INV ID RUS-3  Dr. RUS-3 Doe C: Combination ARM C
#> 4     CHN     N INV ID CHN-11 Dr. CHN-11 Doe     B: Placebo ARM B
#> 5     CHN     N  INV ID CHN-7  Dr. CHN-7 Doe     B: Placebo ARM B
#> 6     CHN     N INV ID CHN-15 Dr. CHN-15 Doe C: Combination ARM C
#>           ACTARM ACTARMCD REGION1 STRATA1 STRATA2    BMRKR1 BMRKR2 ITTFL SAFFL
#> 1      A: Drug X    ARM A    Asia       C      S2 14.424934 MEDIUM     Y     Y
#> 2 C: Combination    ARM C    Asia       C      S1  4.055463    LOW     Y     Y
#> 3 C: Combination    ARM C Eurasia       A      S1  2.803240   HIGH     Y     Y
#> 4     B: Placebo    ARM B    Asia       B      S2 10.262734 MEDIUM     Y     Y
#> 5     B: Placebo    ARM B    Asia       C      S1  6.206763    LOW     Y     Y
#> 6 C: Combination    ARM C    Asia       C      S2  6.906799 MEDIUM     Y     Y
#>   BMEASIFL BEP01FL     RANDDT             TRTSDTM             TRTEDTM
#> 1        Y       Y 2019-02-22 2019-02-24 11:09:18 2021-02-23 22:47:42
#> 2        N       N 2019-02-26 2019-02-26 09:05:00 2021-02-25 20:43:24
#> 3        Y       N 2019-02-24 2019-02-28 03:19:08 2021-02-27 14:57:32
#> 4        Y       Y 2019-02-27 2019-03-01 13:33:03 2021-03-01 01:11:27
#> 5        N       N 2019-03-01 2019-03-02 00:09:16 2021-03-01 11:47:40
#> 6        Y       N 2019-03-05 2019-03-05 15:23:44 2021-02-17 20:23:53
#>         EOSSTT       EOTSTT      EOSDT EOSDY          DCSREAS DTHDT DTHCAUS
#> 1    COMPLETED    COMPLETED 2021-02-23   731             <NA>  <NA>    <NA>
#> 2    COMPLETED    COMPLETED 2021-02-25   731             <NA>  <NA>    <NA>
#> 3    COMPLETED    COMPLETED 2021-02-27   731             <NA>  <NA>    <NA>
#> 4    COMPLETED    COMPLETED 2021-03-01   731             <NA>  <NA>    <NA>
#> 5    COMPLETED    COMPLETED 2021-03-01   731             <NA>  <NA>    <NA>
#> 6 DISCONTINUED DISCONTINUED 2021-02-17   716 LACK OF EFFICACY  <NA>    <NA>
#>   DTHCAT LDDTHELD LDDTHGR1   LSTALVDT study_duration_secs
#> 1   <NA>       NA     <NA> 2021-03-05            63113904
#> 2   <NA>       NA     <NA> 2021-03-15            63113904
#> 3   <NA>       NA     <NA> 2021-03-15            63113904
#> 4   <NA>       NA     <NA> 2021-03-17            63113904
#> 5   <NA>       NA     <NA> 2021-03-25            63113904
#> 6   <NA>       NA     <NA> 2021-03-01            63113904
#> 
#> ...
#>     STUDYID               USUBJID SUBJID SITEID AGE SEX
#> 395 AB12345   AB12345-CHN-9-id-11  id-11  CHN-9  28   F
#> 396 AB12345 AB12345-CHN-16-id-352 id-352 CHN-16  28   M
#> 397 AB12345  AB12345-CHN-1-id-186 id-186  CHN-1  27   M
#> 398 AB12345  AB12345-CHN-1-id-371 id-371  CHN-1  28   F
#> 399 AB12345  AB12345-CHN-1-id-233 id-233  CHN-1  36   F
#> 400 AB12345 AB12345-USA-12-id-131 id-131 USA-12  44   F
#>                                          RACE COUNTRY DTHFL         INVID
#> 395 NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER     CHN     N  INV ID CHN-9
#> 396                                     ASIAN     CHN     N INV ID CHN-16
#> 397                                     ASIAN     CHN     N  INV ID CHN-1
#> 398                                     ASIAN     CHN     Y  INV ID CHN-1
#> 399                 BLACK OR AFRICAN AMERICAN     CHN     N  INV ID CHN-1
#> 400          AMERICAN INDIAN OR ALASKA NATIVE     USA     N INV ID USA-12
#>             INVNAM            ARM ARMCD         ACTARM ACTARMCD       REGION1
#> 395  Dr. CHN-9 Doe     B: Placebo ARM B     B: Placebo    ARM B          Asia
#> 396 Dr. CHN-16 Doe C: Combination ARM C C: Combination    ARM C          Asia
#> 397  Dr. CHN-1 Doe     B: Placebo ARM B     B: Placebo    ARM B          Asia
#> 398  Dr. CHN-1 Doe      A: Drug X ARM A      A: Drug X    ARM A          Asia
#> 399  Dr. CHN-1 Doe C: Combination ARM C C: Combination    ARM C          Asia
#> 400 Dr. USA-12 Doe      A: Drug X ARM A      A: Drug X    ARM A North America
#>     STRATA1 STRATA2   BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL     RANDDT
#> 395       C      S1 4.267916 MEDIUM     Y     Y        Y       Y 2021-01-26
#> 396       A      S2 4.216812 MEDIUM     Y     Y        N       Y 2021-01-30
#> 397       A      S2 4.521179   HIGH     Y     Y        Y       N 2021-01-30
#> 398       A      S2 6.326863    LOW     Y     Y        N       Y 2021-02-07
#> 399       C      S1 4.076818    LOW     Y     Y        N       Y 2021-02-09
#> 400       B      S1 3.726935    LOW     Y     Y        Y       N 2021-02-14
#>                 TRTSDTM             TRTEDTM       EOSSTT       EOTSTT
#> 395 2021-01-27 01:57:07 2023-01-27 13:35:31    COMPLETED    COMPLETED
#> 396 2021-02-02 09:04:52 2023-02-02 20:43:16    COMPLETED    COMPLETED
#> 397 2021-02-02 19:14:46 2023-02-03 06:53:10    COMPLETED    COMPLETED
#> 398 2021-02-09 05:28:52 2021-02-19 19:06:01 DISCONTINUED DISCONTINUED
#> 399 2021-02-11 21:33:08 2021-02-17 09:12:06 DISCONTINUED DISCONTINUED
#> 400 2021-02-17 14:42:27 2023-02-18 02:20:51    COMPLETED    COMPLETED
#>          EOSDT EOSDY            DCSREAS      DTHDT DTHCAUS DTHCAT LDDTHELD
#> 395 2023-01-27   731               <NA>       <NA>    <NA>   <NA>       NA
#> 396 2023-02-02   731               <NA>       <NA>    <NA>   <NA>       NA
#> 397 2023-02-03   731               <NA>       <NA>    <NA>   <NA>       NA
#> 398 2021-02-19    11              DEATH 2021-03-29 MISSING  OTHER       38
#> 399 2021-02-17     6 PHYSICIAN DECISION       <NA>    <NA>   <NA>       NA
#> 400 2023-02-18   731               <NA>       <NA>    <NA>   <NA>       NA
#>     LDDTHGR1   LSTALVDT study_duration_secs
#> 395     <NA> 2023-02-22            63113904
#> 396     <NA> 2023-02-23            63113904
#> 397     <NA> 2023-03-03            63113904
#> 398      >30 2021-03-29            63113904
#> 399     <NA> 2021-03-17            63113904
#> 400     <NA> 2023-03-01            63113904
#> --> Element 2:
#> A TealDataset object containing the following data.frame (1600 rows and 50 columns):
#>   STUDYID              USUBJID SUBJID SITEID AGE SEX                      RACE
#> 1 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38   M BLACK OR AFRICAN AMERICAN
#> 2 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38   M BLACK OR AFRICAN AMERICAN
#> 3 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38   M BLACK OR AFRICAN AMERICAN
#> 4 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38   M BLACK OR AFRICAN AMERICAN
#> 5 AB12345 AB12345-BRA-1-id-134 id-134  BRA-1  47   M                     WHITE
#> 6 AB12345 AB12345-BRA-1-id-134 id-134  BRA-1  47   M                     WHITE
#>   COUNTRY DTHFL        INVID        INVNAM       ARM ARMCD    ACTARM ACTARMCD
#> 1     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#> 2     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#> 3     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#> 4     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#> 5     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#> 6     BRA     N INV ID BRA-1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A
#>         REGION1 STRATA1 STRATA2   BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL
#> 1 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 2 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 3 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 4 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 5 South America       B      S2 6.462991    LOW     Y     Y        Y       N
#> 6 South America       B      S2 6.462991    LOW     Y     Y        Y       N
#>       RANDDT             TRTSDTM             TRTEDTM       EOSSTT       EOTSTT
#> 1 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:47 DISCONTINUED DISCONTINUED
#> 2 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:47 DISCONTINUED DISCONTINUED
#> 3 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:47 DISCONTINUED DISCONTINUED
#> 4 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:47 DISCONTINUED DISCONTINUED
#> 5 2020-11-03 2020-11-04 03:50:33 2022-11-04 15:28:57    COMPLETED    COMPLETED
#> 6 2020-11-03 2020-11-04 03:50:33 2022-11-04 15:28:57    COMPLETED    COMPLETED
#>        EOSDT EOSDY       DCSREAS DTHDT DTHCAUS DTHCAT LDDTHELD LDDTHGR1
#> 1 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 2 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 3 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 4 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 5 2022-11-04   731          <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 6 2022-11-04   731          <NA>  <NA>    <NA>   <NA>       NA     <NA>
#>     LSTALVDT study_duration_secs ASEQ TTESEQ                          PARAM
#> 1 2021-03-07            63113904    2      2 Duration of Confirmed Response
#> 2 2021-03-07            63113904    4      4            Event Free Survival
#> 3 2021-03-07            63113904    3      3               Overall Survival
#> 4 2021-03-07            63113904    1      1      Progression Free Survival
#> 5 2022-11-15            63113904    4      4 Duration of Confirmed Response
#> 6 2022-11-15            63113904    2      2            Event Free Survival
#>   PARAMCD      AVAL AVALU       ADTM ADY CNSR                    EVNTDESC
#> 1    CRSD 1273.8498  DAYS 2020-08-13 158    1 Last Date Known To Be Alive
#> 2     EFS  135.9787  DAYS 2021-02-16 345    0                       Death
#> 3      OS 1759.8678  DAYS 2020-11-05 242    0                       Death
#> 4     PFS  241.1912  DAYS 2020-07-23 137    0         Disease Progression
#> 5    CRSD  544.5728  DAYS 2021-10-25 355    0                       Death
#> 6     EFS  401.2725  DAYS 2021-05-05 182    1 Last Date Known To Be Alive
#>                        CNSDTDSC
#> 1    End of AE Reporting Period
#> 2                              
#> 3                              
#> 4                              
#> 5                              
#> 6 Completion or Discontinuation
#> 
#> ...
#>      STUDYID              USUBJID SUBJID SITEID AGE SEX  RACE COUNTRY DTHFL
#> 1595 AB12345 AB12345-USA-8-id-206 id-206  USA-8  34   F ASIAN     USA     N
#> 1596 AB12345 AB12345-USA-8-id-206 id-206  USA-8  34   F ASIAN     USA     N
#> 1597 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#> 1598 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#> 1599 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#> 1600 AB12345 AB12345-USA-9-id-130 id-130  USA-9  40   M ASIAN     USA     N
#>             INVID        INVNAM            ARM ARMCD         ACTARM ACTARMCD
#> 1595 INV ID USA-8 Dr. USA-8 Doe C: Combination ARM C C: Combination    ARM C
#> 1596 INV ID USA-8 Dr. USA-8 Doe C: Combination ARM C C: Combination    ARM C
#> 1597 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 1598 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 1599 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 1600 INV ID USA-9 Dr. USA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#>            REGION1 STRATA1 STRATA2    BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL
#> 1595 North America       B      S1 14.478057   HIGH     Y     Y        Y
#> 1596 North America       B      S1 14.478057   HIGH     Y     Y        Y
#> 1597 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#> 1598 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#> 1599 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#> 1600 North America       A      S1  1.280525 MEDIUM     Y     Y        N
#>      BEP01FL     RANDDT             TRTSDTM             TRTEDTM    EOSSTT
#> 1595       Y 2019-07-19 2019-07-22 10:19:01 2021-07-21 21:57:25 COMPLETED
#> 1596       Y 2019-07-19 2019-07-22 10:19:01 2021-07-21 21:57:25 COMPLETED
#> 1597       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#> 1598       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#> 1599       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#> 1600       N 2020-07-17 2020-07-18 19:28:19 2022-07-19 07:06:43 COMPLETED
#>         EOTSTT      EOSDT EOSDY DCSREAS DTHDT DTHCAUS DTHCAT LDDTHELD LDDTHGR1
#> 1595 COMPLETED 2021-07-21   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 1596 COMPLETED 2021-07-21   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 1597 COMPLETED 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 1598 COMPLETED 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 1599 COMPLETED 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 1600 COMPLETED 2022-07-19   731    <NA>  <NA>    <NA>   <NA>       NA     <NA>
#>        LSTALVDT study_duration_secs ASEQ TTESEQ                          PARAM
#> 1595 2021-08-01            63113904    1      1               Overall Survival
#> 1596 2021-08-01            63113904    2      2      Progression Free Survival
#> 1597 2022-08-17            63113904    4      4 Duration of Confirmed Response
#> 1598 2022-08-17            63113904    3      3            Event Free Survival
#> 1599 2022-08-17            63113904    1      1               Overall Survival
#> 1600 2022-08-17            63113904    2      2      Progression Free Survival
#>      PARAMCD      AVAL AVALU       ADTM ADY CNSR                    EVNTDESC
#> 1595      OS 1312.4084  DAYS 2019-11-09 110    0                       Death
#> 1596     PFS  429.6270  DAYS 2020-05-03 286    1 Last Date Known To Be Alive
#> 1597    CRSD  832.4600  DAYS 2021-01-23 189    0                       Death
#> 1598     EFS  471.2179  DAYS 2021-01-18 184    1 Last Date Known To Be Alive
#> 1599      OS 1475.7467  DAYS 2020-08-03  16    1       Last Tumor Assessment
#> 1600     PFS  147.4837  DAYS 2020-08-25  38    1       Last Tumor Assessment
#>                           CNSDTDSC
#> 1595                              
#> 1596              Clinical Cut Off
#> 1597                              
#> 1598    End of AE Reporting Period
#> 1599 Completion or Discontinuation
#> 1600 Completion or Discontinuation