Data input for teal app

Function passes datasets to teal application with option to read preprocessing code and reproducibility checking.

cdisc_data(..., code = "", check = FALSE)

Arguments

...
datasets objects

code
(character) preprocessing code.

check
(logical) reproducibility check - whether evaluated preprocessing code gives the same objects as provided in arguments. Check is run only if flag is true and preprocessing code is not empty.

Value

a list of the input data sets

Details

This function checks if there were keys added to all data sets that shall be analyzed inside a teal app.

Examples

library(random.cdisc.data)

ADSL <- cadsl # or: radsl(N = 600, seed = 123)
ADTTE <- cadtte # or: radtte(ADSL, seed = 123)

# basic example
cdisc_data(
  cdisc_dataset("ADSL", ADSL),
  cdisc_dataset("ADTTE", ADTTE),
  code = 'ADSL <- cadsl
          ADTTE <- cadtte')

#> $ADSL
#> $dataname
#> [1] "ADSL"
#> 
#> $data
#> # A tibble: 400 x 31
#>    STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY INVID ARM   ARMCD
#>    <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <chr> <fct> <fct>
#>  1 AB12345 AB1234… id-128 CHN-3     32 M     ASIAN CHN     CHN-3 A: D… ARM A
#>  2 AB12345 AB1234… id-262 CHN-15    35 M     BLAC… CHN     CHN-… C: C… ARM C
#>  3 AB12345 AB1234… id-378 RUS-3     30 F     ASIAN RUS     RUS-3 C: C… ARM C
#>  4 AB12345 AB1234… id-220 CHN-11    26 F     ASIAN CHN     CHN-… B: P… ARM B
#>  5 AB12345 AB1234… id-267 CHN-7     40 M     ASIAN CHN     CHN-7 B: P… ARM B
#>  6 AB12345 AB1234… id-201 CHN-15    49 M     ASIAN CHN     CHN-… C: C… ARM C
#>  7 AB12345 AB1234… id-45  USA-1     34 F     ASIAN USA     USA-1 A: D… ARM A
#>  8 AB12345 AB1234… id-261 USA-1     32 F     ASIAN USA     USA-1 B: P… ARM B
#>  9 AB12345 AB1234… id-173 NGA-11    24 F     BLAC… NGA     NGA-… C: C… ARM C
#> 10 AB12345 AB1234… id-307 CHN-1     24 M     ASIAN CHN     CHN-1 B: P… ARM B
#> # … with 390 more rows, and 20 more variables: ACTARM <fct>, ACTARMCD <fct>,
#> #   STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>,
#> #   SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>,
#> #   TRTEDTM <dttm>, EOSSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>,
#> #   DTHDT <date>, LSTALVDT <date>, study_duration_secs <dbl>
#> 
#> $keys
#> $keys$primary
#> [1] "STUDYID" "USUBJID"
#> 
#> $keys$foreign
#> NULL
#> 
#> $keys$parent
#> NULL
#> 
#> 
#> $labels
#> $labels$dataset_label
#> [1] "Subject Level Analysis Dataset"
#> 
#> $labels$column_labels
#>                                   STUDYID 
#>                        "Study Identifier" 
#>                                   USUBJID 
#>               "Unique Subject Identifier" 
#>                                    SUBJID 
#>        "Subject Identifier for the Study" 
#>                                    SITEID 
#>                   "Study Site Identifier" 
#>                                       AGE 
#>                                     "Age" 
#>                                       SEX 
#>                                     "Sex" 
#>                                      RACE 
#>                                    "Race" 
#>                                   COUNTRY 
#>                                 "Country" 
#>                                     INVID 
#>                 "Investigator Identifier" 
#>                                       ARM 
#>              "Description of Planned Arm" 
#>                                     ARMCD 
#>                        "Planned Arm Code" 
#>                                    ACTARM 
#>               "Description of Actual Arm" 
#>                                  ACTARMCD 
#>                         "Actual Arm Code" 
#>                                   STRATA1 
#>                 "Stratification Factor 1" 
#>                                   STRATA2 
#>                 "Stratification Factor 2" 
#>                                    BMRKR1 
#>             "Continous Level Biomarker 1" 
#>                                    BMRKR2 
#>           "Categorical Level Biomarker 2" 
#>                                     ITTFL 
#>         "Intent-To-Treat Population Flag" 
#>                                     SAFFL 
#>                  "Safety Population Flag" 
#>                                  BMEASIFL 
#>      "Response Evaluable Population Flag" 
#>                                   BEP01FL 
#>     "Biomarker Evaluable Population Flag" 
#>                                    RANDDT 
#>                   "Date of Randomization" 
#>                                   TRTSDTM 
#> "Datetime of First Exposure to Treatment" 
#>                                   TRTEDTM 
#>  "Datetime of Last Exposure to Treatment" 
#>                                    EOSSTT 
#>                     "End of Study Status" 
#>                                     EOSDT 
#>                       "End of Study Date" 
#>                                     EOSDY 
#>               "End of Study Relative Day" 
#>                                   DCSREAS 
#>   "Reason for Discontinuation from Study" 
#>                                     DTHDT 
#>                           "Date of Death" 
#>                                  LSTALVDT 
#>                   "Date Last Known Alive" 
#>                       study_duration_secs 
#>       "NOT A STANDARD BUT NEEDED FOR RCD" 
#> 
#> 
#> attr(,"class")
#> [1] "cdisc_dataset" "dataset"      
#> 
#> $ADTTE
#> $dataname
#> [1] "ADTTE"
#> 
#> $data
#> # A tibble: 1,200 x 42
#>    STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY INVID ARM   ARMCD
#>    <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <chr> <fct> <fct>
#>  1 AB12345 AB1234… id-105 BRA-1     38 M     BLAC… BRA     BRA-1 A: D… ARM A
#>  2 AB12345 AB1234… id-105 BRA-1     38 M     BLAC… BRA     BRA-1 A: D… ARM A
#>  3 AB12345 AB1234… id-105 BRA-1     38 M     BLAC… BRA     BRA-1 A: D… ARM A
#>  4 AB12345 AB1234… id-134 BRA-1     47 M     WHITE BRA     BRA-1 A: D… ARM A
#>  5 AB12345 AB1234… id-134 BRA-1     47 M     WHITE BRA     BRA-1 A: D… ARM A
#>  6 AB12345 AB1234… id-134 BRA-1     47 M     WHITE BRA     BRA-1 A: D… ARM A
#>  7 AB12345 AB1234… id-141 BRA-1     35 F     WHITE BRA     BRA-1 C: C… ARM C
#>  8 AB12345 AB1234… id-141 BRA-1     35 F     WHITE BRA     BRA-1 C: C… ARM C
#>  9 AB12345 AB1234… id-141 BRA-1     35 F     WHITE BRA     BRA-1 C: C… ARM C
#> 10 AB12345 AB1234… id-236 BRA-1     32 M     BLAC… BRA     BRA-1 B: P… ARM B
#> # … with 1,190 more rows, and 31 more variables: ACTARM <fct>, ACTARMCD <fct>,
#> #   STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>,
#> #   SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>,
#> #   TRTEDTM <dttm>, EOSSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>,
#> #   DTHDT <date>, LSTALVDT <date>, study_duration_secs <dbl>, ASEQ <int>,
#> #   TTESEQ <int>, PARAM <fct>, PARAMCD <fct>, AVAL <dbl>, AVALU <fct>,
#> #   ADTM <dttm>, ADY <int>, CNSR <int>, EVNTDESC <chr>, CNSDTDSC <chr>
#> 
#> $keys
#> $keys$primary
#> [1] "STUDYID" "USUBJID" "PARAMCD"
#> 
#> $keys$foreign
#> [1] "STUDYID" "USUBJID"
#> 
#> $keys$parent
#> [1] "ADSL"
#> 
#> 
#> $labels
#> $labels$dataset_label
#> [1] "Time to Event Analysis Dataset"
#> 
#> $labels$column_labels
#>                                   STUDYID 
#>                        "Study Identifier" 
#>                                   USUBJID 
#>               "Unique Subject Identifier" 
#>                                    SUBJID 
#>        "Subject Identifier for the Study" 
#>                                    SITEID 
#>                   "Study Site Identifier" 
#>                                       AGE 
#>                                     "Age" 
#>                                       SEX 
#>                                     "Sex" 
#>                                      RACE 
#>                                    "Race" 
#>                                   COUNTRY 
#>                                 "Country" 
#>                                     INVID 
#>                 "Investigator Identifier" 
#>                                       ARM 
#>              "Description of Planned Arm" 
#>                                     ARMCD 
#>                        "Planned Arm Code" 
#>                                    ACTARM 
#>               "Description of Actual Arm" 
#>                                  ACTARMCD 
#>                         "Actual Arm Code" 
#>                                   STRATA1 
#>                 "Stratification Factor 1" 
#>                                   STRATA2 
#>                 "Stratification Factor 2" 
#>                                    BMRKR1 
#>             "Continous Level Biomarker 1" 
#>                                    BMRKR2 
#>           "Categorical Level Biomarker 2" 
#>                                     ITTFL 
#>         "Intent-To-Treat Population Flag" 
#>                                     SAFFL 
#>                  "Safety Population Flag" 
#>                                  BMEASIFL 
#>      "Response Evaluable Population Flag" 
#>                                   BEP01FL 
#>     "Biomarker Evaluable Population Flag" 
#>                                    RANDDT 
#>                   "Date of Randomization" 
#>                                   TRTSDTM 
#> "Datetime of First Exposure to Treatment" 
#>                                   TRTEDTM 
#>  "Datetime of Last Exposure to Treatment" 
#>                                    EOSSTT 
#>                     "End of Study Status" 
#>                                     EOSDT 
#>                       "End of Study Date" 
#>                                     EOSDY 
#>               "End of Study Relative Day" 
#>                                   DCSREAS 
#>   "Reason for Discontinuation from Study" 
#>                                     DTHDT 
#>                           "Date of Death" 
#>                                  LSTALVDT 
#>                   "Date Last Known Alive" 
#>                       study_duration_secs 
#>       "NOT A STANDARD BUT NEEDED FOR RCD" 
#>                                      ASEQ 
#>                         "Sequence Number" 
#>                                    TTESEQ 
#>                         "Sequence Number" 
#>                                     PARAM 
#>                               "Parameter" 
#>                                   PARAMCD 
#>                          "Parameter Code" 
#>                                      AVAL 
#>                          "Analysis Value" 
#>                                     AVALU 
#>                     "Analysis Value Unit" 
#>                                      ADTM 
#>                       "Analysis Datetime" 
#>                                       ADY 
#>                   "Analysis Relative Day" 
#>                                      CNSR 
#>                                  "Censor" 
#>                                  EVNTDESC 
#>          "Event or Censoring Description" 
#>                                  CNSDTDSC 
#>                 "Censor Date Description" 
#> 
#> 
#> attr(,"class")
#> [1] "cdisc_dataset" "dataset"      
#> 
#> attr(,"code")
#> [1] "ADSL <- cadsl\n          ADTTE <- cadtte"
#> attr(,"class")
#> [1] "cdisc_data"

# Example with keys
cdisc_data(
  cdisc_dataset("ADSL", ADSL, keys = keys(
    primary = c("STUDYID", "USUBJID"),
    foreign = NULL,
    parent = NULL
  )),
  cdisc_dataset("ADTTE", ADTTE, keys = keys(
    primary = c("STUDYID", "USUBJID", "PARAMCD"),
    foreign = c("STUDYID", "USUBJID"),
    parent = "ADSL"
  )),
  code = "ADSL <- radsl(N = 600, seed = 123)
          ADTTE <- radtte(ADSL, seed = 123)",
  check = FALSE
)

#> $ADSL
#> $dataname
#> [1] "ADSL"
#> 
#> $data
#> # A tibble: 400 x 31
#>    STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY INVID ARM   ARMCD
#>    <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <chr> <fct> <fct>
#>  1 AB12345 AB1234… id-128 CHN-3     32 M     ASIAN CHN     CHN-3 A: D… ARM A
#>  2 AB12345 AB1234… id-262 CHN-15    35 M     BLAC… CHN     CHN-… C: C… ARM C
#>  3 AB12345 AB1234… id-378 RUS-3     30 F     ASIAN RUS     RUS-3 C: C… ARM C
#>  4 AB12345 AB1234… id-220 CHN-11    26 F     ASIAN CHN     CHN-… B: P… ARM B
#>  5 AB12345 AB1234… id-267 CHN-7     40 M     ASIAN CHN     CHN-7 B: P… ARM B
#>  6 AB12345 AB1234… id-201 CHN-15    49 M     ASIAN CHN     CHN-… C: C… ARM C
#>  7 AB12345 AB1234… id-45  USA-1     34 F     ASIAN USA     USA-1 A: D… ARM A
#>  8 AB12345 AB1234… id-261 USA-1     32 F     ASIAN USA     USA-1 B: P… ARM B
#>  9 AB12345 AB1234… id-173 NGA-11    24 F     BLAC… NGA     NGA-… C: C… ARM C
#> 10 AB12345 AB1234… id-307 CHN-1     24 M     ASIAN CHN     CHN-1 B: P… ARM B
#> # … with 390 more rows, and 20 more variables: ACTARM <fct>, ACTARMCD <fct>,
#> #   STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>,
#> #   SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>,
#> #   TRTEDTM <dttm>, EOSSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>,
#> #   DTHDT <date>, LSTALVDT <date>, study_duration_secs <dbl>
#> 
#> $keys
#> $keys$primary
#> [1] "STUDYID" "USUBJID"
#> 
#> $keys$foreign
#> NULL
#> 
#> $keys$parent
#> NULL
#> 
#> 
#> $labels
#> $labels$dataset_label
#> [1] "Subject Level Analysis Dataset"
#> 
#> $labels$column_labels
#>                                   STUDYID 
#>                        "Study Identifier" 
#>                                   USUBJID 
#>               "Unique Subject Identifier" 
#>                                    SUBJID 
#>        "Subject Identifier for the Study" 
#>                                    SITEID 
#>                   "Study Site Identifier" 
#>                                       AGE 
#>                                     "Age" 
#>                                       SEX 
#>                                     "Sex" 
#>                                      RACE 
#>                                    "Race" 
#>                                   COUNTRY 
#>                                 "Country" 
#>                                     INVID 
#>                 "Investigator Identifier" 
#>                                       ARM 
#>              "Description of Planned Arm" 
#>                                     ARMCD 
#>                        "Planned Arm Code" 
#>                                    ACTARM 
#>               "Description of Actual Arm" 
#>                                  ACTARMCD 
#>                         "Actual Arm Code" 
#>                                   STRATA1 
#>                 "Stratification Factor 1" 
#>                                   STRATA2 
#>                 "Stratification Factor 2" 
#>                                    BMRKR1 
#>             "Continous Level Biomarker 1" 
#>                                    BMRKR2 
#>           "Categorical Level Biomarker 2" 
#>                                     ITTFL 
#>         "Intent-To-Treat Population Flag" 
#>                                     SAFFL 
#>                  "Safety Population Flag" 
#>                                  BMEASIFL 
#>      "Response Evaluable Population Flag" 
#>                                   BEP01FL 
#>     "Biomarker Evaluable Population Flag" 
#>                                    RANDDT 
#>                   "Date of Randomization" 
#>                                   TRTSDTM 
#> "Datetime of First Exposure to Treatment" 
#>                                   TRTEDTM 
#>  "Datetime of Last Exposure to Treatment" 
#>                                    EOSSTT 
#>                     "End of Study Status" 
#>                                     EOSDT 
#>                       "End of Study Date" 
#>                                     EOSDY 
#>               "End of Study Relative Day" 
#>                                   DCSREAS 
#>   "Reason for Discontinuation from Study" 
#>                                     DTHDT 
#>                           "Date of Death" 
#>                                  LSTALVDT 
#>                   "Date Last Known Alive" 
#>                       study_duration_secs 
#>       "NOT A STANDARD BUT NEEDED FOR RCD" 
#> 
#> 
#> attr(,"class")
#> [1] "cdisc_dataset" "dataset"      
#> 
#> $ADTTE
#> $dataname
#> [1] "ADTTE"
#> 
#> $data
#> # A tibble: 1,200 x 42
#>    STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY INVID ARM   ARMCD
#>    <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <chr> <fct> <fct>
#>  1 AB12345 AB1234… id-105 BRA-1     38 M     BLAC… BRA     BRA-1 A: D… ARM A
#>  2 AB12345 AB1234… id-105 BRA-1     38 M     BLAC… BRA     BRA-1 A: D… ARM A
#>  3 AB12345 AB1234… id-105 BRA-1     38 M     BLAC… BRA     BRA-1 A: D… ARM A
#>  4 AB12345 AB1234… id-134 BRA-1     47 M     WHITE BRA     BRA-1 A: D… ARM A
#>  5 AB12345 AB1234… id-134 BRA-1     47 M     WHITE BRA     BRA-1 A: D… ARM A
#>  6 AB12345 AB1234… id-134 BRA-1     47 M     WHITE BRA     BRA-1 A: D… ARM A
#>  7 AB12345 AB1234… id-141 BRA-1     35 F     WHITE BRA     BRA-1 C: C… ARM C
#>  8 AB12345 AB1234… id-141 BRA-1     35 F     WHITE BRA     BRA-1 C: C… ARM C
#>  9 AB12345 AB1234… id-141 BRA-1     35 F     WHITE BRA     BRA-1 C: C… ARM C
#> 10 AB12345 AB1234… id-236 BRA-1     32 M     BLAC… BRA     BRA-1 B: P… ARM B
#> # … with 1,190 more rows, and 31 more variables: ACTARM <fct>, ACTARMCD <fct>,
#> #   STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>,
#> #   SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>,
#> #   TRTEDTM <dttm>, EOSSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>,
#> #   DTHDT <date>, LSTALVDT <date>, study_duration_secs <dbl>, ASEQ <int>,
#> #   TTESEQ <int>, PARAM <fct>, PARAMCD <fct>, AVAL <dbl>, AVALU <fct>,
#> #   ADTM <dttm>, ADY <int>, CNSR <int>, EVNTDESC <chr>, CNSDTDSC <chr>
#> 
#> $keys
#> $keys$primary
#> [1] "STUDYID" "USUBJID" "PARAMCD"
#> 
#> $keys$foreign
#> [1] "STUDYID" "USUBJID"
#> 
#> $keys$parent
#> [1] "ADSL"
#> 
#> 
#> $labels
#> $labels$dataset_label
#> [1] "Time to Event Analysis Dataset"
#> 
#> $labels$column_labels
#>                                   STUDYID 
#>                        "Study Identifier" 
#>                                   USUBJID 
#>               "Unique Subject Identifier" 
#>                                    SUBJID 
#>        "Subject Identifier for the Study" 
#>                                    SITEID 
#>                   "Study Site Identifier" 
#>                                       AGE 
#>                                     "Age" 
#>                                       SEX 
#>                                     "Sex" 
#>                                      RACE 
#>                                    "Race" 
#>                                   COUNTRY 
#>                                 "Country" 
#>                                     INVID 
#>                 "Investigator Identifier" 
#>                                       ARM 
#>              "Description of Planned Arm" 
#>                                     ARMCD 
#>                        "Planned Arm Code" 
#>                                    ACTARM 
#>               "Description of Actual Arm" 
#>                                  ACTARMCD 
#>                         "Actual Arm Code" 
#>                                   STRATA1 
#>                 "Stratification Factor 1" 
#>                                   STRATA2 
#>                 "Stratification Factor 2" 
#>                                    BMRKR1 
#>             "Continous Level Biomarker 1" 
#>                                    BMRKR2 
#>           "Categorical Level Biomarker 2" 
#>                                     ITTFL 
#>         "Intent-To-Treat Population Flag" 
#>                                     SAFFL 
#>                  "Safety Population Flag" 
#>                                  BMEASIFL 
#>      "Response Evaluable Population Flag" 
#>                                   BEP01FL 
#>     "Biomarker Evaluable Population Flag" 
#>                                    RANDDT 
#>                   "Date of Randomization" 
#>                                   TRTSDTM 
#> "Datetime of First Exposure to Treatment" 
#>                                   TRTEDTM 
#>  "Datetime of Last Exposure to Treatment" 
#>                                    EOSSTT 
#>                     "End of Study Status" 
#>                                     EOSDT 
#>                       "End of Study Date" 
#>                                     EOSDY 
#>               "End of Study Relative Day" 
#>                                   DCSREAS 
#>   "Reason for Discontinuation from Study" 
#>                                     DTHDT 
#>                           "Date of Death" 
#>                                  LSTALVDT 
#>                   "Date Last Known Alive" 
#>                       study_duration_secs 
#>       "NOT A STANDARD BUT NEEDED FOR RCD" 
#>                                      ASEQ 
#>                         "Sequence Number" 
#>                                    TTESEQ 
#>                         "Sequence Number" 
#>                                     PARAM 
#>                               "Parameter" 
#>                                   PARAMCD 
#>                          "Parameter Code" 
#>                                      AVAL 
#>                          "Analysis Value" 
#>                                     AVALU 
#>                     "Analysis Value Unit" 
#>                                      ADTM 
#>                       "Analysis Datetime" 
#>                                       ADY 
#>                   "Analysis Relative Day" 
#>                                      CNSR 
#>                                  "Censor" 
#>                                  EVNTDESC 
#>          "Event or Censoring Description" 
#>                                  CNSDTDSC 
#>                 "Censor Date Description" 
#> 
#> 
#> attr(,"class")
#> [1] "cdisc_dataset" "dataset"      
#> 
#> attr(,"code")
#> [1] "ADSL <- radsl(N = 600, seed = 123)\n          ADTTE <- radtte(ADSL, seed = 123)"
#> attr(,"class")
#> [1] "cdisc_data"