Data input for teal app

Function that creates CDISC dataset object

cdisc_dataset(
  dataname,
  data,
  keys = get_cdisc_keys(dataname),
  label = data_label(data),
  code = character(0),
  vars = list()
)

Arguments

dataname
(character value)
name of dataset.

data
(data.frame)
data must contain fields defined in keys.

keys
(keys)
see keys

label
(character value)
a dataset label

code
(character value)
code to reproduce data

vars
(named list))
In case when this object code depends on the raw_data from the other RelationalDataset object(s) or other constant value, this/these object(s) should be included as named element of the list.

Please note that the order of keys is important.

Keys are not matched by name but by order. In case you want to perform:

SELECT ... FROM ADSL LEFT JOIN ADTTE on ADTTE.STUDYID = ADSL.STUDYID and ADTTE.SUBJECTUNIQUEID = ADSL.USUBJID

the keys need to be

ADSL primary: USUBJID, STUDYID

ADTTE primary: PARAMCD, foreign: SUBJECT, UNIQUEID, STUDYID, parent: ADSL

As you can see the names of foreign ADTTE are different to ADSL keys, but the order and length is equal.

Value

a dataset with connected metadata

Examples

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

cdisc_dataset("ADSL", ADSL)

#> <RelationalDataset>
#>   Inherits from: <NamedDataset>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     colnames: active binding
#>     column_labels: active binding
#>     data: active binding
#>     dim: active binding
#>     get_character_colnames: function () 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_colnames: function () 
#>     get_column_labels: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset_label: function () 
#>     get_factor_colnames: function () 
#>     get_keys: function () 
#>     get_numeric_colnames: function (include_factors = FALSE) 
#>     get_raw_data: function () 
#>     get_row_labels: function () 
#>     get_rownames: function () 
#>     initialize: function (dataname, x, keys, code = character(0), label = character(0), 
#>     is_pulled: function () 
#>     mutate: function (code, vars = list()) 
#>     ncol: active binding
#>     nrow: active binding
#>     raw_data: active binding
#>     row_labels: active binding
#>     rownames: active binding
#>     set_code: function (code) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_vars: function (vars) 
#>     var_names: active binding
#>   Private:
#>     .col_labels: Study Identifier Unique Subject Identifier Subject Ident ...
#>     .colnames: STUDYID USUBJID SUBJID SITEID AGE SEX RACE COUNTRY INVID ...
#>     .dim: 400 33
#>     .keys: keys
#>     .ncol: 33
#>     .nrow: 400
#>     .raw_data: tbl_df, tbl, data.frame
#>     .row_labels: 
#>     .rownames: 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 ...
#>     code: CodeClass, R6
#>     dataname: ADSL
#>     dataset_label: Subject Level Analysis Dataset
#>     execute_code: function (code, vars = list()) 
#>     get_class_colnames: function (class_type = "character", include_factors = FALSE) 
#>     set_dataname: function (dataname) 
#>     vars: list