Data input for teal app

Abstract function that creates dataset object with connected metadata.

dataset(
  dataname,
  data,
  keys = list(primary = NULL, foreign = NULL, parent = NULL)
)

Arguments

dataname
name of dataset

data
data

keys
list of keys -

Please note that the order of keys is important.

Keys are not matched by name but by order. In case you want to perform:

SELECT ... FROM ADSL LEFT JOIN ADTTE on ADTTE.STUDYID = ADSL.STUDYID and ADTTE.SUBJECTUNIQUEID = ADSL.USUBJID

the keys need to be

ADSL primary: USUBJID, STUDYID

ADTTE primary: PARAMCD, foreign: SUBJECTUNIQUEID, STUDYID, parent: ADSL

As you can see the names of foreign ADTTE are different to ADSL keys, but the order and length is equal.

Value

a dataset with connected metadata

Examples

library(random.cdisc.data)

ADSL <-  suppressWarnings(radsl(N = 600, seed = 123))

dataset("ADSL", ADSL)

#> $dataname
#> [1] "ADSL"
#> 
#> $data
#> # A tibble: 600 x 31
#>    STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY INVID ARM   ARMCD
#>    <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <chr> <fct> <fct>
#>  1 AB12345 AB1234… id-541 NGA-5     28 M     ASIAN NGA     NGA-5 B: P… ARM B
#>  2 AB12345 AB1234… id-399 CHN-4     27 F     ASIAN CHN     CHN-4 A: D… ARM A
#>  3 AB12345 AB1234… id-564 CHN-11    35 M     BLAC… CHN     CHN-… B: P… ARM B
#>  4 AB12345 AB1234… id-24  CAN-1     31 F     AMER… CAN     CAN-1 C: C… ARM C
#>  5 AB12345 AB1234… id-382 PAK-4     30 F     ASIAN PAK     PAK-4 C: C… ARM C
#>  6 AB12345 AB1234… id-463 CHN-1     25 M     BLAC… CHN     CHN-1 B: P… ARM B
#>  7 AB12345 AB1234… id-157 USA-1     34 M     WHITE USA     USA-1 A: D… ARM A
#>  8 AB12345 AB1234… id-23  BRA-3     31 M     WHITE BRA     BRA-3 C: C… ARM C
#>  9 AB12345 AB1234… id-343 USA-1     33 M     BLAC… USA     USA-1 C: C… ARM C
#> 10 AB12345 AB1234… id-373 RUS-15    32 F     WHITE RUS     RUS-… B: P… ARM B
#> # … with 590 more rows, and 20 more variables: ACTARM <fct>, ACTARMCD <fct>,
#> #   STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>,
#> #   SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>,
#> #   TRTEDTM <dttm>, EOSSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>,
#> #   DTHDT <date>, LSTALVDT <date>, study_duration_secs <dbl>
#> 
#> $keys
#> $keys$primary
#> NULL
#> 
#> $keys$foreign
#> NULL
#> 
#> $keys$parent
#> NULL
#> 
#> 
#> $column_labels
#>                                   STUDYID 
#>                        "Study Identifier" 
#>                                   USUBJID 
#>               "Unique Subject Identifier" 
#>                                    SUBJID 
#>        "Subject Identifier for the Study" 
#>                                    SITEID 
#>                   "Study Site Identifier" 
#>                                       AGE 
#>                                     "Age" 
#>                                       SEX 
#>                                     "Sex" 
#>                                      RACE 
#>                                    "Race" 
#>                                   COUNTRY 
#>                                 "Country" 
#>                                     INVID 
#>                 "Investigator Identifier" 
#>                                       ARM 
#>              "Description of Planned Arm" 
#>                                     ARMCD 
#>                        "Planned Arm Code" 
#>                                    ACTARM 
#>               "Description of Actual Arm" 
#>                                  ACTARMCD 
#>                         "Actual Arm Code" 
#>                                   STRATA1 
#>                 "Stratification Factor 1" 
#>                                   STRATA2 
#>                 "Stratification Factor 2" 
#>                                    BMRKR1 
#>             "Continous Level Biomarker 1" 
#>                                    BMRKR2 
#>           "Categorical Level Biomarker 2" 
#>                                     ITTFL 
#>         "Intent-To-Treat Population Flag" 
#>                                     SAFFL 
#>                  "Safety Population Flag" 
#>                                  BMEASIFL 
#>      "Response Evaluable Population Flag" 
#>                                   BEP01FL 
#>     "Biomarker Evaluable Population Flag" 
#>                                    RANDDT 
#>                   "Date of Randomization" 
#>                                   TRTSDTM 
#> "Datetime of First Exposure to Treatment" 
#>                                   TRTEDTM 
#>  "Datetime of Last Exposure to Treatment" 
#>                                    EOSSTT 
#>                     "End of Study Status" 
#>                                     EOSDT 
#>                       "End of Study Date" 
#>                                     EOSDY 
#>               "End of Study Relative Day" 
#>                                   DCSREAS 
#>   "Reason for Discontinuation from Study" 
#>                                     DTHDT 
#>                           "Date of Death" 
#>                                  LSTALVDT 
#>                   "Date Last Known Alive" 
#>                       study_duration_secs 
#>       "NOT A STANDARD BUT NEEDED FOR RCD" 
#> 
#> $dataset_label
#> [1] "Subject Level Analysis Dataset"
#> 
#> attr(,"class")
#> [1] "dataset"