R6 Class representing a dataset with its attributes

Any data.frame object can be stored inside this object. Some attributes like colnames, dimension or column names for a specific type will be automatically derived.

Any MultiAssayExperiment object can be stored inside this MAEDataset. Some attributes like colnames, dimension or column names for a specific type will be automatically derived.

Active bindings

ncol

Number of columns

nrow

Number of rows

dim

Dimension c(x, y)

colnames

The column names of the data

rownames

The rownames of the data

raw_data

The data.frame behind this R6 class

data

The data.frame behind this R6 class

var_names

The column names of the data

row_labels

Row labels (can have spaces)

Methods

Public methods

Dataset$new()

Dataset$recreate()

Dataset$print()

Dataset$get_dataset()

Dataset$get_attrs()

Dataset$get_raw_data()

Dataset$get_numeric_colnames()

Dataset$get_character_colnames()

Dataset$get_factor_colnames()

Dataset$get_colnames()

Dataset$get_column_labels()

Dataset$get_rownames()

Dataset$get_row_labels()

Dataset$get_dataname()

Dataset$get_datanames()

Dataset$get_dataset_label()

Dataset$get_keys()

Dataset$get_join_keys()

Dataset$get_hash()

Dataset$get_var_r6()

Dataset$reassign_datasets_vars()

Dataset$set_dataset_label()

Dataset$set_keys()

Dataset$set_join_keys()

Dataset$merge_join_keys()

Dataset$mutate_join_keys()

Dataset$set_vars()

Dataset$set_code()

Dataset$get_code()

Dataset$get_code_class()

Dataset$get_mutate_code_class()

Dataset$get_vars()

Dataset$get_mutate_vars()

Dataset$is_mutate_delayed()

Dataset$mutate()

Dataset$check()

Dataset$check_keys()

Dataset$is_pulled()

Dataset$clone()

Method new()

Create a new object of Dataset class

Usage

Dataset$new(
  dataname,
  x,
  keys = character(0),
  code = character(0),
  label = character(0),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(data.frame)

keys

optional, (character)
Vector with primary keys

code

(character)
A character string defining the code needed to produce the data set in x. initialize() and recreate() accept code as CodeClass also which is needed to preserve the code uniqueness and correct order.

label

(character)
Label to describe the dataset

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Method recreate()

Recreate this Dataset with its current attributes.

Usage

Dataset$recreate(
  dataname = self$get_dataname(),
  x = self$get_raw_data(),
  keys = self$get_keys(),
  code = private$code,
  label = self$get_dataset_label(),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(data.frame)

keys

optional, (character)
Vector with primary keys

code

(character)
A character string defining the code needed to produce the data set in x. initialize() and recreate() accept code as CodeClass also which is needed to preserve the code uniqueness and correct order.

label

(character)
Label to describe the dataset

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Returns

a new object of the Dataset class Prints this Dataset.

Method print()

Usage

Dataset$print(...)

Arguments

...

additional arguments to the printing method

Returns

invisibly self

Method get_dataset()

Performs any delayed mutate calls before returning self.

Usage

Dataset$get_dataset()

Returns

dataset (Dataset)

Method get_attrs()

Get all dataset attributes

Usage

Dataset$get_attrs()

Returns

(named list) with dataset attributes

Method get_raw_data()

Derive the raw data frame inside this object

Usage

Dataset$get_raw_data()

Returns

data.frame or rtable

Method get_numeric_colnames()

Derive the names of all numeric columns

Usage

Dataset$get_numeric_colnames()

Returns

character vector.

Method get_character_colnames()

Derive the names of all character columns

Usage

Dataset$get_character_colnames()

Returns

character vector.

Method get_factor_colnames()

Derive the names of all factor columns

Usage

Dataset$get_factor_colnames()

Returns

character vector.

Method get_colnames()

Derive the column names

Usage

Dataset$get_colnames()

Returns

character vector.

Method get_column_labels()

Derive the column labels

Usage

Dataset$get_column_labels()

Returns

character vector.

Method get_rownames()

Derive the row names

Usage

Dataset$get_rownames()

Returns

character vector.

Method get_row_labels()

Derive the row labels

Usage

Dataset$get_row_labels()

Returns

character vector.

Method get_dataname()

Derive the name which was formerly called dataname

Usage

Dataset$get_dataname()

Returns

character name of the dataset

Method get_datanames()

Derive the dataname

Usage

Dataset$get_datanames()

Returns

character name of the dataset

Method get_dataset_label()

Derive the label which was former called datalabel

Usage

Dataset$get_dataset_label()

Returns

character label of the dataset

Method get_keys()

Get primary keys of dataset

Usage

Dataset$get_keys()

Returns

(character vector) with dataset primary keys

Method get_join_keys()

Get JoinKeys object with keys used for joining.

Usage

Dataset$get_join_keys()

Returns

(JoinKeys)

Method get_hash()

Returns the string representation of the raw data hashed with the MD5 hash algorithm.

Usage

Dataset$get_hash()

Returns

character the hash of the raw data

Method get_var_r6()

Get the list of dependencies that are Dataset or DatasetConnector objects

Usage

Dataset$get_var_r6()

Returns

list

Method reassign_datasets_vars()

Overwrites Dataset or DatasetConnector dependencies of this Dataset with those found in datasets. Reassignment refers only to the provided datasets, other vars remains the same.

Usage

Dataset$reassign_datasets_vars(datasets)

Arguments

datasets

(named list of Dataset(s) or DatasetConnector(s))
objects with valid pointers.

Details

Reassign vars in this object to keep references up to date after deep clone. Update is done based on the objects passed in datasets argument. Overwrites dependencies with names matching the names of the objects passed in datasets.

Returns

NULL invisible

Examples

test_dataset <- teal:::Dataset$new(
  dataname = "iris",
  x = iris,
  vars = list(dep = teal:::Dataset$new("iris2", iris))
)
test_dataset$reassign_datasets_vars(
  list(iris2 = teal:::Dataset$new("iris2", head(iris)))
)

Method set_dataset_label()

Set the label for the dataset

Usage

Dataset$set_dataset_label(label)

Arguments

label

(character)
Label to describe the dataset

Returns

(self) invisibly for chaining

Method set_keys()

Set new keys

Usage

Dataset$set_keys(keys)

Arguments

keys

optional, (character)
Vector with primary keys

Returns

(self) invisibly for chaining.

Method set_join_keys()

set join_keys for a given dataset and object

Usage

Dataset$set_join_keys(x)

Arguments

x

list of JoinKeySet objects (which are created using the join_key function) or single JoinKeySet objects

Returns

(self) invisibly for chaining

Method merge_join_keys()

merge input join key with join key inside of object

Usage

Dataset$merge_join_keys(x)

Arguments

x

list of JoinKeys objects or single JoinKeys object

Returns

(self) invisibly for chaining

Method mutate_join_keys()

mutate the join_keys for a given dataset and object

Usage

Dataset$mutate_join_keys(dataset, val)

Arguments

dataset

(character) dataset for which join_keys are to be set against self

val

(named character) column names used to join

Returns

(self) invisibly for chaining

Method set_vars()

Adds variables which code depends on

Usage

Dataset$set_vars(vars)

Arguments

vars

(named list) contains any R object which code depends on

Returns

(self) invisibly for chaining

Method set_code()

Sets reproducible code

Usage

Dataset$set_code(code)

Arguments

code

(character)
A character string defining the code needed to produce the data set in x. initialize() and recreate() accept code as CodeClass also which is needed to preserve the code uniqueness and correct order.

Returns

(self) invisibly for chaining

Method get_code()

Get code to get data

Usage

Dataset$get_code(deparse = TRUE)

Arguments

deparse

(logical) whether return deparsed form of a call

Returns

optionally deparsed call object

Method get_code_class()

Get internal CodeClass object

Usage

Dataset$get_code_class(nodeps = FALSE)

Arguments

nodeps

(logical(1)) whether CodeClass should not contain the code of the dependent vars the mutate

Returns

\code{CodeClass}

Method get_mutate_code_class()

Get internal CodeClass object

Usage

Dataset$get_mutate_code_class()

Returns

\code{CodeClass}

Method get_vars()

Get internal vars object

Usage

Dataset$get_vars()

Returns

\code{list}

Method get_mutate_vars()

Get internal mutate_vars object

Usage

Dataset$get_mutate_vars()

Returns

\code{list}

Method is_mutate_delayed()

Usage

Dataset$is_mutate_delayed()

Returns

logical

Method mutate()

Mutate dataset by code

Usage

Dataset$mutate(code, vars = list(), force_delay = FALSE)

Arguments

code

(CodeClass) or (character) R expressions to be executed

vars

a named list of R objects that code depends on to execute

force_delay

(logical) used by the containing DatasetConnector objectEither code or script must be provided, but not both.

Returns

(self) invisibly for chaining

Method check()

Check to determine if the raw data is reproducible from the get_code() code.

Usage

Dataset$check()

Returns

TRUE if the dataset generated from evaluating the get_code() code is identical to the raw data, else FALSE. Check if keys has been specified correctly for dataset. Set of keys should distinguish unique rows or be character(0).

Method check_keys()

Usage

Dataset$check_keys(keys = private$.keys)

Arguments

keys

optional, (character)
Vector with primary keys

Returns

TRUE if dataset has been already pulled, else FALSE

Method is_pulled()

Check if dataset has already been pulled.

Usage

Dataset$is_pulled()

Returns

TRUE if dataset has been already pulled, else FALSE

Method clone()

The objects of this class are cloneable with this method.

Usage

Dataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super class

teal::Dataset -> MAEDataset

Methods

Public methods

MAEDataset$new()

MAEDataset$check()

MAEDataset$check_keys()

MAEDataset$print()

MAEDataset$clone()

Inherited methods

teal::Dataset$get_attrs()

teal::Dataset$get_character_colnames()

teal::Dataset$get_code()

teal::Dataset$get_code_class()

teal::Dataset$get_colnames()

teal::Dataset$get_column_labels()

teal::Dataset$get_dataname()

teal::Dataset$get_datanames()

teal::Dataset$get_dataset()

teal::Dataset$get_dataset_label()

teal::Dataset$get_factor_colnames()

teal::Dataset$get_hash()

teal::Dataset$get_join_keys()

teal::Dataset$get_keys()

teal::Dataset$get_mutate_code_class()

teal::Dataset$get_mutate_vars()

teal::Dataset$get_numeric_colnames()

teal::Dataset$get_raw_data()

teal::Dataset$get_row_labels()

teal::Dataset$get_rownames()

teal::Dataset$get_var_r6()

teal::Dataset$get_vars()

teal::Dataset$is_mutate_delayed()

teal::Dataset$is_pulled()

teal::Dataset$merge_join_keys()

teal::Dataset$mutate()

teal::Dataset$mutate_join_keys()

teal::Dataset$reassign_datasets_vars()

teal::Dataset$recreate()

teal::Dataset$set_code()

teal::Dataset$set_dataset_label()

teal::Dataset$set_join_keys()

teal::Dataset$set_keys()

teal::Dataset$set_vars()

Method new()

Create a new object of MAEDataset class

Usage

MAEDataset$new(
  dataname,
  x,
  keys = character(0),
  code = character(0),
  label = character(0),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(MultiAssayExperiment)

keys

optional, (character)
vector with primary keys

code

(character or CodeClass)
A character string defining the code needed to produce the data set in x. initialize() and recreate() accept code as CodeClass also which is needed to preserve the code uniqueness and correct order.

label

(character)
Label to describe the dataset

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Method check()

Check to determine if the raw data is reproducible from the get_code() code.

Usage

MAEDataset$check()

Returns

TRUE if the dataset generated from evaluating the get_code() code is identical to the raw data, else FALSE. Check if keys has been specified correctly for dataset. Set of keys should distinguish unique rows or be character(0).

Method check_keys()

Usage

MAEDataset$check_keys(keys = private$.keys)

Arguments

keys

optional, (character)
vector with primary keys

Returns

TRUE if dataset has been already pulled, else FALSE Prints this MAEDataset.

Method print()

Usage

MAEDataset$print(...)

Arguments

...

additional arguments to the printing method

Returns

invisibly self

Method clone()

The objects of this class are cloneable with this method.

Usage

MAEDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Dataset$reassign_datasets_vars`
## ------------------------------------------------

test_dataset <- teal:::Dataset$new(
  dataname = "iris",
  x = iris,
  vars = list(dep = teal:::Dataset$new("iris2", iris))
)
test_dataset$reassign_datasets_vars(
  list(iris2 = teal:::Dataset$new("iris2", head(iris)))
)