R6 Class representing a dataset with its attributes

Any data.frame object can be stored inside this object. Some attributes like colnames, dimension or column names for a specific type will be automatically derived.

Active bindings

ncol

Number of columns

nrow

Number of rows

dim

Dimension c(x, y)

colnames

The column names of the data

rownames

The rownames of the data

raw_data

The data.frame behind this R6 class

data

The data.frame behind this R6 class

column_labels

for backwards compatibility will be deprecated in future

var_names

The column names of the data

row_labels

Row labels (can have spaces)

Methods

Public methods

Dataset$new()

Dataset$recreate()

Dataset$get_attrs()

Dataset$get_raw_data()

Dataset$get_numeric_colnames()

Dataset$get_character_colnames()

Dataset$get_factor_colnames()

Dataset$get_colnames()

Dataset$get_column_labels()

Dataset$get_rownames()

Dataset$get_row_labels()

Dataset$get_dataname()

Dataset$get_datanames()

Dataset$get_dataset_label()

Dataset$get_keys()

Dataset$set_dataset_label()

Dataset$set_keys()

Dataset$set_vars()

Dataset$set_code()

Dataset$get_code()

Dataset$get_code_class()

Dataset$mutate()

Dataset$check()

Dataset$check_keys()

Dataset$is_pulled()

Dataset$clone()

Method new()

Create a new object of Dataset class

Usage

Dataset$new(
  dataname,
  x,
  keys = character(0),
  code = character(0),
  label = character(0),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(data.frame)

keys

optional, (character)
vector with primary keys

code

(character)
A character string defining the code needed to produce the data set in x. initialize() and recreate() accept code as CodeClass also which is needed to preserve the code uniqueness and correct order.

label

(character)
Label to describe the dataset

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Method recreate()

Recreate a dataset with its current attributes This is useful way to have access to class initialize method basing on class object.

Usage

Dataset$recreate(
  dataname = self$get_dataname(),
  x = self$get_raw_data(),
  keys = self$get_keys(),
  code = self$get_code_class(),
  label = self$get_dataset_label(),
  vars = list()
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(data.frame)

keys

optional, (character)
vector with primary keys

code

(character)
A character string defining the code needed to produce the data set in x. initialize() and recreate() accept code as CodeClass also which is needed to preserve the code uniqueness and correct order.

label

(character)
Label to describe the dataset

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Returns

a new object of Dataset class

Method get_attrs()

Get all dataset attributes

Usage

Dataset$get_attrs()

Returns

(named list) with dataset attributes

Method get_raw_data()

Derive the raw data frame inside this object

Usage

Dataset$get_raw_data()

Returns

data.frame or rtable

Method get_numeric_colnames()

Derive the names of all numeric columns

Usage

Dataset$get_numeric_colnames()

Returns

character vector.

Method get_character_colnames()

Derive the names of all character columns

Usage

Dataset$get_character_colnames()

Returns

character vector.

Method get_factor_colnames()

Derive the names of all factor columns

Usage

Dataset$get_factor_colnames()

Returns

character vector.

Method get_colnames()

Derive the column names

Usage

Dataset$get_colnames()

Returns

character vector.

Method get_column_labels()

Derive the column labels

Usage

Dataset$get_column_labels()

Returns

character vector.

Method get_rownames()

Derive the row names

Usage

Dataset$get_rownames()

Returns

character vector.

Method get_row_labels()

Derive the row labels

Usage

Dataset$get_row_labels()

Returns

character vector.

Method get_dataname()

Derive the name which was formerly called dataname

Usage

Dataset$get_dataname()

Returns

character name of the dataset

Method get_datanames()

Derive the dataname

Usage

Dataset$get_datanames()

Returns

character name of the dataset

Method get_dataset_label()

Derive the label which was former called datalabel

Usage

Dataset$get_dataset_label()

Returns

character label of the dataset

Method get_keys()

Get primary keys of dataset

Usage

Dataset$get_keys()

Returns

(character vector) with dataset primary keys

Method set_dataset_label()

Set the label for the dataset

Usage

Dataset$set_dataset_label(label)

Arguments

label

(character)
Label to describe the dataset

Returns

(self) invisibly for chaining

Method set_keys()

Set new keys

Usage

Dataset$set_keys(keys)

Arguments

keys

optional, (character)
vector with primary keys

Returns

(self) invisibly for chaining.

Method set_vars()

Adds variables which code depends on

Usage

Dataset$set_vars(vars)

Arguments

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Returns

(self) invisibly for chaining

Method set_code()

Sets reproducible code

Usage

Dataset$set_code(code)

Arguments

code

(character)
A character string defining the code needed to produce the data set in x. initialize() and recreate() accept code as CodeClass also which is needed to preserve the code uniqueness and correct order.

Returns

(self) invisibly for chaining

Method get_code()

Get code to get data

Usage

Dataset$get_code(deparse = TRUE)

Arguments

deparse

(logical) whether return deparsed form of a call

Returns

optionally deparsed call object

Method get_code_class()

Get internal CodeClass object

Usage

Dataset$get_code_class()

Returns

\code{CodeClass}

Method mutate()

Mutate dataset by code

Either code or script must be provided, but not both.

Usage

Dataset$mutate(code, vars = list())

Arguments

code

(character)
A character string defining the code needed to produce the data set in x. initialize() and recreate() accept code as CodeClass also which is needed to preserve the code uniqueness and correct order.

vars

(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Returns

(self) invisibly for chaining

Method check()

Check to determine if the raw data is reproducible from the get_code() code.

Usage

Dataset$check()

Returns

TRUE if the dataset generated from evaluating the get_code() code is identical to the raw data, else FALSE. Check if keys has been specified correctly for dataset. Set of keys should distinguish unique rows or be character(0).

Method check_keys()

Usage

Dataset$check_keys(keys = private$.keys)

Arguments

keys

optional, (character)
vector with primary keys

Returns

TRUE if dataset has been already pulled, else FALSE

Method is_pulled()

Check if dataset has already been pulled.

Usage

Dataset$is_pulled()

Returns

TRUE if dataset has been already pulled, else FALSE

Method clone()

The objects of this class are cloneable with this method.

Usage

Dataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.