Create a new DatasetConnector object

dataset_connector(
  dataname,
  pull_callable,
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  vars = list()
)

Arguments

dataname
(character)
A given name for the dataset it may not contain spaces

pull_callable
(CallableFunction)
function with necessary arguments set to fetch data from connection.

keys
optional, (character)
vector of dataset primary keys column names

label
(character)
Label to describe the dataset.

code
(character)
A character string defining code to modify raw_data from this dataset. To modify current dataset code should contain at least one assignment to object defined in dataname argument. For example if dataname = ADSL example code should contain ADSL <- <some R code>. Can't be used simultaneously with script

script
(character)
Alternatively to code - location of the file containing modification code. Can't be used simultaneously with script.

vars
(named list))
In case when this object code depends on other Dataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It's recommended to include Dataset or DatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Value

new DatasetConnector object

Details

Create DatasetConnector from callable_function.

Examples

library(MultiAssayExperiment)

#> Loading required package: SummarizedExperiment

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: DelayedArray

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘matrixStats’

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians

#> Loading required package: BiocParallel

#> 
#> Attaching package: ‘DelayedArray’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following objects are masked from ‘package:base’:
#> 
#>     aperm, apply, rowsum

# data.frame example
pull_fun2 <- callable_function(data.frame)
pull_fun2$set_args(args = list(a = c(1,2,3)))
dataset_connector("test", pull_fun2)

#> <DatasetConnector>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset: function () 
#>     get_dataset_label: function () 
#>     get_error_message: function () 
#>     get_keys: function () 
#>     get_pull_args: function () 
#>     get_pull_callable: function () 
#>     get_raw_data: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     initialize: function (dataname, pull_callable, keys = character(0), label = character(0), 
#>     is_failed: function () 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     pull: function (args = NULL, try = FALSE) 
#>     set_args: function (args) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_ui_input: function (inputs = NULL) 
#>   Private:
#>     dataname: test
#>     dataset: NULL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     get_mutate_code_class: function () 
#>     get_pull_code_class: function (args = NULL) 
#>     keys: 
#>     mutate_code: CodeClass, R6
#>     mutate_vars: list
#>     pull_callable: CallableFunction, Callable, R6
#>     pull_internal: function (args = NULL, try = FALSE) 
#>     pull_vars: list
#>     server: function (input, output, session, data_args = NULL) 
#>     set_dataname: function (dataname) 
#>     set_failure: function (res) 
#>     set_mutate_code: function (code) 
#>     set_mutate_vars: function (vars) 
#>     set_pull_callable: function (pull_callable) 
#>     set_pull_vars: function (pull_vars) 
#>     set_ui: function (ui_args = NULL) 
#>     ui: function (id) 
#>     ui_input: NULL

# MultiAssayExperiment example
pull_fun <- callable_function(
  function() { library("MultiAssayExperiment"); data("miniACC"); return(miniACC) })
dataset_connector(
  "miniacc",
  pull_fun,
  code = 'library("MultiAssayExperiment"); data("miniACC"); return(miniACC)'
)

#> <DatasetConnector>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset: function () 
#>     get_dataset_label: function () 
#>     get_error_message: function () 
#>     get_keys: function () 
#>     get_pull_args: function () 
#>     get_pull_callable: function () 
#>     get_raw_data: function () 
#>     get_server: function () 
#>     get_ui: function (id) 
#>     initialize: function (dataname, pull_callable, keys = character(0), label = character(0), 
#>     is_failed: function () 
#>     is_pulled: function () 
#>     launch: function () 
#>     mutate: function (code, vars = list()) 
#>     pull: function (args = NULL, try = FALSE) 
#>     set_args: function (args) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_ui_input: function (inputs = NULL) 
#>   Private:
#>     dataname: miniacc
#>     dataset: NULL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     get_mutate_code_class: function () 
#>     get_pull_code_class: function (args = NULL) 
#>     keys: 
#>     mutate_code: CodeClass, R6
#>     mutate_vars: list
#>     pull_callable: CallableFunction, Callable, R6
#>     pull_internal: function (args = NULL, try = FALSE) 
#>     pull_vars: list
#>     server: function (input, output, session, data_args = NULL) 
#>     set_dataname: function (dataname) 
#>     set_failure: function (res) 
#>     set_mutate_code: function (code) 
#>     set_mutate_vars: function (vars) 
#>     set_pull_callable: function (pull_callable) 
#>     set_pull_vars: function (pull_vars) 
#>     set_ui: function (ui_args = NULL) 
#>     ui: function (id) 
#>     ui_input: NULL