Data extract dynamic filter specification

This function returns a configuration for the data_extract_input module. This function covers the configuration of filtering datasets (so called filter_spec), which then is used to build the UI element in the teal app.

filter_spec_internal(
  vars_choices,
  vars_selected = NULL,
  vars_label = NULL,
  vars_fixed = FALSE,
  vars_multiple = TRUE,
  choices = NULL,
  selected = NULL,
  label = NULL,
  fixed = FALSE,
  multiple = TRUE,
  sep = if_null(attr(vars_choices, "sep"), " - "),
  drop_keys = FALSE,
  dataname = NULL,
  initialized = FALSE
)

# S3 method for delayed_data
filter_spec_internal(
  vars_choices,
  vars_selected = NULL,
  vars_label = NULL,
  vars_fixed = FALSE,
  vars_multiple = TRUE,
  choices = NULL,
  selected = NULL,
  label = NULL,
  fixed = FALSE,
  multiple = TRUE,
  sep = if_null(attr(vars_choices, "sep"), " - "),
  drop_keys = FALSE,
  dataname = NULL,
  initialized = FALSE
)

# S3 method for default
filter_spec_internal(
  vars_choices,
  vars_selected = NULL,
  vars_label = NULL,
  vars_fixed = FALSE,
  vars_multiple = TRUE,
  choices = NULL,
  selected = NULL,
  label = NULL,
  fixed = FALSE,
  multiple = TRUE,
  sep = if_null(attr(vars_choices, "sep"), " - "),
  drop_keys = FALSE,
  dataname = NULL,
  initialized = FALSE
)

Arguments

vars_choices
(character or delayed_data)
the vector of dataset column names available to build dynamic filter delayed_data objects can be created via variable_choices.

vars_selected
(NULL or named character)
the selected column name out from choices.

vars_label
(character)
the title printed on the UI element generated on the basis of this filter_spec.

vars_fixed
(logical)
if true allow to change the selected variables in the UI element; otherwise, do not allow.

vars_multiple
(logical)
if true allow to select multiple variables in the UI elements; otherwise, do not allow.

choices
(character or numeric or logical or (delayed_data) object. Named character vector to define the choices of a shiny selectInput. These choices will be used to filter the dataset.

These shall be filter values of the vars input separated by the separator(sep). Please watch out that the filter values have to follow the order of the vars input. In the following example we will show how to filter two columns:

vars = c("PARAMCD","AVISIT") and choices = c("CRP - BASELINE", "ALT - BASELINE") will lead to a filtering of (PARAMCD == "CRP" & AVISIT == "BASELINE") | (PARAMCD == "ALT" & AVISIT == "BASELINE").

The sep input has to be " - " in this case.

delayed_data objects can be created via variable_choices or value_choices.

selected
(character or numeric or logical or (delayed_data) object. Named character vector to define the selected values of a shiny selectInput (default values). This value will be displayed inside the shiny app upon start.

label
optional (character). Define a label on top of this specific shiny selectInput.

fixed
(logical)
if true allow to change the initially selected values of the variables; otherwise, do not allow.

multiple
(logical) Whether multiple values shall be allowed in the shiny selectInput.

sep
(character) A separator string to split the choices or selected inputs into the values of the different columns

drop_keys
optional, (logical) whether to drop filter column from the dataset keys, TRUE on default

dataname
(character)
the name of the dataset this filter covers. Set during the initialization of the teal application.

initialized
(logical)
indicates whether this filter was already initialized by data_extract_input_filter_srv. TRUE if this filter was already consumed by the server function; FALSE otherwise.

Value

filter_spec or delayed_filter_spec S3-class object.

See also

Examples

teal:::filter_spec_internal(
  vars_choices = c("PARAMCD", "AVISIT"),
  vars_selected = "PARAMCD",
  vars_multiple = TRUE
)

#> filter_spec with delayed data: 
#> $ vars_choices
#> [1] "PARAMCD" "AVISIT" 
#> $ vars_selected
#> [1] "PARAMCD"
#> $ vars_label
#> NULL
#> $ vars_fixed
#> [1] FALSE
#> $ vars_multiple
#> [1] TRUE
#> $ choices
#> NULL
#> $ selected
#> NULL
#> $ label
#> NULL
#> $ multiple
#> [1] TRUE
#> $ fixed
#> [1] FALSE
#> $ sep
#> [1] " - "
#> $ drop_keys
#> [1] FALSE
#> $ dataname
#> NULL
#> $ initialized
#> [1] FALSE

library(random.cdisc.data)
ADRS <- radrs(cached = TRUE)
teal:::filter_spec_internal(
  vars_choices = variable_choices(ADRS),
  vars_selected = "PARAMCD",
  vars_multiple = TRUE
)

#> filter_spec with delayed data: 
#> $ vars_choices
#> number of choices: 48 
#> 
#> STUDYID: Study Identifier
#> USUBJID: Unique Subject Identifier
#> SUBJID: Subject Identifier for the Study
#> SITEID: Study Site Identifier
#> AGE: Age
#> SEX: Sex
#> RACE: Race
#> COUNTRY: Country
#> DTHFL: Subject Death Flag
#> INVID: Investigator Identifier
#> INVNAM: Investigator Name
#> ARM: Description of Planned Arm
#> ARMCD: Planned Arm Code
#> ACTARM: Description of Actual Arm
#> ACTARMCD: Actual Arm Code
#> REGION1: Geographic Region 1
#> STRATA1: Stratification Factor 1
#> STRATA2: Stratification Factor 2
#> BMRKR1: Continous Level Biomarker 1
#> BMRKR2: Categorical Level Biomarker 2
#> ITTFL: Intent-To-Treat Population Flag
#> SAFFL: Safety Population Flag
#> BMEASIFL: Response Evaluable Population Flag
#> BEP01FL: Biomarker Evaluable Population Flag
#> RANDDT: Date of Randomization
#> TRTSDTM: Datetime of First Exposure to Treatment
#> TRTEDTM: Datetime of Last Exposure to Treatment
#> EOSSTT: End of Study Status
#> EOSDT: End of Study Date
#> EOSDY: End of Study Relative Day
#> DCSREAS: Reason for Discontinuation from Study
#> DTHDT: Date of Death
#> DTHCAUS: Cause of Death
#> DTHCAT: Cause of Death Category
#> LDDTHELD: Elapsed Days from Last Dose to Death
#> LDDTHGR1: Last Dose to Death - Days Elapsed Grp 1
#> LSTALVDT: Date Last Known Alive
#> study_duration_secs: NOT A STANDARD BUT NEEDED FOR RCD
#> ASEQ: Analysis Sequence Number
#> RSSEQ: Sequence Number
#> PARAM: Parameter
#> PARAMCD: Parameter Code
#> AVAL: Analysis Value
#> AVALC: Analysis Value (C)
#> ADTM: Analysis Datetime
#> ADY: Analysis Relative Day
#> AVISIT: Analysis Visit
#> AVISITN: Analysis Visit (N)
#> 
#> $ vars_selected
#> [1] "PARAMCD"
#> $ vars_label
#> NULL
#> $ vars_fixed
#> [1] FALSE
#> $ vars_multiple
#> [1] TRUE
#> $ choices
#> NULL
#> $ selected
#> NULL
#> $ label
#> NULL
#> $ multiple
#> [1] TRUE
#> $ fixed
#> [1] FALSE
#> $ sep
#> [1] " - "
#> $ drop_keys
#> [1] FALSE
#> $ dataname
#> NULL
#> $ initialized
#> [1] FALSE

teal:::filter_spec_internal(
  vars_choices = variable_choices("ADRS"),
  vars_selected = "PARAMCD",
  vars_multiple = TRUE
)

#> filter_spec with delayed data: 
#> $ vars_choices
#>   variable_choices with delayed data: ADRS
#>   $ data
#>   [1] "ADRS"
#>   $ subset
#>   NULL
#>   $ key
#>   NULL
#> $ vars_selected
#> [1] "PARAMCD"
#> $ vars_label
#> NULL
#> $ vars_fixed
#> [1] FALSE
#> $ vars_multiple
#> [1] TRUE
#> $ choices
#> NULL
#> $ selected
#> NULL
#> $ label
#> NULL
#> $ multiple
#> [1] TRUE
#> $ fixed
#> [1] FALSE
#> $ sep
#> [1] " - "
#> $ drop_keys
#> [1] FALSE
#> $ dataname
#> NULL
#> $ initialized
#> [1] FALSE