Create the Server and UI Function For the Shiny App

End-users: This is the most important function for you to start a teal app that is composed out of teal modules.

Notes for developers: This is a wrapper function around the module_teal.R functions. Unless you are an end-user, don't use this function, but instead this module.

init(
  data,
  modules,
  title = NULL,
  filter = list(),
  header = tags$p("Add Title Here"),
  footer = tags$p("Add Footer Here"),
  id = character(0)
)

Arguments

data
(TealData or TealDataset or TealDatasetConnector or list or data.frame)
R6 object as returned by cdisc_data(), teal_data(), cdisc_dataset(), dataset(), dataset_connector() or cdisc_dataset_connector() or a single data.frame or a list of the previous objects or function returning a named list. NOTE: teal does not guarantee reproducibility of the code when names of the list elements do not match the original object names. To ensure reproducibility please use teal_data() or cdisc_data() with check = TRUE enabled.

modules
(list or teal_modules)
nested list of teal_modules or module objects. See modules() and module() for more details.

title
(NULL or character)
The browser window title (defaults to the host URL of the page).

filter
(list)
You can define filters that show when the app starts. The general pattern is: list(iris = list(Species = ..., Sepal.Length = ...), mtcars = ...). An example is: list(iris = list(Species = c("setosa", "versicolor"))). More generally, the filters for the variable, e.g. Species can be specified as follows: list(Species = c("setosa", "versicolor")), list(Species = list(c("setosa", "versicolor"), keep_na = TRUE)), or equivalently with: list(Species = c("setosa", "versicolor", NA)), or for the default filter (not very restrictive): list(Species = default_filter())

Instead of choices above, use the following names:

numerical: range

factor: choices

logical: logical A general example is: list( iris = list(Sepal.Length = default_filter(), Species = c("M", NA)), mtcars = list(mpg = default_filter()) ) Ignored if the app is restored from a bookmarked state.

header
(character or shiny.tag)
the header of the app. Note shiny code placed here (and in the footer argument) will be placed in the app's ui function so code which needs to be placed in the ui function (such as loading css via htmltools::htmlDependency()) should be included here.

footer
(character or shiny.tag)
the footer of the app

id
(character)
module id to embed it, if provided, the server function must be called with shiny::moduleServer(); See the vignette for an example. However, ui_teal_with_splash() is then preferred to this function.

Value

named list with server and ui function

Examples

library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl

options(teal_logging = FALSE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"
  ),
  modules = root_modules(
    module(
      "data source",
      server = function(input, output, session, datasets) {},
      ui = function(id, ...) div(p("information about data source")),
      filters = "all"
    ),
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot(
          hist(datasets$get_data("ADSL", filtered = TRUE)$AGE)
        )
      },
      ui = function(id, ...) {
        ns <- NS(id)
        plotOutput(ns("hist"))
      },
      filters = "ADSL"
    )
  ),
  title = "App title",
  filter = list(ADSL = list(AGE = default_filter())),
  header = tags$h1("Sample App"),
  footer = tags$p("Copyright 2017 - 2020")
)
if (FALSE) {
shinyApp(app$ui, app$server)
# or: to also work with bookmarking
bookmarkableShinyApp(app$ui, app$server)
}

# See the vignette for an example how to embed this app as a module
# into a larger application