Create the Server and UI Function For the Shiny App

End-users: This is the most important function for you to start a teal app that is composed out of teal modules.

init(
  data,
  modules,
  title = NULL,
  filter = list(),
  header = tags$p("Add Title Here"),
  footer = tags$p("Add Footer Here"),
  id = character(0)
)

Arguments

data
(RelationalDataList) R6 object where cdisc_data or teal_data returns such a one.

modules
nested list with one list per module with the following named list elements:

name

string with name shown in menu for the analysis item	server

required, shiny server module function, see \link[shiny]{callModule} for more information	ui

required, shiny ui module function, see \link[shiny]{callModule} for more information	data

required, vector with datasets names that are passed on (filtered) to the server function	options

title
(NULL or character) The browser window title (defaults to the host URL of the page).

filter
(list) You can define filters that show when the app starts. The general pattern is: list(ADSL = list(SEX = ..., AGE = ...), ADAE = ...). An example is: list(ADSL = list(SEX = c("M", "F"))). More generally, the filters for the variable, e.g. SEX can be specified as follows: list(SEX = list(choices = c("M", "F"))), list(SEX = list(choices = c("M", "F"), keep_na = TRUE)), or equivalently with: list(SEX = c("M", "F")), list(SEX = c("M", "F", NA)), or for the default filter (not very restrictive): list(SEX = default_filter())

Instead of choices above, use the following names:

numerical: range

factor: choices

logical: logical A general example is: list( ADSL = list(AGE = default_filter(), SEX = c("M", NA)), ADAE = list(AETOXGR = default_filter()) ) Ignored if the app is restored from a bookmarked state.

header
(character or shiny.tag) the header of the app

footer
(character or shiny.tag) the footer of the app

id
(character) module id to embed it, if provided, the server function must be called with callModule; See the vignette for an example. However, \link{ui_teal_with_splash} is then preferred to this function.

Value

named list with server and ui function

Details

Notes for developers: This is a wrapper function around the module_teal.R functions. Unless you are an end-user, don't use this function, but instead this module.

Examples

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

options(teal_logging = FALSE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- radsl(seed = 1)"
  ),
  modules = root_modules(
    module(
      "data source",
      server = function(input, output, session, datasets) {},
      ui = function(id, ...) div(p("information about data source")),
      filters = "all"
    ),
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot(
          hist(datasets$get_data("ADSL", filtered = TRUE)$AGE)
        )
      },
      ui = function(id, ...) {
        ns <- NS(id)
        plotOutput(ns("hist"))
      },
      filters = "ADSL"
    )
  ),
  title = "App title",
  filter = list(ADSL = list(AGE = default_filter())),
  header = tags$h1("Sample App"),
  footer = tags$p("Copyright 2017 - 2020")
)
if (FALSE) {
shinyApp(app$ui, app$server)
# or: to also work with bookmarking
bookmarkableShinyApp(app$ui, app$server)
}

# See the vignette for an example how to embed this app as a module
# into a larger application