Create the Server and Ui Function for the Shiny App

Creates the server and ui part for a teal shiny app

init(data, modules, filter = NULL, header = tags$p("title here"),
  footer = tags$p("footer here"))

Arguments

data
named list with datasets. Dataset names are case sensitive. The `ASL` data is mandatory.

filter
filter settings. Nested named list, currently with init list element.

header
object of class `shiny.tag` to be used as the header of the app

footer
object of class `shiny.tag` to be used as the footer of the app

analysis
nested list with one list per analysis item with the following named list elements:

name	string with name

shown in menu for the analysis item	server

required, shiny server	module function, see

callModule	for more

information	ui

required, shiny ui module function, see	callModule

for more information	data

required,	vector with datasets names that are passed on (filtered) to the server

function	options

optional, other arguments passed on to the server	function

elements
list with lists defining new pages (as for analysis) or one of the keywords data_table, variable_browser, analysis.

Value

named list with server and ui function

Examples

# NOT RUN {
ASL <- generate_sample_data('ASL')
ARS <- generate_sample_data('ARS')
ATE <- generate_sample_data('ATE')

x <- teal::init(
  data =  list(ASL = ASL, ARS = ARS, ATE = ATE),
  modules = root_modules(
    module(
      "data source",
      server = function(input, output, session, datasets) {},
      ui = function(id) div(p("information about data source")),
      filters = NULL
    ),
    tm_data_table(),
    tm_variable_browser(),
    modules(
      label = "analysis items",
      tm_table(
         label = "demographic table",
         dataname = "ASL",
         xvar = "SEX",
         yvar = "RACE",
         yvar_choices = c("RACE", "AGEGR", "REGION")
      ),
      tm_scatterplot(
         label = "scatterplot",
         dataname = "ASL",
         xvar = "AGE",
         yvar = "BBMI",
         color_by = "_none_",
         color_by_choices = c("_none_", "STUDYID")
      ),
      module(
         label = "survival curves",
         server = function(input, output, session, datasets) {},
         ui = function(id) div(p("Kaplan Meier Curve")),
         filters = "ATE"
      )
    )
  ),
  header = tags$h1("Sample App"),
  footer = tags$p("Copyright 2017")
)

shinyApp(x$ui, x$server)

# }