Load datasets

load_datasets(x, ...)

# S3 method for Dataset
load_datasets(x, ...)

# S3 method for DatasetConnector
load_datasets(x, args = NULL, try = FALSE, ...)

# S3 method for RelationalDataConnector
load_datasets(x, ...)

# S3 method for RelationalData
load_datasets(x, ...)

Arguments

x
(object of class RelationalData, Dataset or DatasetConnector)

...
(not used)
only for support of S3

args
(NULL or named list)
additional dynamic arguments passed to function which loads the data. Applicable only on DatasetConnector)

try
(logical)
whether perform function evaluation inside try clause. Applicable only on DatasetConnector)

Value

If executed in the interactive session shiny app is opened to load the data. If executed in shiny application - it returns shiny server module.

Examples

# Dataset ------
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
x <- dataset("ADSL", x = ADSL)

load_datasets(x)

# DatasetConnector ------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector("ADSL", fun = radsl, cached = FALSE)
load_datasets(adsl)
get_dataset(adsl)

#> A Dataset object containing the following data.frame (400 rows and 38 columns):
#>   STUDYID               USUBJID SUBJID SITEID AGE SEX
#> 1 AB12345  AB12345-CHN-8-id-357 id-357  CHN-8  39   M
#> 2 AB12345 AB12345-CHN-12-id-291 id-291 CHN-12  31   F
#> 3 AB12345 AB12345-PAK-13-id-173 id-173 PAK-13  27   F
#> 4 AB12345 AB12345-NGA-10-id-400 id-400 NGA-10  31   M
#> 5 AB12345   AB12345-PAK-4-id-42  id-42  PAK-4  45   F
#> 6 AB12345   AB12345-CHN-4-id-57  id-57  CHN-4  30   F
#>                               RACE COUNTRY DTHFL         INVID         INVNAM
#> 1 AMERICAN INDIAN OR ALASKA NATIVE     CHN     N  INV ID CHN-8  Dr. CHN-8 Doe
#> 2        BLACK OR AFRICAN AMERICAN     CHN     Y INV ID CHN-12 Dr. CHN-12 Doe
#> 3                            ASIAN     PAK     N INV ID PAK-13 Dr. PAK-13 Doe
#> 4        BLACK OR AFRICAN AMERICAN     NGA     N INV ID NGA-10 Dr. NGA-10 Doe
#> 5                            ASIAN     PAK     N  INV ID PAK-4  Dr. PAK-4 Doe
#> 6                            ASIAN     CHN     N  INV ID CHN-4  Dr. CHN-4 Doe
#>              ARM ARMCD         ACTARM ACTARMCD REGION1 STRATA1 STRATA2   BMRKR1
#> 1 C: Combination ARM C C: Combination    ARM C    Asia       C      S2 6.056490
#> 2 C: Combination ARM C C: Combination    ARM C    Asia       C      S2 2.558474
#> 3     B: Placebo ARM B     B: Placebo    ARM B    Asia       B      S2 9.087296
#> 4 C: Combination ARM C C: Combination    ARM C  Africa       B      S1 5.742336
#> 5     B: Placebo ARM B     B: Placebo    ARM B    Asia       A      S1 4.720375
#> 6      A: Drug X ARM A      A: Drug X    ARM A    Asia       A      S1 1.215811
#>   BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL     RANDDT             TRTSDTM
#> 1 MEDIUM     Y     Y        N       N 2019-02-17 2019-02-21 12:19:55
#> 2 MEDIUM     Y     Y        N       Y 2019-02-21 2019-02-21 17:36:41
#> 3    LOW     Y     Y        N       N 2019-02-23 2019-02-23 07:11:41
#> 4 MEDIUM     Y     Y        Y       N 2019-02-23 2019-02-24 15:24:23
#> 5 MEDIUM     Y     Y        N       N 2019-02-21 2019-02-25 17:29:58
#> 6    LOW     Y     Y        Y       Y 2019-02-24 2019-02-27 12:17:23
#>               TRTEDTM       EOSSTT      EOSDT EOSDY               DCSREAS
#> 1 2021-02-20 23:58:19    COMPLETED 2021-02-20   731                  <NA>
#> 2 2021-02-13 19:38:54 DISCONTINUED 2021-02-13   724                 DEATH
#> 3 2021-02-19 00:44:57 DISCONTINUED 2021-02-19   727 WITHDRAWAL BY SUBJECT
#> 4 2021-02-15 08:49:11 DISCONTINUED 2021-02-15   722    PHYSICIAN DECISION
#> 5 2021-02-25 05:08:22    COMPLETED 2021-02-25   731                  <NA>
#> 6 2021-02-26 23:55:47    COMPLETED 2021-02-26   731                  <NA>
#>        DTHDT DTHCAUS DTHCAT LDDTHELD LDDTHGR1   LSTALVDT study_duration_secs
#> 1       <NA>    <NA>   <NA>       NA     <NA> 2021-03-21            63113904
#> 2 2021-02-25 MISSING  OTHER       12     <=30 2021-02-25            63113904
#> 3       <NA>    <NA>   <NA>       NA     <NA> 2021-03-10            63113904
#> 4       <NA>    <NA>   <NA>       NA     <NA> 2021-03-02            63113904
#> 5       <NA>    <NA>   <NA>       NA     <NA> 2021-03-22            63113904
#> 6       <NA>    <NA>   <NA>       NA     <NA> 2021-03-19            63113904
#> 
#> ...
#>     STUDYID               USUBJID SUBJID SITEID AGE SEX
#> 395 AB12345  AB12345-JPN-1-id-171 id-171  JPN-1  27   M
#> 396 AB12345 AB12345-USA-11-id-318 id-318 USA-11  41   F
#> 397 AB12345 AB12345-CHN-18-id-376 id-376 CHN-18  43   M
#> 398 AB12345   AB12345-USA-3-id-33  id-33  USA-3  38   F
#> 399 AB12345  AB12345-CHN-18-id-44  id-44 CHN-18  42   F
#> 400 AB12345  AB12345-CHN-1-id-177 id-177  CHN-1  31   M
#>                          RACE COUNTRY DTHFL         INVID         INVNAM
#> 395                     WHITE     JPN     N  INV ID JPN-1  Dr. JPN-1 Doe
#> 396                     ASIAN     USA     N INV ID USA-11 Dr. USA-11 Doe
#> 397 BLACK OR AFRICAN AMERICAN     CHN     N INV ID CHN-18 Dr. CHN-18 Doe
#> 398 BLACK OR AFRICAN AMERICAN     USA     N  INV ID USA-3  Dr. USA-3 Doe
#> 399 BLACK OR AFRICAN AMERICAN     CHN     N INV ID CHN-18 Dr. CHN-18 Doe
#> 400                     ASIAN     CHN     N  INV ID CHN-1  Dr. CHN-1 Doe
#>                ARM ARMCD         ACTARM ACTARMCD       REGION1 STRATA1 STRATA2
#> 395 C: Combination ARM C C: Combination    ARM C          Asia       B      S1
#> 396 C: Combination ARM C C: Combination    ARM C North America       B      S2
#> 397     B: Placebo ARM B     B: Placebo    ARM B          Asia       C      S2
#> 398      A: Drug X ARM A      A: Drug X    ARM A North America       B      S2
#> 399      A: Drug X ARM A      A: Drug X    ARM A          Asia       B      S2
#> 400     B: Placebo ARM B     B: Placebo    ARM B          Asia       B      S2
#>       BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL     RANDDT             TRTSDTM
#> 395 3.519679    LOW     Y     Y        Y       Y 2021-02-13 2021-02-13 02:05:55
#> 396 2.529561 MEDIUM     Y     Y        N       N 2021-02-11 2021-02-13 12:21:36
#> 397 2.732193   HIGH     Y     Y        N       Y 2021-02-11 2021-02-13 15:10:07
#> 398 7.137043   HIGH     Y     Y        Y       Y 2021-02-12 2021-02-15 10:03:49
#> 399 2.548159    LOW     Y     Y        Y       N 2021-02-13 2021-02-15 11:31:48
#> 400 3.370379    LOW     Y     Y        N       N 2021-02-14 2021-02-18 10:47:47
#>                 TRTEDTM    EOSSTT      EOSDT EOSDY DCSREAS DTHDT DTHCAUS DTHCAT
#> 395 2023-02-13 13:44:19 COMPLETED 2023-02-13   731    <NA>  <NA>    <NA>   <NA>
#> 396 2023-02-14 00:00:00 COMPLETED 2023-02-14   731    <NA>  <NA>    <NA>   <NA>
#> 397 2023-02-14 02:48:31 COMPLETED 2023-02-14   731    <NA>  <NA>    <NA>   <NA>
#> 398 2023-02-15 21:42:13 COMPLETED 2023-02-15   731    <NA>  <NA>    <NA>   <NA>
#> 399 2023-02-15 23:10:12 COMPLETED 2023-02-15   731    <NA>  <NA>    <NA>   <NA>
#> 400 2023-02-18 22:26:11 COMPLETED 2023-02-18   731    <NA>  <NA>    <NA>   <NA>
#>     LDDTHELD LDDTHGR1   LSTALVDT study_duration_secs
#> 395       NA     <NA> 2023-03-14            63113904
#> 396       NA     <NA> 2023-03-10            63113904
#> 397       NA     <NA> 2023-03-08            63113904
#> 398       NA     <NA> 2023-03-13            63113904
#> 399       NA     <NA> 2023-03-12            63113904
#> 400       NA     <NA> 2023-03-05            63113904

adae <- rcd_cdisc_dataset_connector("ADAE", fun = radae, ADSL = adsl, max_n_aes = 2L)
load_datasets(adae)

# RelationalDataConnector --------
library(random.cdisc.data)
adsl <- rcd_dataset_connector(dataname = "ADSL", fun = radsl, cached = FALSE)
adrs <- rcd_dataset_connector(dataname = "ADRS", fun = radrs, ADSL = adsl)

rdc <- rcd_data(adsl, adrs)

if (FALSE) {
load_datasets(rdc)
}

# RelationalData --------
library(random.cdisc.data)
adsl <- rcd_dataset_connector(dataname = "ADSL", fun = radsl, cached = FALSE)
adlb <- rcd_dataset_connector(dataname = "ADLB", fun = radlb, cached = FALSE)
adrs <- rcd_dataset_connector("ADRS", radrs, ADSL = adsl)

tc <- teal_data(
  rcd_data(adsl, adlb),
  rcd_data(adrs)
)

if (FALSE) {
load_datasets(tc)
}