Load datasets

load_datasets(x, ...)

# S3 method for Dataset
load_datasets(x, ...)

# S3 method for DatasetConnector
load_datasets(x, args = NULL, try = FALSE, ...)

# S3 method for RelationalDataConnector
load_datasets(x, ...)

# S3 method for RelationalData
load_datasets(x, ...)

Arguments

x
(object of class RelationalData, Dataset or DatasetConnector)

...
(not used)
only for support of S3

args
(NULL or named list)
additional dynamic arguments passed to function which loads the data. Applicable only on DatasetConnector)

try
(logical)
whether perform function evaluation inside try clause. Applicable only on DatasetConnector)

Value

If executed in the interactive session shiny app is opened to load the data. If executed in shiny application - it returns shiny server module.

Examples

# Dataset ------
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
x <- dataset("ADSL", x = ADSL)

load_datasets(x)

# DatasetConnector ------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector("ADSL", fun = radsl, cached = FALSE)
load_datasets(adsl)
get_dataset(adsl)

#> <CDISCDataset>
#>   Inherits from: <Dataset>
#>   Public:
#>     check: function () 
#>     check_keys: function (keys = private$.keys) 
#>     clone: function (deep = FALSE) 
#>     colnames: active binding
#>     column_labels: active binding
#>     data: active binding
#>     dim: active binding
#>     get_attrs: function () 
#>     get_character_colnames: function () 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_colnames: function () 
#>     get_column_labels: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset_label: function () 
#>     get_factor_colnames: function () 
#>     get_keys: function () 
#>     get_numeric_colnames: function () 
#>     get_parent: function () 
#>     get_raw_data: function () 
#>     get_row_labels: function () 
#>     get_rownames: function () 
#>     initialize: function (dataname, x, keys, parent, code = character(0), label = character(0), 
#>     is_pulled: function () 
#>     mutate: function (code, vars = list()) 
#>     ncol: active binding
#>     nrow: active binding
#>     raw_data: active binding
#>     recreate: function (dataname = self$get_dataname(), x = self$get_raw_data(), 
#>     row_labels: active binding
#>     rownames: active binding
#>     set_code: function (code) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_parent: function (parent) 
#>     set_vars: function (vars) 
#>     var_names: active binding
#>   Private:
#>     .col_labels: Study Identifier Unique Subject Identifier Subject Ident ...
#>     .colnames: STUDYID USUBJID SUBJID SITEID AGE SEX RACE COUNTRY DTHFL ...
#>     .dim: 400 38
#>     .keys: STUDYID USUBJID
#>     .ncol: 38
#>     .nrow: 400
#>     .raw_data: tbl_df, tbl, data.frame
#>     .row_labels: NULL
#>     .rownames: 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 ...
#>     code: CodeClass, R6
#>     dataname: ADSL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     execute_code: function (code, vars = list()) 
#>     get_class_colnames: function (class_type = "character") 
#>     parent: 
#>     set_dataname: function (dataname) 
#>     vars: list

adae <- rcd_cdisc_dataset_connector("ADAE", fun = radae, ADSL = adsl, max_n_aes = 2L)
load_datasets(adae)

# RelationalDataConnector --------
library(random.cdisc.data)
adsl <- rcd_dataset_connector(dataname = "ADSL", fun = radsl, cached = FALSE)
adrs <- rcd_dataset_connector(dataname = "ADRS", fun = radrs, ADSL = adsl)

rdc <- rcd_data(adsl, adrs)

if (FALSE) {
load_datasets(rdc)
}

# RelationalData --------
library(random.cdisc.data)
adsl <- rcd_dataset_connector(dataname = "ADSL", fun = radsl, cached = FALSE)
adlb <- rcd_dataset_connector(dataname = "ADLB", fun = radlb, cached = FALSE)
adrs <- rcd_dataset_connector("ADRS", radrs, ADSL = adsl)

tc <- teal_data(
  rcd_data(adsl, adlb),
  rcd_data(adrs)
)

if (FALSE) {
load_datasets(tc)
}