Hide, Show Label only or display a pickerInput

Hidden input widgets are useful to have the input[[inputId]] variable on available in the server function but no corresponding visual clutter from input widgets that provide only a single choice.

optionalSelectInput(
  inputId,
  label,
  choices = NULL,
  selected = NULL,
  multiple = FALSE,
  sep = NULL,
  options = list(),
  label_help = NULL,
  fixed = FALSE
)

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

choices
character vector or NULL. If choices is NULL no pickerInput widget is displayed and input[[inputId]] will be "". If choices is of length 1 then a label and character string will be displayed and the pickerInput widget will be hidden. If the length of choices is more than one the pickerInput element will be displayed. If elements of the list are named then that name rather than the value is displayed to the user.

selected
The initially selected value (or multiple values if multiple = TRUE). If not specified then defaults to the first value for single-select lists and no values for multiple select lists.

multiple
Is selection of multiple items allowed?

sep
(character) A separator string to split the choices or selected inputs into the values of the different columns

options
List of options, see pickerOptions for all available options. To limit the number of selection possible, see example below.

label_help
optional an object of class shiny.tag. E.g. an object returned by helpText

fixed
(logical) (optional) whether to block user to select choices

Examples

optionalSelectInput(inputId = "xvar", label = "x variable", choices = "A", selected = "A")

#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="xvar">x variable</label>
#>   <select data-actions-box="false" data-none-selected-text="- Nothing selected -" data-max-options="1" data-show-subtext="true" data-live-search="false" id="xvar" class="selectpicker form-control" multiple="multiple"><option value="A" selected>A</option></select>
#> </div>

optionalSelectInput(inputId = "xvar", label = "x variable", choices = LETTERS[1:5], selected = "A")

#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="xvar">x variable</label>
#>   <select data-actions-box="false" data-none-selected-text="- Nothing selected -" data-max-options="1" data-show-subtext="true" data-live-search="false" id="xvar" class="selectpicker form-control" multiple="multiple"><option value="A" selected>A</option>
#> <option value="B">B</option>
#> <option value="C">C</option>
#> <option value="D">D</option>
#> <option value="E">E</option></select>
#> </div>

optionalSelectInput(inputId = "xvar",
                    label = "x variable",
                    choices = c("A - value A" = "A"),
                    selected = "A")

#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="xvar">x variable</label>
#>   <select data-actions-box="false" data-none-selected-text="- Nothing selected -" data-max-options="1" data-show-subtext="true" data-live-search="false" id="xvar" class="selectpicker form-control" multiple="multiple"><option value="A" selected>A - value A</option></select>
#> </div>

library(random.cdisc.data)
ADRS <- radrs(cached = TRUE)
optionalSelectInput(inputId = "xvar",
                    label = "x variable",
                    choices = choices_labeled(
                      choices = letters,
                      labels = LETTERS,
                      subset = c("a", "b", "c")
                    ),
                    selected = "a")

#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="xvar">x variable</label>
#>   <select data-actions-box="false" data-none-selected-text="- Nothing selected -" data-max-options="1" data-show-subtext="true" data-live-search="false" id="xvar" class="selectpicker form-control" multiple="multiple">
#>     <option value="a" data-content="a &lt;small class=&quot;text-muted&quot;&gt;A&lt;/small&gt;" selected="selected">a</option>
#>     <option value="b" data-content="b &lt;small class=&quot;text-muted&quot;&gt;B&lt;/small&gt;">b</option>
#>     <option value="c" data-content="c &lt;small class=&quot;text-muted&quot;&gt;C&lt;/small&gt;">c</option>
#>   </select>
#> </div>

optionalSelectInput(inputId = "xvar",
                    label = "x variable",
                    choices = variable_choices(data = ADRS, subset = c("AGE", "SEX", "PARAMCD")),
                    selected = "PARAMCD")

#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="xvar">x variable</label>
#>   <select data-actions-box="false" data-none-selected-text="- Nothing selected -" data-max-options="1" data-show-subtext="true" data-live-search="false" id="xvar" class="selectpicker form-control" multiple="multiple">
#>     <option value="AGE" data-content="&lt;i class=&quot;fa fa-sort-numeric-up&quot;&gt;&lt;/i&gt; AGE &lt;small class=&quot;text-muted&quot;&gt;Age&lt;/small&gt;">AGE</option>
#>     <option value="SEX" data-content="&lt;i class=&quot;fa fa-chart-bar&quot;&gt;&lt;/i&gt; SEX &lt;small class=&quot;text-muted&quot;&gt;Sex&lt;/small&gt;">SEX</option>
#>     <option value="PARAMCD" data-content="&lt;i class=&quot;fa fa-chart-bar&quot;&gt;&lt;/i&gt; PARAMCD &lt;small class=&quot;text-muted&quot;&gt;Parameter Code&lt;/small&gt;" selected="selected">PARAMCD</option>
#>   </select>
#> </div>

selected_value <- paste0(lapply(ADRS[1, c("PARAMCD", "AVISIT")], as.character), collapse = " - ")
optionalSelectInput(inputId = "xvar",
                    label = "x variable",
                    choices = value_choices(data = ADRS,
                                            var_choices = c("PARAMCD", "AVISIT"),
                                            var_label = c("PARAM", "AVISIT")),
                    selected = selected_value)

#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="xvar">x variable</label>
#>   <select data-actions-box="false" data-none-selected-text="- Nothing selected -" data-max-options="1" data-show-subtext="true" data-live-search="true" id="xvar" class="selectpicker form-control" multiple="multiple">
#>     <option value="BESRSPI - CYCLE 2 DAY 1" data-content="BESRSPI - CYCLE 2 DAY 1 &lt;small class=&quot;text-muted&quot;&gt;Best Confirmed Overall Response by Investigator - CYCLE 2 DAY 1&lt;/small&gt;" selected="selected">BESRSPI - CYCLE 2 DAY 1: Best Confirmed Overall Response by Investigator - CYCLE 2 DAY 1</option>
#>     <option value="INVET - END OF INDUCTION" data-content="INVET - END OF INDUCTION &lt;small class=&quot;text-muted&quot;&gt;Investigator End Of Induction Response - END OF INDUCTION&lt;/small&gt;">INVET - END OF INDUCTION: Investigator End Of Induction Response - END OF INDUCTION</option>
#>     <option value="OVRINV - SCREENING" data-content="OVRINV - SCREENING &lt;small class=&quot;text-muted&quot;&gt;Overall Response by Investigator - by visit - SCREENING&lt;/small&gt;">OVRINV - SCREENING: Overall Response by Investigator - by visit - SCREENING</option>
#>     <option value="OVRINV - BASELINE" data-content="OVRINV - BASELINE &lt;small class=&quot;text-muted&quot;&gt;Overall Response by Investigator - by visit - BASELINE&lt;/small&gt;">OVRINV - BASELINE: Overall Response by Investigator - by visit - BASELINE</option>
#>     <option value="OVRINV - CYCLE 2 DAY 1" data-content="OVRINV - CYCLE 2 DAY 1 &lt;small class=&quot;text-muted&quot;&gt;Overall Response by Investigator - by visit - CYCLE 2 DAY 1&lt;/small&gt;">OVRINV - CYCLE 2 DAY 1: Overall Response by Investigator - by visit - CYCLE 2 DAY 1</option>
#>     <option value="OVRINV - CYCLE 4 DAY 1" data-content="OVRINV - CYCLE 4 DAY 1 &lt;small class=&quot;text-muted&quot;&gt;Overall Response by Investigator - by visit - CYCLE 4 DAY 1&lt;/small&gt;">OVRINV - CYCLE 4 DAY 1: Overall Response by Investigator - by visit - CYCLE 4 DAY 1</option>
#>     <option value="OVRINV - END OF INDUCTION" data-content="OVRINV - END OF INDUCTION &lt;small class=&quot;text-muted&quot;&gt;Overall Response by Investigator - by visit - END OF INDUCTION&lt;/small&gt;">OVRINV - END OF INDUCTION: Overall Response by Investigator - by visit - END OF INDUCTION</option>
#>     <option value="OVRINV - FOLLOW UP" data-content="OVRINV - FOLLOW UP &lt;small class=&quot;text-muted&quot;&gt;Overall Response by Investigator - by visit - FOLLOW UP&lt;/small&gt;">OVRINV - FOLLOW UP: Overall Response by Investigator - by visit - FOLLOW UP</option>
#>     <option value="BESRSPI - CYCLE 4 DAY 1" data-content="BESRSPI - CYCLE 4 DAY 1 &lt;small class=&quot;text-muted&quot;&gt;Best Confirmed Overall Response by Investigator - CYCLE 4 DAY 1&lt;/small&gt;">BESRSPI - CYCLE 4 DAY 1: Best Confirmed Overall Response by Investigator - CYCLE 4 DAY 1</option>
#>     <option value="BESRSPI - END OF INDUCTION" data-content="BESRSPI - END OF INDUCTION &lt;small class=&quot;text-muted&quot;&gt;Best Confirmed Overall Response by Investigator - END OF INDUCTION&lt;/small&gt;">BESRSPI - END OF INDUCTION: Best Confirmed Overall Response by Investigator - END OF INDUCTION</option>
#>     <option value="BESRSPI - FOLLOW UP" data-content="BESRSPI - FOLLOW UP &lt;small class=&quot;text-muted&quot;&gt;Best Confirmed Overall Response by Investigator - FOLLOW UP&lt;/small&gt;">BESRSPI - FOLLOW UP: Best Confirmed Overall Response by Investigator - FOLLOW UP</option>
#>     <option value="BESRSPI - SCREENING" data-content="BESRSPI - SCREENING &lt;small class=&quot;text-muted&quot;&gt;Best Confirmed Overall Response by Investigator - SCREENING&lt;/small&gt;">BESRSPI - SCREENING: Best Confirmed Overall Response by Investigator - SCREENING</option>
#>   </select>
#> </div>