Hide, Show Label only or display a selectInput

Hidden input widgets are useful to have the input[[inputId]] variable on available in the server funciont but no corresponding visual clutter from input widgets that provide only a single choice.

optionalSelectInput(inputId, label, choices, selected, ...,
  label_help = NULL)

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

choices
character vector or NULL. If choices is NULL no selectInput widget is displayed and input[[inputId]] will be "". If choices is of length 1 then a label and character string will be displayed and the selectInput widget will be hidden. If the length of choices is more than one the selectInput element will be displayed.

selected
The initially selected value (or multiple values if multiple = TRUE). If not specified then defaults to the first value for single-select lists and no values for multiple select lists.

...
Arguments passed to selectInput().

label_help
optional an object of class shiny.tag. E.g. an object returned by helpText

Examples

optionalSelectInput("xvar", "x variable", 'A', 'A')

#> <div>
#>   <div class="form-group shiny-input-container hidden">
#>     <label class="control-label" for="xvar">x variable</label>
#>     <div>
#>       <select id="xvar"><option value="A" selected>A</option></select>
#>       <script type="application/json" data-for="xvar" data-nonempty="">{}</script>
#>     </div>
#>   </div>
#>   <span>
#>     <label>x variable:</label>
#>     A
#>   </span>
#> </div>

optionalSelectInput("xvar", "x variable", LETTERS[1:5], 'A')

#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="xvar">x variable</label>
#>   <div>
#>     <select id="xvar"><option value="A" selected>A</option>
#> <option value="B">B</option>
#> <option value="C">C</option>
#> <option value="D">D</option>
#> <option value="E">E</option></select>
#>     <script type="application/json" data-for="xvar" data-nonempty="">{}</script>
#>   </div>
#> </div>