RCD DatasetConnector

rcd_dataset_connector(
  dataname,
  fun,
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  ...
)

Arguments

dataname
(character)
A given name for the dataset it may not contain spaces

fun
(function)
any R function which generates data.frame, especially functions from random.cdisc.data like radsl

keys
optional, (character)
vector of dataset primary keys column names

label
(character)
Label to describe the dataset.

code
(character)
A character string defining code to modify raw_data from this dataset. To modify current dataset code should contain at least one assignment to object defined in dataname argument. For example if dataname = ADSL example code should contain ADSL <- <some R code>. Can't be used simultaneously with script

script
(character)
Alternatively to code - location of the file containing modification code. Can't be used simultaneously with script.

...
(optional)
Additional arguments applied to pull function. In case when this object code depends on the raw_data from the other Dataset, DatasetConnector object(s) or other constant value, this/these object(s) should be included. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

Details

Create a DatasetConnector from function in random.cdisc.data.

Examples

library(random.cdisc.data)
x <- rcd_dataset_connector(
  dataname = "ADSL",
  fun = radsl,
  cached = TRUE
)
x$get_code()

#> [1] "ADSL <- radsl(cached = TRUE)"

load_dataset(x)
get_dataset(x)

#> <Dataset>
#>   Public:
#>     check: function () 
#>     check_keys: function (keys = private$.keys) 
#>     clone: function (deep = FALSE) 
#>     colnames: active binding
#>     column_labels: active binding
#>     data: active binding
#>     dim: active binding
#>     get_attrs: function () 
#>     get_character_colnames: function () 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_colnames: function () 
#>     get_column_labels: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset_label: function () 
#>     get_factor_colnames: function () 
#>     get_keys: function () 
#>     get_numeric_colnames: function () 
#>     get_raw_data: function () 
#>     get_row_labels: function () 
#>     get_rownames: function () 
#>     initialize: function (dataname, x, keys = character(0), code = character(0), 
#>     is_pulled: function () 
#>     mutate: function (code, vars = list()) 
#>     ncol: active binding
#>     nrow: active binding
#>     raw_data: active binding
#>     recreate: function (dataname = self$get_dataname(), x = self$get_raw_data(), 
#>     row_labels: active binding
#>     rownames: active binding
#>     set_code: function (code) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_vars: function (vars) 
#>     var_names: active binding
#>   Private:
#>     .col_labels: Study Identifier Unique Subject Identifier Subject Ident ...
#>     .colnames: STUDYID USUBJID SUBJID SITEID AGE SEX RACE COUNTRY DTHFL ...
#>     .dim: 400 38
#>     .keys: 
#>     .ncol: 38
#>     .nrow: 400
#>     .raw_data: tbl_df, tbl, data.frame
#>     .row_labels: NULL
#>     .rownames: 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 ...
#>     code: CodeClass, R6
#>     dataname: ADSL
#>     dataset_label: 
#>     deep_clone: function (name, value) 
#>     execute_code: function (code, vars = list()) 
#>     get_class_colnames: function (class_type = "character") 
#>     set_dataname: function (dataname) 
#>     vars: list

x$get_raw_data()

#> # A tibble: 400 x 38
#>    STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY DTHFL INVID INVNAM
#>    <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <fct> <chr> <chr> 
#>  1 AB12345 AB1234… id-128 CHN-3     32 M     ASIAN CHN     N     INV … Dr. C…
#>  2 AB12345 AB1234… id-262 CHN-15    35 M     BLAC… CHN     N     INV … Dr. C…
#>  3 AB12345 AB1234… id-378 RUS-3     30 F     ASIAN RUS     N     INV … Dr. R…
#>  4 AB12345 AB1234… id-220 CHN-11    26 F     ASIAN CHN     N     INV … Dr. C…
#>  5 AB12345 AB1234… id-267 CHN-7     40 M     ASIAN CHN     N     INV … Dr. C…
#>  6 AB12345 AB1234… id-201 CHN-15    49 M     ASIAN CHN     N     INV … Dr. C…
#>  7 AB12345 AB1234… id-45  USA-1     34 F     ASIAN USA     N     INV … Dr. U…
#>  8 AB12345 AB1234… id-261 USA-1     32 F     ASIAN USA     N     INV … Dr. U…
#>  9 AB12345 AB1234… id-173 NGA-11    24 F     BLAC… NGA     N     INV … Dr. N…
#> 10 AB12345 AB1234… id-307 CHN-1     24 M     ASIAN CHN     N     INV … Dr. C…
#> # … with 390 more rows, and 27 more variables: ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>,
#> #   BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>,
#> #   BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>, EOSSTT <fct>,
#> #   EOSDT <date>, EOSDY <int>, DCSREAS <fct>, DTHDT <date>, DTHCAUS <fct>,
#> #   DTHCAT <fct>, LDDTHELD <int>, LDDTHGR1 <fct>, LSTALVDT <date>,
#> #   study_duration_secs <dbl>