Create a new RelationalDatasetConnector object

Create RelationalDatasetConnector from callable_function.

Please note that the script has to end with a call creating desired object. The error will be raised otherwise.

rcd_dataset_connector - Create a RelationalDatasetConnector from any R function.

rds_dataset_connector - Create a RelationalDatasetConnector from RDS file.

script_dataset_connector - Create a RelationalDatasetConnector from .R file.

code_dataset_connector - Create a RelationalDatasetConnector from a string.

rice_dataset_connector - Create a RelationalDatasetConnector from RICE dataset.

teradata_dataset_connector - Create a RelationalDatasetConnector from TERADATA dataset

rds_cdisc_dataset_connector - Create a RelationalDatasetConnector from RDS file with keys automatically assigned by dataname

rcd_cdisc_dataset_connector - Create a DatasetConnector from any R function with keys assigned automatically by dataname.

rice_cdisc_dataset_connector - Create a RelationalDatasetConnector from RICE dataset with keys assigned automatically by dataname.

script_cdisc_dataset_connector - Create a RelationalDatasetConnector from script file with keys assigned automatically by dataname.

code_cdisc_dataset_connector - Create a RelationalDatasetConnector from code string with keys assigned automatically by dataname.

teradata_cdisc_dataset_connector - Create a RelationalDatasetConnector from TERADATA dataset with keys assigned automatically by dataname.

relational_dataset_connector(
  dataname,
  pull_fun,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  vars = list()
)

relational_dataset_connector_file(path)

rcd_dataset_connector(
  dataname,
  fun,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rds_dataset_connector(
  dataname,
  file,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

script_dataset_connector(
  dataname,
  file,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

code_dataset_connector(
  dataname,
  pull_code,
  keys,
  code = character(0),
  label = character(0),
  ...
)

rice_dataset_connector(
  dataname,
  path,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

teradata_dataset_connector(
  dataname,
  table,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rds_cdisc_dataset_connector(
  dataname,
  file,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rcd_cdisc_dataset_connector(
  dataname,
  fun,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rice_cdisc_dataset_connector(
  dataname,
  path,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

script_cdisc_dataset_connector(
  dataname,
  file,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

code_cdisc_dataset_connector(
  dataname,
  pull_code = character(0),
  code = character(0),
  label = character(0),
  ...
)

teradata_cdisc_dataset_connector(
  dataname,
  table,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

Arguments

dataname
(character)
A given name for the dataset it may not contain spaces

pull_fun
(CallableFunction)
function with necessary arguments set to fetch data from connection.

keys
(keys)
object of S3 class keys containing foreign, primary keys and parent information

code
(character)
A character string defining code to modify raw_data from this dataset. To modify current dataset code should contain at least one assignment to object defined in dataname argument. For example if dataname = ADSL example code should contain ADSL <- <some R code>. Can't be used simultaneously with script

script
(character)
Alternatively to code - location of the file containing modification code. Can't be used simultaneously with script.

label
(character)
Label to describe the dataset.

vars
(list)
In case when this object code depends on the raw_data from the other NamedDataset, NamedDatasetConnector object(s) or other constant value, this/these object(s) should be included.

path
(character)
path to the file

fun
(function)
any R function which generates data.frame, especially functions from random.cdisc.data like radsl

...
(optional)
additional arguments applied to pull function

file
(character)
file location containing code to be evaluated in connector. Object obtained in the last call from file will be returned to the connector - same as source(file = file)$value

pull_code
(character)
function call with arguments to obtain dataset.

table
(character) table name

Value

new RelationalDatasetConnector object

RelationalDatasetConnector object

Note

Do not include assignment in code argument. code string should be a single function call which returns an object. Use mutate_code to make additional transformations.

Examples

# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(random.cdisc.data)

     pull_fun <- callable_function(radsl)
     pull_fun$set_args(list(cached = TRUE))
     relational_dataset_connector(\"ADSL\", pull_fun, get_cdisc_keys(\"ADSL\"))"
  ),
  con = file_example
)
x <- relational_dataset_connector_file(file_example)
get_code(x)

#> [1] "ADSL <- radsl(cached = TRUE)"

library(random.cdisc.data)
x <- rcd_dataset_connector(
  dataname = "ADSL",
  fun = radsl,
  keys = get_cdisc_keys("ADSL"),
  cached = TRUE
)
x$get_code()

#> [1] "ADSL <- radsl(cached = TRUE)"

load_dataset(x)
get_dataset(x)

#> <RelationalDataset>
#>   Inherits from: <NamedDataset>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     colnames: active binding
#>     column_labels: active binding
#>     data: active binding
#>     dim: active binding
#>     get_character_colnames: function () 
#>     get_code: function (deparse = TRUE) 
#>     get_code_class: function () 
#>     get_colnames: function () 
#>     get_column_labels: function () 
#>     get_dataname: function () 
#>     get_datanames: function () 
#>     get_dataset_label: function () 
#>     get_factor_colnames: function () 
#>     get_keys: function () 
#>     get_numeric_colnames: function (include_factors = FALSE) 
#>     get_raw_data: function () 
#>     get_row_labels: function () 
#>     get_rownames: function () 
#>     initialize: function (dataname, x, keys, code = character(0), label = character(0), 
#>     is_pulled: function () 
#>     mutate: function (code, vars = list()) 
#>     ncol: active binding
#>     nrow: active binding
#>     raw_data: active binding
#>     row_labels: active binding
#>     rownames: active binding
#>     set_code: function (code) 
#>     set_dataset_label: function (label) 
#>     set_keys: function (keys) 
#>     set_vars: function (vars) 
#>     var_names: active binding
#>   Private:
#>     .col_labels: Study Identifier Unique Subject Identifier Subject Ident ...
#>     .colnames: STUDYID USUBJID SUBJID SITEID AGE SEX RACE COUNTRY INVID ...
#>     .dim: 400 33
#>     .keys: keys
#>     .ncol: 33
#>     .nrow: 400
#>     .raw_data: tbl_df, tbl, data.frame
#>     .row_labels: 
#>     .rownames: 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 ...
#>     code: CodeClass, R6
#>     dataname: ADSL
#>     dataset_label: 
#>     execute_code: function (code, vars = list()) 
#>     get_class_colnames: function (class_type = "character", include_factors = FALSE) 
#>     set_dataname: function (dataname) 
#>     vars: list

x$get_raw_data()

#> # A tibble: 400 x 33
#>    STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY INVID INVNAM ARM  
#>    <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <chr> <chr>  <fct>
#>  1 AB12345 AB1234… id-128 CHN-3     32 M     ASIAN CHN     INV … Dr. C… A: D…
#>  2 AB12345 AB1234… id-262 CHN-15    35 M     BLAC… CHN     INV … Dr. C… C: C…
#>  3 AB12345 AB1234… id-378 RUS-3     30 F     ASIAN RUS     INV … Dr. R… C: C…
#>  4 AB12345 AB1234… id-220 CHN-11    26 F     ASIAN CHN     INV … Dr. C… B: P…
#>  5 AB12345 AB1234… id-267 CHN-7     40 M     ASIAN CHN     INV … Dr. C… B: P…
#>  6 AB12345 AB1234… id-201 CHN-15    49 M     ASIAN CHN     INV … Dr. C… C: C…
#>  7 AB12345 AB1234… id-45  USA-1     34 F     ASIAN USA     INV … Dr. U… A: D…
#>  8 AB12345 AB1234… id-261 USA-1     32 F     ASIAN USA     INV … Dr. U… B: P…
#>  9 AB12345 AB1234… id-173 NGA-11    24 F     BLAC… NGA     INV … Dr. N… C: C…
#> 10 AB12345 AB1234… id-307 CHN-1     24 M     ASIAN CHN     INV … Dr. C… B: P…
#> # … with 390 more rows, and 22 more variables: ARMCD <fct>, ACTARM <fct>,
#> #   ACTARMCD <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>, EOSSTT <fct>, EOSDT <date>,
#> #   EOSDY <int>, DCSREAS <fct>, DTHDT <date>, LSTALVDT <date>,
#> #   study_duration_secs <dbl>

if (FALSE) {
x <- rds_dataset_connector(
  dataname = "ADSL",
  file = "path/to/file.RDS",
  keys = get_cdisc_keys("ADSL")
)
x$get_code()
}
if (FALSE) {
x <- script_dataset_connector(
  dataname = "ADSL",
  file = "path/to/script.R",
  keys = get_cdisc_keys("ADSL")
)
x$get_code()
}
library(random.cdisc.data)
x <- code_dataset_connector(
  dataname = "ADSL",
  keys = get_cdisc_keys("ADSL"),
  pull_code = "radsl(cached = TRUE)"
)

x$get_code()

#> [1] "ADSL <- radsl(cached = TRUE)"

mutate_dataset(x, code = "ADSL$new_variable <- 1")
x$get_code()

#> [1] "ADSL <- radsl(cached = TRUE)\nADSL$new_variable <- 1"

x <- rice_dataset_connector(
  dataname = "ADSL",
  path = "/path/to/file.sas7bdat",
  keys = get_cdisc_keys("ADSL")
)
x$get_code()

#> [1] "ADSL <- rice::rice_read(node = \"/path/to/file.sas7bdat\", prolong = TRUE, quiet = TRUE)"

if (FALSE) {
load_dataset(x)
get_dataset(x)
x$get_raw_data()
}