Create a standard ui layout with output on the right and an encoding panel on the left

This is the layout used for the teal modules in the beam package

standard_layout(output, encoding = NULL, forms = NULL, pre_output = NULL,
  post_output = NULL)

Arguments

output
shiny.tag object with the output element (table, plot, listing) such as for example returned by plotOutput.

encoding
a shiny.tag object containing the encoding elements. If this element is NULL then no encoding side panel on the right is created.

forms
a tagList of forms (e.g. actionButton) that are placed below the encodings panel

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

Value

an object of class shiny.tag with the ui code.

Examples

# NOT RUN {
library(random.cdisc.data)

ui_test <- function(id) {
 ns <- NS(id)
 standard_layout(
   output = plotOutput(ns("plot")),
   encoding = helpText("Ecoding Panel"),
   forms = tags$div(
      actionButton(ns("show_r_code"), "Show R Code"),
      actionButton(ns("export_plot"), "Export Plot")
   ),
   pre_output = helpText("This is a plot of the", tags$code("iris"), "data"),
   post_output = helpText("More information can be added here.")
 )
}

srv_test <- function(input, output, session, datasets) {
   output$plot <- renderPlot({
      with(iris, plot(Sepal.Length, Petal.Length, col = Species))
   })
}

ui_test2 <- function(id) {
 ns <- NS(id)
 standard_layout(
   output = plotOutput(ns("plot")),
   encoding = NULL,
   forms = tags$div(
      actionButton(ns("show_r_code"), "Show R Code"),
      actionButton(ns("export_plot"), "Export Plot")
   ),
   pre_output = helpText("This is a plot of the", tags$code("iris"), "data"),
   post_output = helpText("More information can be added here.")
 )
}

srv_test2 <- function(input, output, session, datasets) {
   output$plot <- renderPlot({
      with(iris, plot(Sepal.Length, Petal.Length, col = Species))
   })
}

x <- teal::init(
   data = list(ASL = radsl()),
   modules = root_modules(
      module(
         "example",
         ui = ui_test,
         server = srv_test,
         filters = "ASL"
      ),
      module(
         "example no encoding",
         ui = ui_test2,
         server = srv_test2,
         filters = "ASL"
      )
   )
)

shinyApp(x$ui, x$server)

# }