teal: Interactive Exploration of Clinical Trials Data

The teal package provides a shiny based framework for creating an interactive data analysis environment.

Details

To learn mode about the package either read the project website at Project Website or read the init manual pages.

Author

Maintainer: NEST basel.nestcicd@roche.com

Authors:

Adrian Waddell adrian.waddell@roche.com

Pawel Rucki pawel.rucki@contractors.roche.com

Dawid Kaledkowski dawid.kaledkowski@contractors.roche.com

Sebastian Wolf sebastian.wolf.sw1@roche.com

Maximilian Mordig maximilian_oliver.mordig@roche.com

Bartosz Baranowski bartosz.baranowski@roche.com

Ondrej Slama ondrej.slama@roche.com

Nikolas Burkoff nikolas.burkoff@roche.com

Maciej Nasinski maciej.nasinski@contractors.roche.com

Other contributors:

Tadeusz Lewandowski tadeusz.lewandowski@roche.com [contributor]

Roche [copyright holder, funder]