teal: Interactive exploration of clinical trials data

The teal package provides a shiny based framework for creating an interactive data analysis environment.

Details

To learn mode about the package, visit the project website or read the init() manual page.

See also

Author

Maintainer: Dawid Kaledkowski dawid.kaledkowski@roche.com (ORCID)

Authors:

Pawel Rucki pawel.rucki@roche.com

Aleksander Chlebowski aleksander.chlebowski@contractors.roche.com (ORCID)

Andre Verissimo andre.verissimo@roche.com (ORCID)

Kartikeya Kirar kartikeya.kirar@businesspartner.roche.com

Vedha Viyash vedha.viyash@roche.com

Marcin Kosinski marcin.kosinski.mk1@roche.com

Adrian Waddell adrian.waddell@gene.com

Nikolas Burkoff

Mahmoud Hallal

Maciej Nasinski

Konrad Pagacz

Junlue Zhao

Tadeusz Lewandowski

Other contributors:

Chendi Liao chendi.liao@roche.com [reviewer]

Dony Unardi unardid@gene.com [reviewer]

F. Hoffmann-La Roche AG [copyright holder, funder]

Maximilian Mordig [contributor]