Create a UI of nested tabs of teal_modules

Each teal_modules is translated to a tabsetPanel and each of its children is put into one tab. The UI of a teal_module is obtained by calling the ui function on it.

ui_nested_tabs(id, modules, datasets)

Arguments

id
module id

modules
teal_module or teal_modules object, the latter can be used for nested tabs, see ui_nested_tabs

datasets
FilteredData object to store filter state and filtered datasets, shared across modules

Value

depending on class of modules:

teal_module: instantiated UI of the module

teal_modules: tabsetPanel with each tab corresponding to recursively calling this function on it

Details

The datasets argument is required to resolve the teal arguments in an isolated context (with respect to reactivity)

Examples

mods <- teal:::get_dummy_modules()
datasets <- teal:::get_dummy_datasets()
app <- shinyApp(
  ui = function() {
    tagList(
      teal:::include_teal_css_js(),
      textOutput("info"),
      fluidPage( # needed for nice tabs
        teal:::ui_nested_tabs("dummy", modules = mods, datasets = datasets)
      )
    )
  },
  server = function(input, output, session) {
    active_module <- callModule(
      teal:::srv_nested_tabs, "dummy", datasets = datasets, modules = mods
    )
    output$info <- renderText({
      paste0("The currently active tab name is ", active_module()$label)
    })
  }
)
if (FALSE) {
runApp(app)
}