Wrapper on choices_labeled to label variables basing on existing labels in data

variable_choices(data, subset = NULL, fill = FALSE, key = NULL)

# S3 method for character
variable_choices(data, subset = NULL, fill = FALSE, key = NULL)

# S3 method for data.frame
variable_choices(data, subset = NULL, fill = FALSE, key = NULL)

# S3 method for Dataset
variable_choices(data, subset = NULL, fill = FALSE, key = get_keys(data))

# S3 method for DatasetConnector
variable_choices(data, subset = NULL, fill = FALSE, key = get_keys(data))

Arguments

data
(data.frame, character, Dataset, DatasetConnector) If data.frame, then data to extract labels from If character, then name of the dataset to extract data from once available If Dataset or DatasetConnector, then raw data to extract labels from.

subset
(character or function) If character, then a vector of column names. If function, then this function is used to determine the possible columns (e.g. all factor columns). In this case, the function must take only single argument "data" and return a character vector. See examples for more details.

fill
boolean in case the label attribute does not exist if TRUE the variable names is returned, otherwise NA

key
(character) vector with names of the variables, which are part of the primary key of the data argument. This is an optional argument, which allows to identify variables associated with the primary key and display the appropriate icon for them in the optionalSelectInput widget.

Value

named character vector with additional attributes or delayed_data object

Examples

library(random.cdisc.data)
ADRS <- radrs(cached = TRUE)

variable_choices(ADRS)

#> number of choices: 48 
#> 
#> STUDYID: Study Identifier
#> USUBJID: Unique Subject Identifier
#> SUBJID: Subject Identifier for the Study
#> SITEID: Study Site Identifier
#> AGE: Age
#> SEX: Sex
#> RACE: Race
#> COUNTRY: Country
#> DTHFL: Subject Death Flag
#> INVID: Investigator Identifier
#> INVNAM: Investigator Name
#> ARM: Description of Planned Arm
#> ARMCD: Planned Arm Code
#> ACTARM: Description of Actual Arm
#> ACTARMCD: Actual Arm Code
#> REGION1: Geographic Region 1
#> STRATA1: Stratification Factor 1
#> STRATA2: Stratification Factor 2
#> BMRKR1: Continous Level Biomarker 1
#> BMRKR2: Categorical Level Biomarker 2
#> ITTFL: Intent-To-Treat Population Flag
#> SAFFL: Safety Population Flag
#> BMEASIFL: Response Evaluable Population Flag
#> BEP01FL: Biomarker Evaluable Population Flag
#> RANDDT: Date of Randomization
#> TRTSDTM: Datetime of First Exposure to Treatment
#> TRTEDTM: Datetime of Last Exposure to Treatment
#> EOSSTT: End of Study Status
#> EOSDT: End of Study Date
#> EOSDY: End of Study Relative Day
#> DCSREAS: Reason for Discontinuation from Study
#> DTHDT: Date of Death
#> DTHCAUS: Cause of Death
#> DTHCAT: Cause of Death Category
#> LDDTHELD: Elapsed Days from Last Dose to Death
#> LDDTHGR1: Last Dose to Death - Days Elapsed Grp 1
#> LSTALVDT: Date Last Known Alive
#> study_duration_secs: NOT A STANDARD BUT NEEDED FOR RCD
#> ASEQ: Analysis Sequence Number
#> RSSEQ: Sequence Number
#> PARAM: Parameter
#> PARAMCD: Parameter Code
#> AVAL: Analysis Value
#> AVALC: Analysis Value (C)
#> ADTM: Analysis Datetime
#> ADY: Analysis Relative Day
#> AVISIT: Analysis Visit
#> AVISITN: Analysis Visit (N)
#> 

variable_choices(ADRS, subset = c("PARAM", "PARAMCD"))

#> number of choices: 2 
#> 
#> PARAM: Parameter
#> PARAMCD: Parameter Code
#> 

variable_choices(ADRS, subset = c("", "PARAM", "PARAMCD"))

#> number of choices: 3 
#> 
#> : 
#> PARAM: Parameter
#> PARAMCD: Parameter Code
#> 

variable_choices(ADRS, subset = c("", "PARAM", "PARAMCD"), key = get_cdisc_keys("ADRS"))

#> number of choices: 3 
#> 
#> : 
#> PARAM: Parameter
#> PARAMCD: Parameter Code
#> 

# delayed version
variable_choices("ADRS", subset = c("USUBJID", "STUDYID"))

#> variable_choices with delayed data: ADRS
#> $ data
#> [1] "ADRS"
#> $ subset
#> [1] "USUBJID" "STUDYID"
#> $ key
#> NULL

# also works with Dataset and DatasetConnector
ADRS_dataset <- dataset("ADRS", ADRS, key = get_cdisc_keys("ADRS"))
variable_choices(ADRS_dataset)

#> number of choices: 48 
#> 
#> STUDYID: Study Identifier
#> USUBJID: Unique Subject Identifier
#> SUBJID: Subject Identifier for the Study
#> SITEID: Study Site Identifier
#> AGE: Age
#> SEX: Sex
#> RACE: Race
#> COUNTRY: Country
#> DTHFL: Subject Death Flag
#> INVID: Investigator Identifier
#> INVNAM: Investigator Name
#> ARM: Description of Planned Arm
#> ARMCD: Planned Arm Code
#> ACTARM: Description of Actual Arm
#> ACTARMCD: Actual Arm Code
#> REGION1: Geographic Region 1
#> STRATA1: Stratification Factor 1
#> STRATA2: Stratification Factor 2
#> BMRKR1: Continous Level Biomarker 1
#> BMRKR2: Categorical Level Biomarker 2
#> ITTFL: Intent-To-Treat Population Flag
#> SAFFL: Safety Population Flag
#> BMEASIFL: Response Evaluable Population Flag
#> BEP01FL: Biomarker Evaluable Population Flag
#> RANDDT: Date of Randomization
#> TRTSDTM: Datetime of First Exposure to Treatment
#> TRTEDTM: Datetime of Last Exposure to Treatment
#> EOSSTT: End of Study Status
#> EOSDT: End of Study Date
#> EOSDY: End of Study Relative Day
#> DCSREAS: Reason for Discontinuation from Study
#> DTHDT: Date of Death
#> DTHCAUS: Cause of Death
#> DTHCAT: Cause of Death Category
#> LDDTHELD: Elapsed Days from Last Dose to Death
#> LDDTHGR1: Last Dose to Death - Days Elapsed Grp 1
#> LSTALVDT: Date Last Known Alive
#> study_duration_secs: NOT A STANDARD BUT NEEDED FOR RCD
#> ASEQ: Analysis Sequence Number
#> RSSEQ: Sequence Number
#> PARAM: Parameter
#> PARAMCD: Parameter Code
#> AVAL: Analysis Value
#> AVALC: Analysis Value (C)
#> ADTM: Analysis Datetime
#> ADY: Analysis Relative Day
#> AVISIT: Analysis Visit
#> AVISITN: Analysis Visit (N)
#> 

ADRS_conn <- dataset_connector(
  "ADRS",
  pull_callable = callable_code("radrs(cached = TRUE)"),
  key = get_cdisc_keys("ADRS"))
variable_choices(ADRS_conn)

#> variable_choices with delayed data: ADRS
#> $ data
#> [1] "ADRS"
#> $ subset
#> NULL
#> $ key
#> [1] "STUDYID" "USUBJID" "PARAMCD" "AVISIT" 

# functional subset (with delayed data) - return only factor variables
variable_choices("ADRS", subset = function(data) {
  idx <- vapply(data, is.factor, logical(1))
  return(names(data)[idx])
})

#> variable_choices with delayed data: ADRS
#> $ data
#> [1] "ADRS"
#> $ subset
#> function(data) {
#>   idx <- vapply(data, is.factor, logical(1))
#>   return(names(data)[idx])
#> }
#> <environment: 0x55f8fb033150>
#> $ key
#> NULL